Data.Promotion.Prelude

Description

Mimics the Haskell Prelude, but with promoted types.

Standard types, classes and related functions

Basic data types

type family If cond tru fls :: k

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If k True tru fls = tru

If k False tru fls = fls

type family Not a :: Bool

Equations

Not False = TrueSym0

Not True = FalseSym0

type family a :&& a :: Bool infixr 3

Equations

False :&& _z_1627785679 = FalseSym0

True :&& x = x

type family a :|| a :: Bool infixr 2

Equations

False :|| x = x

True :|| _z_1627785667 = TrueSym0

type family Otherwise :: Bool

Equations

Otherwise = TrueSym0

maybe_ :: forall b a. b -> (a -> b) -> Maybe a -> b

type family Maybe_ a a a :: b

Equations

Maybe_ n _z_1627933303 Nothing = n

Maybe_ _z_1627933306 f (Just x) = Apply f x

either_ :: forall a c b. (a -> c) -> (b -> c) -> Either a b -> c

type family Either_ a a a :: c

Equations

Either_ f _z_1627916125 (Left x) = Apply f x

Either_ _z_1627916129 g (Right y) = Apply g y

data Symbol :: *

(Kind) This is the kind of type-level symbols.

type family Fst a :: a

Equations

Fst `(x, _z_1627926696)` = x

type family Snd a :: b

Equations

Snd `(_z_1627926687, y)` = y

type family Curry a a a :: c

Equations

Curry f x y = Apply f (Apply (Apply Tuple2Sym0 x) y)

type family Uncurry a a :: c

Equations

Uncurry f p = Apply (Apply f (Apply FstSym0 p)) (Apply SndSym0 p)

Error reporting

type family Error str :: k

The promotion of error. This version is more poly-kinded for easier use.

data ErrorSym0 l

Promoted equality

module Data.Promotion.Prelude.Eq

Promoted comparisons

module Data.Promotion.Prelude.Ord

Promoted enumerations

As a matter of convenience, the promoted Prelude does not export promoted succ and pred, due to likely conflicts with unary numbers. Please import Enum directly if you want these.

module Data.Promotion.Prelude.Enum

Promoted numbers

module Data.Promotion.Prelude.Num

Miscellaneous functions

type family Id a :: a

Equations

Id x = x

type family Const a a :: a

Equations

Const x _z_1627883774 = x

type family (a :. a) a :: c infixr 9

Equations

(f :. g) a_1627883737 = Apply (Apply (Apply (Apply Lambda_1627883742Sym0 f) g) a_1627883737) a_1627883737

type family f $ x :: b infixr 0

type family f $! x :: b infixr 0

type family Flip a a a :: c

Equations

Flip f x y = Apply (Apply f y) x

type family AsTypeOf a a :: a

Equations

AsTypeOf a_1627883777 a_1627883779 = Apply (Apply ConstSym0 a_1627883777) a_1627883779

type family Until a a a :: a

Equations

Until p f a_1627931020 = Apply (Let1627931025GoSym3 p f a_1627931020) a_1627931020

type family Seq a a :: b infixr 0

Equations

Seq _z_1627883700 x = x

List operations

type family Map a a :: [b]

Equations

Map _z_1627883819 `[]` = `[]`

Map f ((:) x xs) = Apply (Apply (:$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

type family a :++ a :: [a] infixr 5

Equations

`[]` :++ ys = ys

((:) x xs) :++ ys = Apply (Apply (:$) x) (Apply (Apply (:++$) xs) ys)

type family Filter a a :: [a]

Equations

Filter _p `[]` = `[]`

Filter p ((:) x xs) = Case_1628026735 p x xs (Let1628026722Scrutinee_1628025925Sym3 p x xs)

type family Head a :: a

Equations

Head ((:) a _z_1628029340) = a

Head `[]` = Apply ErrorSym0 "Data.Singletons.List.head: empty list"

type family Last a :: a

Equations

Last `[]` = Apply ErrorSym0 "Data.Singletons.List.last: empty list"

Last `[x]` = x

Last ((:) _z_1628029330 ((:) x xs)) = Apply LastSym0 (Apply (Apply (:$) x) xs)

type family Tail a :: [a]

Equations

Tail ((:) _z_1628029321 t) = t

Tail `[]` = Apply ErrorSym0 "Data.Singletons.List.tail: empty list"

type family Init a :: [a]

Equations

Init `[]` = Apply ErrorSym0 "Data.Singletons.List.init: empty list"

Init ((:) x xs) = Apply (Apply (Let1628029290Init'Sym2 x xs) x) xs

type family Null a :: Bool

Equations

Null `[]` = TrueSym0

Null ((:) _z_1628029219 _z_1628029222) = FalseSym0

type family Length a :: Nat

Equations

Length `[]` = FromInteger 0

Length ((:) _z_1628026134 xs) = Apply (Apply (:+$) (FromInteger 1)) (Apply LengthSym0 xs)

type family a :!! a :: a

Equations

`[]` :!! _z_1628026086 = Apply ErrorSym0 "Data.Singletons.List.!!: index too large"

((:) x xs) :!! n = Case_1628026105 x xs n (Let1628026092Scrutinee_1628025965Sym3 x xs n)

type family Reverse a :: [a]

Equations

Reverse l = Apply (Apply (Let1628029185RevSym1 l) l) `[]`

Reducing lists (folds)

type family Foldl a a a :: b

Equations

Foldl f z0 xs0 = Apply (Apply (Let1627751476LgoSym3 f z0 xs0) z0) xs0

type family Foldl1 a a :: a

Equations

Foldl1 f ((:) x xs) = Apply (Apply (Apply FoldlSym0 f) x) xs

Foldl1 _z_1628028429 `[]` = Apply ErrorSym0 "Data.Singletons.List.foldl1: empty list"

type family Foldr a a a :: b

Equations

Foldr k z a_1627883840 = Apply (Let1627883845GoSym3 k z a_1627883840) a_1627883840

type family Foldr1 a a :: a

Equations

Foldr1 _z_1628028387 `[x]` = x

Foldr1 f ((:) x ((:) wild_1628025871 wild_1628025873)) = Apply (Apply f x) (Apply (Apply Foldr1Sym0 f) (Let1628028395XsSym4 f x wild_1628025871 wild_1628025873))

Foldr1 _z_1628028414 `[]` = Apply ErrorSym0 "Data.Singletons.List.foldr1: empty list"

Special folds

type family And a :: Bool

Equations

And `[]` = TrueSym0

And ((:) x xs) = Apply (Apply (:&&$) x) (Apply AndSym0 xs)

type family Or a :: Bool

Equations

Or `[]` = FalseSym0

Or ((:) x xs) = Apply (Apply (:||$) x) (Apply OrSym0 xs)

any_ :: forall a. (a -> Bool) -> [a] -> Bool

type family Any_ a a :: Bool

Equations

Any_ _z_1628014649 `[]` = FalseSym0

Any_ p ((:) x xs) = Apply (Apply (:||$) (Apply p x)) (Apply (Apply Any_Sym0 p) xs)

type family All a a :: Bool

Equations

All _z_1628028340 `[]` = TrueSym0

All p ((:) x xs) = Apply (Apply (:&&$) (Apply p x)) (Apply (Apply AllSym0 p) xs)

type family Sum a :: a

Equations

Sum l = Apply (Apply (Let1628026167Sum'Sym1 l) l) (FromInteger 0)

type family Product a :: a

Equations

Product l = Apply (Apply (Let1628026143ProdSym1 l) l) (FromInteger 1)

type family Concat a :: [a]

Equations

Concat a_1628028371 = Apply (Apply (Apply FoldrSym0 (:++$)) `[]`) a_1628028371

type family ConcatMap a a :: [b]

Equations

ConcatMap f a_1628028367 = Apply (Apply (Apply FoldrSym0 (Apply (Apply (:.$) (:++$)) f)) `[]`) a_1628028367

type family Maximum a :: a

Equations

Maximum `[]` = Apply ErrorSym0 "Data.Singletons.List.maximum: empty list"

Maximum ((:) wild_1628025951 wild_1628025953) = Apply (Apply Foldl1Sym0 MaxSym0) (Let1628028612XsSym2 wild_1628025951 wild_1628025953)

type family Minimum a :: a

Equations

Minimum `[]` = Apply ErrorSym0 "Data.Singletons.List.minimum: empty list"

Minimum ((:) wild_1628025955 wild_1628025957) = Apply (Apply Foldl1Sym0 MinSym0) (Let1628028626XsSym2 wild_1628025955 wild_1628025957)

Building lists

Scans

type family Scanl a a a :: [b]

Equations

Scanl f q ls = Apply (Apply (:$) q) (Case_1628028311 f q ls ls)

type family Scanl1 a a :: [a]

Equations

Scanl1 f ((:) x xs) = Apply (Apply (Apply ScanlSym0 f) x) xs

Scanl1 _z_1628028328 `[]` = `[]`

type family Scanr a a a :: [b]

Equations

Scanr _z_1628028261 q0 `[]` = Apply (Apply (:$) q0) `[]`

Scanr f q0 ((:) x xs) = Case_1628028288 f q0 x xs (Let1628028269Scrutinee_1628025875Sym4 f q0 x xs)

type family Scanr1 a a :: [a]

Equations

Scanr1 _z_1628028192 `[]` = `[]`

Scanr1 _z_1628028195 `[x]` = Apply (Apply (:$) x) `[]`

Scanr1 f ((:) x ((:) wild_1628025879 wild_1628025881)) = Case_1628028241 f x wild_1628025879 wild_1628025881 (Let1628028222Scrutinee_1628025877Sym4 f x wild_1628025879 wild_1628025881)

Infinite lists

type family Replicate a a :: [a]

Equations

Replicate n x = Case_1628026127 n x (Let1628026119Scrutinee_1628025963Sym2 n x)

Sublists

type family Take a a :: [a]

Equations

Take _z_1628026314 `[]` = `[]`

Take n ((:) x xs) = Case_1628026333 n x xs (Let1628026320Scrutinee_1628025947Sym3 n x xs)

type family Drop a a :: [a]

Equations

Drop _z_1628026283 `[]` = `[]`

Drop n ((:) x xs) = Case_1628026302 n x xs (Let1628026289Scrutinee_1628025949Sym3 n x xs)

type family SplitAt a a :: ([a], [a])

Equations

SplitAt n xs = Apply (Apply Tuple2Sym0 (Apply (Apply TakeSym0 n) xs)) (Apply (Apply DropSym0 n) xs)

type family TakeWhile a a :: [a]

Equations

TakeWhile _z_1628026687 `[]` = `[]`

TakeWhile p ((:) x xs) = Case_1628026706 p x xs (Let1628026693Scrutinee_1628025937Sym3 p x xs)

type family DropWhile a a :: [a]

Equations

DropWhile _z_1628026643 `[]` = `[]`

DropWhile p ((:) x xs') = Case_1628026675 p x xs' (Let1628026662Scrutinee_1628025939Sym3 p x xs')

type family Span a a :: ([a], [a])

Equations

Span _z_1628026461 `[]` = Apply (Apply Tuple2Sym0 (Let1628026464XsSym1 _z_1628026461)) (Let1628026464XsSym1 _z_1628026461)

Span p ((:) x xs') = Case_1628026497 p x xs' (Let1628026484Scrutinee_1628025943Sym3 p x xs')

type family Break a a :: ([a], [a])

Equations

Break _z_1628026356 `[]` = Apply (Apply Tuple2Sym0 (Let1628026359XsSym1 _z_1628026356)) (Let1628026359XsSym1 _z_1628026356)

Break p ((:) x xs') = Case_1628026392 p x xs' (Let1628026379Scrutinee_1628025945Sym3 p x xs')

Searching lists

type family Elem a a :: Bool

Equations

Elem _z_1628027703 `[]` = FalseSym0

Elem x ((:) y ys) = Apply (Apply (:||$) (Apply (Apply (:==$) x) y)) (Apply (Apply ElemSym0 x) ys)

type family NotElem a a :: Bool

Equations

NotElem _z_1628027688 `[]` = TrueSym0

NotElem x ((:) y ys) = Apply (Apply (:&&$) (Apply (Apply (:/=$) x) y)) (Apply (Apply NotElemSym0 x) ys)

type family Lookup a a :: Maybe b

Equations

Lookup _key `[]` = NothingSym0

Lookup key ((:) `(x, y)` xys) = Case_1628026271 key x y xys (Let1628026252Scrutinee_1628025959Sym4 key x y xys)

Zipping and unzipping lists

type family Zip a a :: [(a, b)]

Equations

Zip ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply Tuple2Sym0 x) y)) (Apply (Apply ZipSym0 xs) ys)

Zip `[]` `[]` = `[]`

Zip ((:) _z_1628027565 _z_1628027568) `[]` = `[]`

Zip `[]` ((:) _z_1628027571 _z_1628027574) = `[]`

type family Zip3 a a a :: [(a, b, c)]

Equations

Zip3 ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply Tuple3Sym0 a) b) c)) (Apply (Apply (Apply Zip3Sym0 as) bs) cs)

Zip3 `[]` `[]` `[]` = `[]`

Zip3 `[]` `[]` ((:) _z_1628027498 _z_1628027501) = `[]`

Zip3 `[]` ((:) _z_1628027504 _z_1628027507) `[]` = `[]`

Zip3 `[]` ((:) _z_1628027510 _z_1628027513) ((:) _z_1628027516 _z_1628027519) = `[]`

Zip3 ((:) _z_1628027522 _z_1628027525) `[]` `[]` = `[]`

Zip3 ((:) _z_1628027528 _z_1628027531) `[]` ((:) _z_1628027534 _z_1628027537) = `[]`

Zip3 ((:) _z_1628027540 _z_1628027543) ((:) _z_1628027546 _z_1628027549) `[]` = `[]`

type family ZipWith a a a :: [c]

Equations

ZipWith f ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply f x) y)) (Apply (Apply (Apply ZipWithSym0 f) xs) ys)

ZipWith _z_1628027456 `[]` `[]` = `[]`

ZipWith _z_1628027459 ((:) _z_1628027462 _z_1628027465) `[]` = `[]`

ZipWith _z_1628027468 `[]` ((:) _z_1628027471 _z_1628027474) = `[]`

type family ZipWith3 a a a a :: [d]

Equations

ZipWith3 z ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply z a) b) c)) (Apply (Apply (Apply (Apply ZipWith3Sym0 z) as) bs) cs)

ZipWith3 _z_1628027361 `[]` `[]` `[]` = `[]`

ZipWith3 _z_1628027364 `[]` `[]` ((:) _z_1628027367 _z_1628027370) = `[]`

ZipWith3 _z_1628027373 `[]` ((:) _z_1628027376 _z_1628027379) `[]` = `[]`

ZipWith3 _z_1628027382 `[]` ((:) _z_1628027385 _z_1628027388) ((:) _z_1628027391 _z_1628027394) = `[]`

ZipWith3 _z_1628027397 ((:) _z_1628027400 _z_1628027403) `[]` `[]` = `[]`

ZipWith3 _z_1628027406 ((:) _z_1628027409 _z_1628027412) `[]` ((:) _z_1628027415 _z_1628027418) = `[]`

ZipWith3 _z_1628027421 ((:) _z_1628027424 _z_1628027427) ((:) _z_1628027430 _z_1628027433) `[]` = `[]`

type family Unzip a :: ([a], [b])

Equations

Unzip xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_1628027307Sym0 xs)) (Apply (Apply Tuple2Sym0 `[]`) `[]`)) xs

type family Unzip3 a :: ([a], [b], [c])

Equations

Unzip3 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_1628027275Sym0 xs)) (Apply (Apply (Apply Tuple3Sym0 `[]`) `[]`) `[]`)) xs

Other datatypes

data Proxy t :: k -> *

A concrete, poly-kinded proxy type

Constructors

Proxy

data KProxy t :: * -> *

A concrete, promotable proxy type, for use at the kind level There are no instances for this because it is intended at the kind level only

Constructors

KProxy

Defunctionalization symbols

type FalseSym0 = False

type TrueSym0 = True

data NotSym0 l

type NotSym1 t = Not t

data (:&&$) l

data l :&&$$ l

type (:&&$$$) t t = (:&&) t t

data (:||$) l

data l :||$$ l

type (:||$$$) t t = (:||) t t

type OtherwiseSym0 = Otherwise

type NothingSym0 = Nothing

data JustSym0 l

type JustSym1 t = Just t

data Maybe_Sym0 l

data Maybe_Sym1 l l

data Maybe_Sym2 l l l

type Maybe_Sym3 t t t = Maybe_ t t t

data LeftSym0 l

type LeftSym1 t = Left t

data RightSym0 l

type RightSym1 t = Right t

data Either_Sym0 l

data Either_Sym1 l l

data Either_Sym2 l l l

type Either_Sym3 t t t = Either_ t t t

type Tuple0Sym0 = `()`

data Tuple2Sym0 l

data Tuple2Sym1 l l

type Tuple2Sym2 t t = `(t, t)`

data Tuple3Sym0 l

data Tuple3Sym1 l l

data Tuple3Sym2 l l l

type Tuple3Sym3 t t t = `(t, t, t)`

data Tuple4Sym0 l

data Tuple4Sym1 l l

data Tuple4Sym2 l l l

data Tuple4Sym3 l l l l

type Tuple4Sym4 t t t t = `(t, t, t, t)`

data Tuple5Sym0 l

data Tuple5Sym1 l l

data Tuple5Sym2 l l l

data Tuple5Sym3 l l l l

data Tuple5Sym4 l l l l l

type Tuple5Sym5 t t t t t = `(t, t, t, t, t)`

data Tuple6Sym0 l

data Tuple6Sym1 l l

data Tuple6Sym2 l l l

data Tuple6Sym3 l l l l

data Tuple6Sym4 l l l l l

data Tuple6Sym5 l l l l l l

type Tuple6Sym6 t t t t t t = `(t, t, t, t, t, t)`

data Tuple7Sym0 l

data Tuple7Sym1 l l

data Tuple7Sym2 l l l

data Tuple7Sym3 l l l l

data Tuple7Sym4 l l l l l

data Tuple7Sym5 l l l l l l

data Tuple7Sym6 l l l l l l l

type Tuple7Sym7 t t t t t t t = `(t, t, t, t, t, t, t)`

data FstSym0 l

type FstSym1 t = Fst t

data SndSym0 l

type SndSym1 t = Snd t

data CurrySym0 l

data CurrySym1 l l

data CurrySym2 l l l

type CurrySym3 t t t = Curry t t t

data UncurrySym0 l

data UncurrySym1 l l

type UncurrySym2 t t = Uncurry t t

data (:^$) l

data l :^$$ l

data IdSym0 l

type IdSym1 t = Id t

data ConstSym0 l

data ConstSym1 l l

type ConstSym2 t t = Const t t

data (:.$) l

data l :.$$ l

data (l :.$$$ l) l

data ($$) :: TyFun (TyFun a b -> *) (TyFun a b -> *) -> *

data ($$$) :: (TyFun a b -> *) -> TyFun a b -> *

type ($$$$) a b = ($) a b

data ($!$) :: TyFun (TyFun a b -> *) (TyFun a b -> *) -> *

data ($!$$) :: (TyFun a b -> *) -> TyFun a b -> *

type ($!$$$) a b = ($!) a b

data FlipSym0 l

data FlipSym1 l l

data FlipSym2 l l l

data AsTypeOfSym0 l

data AsTypeOfSym1 l l

type AsTypeOfSym2 t t = AsTypeOf t t

data SeqSym0 l

data SeqSym1 l l

type SeqSym2 t t = Seq t t

data (:$) l

data l :$$ l

type (:$$$) t t = (:) t t

type NilSym0 = `[]`

data MapSym0 l

data MapSym1 l l

type MapSym2 t t = Map t t

data ReverseSym0 l

type ReverseSym1 t = Reverse t

data l :++$$ l

data (:++$) l

data HeadSym0 l

type HeadSym1 t = Head t

data LastSym0 l

type LastSym1 t = Last t

data TailSym0 l

type TailSym1 t = Tail t

data InitSym0 l

type InitSym1 t = Init t

data NullSym0 l

type NullSym1 t = Null t

data FoldlSym0 l

data FoldlSym1 l l

data FoldlSym2 l l l

type FoldlSym3 t t t = Foldl t t t

data Foldl1Sym0 l

data Foldl1Sym1 l l

type Foldl1Sym2 t t = Foldl1 t t

data FoldrSym0 l

data FoldrSym1 l l

data FoldrSym2 l l l

type FoldrSym3 t t t = Foldr t t t

data Foldr1Sym0 l

data Foldr1Sym1 l l

type Foldr1Sym2 t t = Foldr1 t t

data ConcatSym0 l

type ConcatSym1 t = Concat t

data ConcatMapSym0 l

data ConcatMapSym1 l l

type ConcatMapSym2 t t = ConcatMap t t

data MaximumBySym0 l

data MaximumBySym1 l l

type MaximumBySym2 t t = MaximumBy t t

data MinimumBySym0 l

data MinimumBySym1 l l

type MinimumBySym2 t t = MinimumBy t t

data AndSym0 l

type AndSym1 t = And t

data OrSym0 l

type OrSym1 t = Or t

data Any_Sym0 l

data Any_Sym1 l l

type Any_Sym2 t t = Any_ t t

data AllSym0 l

data AllSym1 l l

type AllSym2 t t = All t t

data ScanlSym0 l

data ScanlSym1 l l

data ScanlSym2 l l l

type ScanlSym3 t t t = Scanl t t t

data Scanl1Sym0 l

data Scanl1Sym1 l l

type Scanl1Sym2 t t = Scanl1 t t

data ScanrSym0 l

data ScanrSym1 l l

data ScanrSym2 l l l

type ScanrSym3 t t t = Scanr t t t

data Scanr1Sym0 l

data Scanr1Sym1 l l

type Scanr1Sym2 t t = Scanr1 t t

data ElemSym0 l

data ElemSym1 l l

type ElemSym2 t t = Elem t t

data NotElemSym0 l

data NotElemSym1 l l

type NotElemSym2 t t = NotElem t t

data ZipSym0 l

data ZipSym1 l l

type ZipSym2 t t = Zip t t

data Zip3Sym0 l

data Zip3Sym1 l l

data Zip3Sym2 l l l

type Zip3Sym3 t t t = Zip3 t t t

data ZipWithSym0 l

data ZipWithSym1 l l

data ZipWithSym2 l l l

type ZipWithSym3 t t t = ZipWith t t t

data ZipWith3Sym0 l

data ZipWith3Sym1 l l

data ZipWith3Sym2 l l l

data ZipWith3Sym3 l l l l

data UnzipSym0 l

type UnzipSym1 t = Unzip t

data UntilSym0 l

data UntilSym1 l l

data UntilSym2 l l l

type UntilSym3 t t t = Until t t t

data LengthSym0 l

type LengthSym1 t = Length t

data SumSym0 l

type SumSym1 t = Sum t

data ProductSym0 l

type ProductSym1 t = Product t

data ReplicateSym0 l

data ReplicateSym1 l l

type ReplicateSym2 t t = Replicate t t

data TakeSym0 l

data TakeSym1 l l

type TakeSym2 t t = Take t t

data DropSym0 l

data DropSym1 l l

type DropSym2 t t = Drop t t

data SplitAtSym0 l

data SplitAtSym1 l l

type SplitAtSym2 t t = SplitAt t t

data TakeWhileSym0 l

data TakeWhileSym1 l l

type TakeWhileSym2 t t = TakeWhile t t

data DropWhileSym0 l

data DropWhileSym1 l l

type DropWhileSym2 t t = DropWhile t t

data SpanSym0 l

data SpanSym1 l l

type SpanSym2 t t = Span t t

data BreakSym0 l

data BreakSym1 l l

type BreakSym2 t t = Break t t

data LookupSym0 l

data LookupSym1 l l

type LookupSym2 t t = Lookup t t

data FilterSym0 l

data FilterSym1 l l

type FilterSym2 t t = Filter t t

data (:!!$) l

data l :!!$$ l

type (:!!$$$) t t = (:!!) t t