Data.Promotion.Prelude.Either

Description

Defines promoted functions and datatypes relating to Either, including a promoted version of all the definitions in Data.Either.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Either. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Promoted functions from Data.Either

either_ :: forall a c b. (a -> c) -> (b -> c) -> Either a b -> c

type family Either_ a a a :: c

Equations

Either_ f _z_1627916125 (Left x) = Apply f x

Either_ _z_1627916129 g (Right y) = Apply g y

The preceding two definitions are derived from the function either in Data.Either. The extra underscore is to avoid name clashes with the type Either.

type family Lefts a :: [a]

Equations

Lefts `[]` = `[]`

Lefts ((:) (Left x) xs) = Apply (Apply (:$) x) (Apply LeftsSym0 xs)

Lefts ((:) (Right _z_1627917585) xs) = Apply LeftsSym0 xs

type family Rights a :: [b]

Equations

Rights `[]` = `[]`

Rights ((:) (Left _z_1627917573) xs) = Apply RightsSym0 xs

Rights ((:) (Right x) xs) = Apply (Apply (:$) x) (Apply RightsSym0 xs)

type family PartitionEithers a :: ([a], [b])

Equations

PartitionEithers a_1627917527 = Apply (Apply (Apply FoldrSym0 (Apply (Apply Either_Sym0 (Let1627917534LeftSym1 a_1627917527)) (Let1627917534RightSym1 a_1627917527))) (Apply (Apply Tuple2Sym0 `[]`) `[]`)) a_1627917527

type family IsLeft a :: Bool

Equations

IsLeft (Left _z_1627917521) = TrueSym0

IsLeft (Right _z_1627917524) = FalseSym0

type family IsRight a :: Bool

Equations

IsRight (Left _z_1627917511) = FalseSym0

IsRight (Right _z_1627917514) = TrueSym0

Defunctionalization symbols

data LeftSym0 l

type LeftSym1 t = Left t

data RightSym0 l

type RightSym1 t = Right t

data Either_Sym0 l

data Either_Sym1 l l

data Either_Sym2 l l l

type Either_Sym3 t t t = Either_ t t t

data LeftsSym0 l

type LeftsSym1 t = Lefts t

data RightsSym0 l

type RightsSym1 t = Rights t

data IsLeftSym0 l

type IsLeftSym1 t = IsLeft t

data IsRightSym0 l

type IsRightSym1 t = IsRight t