Data.Promotion.Prelude.Maybe

Description

Defines promoted functions and datatypes relating to Maybe, including a promoted version of all the definitions in Data.Maybe.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Maybe. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Promoted functions from Data.Maybe

maybe_ :: forall b a. b -> (a -> b) -> Maybe a -> b

type family Maybe_ a a a :: b

Equations

Maybe_ n _z_1627933303 Nothing = n

Maybe_ _z_1627933306 f (Just x) = Apply f x

The preceding two definitions is derived from the function maybe in Data.Maybe. The extra underscore is to avoid name clashes with the type Maybe.

type family IsJust a :: Bool

Equations

IsJust Nothing = FalseSym0

IsJust (Just _z_1627934735) = TrueSym0

type family IsNothing a :: Bool

Equations

IsNothing Nothing = TrueSym0

IsNothing (Just _z_1627934728) = FalseSym0

type family FromJust a :: a

Equations

FromJust Nothing = Apply ErrorSym0 "Maybe.fromJust: Nothing"

FromJust (Just x) = x

type family FromMaybe a a :: a

Equations

FromMaybe d x = Case_1627934715 d x x

type family MaybeToList a :: [a]

Equations

MaybeToList Nothing = `[]`

MaybeToList (Just x) = Apply (Apply (:$) x) `[]`

type family ListToMaybe a :: Maybe a

Equations

ListToMaybe `[]` = NothingSym0

ListToMaybe ((:) a _z_1627934696) = Apply JustSym0 a

type family CatMaybes a :: [a]

Equations

CatMaybes `[]` = `[]`

CatMaybes ((:) (Just x) xs) = Apply (Apply (:$) x) (Apply CatMaybesSym0 xs)

CatMaybes ((:) Nothing xs) = Apply CatMaybesSym0 xs

type family MapMaybe a a :: [b]

Equations

MapMaybe _z_1627934648 `[]` = `[]`

MapMaybe f ((:) x xs) = Case_1627934680 f x xs (Let1627934667Scrutinee_1627934637Sym3 f x xs)

Defunctionalization symbols

type NothingSym0 = Nothing

data JustSym0 l

type JustSym1 t = Just t

data Maybe_Sym0 l

data Maybe_Sym1 l l

data Maybe_Sym2 l l l

type Maybe_Sym3 t t t = Maybe_ t t t

data IsJustSym0 l

type IsJustSym1 t = IsJust t

data IsNothingSym0 l

type IsNothingSym1 t = IsNothing t

data FromJustSym0 l

type FromJustSym1 t = FromJust t

data FromMaybeSym0 l

data FromMaybeSym1 l l

type FromMaybeSym2 t t = FromMaybe t t

data MaybeToListSym0 l

type MaybeToListSym1 t = MaybeToList t

data ListToMaybeSym0 l

type ListToMaybeSym1 t = ListToMaybe t

data CatMaybesSym0 l

type CatMaybesSym1 t = CatMaybes t

data MapMaybeSym0 l

data MapMaybeSym1 l l

type MapMaybeSym2 t t = MapMaybe t t