Data.Promotion.TH

Description

This module contains everything you need to promote your own functions via Template Haskell.

Primary Template Haskell generation functions

promote :: DsMonad q => q [Dec] -> q [Dec]

Promote every declaration given to the type level, retaining the originals.

promoteOnly :: DsMonad q => q [Dec] -> q [Dec]

Promote each declaration, discarding the originals. Note that a promoted datatype uses the same definition as an original datatype, so this will not work with datatypes. Classes, instances, and functions are all fine.

genDefunSymbols :: DsMonad q => [Name] -> q [Dec]

Generate defunctionalization symbols for existing type family

genPromotions :: DsMonad q => [Name] -> q [Dec]

Generate promoted definitions from a type that is already defined. This is generally only useful with classes.

Functions to generate Eq instances

promoteEqInstances :: DsMonad q => [Name] -> q [Dec]

Produce instances for '(:==)' (type-level equality) from the given types

promoteEqInstance :: DsMonad q => Name -> q [Dec]

Produce an instance for '(:==)' (type-level equality) from the given type

Functions to generate Ord instances

promoteOrdInstances :: DsMonad q => [Name] -> q [Dec]

Produce instances for POrd from the given types

promoteOrdInstance :: DsMonad q => Name -> q [Dec]

Produce an instance for POrd from the given type

Functions to generate Bounded instances

promoteBoundedInstances :: DsMonad q => [Name] -> q [Dec]

Produce instances for PBounded from the given types

promoteBoundedInstance :: DsMonad q => Name -> q [Dec]

Produce an instance for PBounded from the given type

Functions to generate Enum instances

promoteEnumInstances :: DsMonad q => [Name] -> q [Dec]

Produce instances for PEnum from the given types

promoteEnumInstance :: DsMonad q => Name -> q [Dec]

Produce an instance for PEnum from the given type

defunctionalization

data TyFun :: * -> * -> *

Representation of the kind of a type-level function. The difference between term-level arrows and this type-level arrow is that at the term level applications can be unsaturated, whereas at the type level all applications have to be fully saturated.

type family Apply f x :: k2

Type level function application

type (@@) a b = Apply a b infixl 9

An infix synonym for Apply

Auxiliary definitions

These definitions might be mentioned in code generated by Template Haskell, so they must be in scope.

class (kproxy ~ KProxy) => PEq kproxy

The promoted analogue of Eq. If you supply no definition for '(:==)', then it defaults to a use of '(==)', from Data.Type.Equality.

Associated Types

type x :== y :: Bool infix 4

type x :/= y :: Bool infix 4

type family If cond tru fls :: k

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If k True tru fls = tru

If k False tru fls = fls

type family a :&& a :: Bool infixr 3

Equations

False :&& _z_1627785679 = FalseSym0

True :&& x = x

class (PEq (KProxy :: KProxy a), kproxy ~ KProxy) => POrd kproxy

Associated Types

type Compare arg arg :: Ordering

type arg :< arg :: Bool infix 4

type arg :<= arg :: Bool infix 4

type arg :> arg :: Bool infix 4

type arg :>= arg :: Bool infix 4

type Max arg arg :: a

type Min arg arg :: a

type family Any :: k

The type constructor Any is type to which you can unsafely coerce any lifted type, and back.

It is lifted, and hence represented by a pointer

It does not claim to be a data type, and that's important for the code generator, because the code gen may enter a data value but never enters a function value.

It's also used to instantiate un-constrained type variables after type checking. For example, length has type

length :: forall a. [a] -> Int

and the list datacon for the empty list has type

[] :: forall a. [a]

In order to compose these two terms as length [] a type application is required, but there is no constraint on the choice. In this situation GHC uses Any:

length (Any *) ([] (Any *))

Above, we print kinds explicitly, as if with -fprint-explicit-kinds.

Note that Any is kind polymorphic; its kind is thus forall k. k.

data Proxy t :: k -> *

A concrete, poly-kinded proxy type

Constructors

Proxy

data KProxy t :: * -> *

A concrete, promotable proxy type, for use at the kind level There are no instances for this because it is intended at the kind level only

Constructors

KProxy

type family ThenCmp a a :: Ordering

Equations

ThenCmp EQ x = x

ThenCmp LT _z_1627819454 = LTSym0

ThenCmp GT _z_1627819457 = GTSym0

type family Foldl a a a :: b

Equations

Foldl f z0 xs0 = Apply (Apply (Let1627751476LgoSym3 f z0 xs0) z0) xs0

type family Error str :: k

The promotion of error. This version is more poly-kinded for easier use.

data ErrorSym0 l

type TrueSym0 = True

type FalseSym0 = False

type LTSym0 = LT

type EQSym0 = EQ

type GTSym0 = GT

type Tuple0Sym0 = `()`

data Tuple2Sym0 l

data Tuple2Sym1 l l

type Tuple2Sym2 t t = `(t, t)`

data Tuple3Sym0 l

data Tuple3Sym1 l l

data Tuple3Sym2 l l l

type Tuple3Sym3 t t t = `(t, t, t)`

data Tuple4Sym0 l

data Tuple4Sym1 l l

data Tuple4Sym2 l l l

data Tuple4Sym3 l l l l

type Tuple4Sym4 t t t t = `(t, t, t, t)`

data Tuple5Sym0 l

data Tuple5Sym1 l l

data Tuple5Sym2 l l l

data Tuple5Sym3 l l l l

data Tuple5Sym4 l l l l l

type Tuple5Sym5 t t t t t = `(t, t, t, t, t)`

data Tuple6Sym0 l

data Tuple6Sym1 l l

data Tuple6Sym2 l l l

data Tuple6Sym3 l l l l

data Tuple6Sym4 l l l l l

data Tuple6Sym5 l l l l l l

type Tuple6Sym6 t t t t t t = `(t, t, t, t, t, t)`

data Tuple7Sym0 l

data Tuple7Sym1 l l

data Tuple7Sym2 l l l

data Tuple7Sym3 l l l l

data Tuple7Sym4 l l l l l

data Tuple7Sym5 l l l l l l

data Tuple7Sym6 l l l l l l l

type Tuple7Sym7 t t t t t t t = `(t, t, t, t, t, t, t)`

data ThenCmpSym0 l

data FoldlSym0 l

class SuppressUnusedWarnings t where

This class (which users should never see) is to be instantiated in order to use an otherwise-unused data constructor, such as the "kind-inference" data constructor for defunctionalization symbols.

Methods

suppressUnusedWarnings :: Proxy t -> ()