Data.Singletons

Description

This module exports the basic definitions to use singletons. For routine use, consider importing Prelude, which exports constructors for singletons based on types in the Prelude.

You may also want to read http://www.cis.upenn.edu/~eir/packages/singletons/README.html and the original paper presenting this library, available at http://www.cis.upenn.edu/~eir/papers/2012/singletons/paper.pdf.

Main singleton definitions

data family Sing a

The singleton kind-indexed data family.

See also Sing for exported constructors

class SingI a where

A SingI constraint is essentially an implicitly-passed singleton. If you need to satisfy this constraint with an explicit singleton, please see withSingI.

Methods

sing :: Sing a

Produce the singleton explicitly. You will likely need the ScopedTypeVariables extension to use this method the way you want.

class kparam ~ KProxy => SingKind kparam where

The SingKind class is essentially a kind class. It classifies all kinds for which singletons are defined. The class supports converting between a singleton type and the base (unrefined) type which it is built from.

Associated Types

type DemoteRep kparam :: *

Get a base type from a proxy for the promoted kind. For example, DemoteRep ('KProxy :: KProxy Bool) will be the type Bool.

Methods

fromSing :: Sing (a :: k) -> DemoteRep kparam

Convert a singleton to its unrefined version.

toSing :: DemoteRep kparam -> SomeSing kparam

Convert an unrefined type to an existentially-quantified singleton type.

Working with singletons

type KindOf a = (KProxy :: KProxy k)

Convenient synonym to refer to the kind of a type variable: type KindOf (a :: k) = ('KProxy :: KProxy k)

type Demote a = DemoteRep (KProxy :: KProxy k)

Convenient abbreviation for DemoteRep: type Demote (a :: k) = DemoteRep ('KProxy :: KProxy k)

data SingInstance a where

A SingInstance wraps up a SingI instance for explicit handling.

Constructors

SingInstance :: SingI a => SingInstance a

data SomeSing kproxy where

An existentially-quantified singleton. This type is useful when you want a singleton type, but there is no way of knowing, at compile-time, what the type index will be. To make use of this type, you will generally have to use a pattern-match:

 foo :: Bool -> ...
 foo b = case toSing b of
           SomeSing sb -> {- fancy dependently-typed code with sb -}

An example like the one above may be easier to write using withSomeSing.

Constructors

SomeSing :: Sing (a :: k) -> SomeSing (KProxy :: KProxy k)

singInstance :: forall a. Sing a -> SingInstance a

Get an implicit singleton (a SingI instance) from an explicit one.

withSingI :: Sing n -> (SingI n => r) -> r

Convenience function for creating a context with an implicit singleton available.

withSomeSing

Arguments

:: SingKind (KProxy :: KProxy k)

=> DemoteRep (KProxy :: KProxy k)
The original datatype

-> (forall a. Sing a -> r)
Function expecting a singleton

-> r

Convert a normal datatype (like Bool) to a singleton for that datatype, passing it into a continuation.

singByProxy :: SingI a => proxy a -> Sing a

Allows creation of a singleton when a proxy is at hand.

singByProxy# :: SingI a => Proxy# a -> Sing a

Allows creation of a singleton when a proxy# is at hand.

withSing :: SingI a => (Sing a -> b) -> b

A convenience function useful when we need to name a singleton value multiple times. Without this function, each use of sing could potentially refer to a different singleton, and one has to use type signatures (often with ScopedTypeVariables) to ensure that they are the same.

singThat :: forall a. (SingKind (KProxy :: KProxy k), SingI a) => (Demote a -> Bool) -> Maybe (Sing a)

A convenience function that names a singleton satisfying a certain property. If the singleton does not satisfy the property, then the function returns Nothing. The property is expressed in terms of the underlying representation of the singleton.

Auxiliary functions

bugInGHC :: forall a. a

GHC 7.8 sometimes warns about incomplete pattern matches when no such patterns are possible, due to GADT constraints. See the bug report at https://ghc.haskell.org/trac/ghc/ticket/3927. In such cases, it's useful to have a catch-all pattern that then has bugInGHC as its right-hand side.

type family Error str :: k

The promotion of error

sError :: Sing (str :: Symbol) -> a

The singleton for error

data KProxy t

A concrete, promotable proxy type, for use at the kind level There are no instances for this because it is intended at the kind level only

Constructors

KProxy

data Proxy t

A concrete, poly-kinded proxy type

Constructors

Proxy