Data.Singletons.Decide

Description

Defines the class SDecide, allowing for decidable equality over singletons.

The SDecide class

class kparam ~ KProxy => SDecide kparam where

Members of the SDecide kind class support decidable equality. Instances of this class are generated alongside singleton definitions for datatypes that derive an Eq instance.

Methods

(%~) :: forall a b. Sing a -> Sing b -> Decision (a :~: b)

Compute a proof or disproof of equality, given two singletons.

Supporting definitions

data a :~: b where

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: :~: k a1 a1

data Void

A logically uninhabited data type.

type Refuted a = a -> Void

Because we can never create a value of type Void, a function that type-checks at a -> Void shows that objects of type a can never exist. Thus, we say that a is Refuted

data Decision a

A Decision about a type a is either a proof of existence or a proof that a cannot exist.

Constructors

Proved a
Witness for a

Disproved (Refuted a)
Proof that no a exists