Data.Singletons.Maybe

Description

Defines functions and datatypes relating to the singleton for Maybe, including a singletons version of all the definitions in Data.Maybe.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Maybe. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Documentation

data family Sing a

The singleton kind-indexed data family.

Though Haddock doesn't show it, the Sing instance above declares constructors

SNothing :: Sing Nothing
SJust    :: Sing a -> Sing (Just a)

type SMaybe z = Sing z

SBool is a kind-restricted synonym for Sing: type SMaybe (a :: Maybe k) = Sing a

Singletons from Data.Maybe

type family Maybe_ a a a :: b

Equations

Maybe_ n z Nothing = n

Maybe_ z f (Just x) = f x

sMaybe_ :: forall t t t. Sing t -> (forall t. Sing t -> Sing (t t)) -> Sing t -> Sing (Maybe_ t t t)

The preceding two definitions are derived from the function maybe in Data.Maybe. The extra underscore is to avoid name clashes with the type Maybe.

type family IsJust a :: Bool

Equations

IsJust Nothing = False

IsJust (Just z) = True

sIsJust :: forall t. Sing t -> Sing (IsJust t)

type family IsNothing a :: Bool

Equations

IsNothing Nothing = True

IsNothing (Just z) = False

sIsNothing :: forall t. Sing t -> Sing (IsNothing t)

type family FromJust a :: a

Equations

FromJust Nothing = Error "Maybe.fromJust: Nothing"

FromJust (Just x) = x

sFromJust :: forall t. Sing t -> Sing (FromJust t)

type family FromMaybe a a :: a

Equations

FromMaybe d Nothing = d

FromMaybe z (Just v) = v

sFromMaybe :: forall t t. Sing t -> Sing t -> Sing (FromMaybe t t)

type family MaybeToList a :: [a]

Equations

MaybeToList Nothing = `[]`

MaybeToList (Just x) = `[x]`

sMaybeToList :: forall t. Sing t -> Sing (MaybeToList t)

type family ListToMaybe a :: Maybe a

Equations

ListToMaybe [] = Nothing

ListToMaybe ((:) a z) = Just a

sListToMaybe :: forall t. Sing t -> Sing (ListToMaybe t)

type family CatMaybes a :: [a]

Equations

CatMaybes [] = `[]`

CatMaybes ((:) (Just x) xs) = (:) x (CatMaybes xs)

CatMaybes ((:) Nothing xs) = CatMaybes xs

sCatMaybes :: forall t. Sing t -> Sing (CatMaybes t)

type family MapMaybe a a :: [b]

Equations

MapMaybe z [] = `[]`

MapMaybe f ((:) x xs) = (:++) (MaybeToList (f x)) (MapMaybe f xs)

sMapMaybe :: forall t t. (forall t. Sing t -> Sing (t t)) -> Sing t -> Sing (MapMaybe t t)