Data.Singletons.Prelude.Eq

Description

Defines the SEq singleton version of the Eq type class.

Documentation

class PEq a

The promoted analogue of Eq. If you supply no definition for (==), then it defaults to a use of DefaultEq.

Associated Types

type (x :: a) == (y :: a) :: Bool infix 4

type (==) (x :: a) (y :: a) = x `DefaultEq` y

type (x :: a) /= (y :: a) :: Bool infix 4

type (/=) (x :: a) (y :: a) = Not (x == y)

class SEq k where

The singleton analogue of Eq. Unlike the definition for Eq, it is required that instances define a body for (%==). You may also supply a body for (%/=).

Minimal complete definition

(%==)

Methods

(%==) :: forall (a :: k) (b :: k). Sing a -> Sing b -> Sing (a == b) infix 4

Boolean equality on singletons

(%/=) :: forall (a :: k) (b :: k). Sing a -> Sing b -> Sing (a /= b) infix 4

Boolean disequality on singletons

default (%/=) :: forall (a :: k) (b :: k). (a /= b) ~ Not (a == b) => Sing a -> Sing b -> Sing (a /= b)

type family DefaultEq (a :: k) (b :: k) :: Bool where ...

A sensible way to compute Boolean equality for types of any kind. Note that this definition is slightly different from the (==) type family from Data.Type.Equality in base, as (==) attempts to distinguish applications of type constructors from other types. As a result, a == a does not reduce to True for every a, but DefaultEq a a does reduce to True for every a. The latter behavior is more desirable for singletons' purposes, so we use it instead of (==).

Equations

DefaultEq a a = 'True

DefaultEq a b = 'False

Defunctionalization symbols

data (==@#@$) :: forall a6989586621679381571. (~>) a6989586621679381571 ((~>) a6989586621679381571 Bool) infix 4

data (==@#@$$) (x6989586621679381572 :: a6989586621679381571) :: (~>) a6989586621679381571 Bool infix 4

type (==@#@$$$) (x6989586621679381572 :: a6989586621679381571) (y6989586621679381573 :: a6989586621679381571) = (==) x6989586621679381572 y6989586621679381573

data (/=@#@$) :: forall a6989586621679381571. (~>) a6989586621679381571 ((~>) a6989586621679381571 Bool) infix 4

data (/=@#@$$) (x6989586621679381574 :: a6989586621679381571) :: (~>) a6989586621679381571 Bool infix 4

type (/=@#@$$$) (x6989586621679381574 :: a6989586621679381571) (y6989586621679381575 :: a6989586621679381571) = (/=) x6989586621679381574 y6989586621679381575

data DefaultEqSym0 :: forall k6989586621679381565. (~>) k6989586621679381565 ((~>) k6989586621679381565 Bool)

data DefaultEqSym1 (a6989586621679381566 :: k6989586621679381565) :: (~>) k6989586621679381565 Bool

type DefaultEqSym2 (a6989586621679381566 :: k6989586621679381565) (b6989586621679381567 :: k6989586621679381565) = DefaultEq a6989586621679381566 b6989586621679381567