Data.Singletons.Prelude.Function

Description

Defines singleton versions of the definitions in Data.Function.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Function. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Prelude re-exports

type family Id a where ...

Equations

Id x = x

sId :: forall a (t :: a). Sing t -> Sing (Apply IdSym0 t :: a)

type family Const a a where ...

Equations

Const x _ = x

sConst :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply ConstSym0 t) t :: a)

type family (a . a) a where ... infixr 9

Equations

(f . g) a_6989586621679534299 = Apply (Apply (Apply (Apply Lambda_6989586621679534311Sym0 f) g) a_6989586621679534299) a_6989586621679534299

(%.) :: forall b c a (t :: (~>) b c) (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (.@#@$) t) t) t :: c) infixr 9

type family Flip a a a where ...

Equations

Flip f x y = Apply (Apply f y) x

sFlip :: forall a b c (t :: (~>) a ((~>) b c)) (t :: b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FlipSym0 t) t) t :: c)

type family a $ a where ... infixr 0

Equations

f $ x = Apply f x

(%$) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($@#@$) t) t :: b) infixr 0

Other combinators

type family a & a where ... infixl 1

Equations

x & f = Apply f x

(%&) :: forall a b (t :: a) (t :: (~>) a b). Sing t -> Sing t -> Sing (Apply (Apply (&@#@$) t) t :: b) infixl 1

type family On a a a a where ... infixl 0

Equations

On ty f a_6989586621679745822 a_6989586621679745824 = Apply (Apply (Apply (Apply (Apply (Apply Lambda_6989586621679745839Sym0 ty) f) a_6989586621679745822) a_6989586621679745824) a_6989586621679745822) a_6989586621679745824

sOn :: forall b c a (t :: (~>) b ((~>) b c)) (t :: (~>) a b) (t :: a) (t :: a). Sing t -> Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (Apply OnSym0 t) t) t) t :: c) infixl 0

Defunctionalization symbols

data IdSym0 a6989586621679534323

type IdSym1 (a6989586621679534323 :: a) = Id a6989586621679534323 :: a

data ConstSym0 a6989586621679534318

data ConstSym1 a6989586621679534318 a6989586621679534319

type ConstSym2 (a6989586621679534318 :: a) (a6989586621679534319 :: b) = Const a6989586621679534318 a6989586621679534319 :: a

data (.@#@$) a6989586621679534305 infixr 9

data a6989586621679534305 .@#@$$ a6989586621679534306 infixr 9

data (a6989586621679534305 .@#@$$$ a6989586621679534306) a6989586621679534307 infixr 9

type (.@#@$$$$) (a6989586621679534305 :: (~>) b c) (a6989586621679534306 :: (~>) a b) (a6989586621679534307 :: a) = (.) a6989586621679534305 a6989586621679534306 a6989586621679534307 :: c infixr 9

data FlipSym0 a6989586621679534293

data FlipSym1 a6989586621679534293 a6989586621679534294

data FlipSym2 a6989586621679534293 a6989586621679534294 a6989586621679534295

type FlipSym3 (a6989586621679534293 :: (~>) a ((~>) b c)) (a6989586621679534294 :: b) (a6989586621679534295 :: a) = Flip a6989586621679534293 a6989586621679534294 a6989586621679534295 :: c

data ($@#@$) a6989586621679534274 infixr 0

data a6989586621679534274 $@#@$$ a6989586621679534275 infixr 0

type ($@#@$$$) (a6989586621679534274 :: (~>) a b) (a6989586621679534275 :: a) = ($) a6989586621679534274 a6989586621679534275 :: b infixr 0

data (&@#@$) a6989586621679745818 infixl 1

data a6989586621679745818 &@#@$$ a6989586621679745819 infixl 1

type (&@#@$$$) (a6989586621679745818 :: a) (a6989586621679745819 :: (~>) a b) = (&) a6989586621679745818 a6989586621679745819 :: b infixl 1

data OnSym0 a6989586621679745831 infixl 0

data OnSym1 a6989586621679745831 a6989586621679745832 infixl 0

data OnSym2 a6989586621679745831 a6989586621679745832 a6989586621679745833 infixl 0

data OnSym3 a6989586621679745831 a6989586621679745832 a6989586621679745833 a6989586621679745834 infixl 0

type OnSym4 (a6989586621679745831 :: (~>) b ((~>) b c)) (a6989586621679745832 :: (~>) a b) (a6989586621679745833 :: a) (a6989586621679745834 :: a) = On a6989586621679745831 a6989586621679745832 a6989586621679745833 a6989586621679745834 :: c infixl 0