Data.Singletons.Prelude.List

Description

Defines functions and datatypes relating to the singleton for '[]', including a singletons version of a few of the definitions in Data.List.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.List. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The singleton for lists

data family Sing a

The singleton kind-indexed data family.

Though Haddock doesn't show it, the Sing instance above declares constructors

SNil  :: Sing '[]
SCons :: Sing (h :: k) -> Sing (t :: [k]) -> Sing (h ': t)

type SList = (Sing :: [a] -> *)

SList is a kind-restricted synonym for Sing: type SList (a :: [k]) = Sing a

Basic functions

type family a :++ a :: [a] infixr 5

Equations

`[]` :++ ys = ys

((:) x xs) :++ ys = Apply (Apply (:$) x) (Apply (Apply (:++$) xs) ys)

(%:++) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:++$) t) t :: [a]) infixr 5

type family Head a :: a

Equations

Head ((:) a _z_1628029340) = a

Head `[]` = Apply ErrorSym0 "Data.Singletons.List.head: empty list"

sHead :: forall t. Sing t -> Sing (Apply HeadSym0 t :: a)

type family Last a :: a

Equations

Last `[]` = Apply ErrorSym0 "Data.Singletons.List.last: empty list"

Last `[x]` = x

Last ((:) _z_1628029330 ((:) x xs)) = Apply LastSym0 (Apply (Apply (:$) x) xs)

sLast :: forall t. Sing t -> Sing (Apply LastSym0 t :: a)

type family Tail a :: [a]

Equations

Tail ((:) _z_1628029321 t) = t

Tail `[]` = Apply ErrorSym0 "Data.Singletons.List.tail: empty list"

sTail :: forall t. Sing t -> Sing (Apply TailSym0 t :: [a])

type family Init a :: [a]

Equations

Init `[]` = Apply ErrorSym0 "Data.Singletons.List.init: empty list"

Init ((:) x xs) = Apply (Apply (Let1628029290Init'Sym2 x xs) x) xs

sInit :: forall t. Sing t -> Sing (Apply InitSym0 t :: [a])

type family Null a :: Bool

Equations

Null `[]` = TrueSym0

Null ((:) _z_1628029219 _z_1628029222) = FalseSym0

sNull :: forall t. Sing t -> Sing (Apply NullSym0 t :: Bool)

type family Length a :: Nat

Equations

Length `[]` = FromInteger 0

Length ((:) _z_1628026134 xs) = Apply (Apply (:+$) (FromInteger 1)) (Apply LengthSym0 xs)

sLength :: forall t. Sing t -> Sing (Apply LengthSym0 t :: Nat)

List transformations

type family Map a a :: [b]

Equations

Map _z_1627883819 `[]` = `[]`

Map f ((:) x xs) = Apply (Apply (:$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

sMap :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: [b])

type family Reverse a :: [a]

Equations

Reverse l = Apply (Apply (Let1628029185RevSym1 l) l) `[]`

sReverse :: forall t. Sing t -> Sing (Apply ReverseSym0 t :: [a])

type family Intersperse a a :: [a]

Equations

Intersperse _z_1628029163 `[]` = `[]`

Intersperse sep ((:) x xs) = Apply (Apply (:$) x) (Apply (Apply PrependToAllSym0 sep) xs)

sIntersperse :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply IntersperseSym0 t) t :: [a])

type family Intercalate a a :: [a]

Equations

Intercalate xs xss = Apply ConcatSym0 (Apply (Apply IntersperseSym0 xs) xss)

sIntercalate :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply IntercalateSym0 t) t :: [a])

type family Transpose a :: [[a]]

Equations

Transpose `[]` = `[]`

Transpose ((:) `[]` xss) = Apply TransposeSym0 xss

Transpose ((:) ((:) x xs) xss) = Apply (Apply (:$) (Apply (Apply (:$) x) (Apply (Apply MapSym0 HeadSym0) xss))) (Apply TransposeSym0 (Apply (Apply (:$) xs) (Apply (Apply MapSym0 TailSym0) xss)))

sTranspose :: forall t. Sing t -> Sing (Apply TransposeSym0 t :: [[a]])

type family Subsequences a :: [[a]]

Equations

Subsequences xs = Apply (Apply (:$) `[]`) (Apply NonEmptySubsequencesSym0 xs)

sSubsequences :: forall t. Sing t -> Sing (Apply SubsequencesSym0 t :: [[a]])

type family Permutations a :: [[a]]

Equations

Permutations xs0 = Apply (Apply (:$) xs0) (Apply (Apply (Let1628028738PermsSym1 xs0) xs0) `[]`)

sPermutations :: forall t. Sing t -> Sing (Apply PermutationsSym0 t :: [[a]])

Reducing lists (folds)

type family Foldl a a a :: b

Equations

Foldl f z0 xs0 = Apply (Apply (Let1627751476LgoSym3 f z0 xs0) z0) xs0

sFoldl :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldlSym0 t) t) t :: b)

type family Foldl' a a a :: b

Equations

Foldl' f z0 xs0 = Apply (Apply (Let1628028652LgoSym3 f z0 xs0) z0) xs0

sFoldl' :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Foldl'Sym0 t) t) t :: b)

type family Foldl1 a a :: a

Equations

Foldl1 f ((:) x xs) = Apply (Apply (Apply FoldlSym0 f) x) xs

Foldl1 _z_1628028429 `[]` = Apply ErrorSym0 "Data.Singletons.List.foldl1: empty list"

sFoldl1 :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Foldl1Sym0 t) t :: a)

type family Foldl1' a a :: a

Equations

Foldl1' f ((:) x xs) = Apply (Apply (Apply Foldl'Sym0 f) x) xs

Foldl1' _z_1628028730 `[]` = Apply ErrorSym0 "Data.Singletons.List.foldl1': empty list"

sFoldl1' :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Foldl1'Sym0 t) t :: a)

type family Foldr a a a :: b

Equations

Foldr k z a_1627883840 = Apply (Let1627883845GoSym3 k z a_1627883840) a_1627883840

sFoldr :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldrSym0 t) t) t :: b)

type family Foldr1 a a :: a

Equations

Foldr1 _z_1628028387 `[x]` = x

Foldr1 f ((:) x ((:) wild_1628025871 wild_1628025873)) = Apply (Apply f x) (Apply (Apply Foldr1Sym0 f) (Let1628028395XsSym4 f x wild_1628025871 wild_1628025873))

Foldr1 _z_1628028414 `[]` = Apply ErrorSym0 "Data.Singletons.List.foldr1: empty list"

sFoldr1 :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Foldr1Sym0 t) t :: a)

Special folds

type family Concat a :: [a]

Equations

Concat a_1628028371 = Apply (Apply (Apply FoldrSym0 (:++$)) `[]`) a_1628028371

sConcat :: forall t. Sing t -> Sing (Apply ConcatSym0 t :: [a])

type family ConcatMap a a :: [b]

Equations

ConcatMap f a_1628028367 = Apply (Apply (Apply FoldrSym0 (Apply (Apply (:.$) (:++$)) f)) `[]`) a_1628028367

sConcatMap :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply ConcatMapSym0 t) t :: [b])

type family And a :: Bool

Equations

And `[]` = TrueSym0

And ((:) x xs) = Apply (Apply (:&&$) x) (Apply AndSym0 xs)

sAnd :: forall t. Sing t -> Sing (Apply AndSym0 t :: Bool)

type family Or a :: Bool

Equations

Or `[]` = FalseSym0

Or ((:) x xs) = Apply (Apply (:||$) x) (Apply OrSym0 xs)

sOr :: forall t. Sing t -> Sing (Apply OrSym0 t :: Bool)

type family Any_ a a :: Bool

Equations

Any_ _z_1628014649 `[]` = FalseSym0

Any_ p ((:) x xs) = Apply (Apply (:||$) (Apply p x)) (Apply (Apply Any_Sym0 p) xs)

sAny_ :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Any_Sym0 t) t :: Bool)

type family All a a :: Bool

Equations

All _z_1628028340 `[]` = TrueSym0

All p ((:) x xs) = Apply (Apply (:&&$) (Apply p x)) (Apply (Apply AllSym0 p) xs)

sAll :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply AllSym0 t) t :: Bool)

type family Sum a :: a

Equations

Sum l = Apply (Apply (Let1628026167Sum'Sym1 l) l) (FromInteger 0)

sSum :: forall t. SNum (KProxy :: KProxy a) => Sing t -> Sing (Apply SumSym0 t :: a)

type family Product a :: a

Equations

Product l = Apply (Apply (Let1628026143ProdSym1 l) l) (FromInteger 1)

sProduct :: forall t. SNum (KProxy :: KProxy a) => Sing t -> Sing (Apply ProductSym0 t :: a)

type family Maximum a :: a

Equations

Maximum `[]` = Apply ErrorSym0 "Data.Singletons.List.maximum: empty list"

Maximum ((:) wild_1628025951 wild_1628025953) = Apply (Apply Foldl1Sym0 MaxSym0) (Let1628028612XsSym2 wild_1628025951 wild_1628025953)

sMaximum :: forall t. SOrd (KProxy :: KProxy a) => Sing t -> Sing (Apply MaximumSym0 t :: a)

type family Minimum a :: a

Equations

Minimum `[]` = Apply ErrorSym0 "Data.Singletons.List.minimum: empty list"

Minimum ((:) wild_1628025955 wild_1628025957) = Apply (Apply Foldl1Sym0 MinSym0) (Let1628028626XsSym2 wild_1628025955 wild_1628025957)

sMinimum :: forall t. SOrd (KProxy :: KProxy a) => Sing t -> Sing (Apply MinimumSym0 t :: a)

any_ :: forall a. (a -> Bool) -> [a] -> Bool

Building lists

Scans

type family Scanl a a a :: [b]

Equations

Scanl f q ls = Apply (Apply (:$) q) (Case_1628028311 f q ls ls)

sScanl :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanlSym0 t) t) t :: [b])

type family Scanl1 a a :: [a]

Equations

Scanl1 f ((:) x xs) = Apply (Apply (Apply ScanlSym0 f) x) xs

Scanl1 _z_1628028328 `[]` = `[]`

sScanl1 :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Scanl1Sym0 t) t :: [a])

type family Scanr a a a :: [b]

Equations

Scanr _z_1628028261 q0 `[]` = Apply (Apply (:$) q0) `[]`

Scanr f q0 ((:) x xs) = Case_1628028288 f q0 x xs (Let1628028269Scrutinee_1628025875Sym4 f q0 x xs)

sScanr :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanrSym0 t) t) t :: [b])

type family Scanr1 a a :: [a]

Equations

Scanr1 _z_1628028192 `[]` = `[]`

Scanr1 _z_1628028195 `[x]` = Apply (Apply (:$) x) `[]`

Scanr1 f ((:) x ((:) wild_1628025879 wild_1628025881)) = Case_1628028241 f x wild_1628025879 wild_1628025881 (Let1628028222Scrutinee_1628025877Sym4 f x wild_1628025879 wild_1628025881)

sScanr1 :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Scanr1Sym0 t) t :: [a])

Accumulating maps

type family MapAccumL a a a :: (acc, [y])

Equations

MapAccumL _z_1628028026 s `[]` = Apply (Apply Tuple2Sym0 s) `[]`

MapAccumL f s ((:) x xs) = Apply (Apply Tuple2Sym0 (Let1628028034S''Sym4 f s x xs)) (Apply (Apply (:$) (Let1628028034YSym4 f s x xs)) (Let1628028034YsSym4 f s x xs))

sMapAccumL :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply MapAccumLSym0 t) t) t :: (acc, [y]))

type family MapAccumR a a a :: (acc, [y])

Equations

MapAccumR _z_1628027854 s `[]` = Apply (Apply Tuple2Sym0 s) `[]`

MapAccumR f s ((:) x xs) = Apply (Apply Tuple2Sym0 (Let1628027862S''Sym4 f s x xs)) (Apply (Apply (:$) (Let1628027862YSym4 f s x xs)) (Let1628027862YsSym4 f s x xs))

sMapAccumR :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply MapAccumRSym0 t) t) t :: (acc, [y]))

Cyclical lists

type family Replicate a a :: [a]

Equations

Replicate n x = Case_1628026127 n x (Let1628026119Scrutinee_1628025963Sym2 n x)

sReplicate :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply ReplicateSym0 t) t :: [a])

Unfolding

type family Unfoldr a a :: [a]

Equations

Unfoldr f b = Case_1628027834 f b (Let1628027826Scrutinee_1628025883Sym2 f b)

sUnfoldr :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply UnfoldrSym0 t) t :: [a])

Sublists

Extracting sublists

type family Take a a :: [a]

Equations

Take _z_1628026314 `[]` = `[]`

Take n ((:) x xs) = Case_1628026333 n x xs (Let1628026320Scrutinee_1628025947Sym3 n x xs)

sTake :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply TakeSym0 t) t :: [a])

type family Drop a a :: [a]

Equations

Drop _z_1628026283 `[]` = `[]`

Drop n ((:) x xs) = Case_1628026302 n x xs (Let1628026289Scrutinee_1628025949Sym3 n x xs)

sDrop :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply DropSym0 t) t :: [a])

type family SplitAt a a :: ([a], [a])

Equations

SplitAt n xs = Apply (Apply Tuple2Sym0 (Apply (Apply TakeSym0 n) xs)) (Apply (Apply DropSym0 n) xs)

sSplitAt :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply SplitAtSym0 t) t :: ([a], [a]))

type family TakeWhile a a :: [a]

Equations

TakeWhile _z_1628026687 `[]` = `[]`

TakeWhile p ((:) x xs) = Case_1628026706 p x xs (Let1628026693Scrutinee_1628025937Sym3 p x xs)

sTakeWhile :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply TakeWhileSym0 t) t :: [a])

type family DropWhile a a :: [a]

Equations

DropWhile _z_1628026643 `[]` = `[]`

DropWhile p ((:) x xs') = Case_1628026675 p x xs' (Let1628026662Scrutinee_1628025939Sym3 p x xs')

sDropWhile :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply DropWhileSym0 t) t :: [a])

type family DropWhileEnd a a :: [a]

Equations

DropWhileEnd p a_1628029234 = Apply (Apply (Apply FoldrSym0 (Apply (Apply Lambda_1628029238Sym0 p) a_1628029234)) `[]`) a_1628029234

sDropWhileEnd :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply DropWhileEndSym0 t) t :: [a])

type family Span a a :: ([a], [a])

Equations

Span _z_1628026461 `[]` = Apply (Apply Tuple2Sym0 (Let1628026464XsSym1 _z_1628026461)) (Let1628026464XsSym1 _z_1628026461)

Span p ((:) x xs') = Case_1628026497 p x xs' (Let1628026484Scrutinee_1628025943Sym3 p x xs')

sSpan :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply SpanSym0 t) t :: ([a], [a]))

type family Break a a :: ([a], [a])

Equations

Break _z_1628026356 `[]` = Apply (Apply Tuple2Sym0 (Let1628026359XsSym1 _z_1628026356)) (Let1628026359XsSym1 _z_1628026356)

Break p ((:) x xs') = Case_1628026392 p x xs' (Let1628026379Scrutinee_1628025945Sym3 p x xs')

sBreak :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply BreakSym0 t) t :: ([a], [a]))

type family Group a :: [[a]]

Equations

Group xs = Apply (Apply GroupBySym0 (:==$)) xs

sGroup :: forall t. SEq (KProxy :: KProxy a) => Sing t -> Sing (Apply GroupSym0 t :: [[a]])

type family Inits a :: [[a]]

Equations

Inits xs = Apply (Apply (:$) `[]`) (Case_1628027810 xs xs)

sInits :: forall t. Sing t -> Sing (Apply InitsSym0 t :: [[a]])

type family Tails a :: [[a]]

Equations

Tails xs = Apply (Apply (:$) xs) (Case_1628027787 xs xs)

sTails :: forall t. Sing t -> Sing (Apply TailsSym0 t :: [[a]])

Predicates

type family IsPrefixOf a a :: Bool

Equations

IsPrefixOf `[]` `[]` = TrueSym0

IsPrefixOf `[]` ((:) _z_1628027766 _z_1628027769) = TrueSym0

IsPrefixOf ((:) _z_1628027772 _z_1628027775) `[]` = FalseSym0

IsPrefixOf ((:) x xs) ((:) y ys) = Apply (Apply (:&&$) (Apply (Apply (:==$) x) y)) (Apply (Apply IsPrefixOfSym0 xs) ys)

sIsPrefixOf :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply IsPrefixOfSym0 t) t :: Bool)

type family IsSuffixOf a a :: Bool

Equations

IsSuffixOf x y = Apply (Apply IsPrefixOfSym0 (Apply ReverseSym0 x)) (Apply ReverseSym0 y)

sIsSuffixOf :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply IsSuffixOfSym0 t) t :: Bool)

type family IsInfixOf a a :: Bool

Equations

IsInfixOf needle haystack = Apply (Apply Any_Sym0 (Apply IsPrefixOfSym0 needle)) (Apply TailsSym0 haystack)

sIsInfixOf :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply IsInfixOfSym0 t) t :: Bool)

Searching lists

Searching by equality

type family Elem a a :: Bool

Equations

Elem _z_1628027703 `[]` = FalseSym0

Elem x ((:) y ys) = Apply (Apply (:||$) (Apply (Apply (:==$) x) y)) (Apply (Apply ElemSym0 x) ys)

sElem :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply ElemSym0 t) t :: Bool)

type family NotElem a a :: Bool

Equations

NotElem _z_1628027688 `[]` = TrueSym0

NotElem x ((:) y ys) = Apply (Apply (:&&$) (Apply (Apply (:/=$) x) y)) (Apply (Apply NotElemSym0 x) ys)

sNotElem :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply NotElemSym0 t) t :: Bool)

type family Lookup a a :: Maybe b

Equations

Lookup _key `[]` = NothingSym0

Lookup key ((:) `(x, y)` xys) = Case_1628026271 key x y xys (Let1628026252Scrutinee_1628025959Sym4 key x y xys)

sLookup :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply LookupSym0 t) t :: Maybe b)

Searching with a predicate

type family Find a a :: Maybe a

Equations

Find p a_1628026747 = Apply (Apply (Apply (:.$) ListToMaybeSym0) (Apply FilterSym0 p)) a_1628026747

sFind :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply FindSym0 t) t :: Maybe a)

type family Filter a a :: [a]

Equations

Filter _p `[]` = `[]`

Filter p ((:) x xs) = Case_1628026735 p x xs (Let1628026722Scrutinee_1628025925Sym3 p x xs)

sFilter :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply FilterSym0 t) t :: [a])

type family Partition a a :: ([a], [a])

Equations

Partition p xs = Apply (Apply (Apply FoldrSym0 (Apply SelectSym0 p)) (Apply (Apply Tuple2Sym0 `[]`) `[]`)) xs

sPartition :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply PartitionSym0 t) t :: ([a], [a]))

Indexing lists

type family a :!! a :: a

Equations

`[]` :!! _z_1628026086 = Apply ErrorSym0 "Data.Singletons.List.!!: index too large"

((:) x xs) :!! n = Case_1628026105 x xs n (Let1628026092Scrutinee_1628025965Sym3 x xs n)

(%:!!) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:!!$) t) t :: a)

type family ElemIndex a a :: Maybe Nat

Equations

ElemIndex x a_1628027675 = Apply (Apply FindIndexSym0 (Apply (:==$) x)) a_1628027675

sElemIndex :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply ElemIndexSym0 t) t :: Maybe Nat)

type family ElemIndices a a :: [Nat]

Equations

ElemIndices x a_1628027649 = Apply (Apply FindIndicesSym0 (Apply (:==$) x)) a_1628027649

sElemIndices :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply ElemIndicesSym0 t) t :: [Nat])

type family FindIndex a a :: Maybe Nat

Equations

FindIndex p a_1628027662 = Apply (Apply (Apply (:.$) ListToMaybeSym0) (Apply FindIndicesSym0 p)) a_1628027662

sFindIndex :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply FindIndexSym0 t) t :: Maybe Nat)

type family FindIndices a a :: [Nat]

Equations

FindIndices p xs = Apply (Apply MapSym0 SndSym0) (Apply (Apply FilterSym0 (Apply (Apply Lambda_1628027617Sym0 p) xs)) (Apply (Apply ZipSym0 xs) (Apply (Apply (Let1628027588BuildListSym2 p xs) (FromInteger 0)) xs)))

sFindIndices :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply FindIndicesSym0 t) t :: [Nat])

Zipping and unzipping lists

type family Zip a a :: [(a, b)]

Equations

Zip ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply Tuple2Sym0 x) y)) (Apply (Apply ZipSym0 xs) ys)

Zip `[]` `[]` = `[]`

Zip ((:) _z_1628027565 _z_1628027568) `[]` = `[]`

Zip `[]` ((:) _z_1628027571 _z_1628027574) = `[]`

sZip :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply ZipSym0 t) t :: [(a, b)])

type family Zip3 a a a :: [(a, b, c)]

Equations

Zip3 ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply Tuple3Sym0 a) b) c)) (Apply (Apply (Apply Zip3Sym0 as) bs) cs)

Zip3 `[]` `[]` `[]` = `[]`

Zip3 `[]` `[]` ((:) _z_1628027498 _z_1628027501) = `[]`

Zip3 `[]` ((:) _z_1628027504 _z_1628027507) `[]` = `[]`

Zip3 `[]` ((:) _z_1628027510 _z_1628027513) ((:) _z_1628027516 _z_1628027519) = `[]`

Zip3 ((:) _z_1628027522 _z_1628027525) `[]` `[]` = `[]`

Zip3 ((:) _z_1628027528 _z_1628027531) `[]` ((:) _z_1628027534 _z_1628027537) = `[]`

Zip3 ((:) _z_1628027540 _z_1628027543) ((:) _z_1628027546 _z_1628027549) `[]` = `[]`

sZip3 :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Zip3Sym0 t) t) t :: [(a, b, c)])

type family ZipWith a a a :: [c]

Equations

ZipWith f ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply f x) y)) (Apply (Apply (Apply ZipWithSym0 f) xs) ys)

ZipWith _z_1628027456 `[]` `[]` = `[]`

ZipWith _z_1628027459 ((:) _z_1628027462 _z_1628027465) `[]` = `[]`

ZipWith _z_1628027468 `[]` ((:) _z_1628027471 _z_1628027474) = `[]`

sZipWith :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ZipWithSym0 t) t) t :: [c])

type family ZipWith3 a a a a :: [d]

Equations

ZipWith3 z ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply z a) b) c)) (Apply (Apply (Apply (Apply ZipWith3Sym0 z) as) bs) cs)

ZipWith3 _z_1628027361 `[]` `[]` `[]` = `[]`

ZipWith3 _z_1628027364 `[]` `[]` ((:) _z_1628027367 _z_1628027370) = `[]`

ZipWith3 _z_1628027373 `[]` ((:) _z_1628027376 _z_1628027379) `[]` = `[]`

ZipWith3 _z_1628027382 `[]` ((:) _z_1628027385 _z_1628027388) ((:) _z_1628027391 _z_1628027394) = `[]`

ZipWith3 _z_1628027397 ((:) _z_1628027400 _z_1628027403) `[]` `[]` = `[]`

ZipWith3 _z_1628027406 ((:) _z_1628027409 _z_1628027412) `[]` ((:) _z_1628027415 _z_1628027418) = `[]`

ZipWith3 _z_1628027421 ((:) _z_1628027424 _z_1628027427) ((:) _z_1628027430 _z_1628027433) `[]` = `[]`

sZipWith3 :: forall t t t t. Sing t -> Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (Apply ZipWith3Sym0 t) t) t) t :: [d])

type family Unzip a :: ([a], [b])

Equations

Unzip xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_1628027307Sym0 xs)) (Apply (Apply Tuple2Sym0 `[]`) `[]`)) xs

sUnzip :: forall t. Sing t -> Sing (Apply UnzipSym0 t :: ([a], [b]))

type family Unzip3 a :: ([a], [b], [c])

Equations

Unzip3 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_1628027275Sym0 xs)) (Apply (Apply (Apply Tuple3Sym0 `[]`) `[]`) `[]`)) xs

sUnzip3 :: forall t. Sing t -> Sing (Apply Unzip3Sym0 t :: ([a], [b], [c]))

type family Unzip4 a :: ([a], [b], [c], [d])

Equations

Unzip4 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_1628027241Sym0 xs)) (Apply (Apply (Apply (Apply Tuple4Sym0 `[]`) `[]`) `[]`) `[]`)) xs

sUnzip4 :: forall t. Sing t -> Sing (Apply Unzip4Sym0 t :: ([a], [b], [c], [d]))

type family Unzip5 a :: ([a], [b], [c], [d], [e])

Equations

Unzip5 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_1628027205Sym0 xs)) (Apply (Apply (Apply (Apply (Apply Tuple5Sym0 `[]`) `[]`) `[]`) `[]`) `[]`)) xs

sUnzip5 :: forall t. Sing t -> Sing (Apply Unzip5Sym0 t :: ([a], [b], [c], [d], [e]))

type family Unzip6 a :: ([a], [b], [c], [d], [e], [f])

Equations

Unzip6 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_1628027167Sym0 xs)) (Apply (Apply (Apply (Apply (Apply (Apply Tuple6Sym0 `[]`) `[]`) `[]`) `[]`) `[]`) `[]`)) xs

sUnzip6 :: forall t. Sing t -> Sing (Apply Unzip6Sym0 t :: ([a], [b], [c], [d], [e], [f]))

type family Unzip7 a :: ([a], [b], [c], [d], [e], [f], [g])

Equations

Unzip7 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_1628027127Sym0 xs)) (Apply (Apply (Apply (Apply (Apply (Apply (Apply Tuple7Sym0 `[]`) `[]`) `[]`) `[]`) `[]`) `[]`) `[]`)) xs

sUnzip7 :: forall t. Sing t -> Sing (Apply Unzip7Sym0 t :: ([a], [b], [c], [d], [e], [f], [g]))

Special lists

"Set" operations

type family Nub a :: [a]

Equations

Nub l = Apply (Apply (Let1628027714Nub'Sym1 l) l) `[]`

sNub :: forall t. SEq (KProxy :: KProxy a) => Sing t -> Sing (Apply NubSym0 t :: [a])

type family Delete a a :: [a]

Equations

Delete a_1628027092 a_1628027094 = Apply (Apply (Apply DeleteBySym0 (:==$)) a_1628027092) a_1628027094

sDelete :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply DeleteSym0 t) t :: [a])

type family a :\\ a :: [a] infix 5

Equations

a_1628027107 :\\ a_1628027109 = Apply (Apply (Apply FoldlSym0 (Apply FlipSym0 DeleteSym0)) a_1628027107) a_1628027109

(%:\\) :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply (:\\$) t) t :: [a]) infix 5

type family Union a a :: [a]

Equations

Union a_1628027077 a_1628027079 = Apply (Apply (Apply UnionBySym0 (:==$)) a_1628027077) a_1628027079

sUnion :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply UnionSym0 t) t :: [a])

type family Intersect a a :: [a]

Equations

Intersect a_1628026880 a_1628026882 = Apply (Apply (Apply IntersectBySym0 (:==$)) a_1628026880) a_1628026882

sIntersect :: forall t t. SEq (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply IntersectSym0 t) t :: [a])

Ordered lists

type family Insert a a :: [a]

Equations

Insert e ls = Apply (Apply (Apply InsertBySym0 CompareSym0) e) ls

sInsert :: forall t t. SOrd (KProxy :: KProxy a) => Sing t -> Sing t -> Sing (Apply (Apply InsertSym0 t) t :: [a])

type family Sort a :: [a]

Equations

Sort a_1628026983 = Apply (Apply SortBySym0 CompareSym0) a_1628026983

sSort :: forall t. SOrd (KProxy :: KProxy a) => Sing t -> Sing (Apply SortSym0 t :: [a])

Generalized functions

The "By" operations

User-supplied equality (replacing an Eq context)

The predicate is assumed to define an equivalence.

type family NubBy a a :: [a]

Equations

NubBy eq l = Apply (Apply (Let1628026021NubBy'Sym2 eq l) l) `[]`

sNubBy :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply NubBySym0 t) t :: [a])

type family DeleteBy a a a :: [a]

Equations

DeleteBy _z_1628027005 _z_1628027008 `[]` = `[]`

DeleteBy eq x ((:) y ys) = Case_1628027034 eq x y ys (Let1628027015Scrutinee_1628025909Sym4 eq x y ys)

sDeleteBy :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply DeleteBySym0 t) t) t :: [a])

type family DeleteFirstsBy a a a :: [a]

Equations

DeleteFirstsBy eq a_1628027052 a_1628027054 = Apply (Apply (Apply FoldlSym0 (Apply FlipSym0 (Apply DeleteBySym0 eq))) a_1628027052) a_1628027054

sDeleteFirstsBy :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply DeleteFirstsBySym0 t) t) t :: [a])

type family UnionBy a a a :: [a]

Equations

UnionBy eq xs ys = Apply (Apply (:++$) xs) (Apply (Apply (Apply FoldlSym0 (Apply FlipSym0 (Apply DeleteBySym0 eq))) (Apply (Apply NubBySym0 eq) ys)) xs)

sUnionBy :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply UnionBySym0 t) t) t :: [a])

type family IntersectBy a a a :: [a]

Equations

IntersectBy _z_1628026766 `[]` `[]` = `[]`

IntersectBy _z_1628026769 `[]` ((:) _z_1628026772 _z_1628026775) = `[]`

IntersectBy _z_1628026778 ((:) _z_1628026781 _z_1628026784) `[]` = `[]`

IntersectBy eq ((:) wild_1628025929 wild_1628025931) ((:) wild_1628025933 wild_1628025935) = Apply (Apply FilterSym0 (Apply (Apply (Apply (Apply (Apply Lambda_1628026843Sym0 eq) wild_1628025929) wild_1628025931) wild_1628025933) wild_1628025935)) (Let1628026792XsSym5 eq wild_1628025929 wild_1628025931 wild_1628025933 wild_1628025935)

sIntersectBy :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply IntersectBySym0 t) t) t :: [a])

type family GroupBy a a :: [[a]]

Equations

GroupBy _z_1628026566 `[]` = `[]`

GroupBy eq ((:) x xs) = Apply (Apply (:$) (Apply (Apply (:$) x) (Let1628026572YsSym3 eq x xs))) (Apply (Apply GroupBySym0 eq) (Let1628026572ZsSym3 eq x xs))

sGroupBy :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply GroupBySym0 t) t :: [[a]])

User-supplied comparison (replacing an Ord context)

The function is assumed to define a total ordering.

type family SortBy a a :: [a]

Equations

SortBy cmp a_1628026979 = Apply (Apply (Apply FoldrSym0 (Apply InsertBySym0 cmp)) `[]`) a_1628026979

sSortBy :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply SortBySym0 t) t :: [a])

type family InsertBy a a a :: [a]

Equations

InsertBy _z_1628026910 x `[]` = Apply (Apply (:$) x) `[]`

InsertBy cmp x ((:) y ys') = Case_1628026956 cmp x y ys' (Let1628026937Scrutinee_1628025911Sym4 cmp x y ys')

sInsertBy :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply InsertBySym0 t) t) t :: [a])

type family MaximumBy a a :: a

Equations

MaximumBy _z_1628028441 `[]` = Apply ErrorSym0 "Data.Singletons.List.maximumBy: empty list"

MaximumBy cmp ((:) wild_1628025915 wild_1628025917) = Apply (Apply Foldl1Sym0 (Let1628028460MaxBySym3 cmp wild_1628025915 wild_1628025917)) (Let1628028447XsSym3 cmp wild_1628025915 wild_1628025917)

sMaximumBy :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply MaximumBySym0 t) t :: a)

type family MinimumBy a a :: a

Equations

MinimumBy _z_1628028528 `[]` = Apply ErrorSym0 "Data.Singletons.List.minimumBy: empty list"

MinimumBy cmp ((:) wild_1628025921 wild_1628025923) = Apply (Apply Foldl1Sym0 (Let1628028547MinBySym3 cmp wild_1628025921 wild_1628025923)) (Let1628028534XsSym3 cmp wild_1628025921 wild_1628025923)

sMinimumBy :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply MinimumBySym0 t) t :: a)

The "generic" operations

The prefix `generic' indicates an overloaded function that is a generalized version of a Prelude function.

type family GenericLength a :: i

Equations

GenericLength `[]` = FromInteger 0

GenericLength ((:) _z_1628025981 xs) = Apply (Apply (:+$) (FromInteger 1)) (Apply GenericLengthSym0 xs)

sGenericLength :: forall t. SNum (KProxy :: KProxy i) => Sing t -> Sing (Apply GenericLengthSym0 t :: i)

Defunctionalization symbols

type NilSym0 = `[]`

data (:$) l

data l :$$ l

type (:$$$) t t = (:) t t

type (:++$$$) t t = (:++) t t

data l :++$$ l

data (:++$) l

data HeadSym0 l

type HeadSym1 t = Head t

data LastSym0 l

type LastSym1 t = Last t

data TailSym0 l

type TailSym1 t = Tail t

data InitSym0 l

type InitSym1 t = Init t

data NullSym0 l

type NullSym1 t = Null t

data LengthSym0 l

type LengthSym1 t = Length t

data MapSym0 l

data MapSym1 l l

type MapSym2 t t = Map t t

data ReverseSym0 l

type ReverseSym1 t = Reverse t

data IntersperseSym0 l

data IntersperseSym1 l l

type IntersperseSym2 t t = Intersperse t t

data IntercalateSym0 l

data IntercalateSym1 l l

type IntercalateSym2 t t = Intercalate t t

data TransposeSym0 l

type TransposeSym1 t = Transpose t

data SubsequencesSym0 l

type SubsequencesSym1 t = Subsequences t

data PermutationsSym0 l

type PermutationsSym1 t = Permutations t

data FoldlSym0 l

data FoldlSym1 l l

data FoldlSym2 l l l

type FoldlSym3 t t t = Foldl t t t

data Foldl'Sym0 l

data Foldl'Sym1 l l

data Foldl'Sym2 l l l

type Foldl'Sym3 t t t = Foldl' t t t

data Foldl1Sym0 l

data Foldl1Sym1 l l

type Foldl1Sym2 t t = Foldl1 t t

data Foldl1'Sym0 l

data Foldl1'Sym1 l l

type Foldl1'Sym2 t t = Foldl1' t t

data FoldrSym0 l

data FoldrSym1 l l

data FoldrSym2 l l l

type FoldrSym3 t t t = Foldr t t t

data Foldr1Sym0 l

data Foldr1Sym1 l l

type Foldr1Sym2 t t = Foldr1 t t

data ConcatSym0 l

type ConcatSym1 t = Concat t

data ConcatMapSym0 l

data ConcatMapSym1 l l

type ConcatMapSym2 t t = ConcatMap t t

data AndSym0 l

type AndSym1 t = And t

data OrSym0 l

type OrSym1 t = Or t

data Any_Sym0 l

data Any_Sym1 l l

type Any_Sym2 t t = Any_ t t

data AllSym0 l

data AllSym1 l l

type AllSym2 t t = All t t

data SumSym0 l

type SumSym1 t = Sum t

data ProductSym0 l

type ProductSym1 t = Product t

data MaximumSym0 l

type MaximumSym1 t = Maximum t

data MinimumSym0 l

type MinimumSym1 t = Minimum t

data ScanlSym0 l

data ScanlSym1 l l

data ScanlSym2 l l l

type ScanlSym3 t t t = Scanl t t t

data Scanl1Sym0 l

data Scanl1Sym1 l l

type Scanl1Sym2 t t = Scanl1 t t

data ScanrSym0 l

data ScanrSym1 l l

data ScanrSym2 l l l

type ScanrSym3 t t t = Scanr t t t

data Scanr1Sym0 l

data Scanr1Sym1 l l

type Scanr1Sym2 t t = Scanr1 t t

data MapAccumLSym0 l

data MapAccumLSym1 l l

data MapAccumLSym2 l l l

type MapAccumLSym3 t t t = MapAccumL t t t

data MapAccumRSym0 l

data MapAccumRSym1 l l

data MapAccumRSym2 l l l

type MapAccumRSym3 t t t = MapAccumR t t t

data ReplicateSym0 l

data ReplicateSym1 l l

type ReplicateSym2 t t = Replicate t t

data UnfoldrSym0 l

data UnfoldrSym1 l l

type UnfoldrSym2 t t = Unfoldr t t

data TakeSym0 l

data TakeSym1 l l

type TakeSym2 t t = Take t t

data DropSym0 l

data DropSym1 l l

type DropSym2 t t = Drop t t

data SplitAtSym0 l

data SplitAtSym1 l l

type SplitAtSym2 t t = SplitAt t t

data TakeWhileSym0 l

data TakeWhileSym1 l l

type TakeWhileSym2 t t = TakeWhile t t

data DropWhileSym0 l

data DropWhileSym1 l l

type DropWhileSym2 t t = DropWhile t t

data DropWhileEndSym0 l

data DropWhileEndSym1 l l

type DropWhileEndSym2 t t = DropWhileEnd t t

data SpanSym0 l

data SpanSym1 l l

type SpanSym2 t t = Span t t

data BreakSym0 l

data BreakSym1 l l

type BreakSym2 t t = Break t t

data GroupSym0 l

type GroupSym1 t = Group t

data InitsSym0 l

type InitsSym1 t = Inits t

data TailsSym0 l

type TailsSym1 t = Tails t

data IsPrefixOfSym0 l

data IsPrefixOfSym1 l l

type IsPrefixOfSym2 t t = IsPrefixOf t t

data IsSuffixOfSym0 l

data IsSuffixOfSym1 l l

type IsSuffixOfSym2 t t = IsSuffixOf t t

data IsInfixOfSym0 l

data IsInfixOfSym1 l l

type IsInfixOfSym2 t t = IsInfixOf t t

data ElemSym0 l

data ElemSym1 l l

type ElemSym2 t t = Elem t t

data NotElemSym0 l

data NotElemSym1 l l

type NotElemSym2 t t = NotElem t t

data LookupSym0 l

data LookupSym1 l l

type LookupSym2 t t = Lookup t t

data FindSym0 l

data FindSym1 l l

type FindSym2 t t = Find t t

data FilterSym0 l

data FilterSym1 l l

type FilterSym2 t t = Filter t t

data PartitionSym0 l

data PartitionSym1 l l

type PartitionSym2 t t = Partition t t

data (:!!$) l

data l :!!$$ l

type (:!!$$$) t t = (:!!) t t

data ElemIndexSym0 l

data ElemIndexSym1 l l

type ElemIndexSym2 t t = ElemIndex t t

data ElemIndicesSym0 l

data ElemIndicesSym1 l l

type ElemIndicesSym2 t t = ElemIndices t t

data FindIndexSym0 l

data FindIndexSym1 l l

type FindIndexSym2 t t = FindIndex t t

data FindIndicesSym0 l

data FindIndicesSym1 l l

type FindIndicesSym2 t t = FindIndices t t

data ZipSym0 l

data ZipSym1 l l

type ZipSym2 t t = Zip t t

data Zip3Sym0 l

data Zip3Sym1 l l

data Zip3Sym2 l l l

type Zip3Sym3 t t t = Zip3 t t t

data ZipWithSym0 l

data ZipWithSym1 l l

data ZipWithSym2 l l l

type ZipWithSym3 t t t = ZipWith t t t

data ZipWith3Sym0 l

data ZipWith3Sym1 l l

data ZipWith3Sym2 l l l

data ZipWith3Sym3 l l l l

type ZipWith3Sym4 t t t t = ZipWith3 t t t t

data UnzipSym0 l

type UnzipSym1 t = Unzip t

data Unzip3Sym0 l

type Unzip3Sym1 t = Unzip3 t

data Unzip4Sym0 l

type Unzip4Sym1 t = Unzip4 t

data Unzip5Sym0 l

type Unzip5Sym1 t = Unzip5 t

data Unzip6Sym0 l

type Unzip6Sym1 t = Unzip6 t

data Unzip7Sym0 l

type Unzip7Sym1 t = Unzip7 t

data NubSym0 l

type NubSym1 t = Nub t

data DeleteSym0 l

data DeleteSym1 l l

type DeleteSym2 t t = Delete t t

data (:\\$) l

data l :\\$$ l

type (:\\$$$) t t = (:\\) t t

data UnionSym0 l

data UnionSym1 l l

type UnionSym2 t t = Union t t

data IntersectSym0 l

data IntersectSym1 l l

type IntersectSym2 t t = Intersect t t

data InsertSym0 l

data InsertSym1 l l

type InsertSym2 t t = Insert t t

data SortSym0 l

type SortSym1 t = Sort t

data NubBySym0 l

data NubBySym1 l l

type NubBySym2 t t = NubBy t t

data DeleteBySym0 l

data DeleteBySym1 l l

data DeleteBySym2 l l l

type DeleteBySym3 t t t = DeleteBy t t t

data DeleteFirstsBySym0 l

data DeleteFirstsBySym1 l l

data DeleteFirstsBySym2 l l l

type DeleteFirstsBySym3 t t t = DeleteFirstsBy t t t

data UnionBySym0 l

data UnionBySym1 l l

data UnionBySym2 l l l

type UnionBySym3 t t t = UnionBy t t t

data IntersectBySym0 l

data IntersectBySym1 l l

data IntersectBySym2 l l l

type IntersectBySym3 t t t = IntersectBy t t t

data GroupBySym0 l

data GroupBySym1 l l

type GroupBySym2 t t = GroupBy t t

data SortBySym0 l

data SortBySym1 l l

type SortBySym2 t t = SortBy t t

data InsertBySym0 l

data InsertBySym1 l l

data InsertBySym2 l l l

type InsertBySym3 t t t = InsertBy t t t

data MaximumBySym0 l

data MaximumBySym1 l l

type MaximumBySym2 t t = MaximumBy t t

data MinimumBySym0 l

data MinimumBySym1 l l

type MinimumBySym2 t t = MinimumBy t t

data GenericLengthSym0 l

type GenericLengthSym1 t = GenericLength t