Data.Singletons.Prelude.List.NonEmpty

Description

Defines functions and datatypes relating to the singleton for NonEmpty, including a singletons version of all the definitions in Data.List.NonEmpty.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.List.NonEmpty. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The NonEmpty singleton

type family Sing

The singleton kind-indexed type family.

data SNonEmpty z where

Constructors

(:%|) :: forall (a :: Type) (n :: a) (n :: [a]). (Sing n) -> (Sing n) -> SNonEmpty ('(:|) n n :: NonEmpty (a :: Type)) infixr 5

Non-empty stream transformations

type family Map a a where ...

Equations

Map f ('(:|) a as) = Apply (Apply (:|@#@$) (Apply f a)) (Apply (Apply ListmapSym0 f) as)

sMap :: forall a b (t :: (~>) a b) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: NonEmpty b)

type family Intersperse a a where ...

Equations

Intersperse a ('(:|) b bs) = Apply (Apply (:|@#@$) b) (Case_6989586621681176003 a b bs bs)

sIntersperse :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply IntersperseSym0 t) t :: NonEmpty a)

type family Scanl a a a where ...

Equations

Scanl f z a_6989586621681176034 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply ListscanlSym0 f) z)) a_6989586621681176034

sScanl :: forall b a (t :: (~>) b ((~>) a b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanlSym0 t) t) t :: NonEmpty b)

type family Scanr a a a where ...

Equations

Scanr f z a_6989586621681176022 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply ListscanrSym0 f) z)) a_6989586621681176022

sScanr :: forall a b (t :: (~>) a ((~>) b b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanrSym0 t) t) t :: NonEmpty b)

type family Scanl1 a a where ...

Equations

Scanl1 f ('(:|) a as) = Apply FromListSym0 (Apply (Apply (Apply ListscanlSym0 f) a) as)

sScanl1 :: forall a (t :: (~>) a ((~>) a a)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply Scanl1Sym0 t) t :: NonEmpty a)

type family Scanr1 a a where ...

Equations

Scanr1 f ('(:|) a as) = Apply FromListSym0 (Apply (Apply Listscanr1Sym0 f) (Apply (Apply (:@#@$) a) as))

sScanr1 :: forall a (t :: (~>) a ((~>) a a)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply Scanr1Sym0 t) t :: NonEmpty a)

type family Transpose a where ...

Equations

Transpose a_6989586621681175703 = Apply (Apply (Apply (.@#@$) (Apply FmapSym0 FromListSym0)) (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply (.@#@$) ListtransposeSym0) (Apply (Apply (.@#@$) ToListSym0) (Apply FmapSym0 ToListSym0))))) a_6989586621681175703

sTranspose :: forall a (t :: NonEmpty (NonEmpty a)). Sing t -> Sing (Apply TransposeSym0 t :: NonEmpty (NonEmpty a))

type family SortBy a a where ...

Equations

SortBy f a_6989586621681175694 = Apply (Apply LiftSym0 (Apply ListsortBySym0 f)) a_6989586621681175694

sSortBy :: forall a (t :: (~>) a ((~>) a Ordering)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SortBySym0 t) t :: NonEmpty a)

type family SortWith a a where ...

Equations

SortWith a_6989586621681175683 a_6989586621681175685 = Apply (Apply (Apply (Apply (.@#@$) SortBySym0) ComparingSym0) a_6989586621681175683) a_6989586621681175685

sSortWith :: forall a o (t :: (~>) a o) (t :: NonEmpty a). SOrd o => Sing t -> Sing t -> Sing (Apply (Apply SortWithSym0 t) t :: NonEmpty a)

type family Length a where ...

Equations

Length ('(:|) _ xs) = Apply (Apply (+@#@$) (FromInteger 1)) (Apply ListlengthSym0 xs)

sLength :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply LengthSym0 t :: Nat)

type family Head a where ...

Equations

Head ('(:|) a _) = a

sHead :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply HeadSym0 t :: a)

type family Tail a where ...

Equations

Tail ('(:|) _ as) = as

sTail :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply TailSym0 t :: [a])

type family Last a where ...

Equations

Last ('(:|) a as) = Apply ListlastSym0 (Apply (Apply (:@#@$) a) as)

sLast :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply LastSym0 t :: a)

type family Init a where ...

Equations

Init ('(:|) a as) = Apply ListinitSym0 (Apply (Apply (:@#@$) a) as)

sInit :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply InitSym0 t :: [a])

type family a <| a where ...

Equations

a <| ('(:|) b bs) = Apply (Apply (:|@#@$) a) (Apply (Apply (:@#@$) b) bs)

(%<|) :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply (<|@#@$) t) t :: NonEmpty a)

type family Cons a a where ...

Equations

Cons a_6989586621681176100 a_6989586621681176102 = Apply (Apply (<|@#@$) a_6989586621681176100) a_6989586621681176102

sCons :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply ConsSym0 t) t :: NonEmpty a)

type family Uncons a where ...

Equations

Uncons ('(:|) a as) = Apply (Apply Tuple2Sym0 a) (Apply NonEmpty_Sym0 as)

sUncons :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply UnconsSym0 t :: (a, Maybe (NonEmpty a)))

type family Unfoldr a a where ...

Equations

Unfoldr f a = Case_6989586621681176157 f a (Let6989586621681176155Scrutinee_6989586621681174570Sym2 f a)

sUnfoldr :: forall a b (t :: (~>) a (b, Maybe a)) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply UnfoldrSym0 t) t :: NonEmpty b)

type family Sort a where ...

Equations

Sort a_6989586621681176094 = Apply (Apply LiftSym0 ListsortSym0) a_6989586621681176094

sSort :: forall a (t :: NonEmpty a). SOrd a => Sing t -> Sing (Apply SortSym0 t :: NonEmpty a)

type family Reverse a where ...

Equations

Reverse a_6989586621681175989 = Apply (Apply LiftSym0 ListreverseSym0) a_6989586621681175989

sReverse :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply ReverseSym0 t :: NonEmpty a)

type family Inits a where ...

Equations

Inits a_6989586621681176061 = Apply (Apply (Apply (.@#@$) FromListSym0) ListinitsSym0) a_6989586621681176061

sInits :: forall a (t :: [a]). Sing t -> Sing (Apply InitsSym0 t :: NonEmpty [a])

type family Tails a where ...

Equations

Tails a_6989586621681176055 = Apply (Apply (Apply (.@#@$) FromListSym0) ListtailsSym0) a_6989586621681176055

sTails :: forall a (t :: [a]). Sing t -> Sing (Apply TailsSym0 t :: NonEmpty [a])

type family Unfold a a where ...

Equations

Unfold f a = Case_6989586621681176181 f a (Let6989586621681176179Scrutinee_6989586621681174560Sym2 f a)

sUnfold :: forall a b (t :: (~>) a (b, Maybe a)) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply UnfoldSym0 t) t :: NonEmpty b)

type family Insert a a where ...

Equations

Insert a a_6989586621681176046 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply ListinsertSym0 a)) a_6989586621681176046

sInsert :: forall a (t :: a) (t :: [a]). SOrd a => Sing t -> Sing t -> Sing (Apply (Apply InsertSym0 t) t :: NonEmpty a)

type family Take a a where ...

Equations

Take n a_6989586621681175980 = Apply (Apply (Apply (.@#@$) (Apply ListtakeSym0 n)) ToListSym0) a_6989586621681175980

sTake :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply TakeSym0 t) t :: [a])

type family Drop a a where ...

Equations

Drop n a_6989586621681175971 = Apply (Apply (Apply (.@#@$) (Apply ListdropSym0 n)) ToListSym0) a_6989586621681175971

sDrop :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply DropSym0 t) t :: [a])

type family SplitAt a a where ...

Equations

SplitAt n a_6989586621681175962 = Apply (Apply (Apply (.@#@$) (Apply ListsplitAtSym0 n)) ToListSym0) a_6989586621681175962

sSplitAt :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SplitAtSym0 t) t :: ([a], [a]))

type family TakeWhile a a where ...

Equations

TakeWhile p a_6989586621681175953 = Apply (Apply (Apply (.@#@$) (Apply ListtakeWhileSym0 p)) ToListSym0) a_6989586621681175953

sTakeWhile :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply TakeWhileSym0 t) t :: [a])

type family DropWhile a a where ...

Equations

DropWhile p a_6989586621681175944 = Apply (Apply (Apply (.@#@$) (Apply ListdropWhileSym0 p)) ToListSym0) a_6989586621681175944

sDropWhile :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply DropWhileSym0 t) t :: [a])

type family Span a a where ...

Equations

Span p a_6989586621681175935 = Apply (Apply (Apply (.@#@$) (Apply ListspanSym0 p)) ToListSym0) a_6989586621681175935

sSpan :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SpanSym0 t) t :: ([a], [a]))

type family Break a a where ...

Equations

Break p a_6989586621681175926 = Apply (Apply SpanSym0 (Apply (Apply (.@#@$) NotSym0) p)) a_6989586621681175926

sBreak :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply BreakSym0 t) t :: ([a], [a]))

type family Filter a a where ...

Equations

Filter p a_6989586621681175917 = Apply (Apply (Apply (.@#@$) (Apply ListfilterSym0 p)) ToListSym0) a_6989586621681175917

sFilter :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply FilterSym0 t) t :: [a])

type family Partition a a where ...

Equations

Partition p a_6989586621681175908 = Apply (Apply (Apply (.@#@$) (Apply ListpartitionSym0 p)) ToListSym0) a_6989586621681175908

sPartition :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply PartitionSym0 t) t :: ([a], [a]))

type family Group a where ...

Equations

Group a_6989586621681175902 = Apply (Apply GroupBySym0 (==@#@$)) a_6989586621681175902

sGroup :: forall a (t :: [a]). SEq a => Sing t -> Sing (Apply GroupSym0 t :: [NonEmpty a])

type family GroupBy a a where ...

Equations

GroupBy eq0 a_6989586621681175868 = Apply (Apply (Let6989586621681175877GoSym2 eq0 a_6989586621681175868) eq0) a_6989586621681175868

sGroupBy :: forall a (t :: (~>) a ((~>) a Bool)) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply GroupBySym0 t) t :: [NonEmpty a])

type family GroupWith a a where ...

Equations

GroupWith f a_6989586621681175859 = Apply (Apply GroupBySym0 (Apply (Apply OnSym0 (==@#@$)) f)) a_6989586621681175859

sGroupWith :: forall a b (t :: (~>) a b) (t :: [a]). SEq b => Sing t -> Sing t -> Sing (Apply (Apply GroupWithSym0 t) t :: [NonEmpty a])

type family GroupAllWith a a where ...

Equations

GroupAllWith f a_6989586621681175850 = Apply (Apply (Apply (.@#@$) (Apply GroupWithSym0 f)) (Apply ListsortBySym0 (Apply (Apply OnSym0 CompareSym0) f))) a_6989586621681175850

sGroupAllWith :: forall a b (t :: (~>) a b) (t :: [a]). SOrd b => Sing t -> Sing t -> Sing (Apply (Apply GroupAllWithSym0 t) t :: [NonEmpty a])

type family Group1 a where ...

Equations

Group1 a_6989586621681175844 = Apply (Apply GroupBy1Sym0 (==@#@$)) a_6989586621681175844

sGroup1 :: forall a (t :: NonEmpty a). SEq a => Sing t -> Sing (Apply Group1Sym0 t :: NonEmpty (NonEmpty a))

type family GroupBy1 a a where ...

Equations

GroupBy1 eq ('(:|) x xs) = Apply (Apply (:|@#@$) (Apply (Apply (:|@#@$) x) (Let6989586621681175826YsSym3 eq x xs))) (Apply (Apply GroupBySym0 eq) (Let6989586621681175826ZsSym3 eq x xs))

sGroupBy1 :: forall a (t :: (~>) a ((~>) a Bool)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply GroupBy1Sym0 t) t :: NonEmpty (NonEmpty a))

type family GroupWith1 a a where ...

Equations

GroupWith1 f a_6989586621681175809 = Apply (Apply GroupBy1Sym0 (Apply (Apply OnSym0 (==@#@$)) f)) a_6989586621681175809

sGroupWith1 :: forall a b (t :: (~>) a b) (t :: NonEmpty a). SEq b => Sing t -> Sing t -> Sing (Apply (Apply GroupWith1Sym0 t) t :: NonEmpty (NonEmpty a))

type family GroupAllWith1 a a where ...

Equations

GroupAllWith1 f a_6989586621681175800 = Apply (Apply (Apply (.@#@$) (Apply GroupWith1Sym0 f)) (Apply SortWithSym0 f)) a_6989586621681175800

sGroupAllWith1 :: forall a b (t :: (~>) a b) (t :: NonEmpty a). SOrd b => Sing t -> Sing t -> Sing (Apply (Apply GroupAllWith1Sym0 t) t :: NonEmpty (NonEmpty a))

type family IsPrefixOf a a where ...

Equations

IsPrefixOf '[] _ = TrueSym0

IsPrefixOf ('(:) y ys) ('(:|) x xs) = Apply (Apply (&&@#@$) (Apply (Apply (==@#@$) y) x)) (Apply (Apply ListisPrefixOfSym0 ys) xs)

sIsPrefixOf :: forall a (t :: [a]) (t :: NonEmpty a). SEq a => Sing t -> Sing t -> Sing (Apply (Apply IsPrefixOfSym0 t) t :: Bool)

type family Nub a where ...

Equations

Nub a_6989586621681175721 = Apply (Apply NubBySym0 (==@#@$)) a_6989586621681175721

sNub :: forall a (t :: NonEmpty a). SEq a => Sing t -> Sing (Apply NubSym0 t :: NonEmpty a)

type family NubBy a a where ...

Equations

NubBy eq ('(:|) a as) = Apply (Apply (:|@#@$) a) (Apply (Apply ListnubBySym0 eq) (Apply (Apply ListfilterSym0 (Apply (Apply (Apply Lambda_6989586621681175717Sym0 eq) a) as)) as))

sNubBy :: forall a (t :: (~>) a ((~>) a Bool)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply NubBySym0 t) t :: NonEmpty a)

type family a !! a where ...

Equations

arg_6989586621681174582 !! arg_6989586621681174584 = Case_6989586621681175779 arg_6989586621681174582 arg_6989586621681174584 (Apply (Apply Tuple2Sym0 arg_6989586621681174582) arg_6989586621681174584)

(%!!) :: forall a (t :: NonEmpty a) (t :: Nat). Sing t -> Sing t -> Sing (Apply (Apply (!!@#@$) t) t :: a)

type family Zip a a where ...

Equations

Zip ('(:|) x xs) ('(:|) y ys) = Apply (Apply (:|@#@$) (Apply (Apply Tuple2Sym0 x) y)) (Apply (Apply ListzipSym0 xs) ys)

sZip :: forall a b (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing (Apply (Apply ZipSym0 t) t :: NonEmpty (a, b))

type family ZipWith a a a where ...

Equations

ZipWith f ('(:|) x xs) ('(:|) y ys) = Apply (Apply (:|@#@$) (Apply (Apply f x) y)) (Apply (Apply (Apply ListzipWithSym0 f) xs) ys)

sZipWith :: forall a b c (t :: (~>) a ((~>) b c)) (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ZipWithSym0 t) t) t :: NonEmpty c)

type family Unzip a where ...

Equations

Unzip ('(:|) '(a, b) asbs) = Apply (Apply Tuple2Sym0 (Apply (Apply (:|@#@$) a) (Let6989586621681175733AsSym3 a b asbs))) (Apply (Apply (:|@#@$) b) (Let6989586621681175733BsSym3 a b asbs))

sUnzip :: forall a b (t :: NonEmpty (a, b)). Sing t -> Sing (Apply UnzipSym0 t :: (NonEmpty a, NonEmpty b))

type family FromList a where ...

Equations

FromList ('(:) a as) = Apply (Apply (:|@#@$) a) as

FromList '[] = Apply ErrorSym0 "NonEmpty.fromList: empty list"

sFromList :: forall a (t :: [a]). Sing t -> Sing (Apply FromListSym0 t :: NonEmpty a)

type family ToList a where ...

Equations

ToList ('(:|) a as) = Apply (Apply (:@#@$) a) as

sToList :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply ToListSym0 t :: [a])

type family NonEmpty_ a where ...

Equations

NonEmpty_ '[] = NothingSym0

NonEmpty_ ('(:) a as) = Apply JustSym0 (Apply (Apply (:|@#@$) a) as)

sNonEmpty_ :: forall a (t :: [a]). Sing t -> Sing (Apply NonEmpty_Sym0 t :: Maybe (NonEmpty a))

type family Xor a where ...

Equations

Xor ('(:|) x xs) = Apply (Apply (Apply FoldrSym0 (Let6989586621681176192Xor'Sym2 x xs)) x) xs

sXor :: forall (t :: NonEmpty Bool). Sing t -> Sing (Apply XorSym0 t :: Bool)

Defunctionalization symbols

data (:|@#@$) a6989586621679304457 infixr 5

data a6989586621679304457 :|@#@$$ a6989586621679304458 infixr 5

type (:|@#@$$$) (a6989586621679304457 :: a) (a6989586621679304458 :: [a]) = '(:|) a6989586621679304457 a6989586621679304458 :: NonEmpty (a :: Type) infixr 5

data MapSym0 a6989586621681176070

data MapSym1 a6989586621681176070 a6989586621681176071

type MapSym2 (a6989586621681176070 :: (~>) a b) (a6989586621681176071 :: NonEmpty a) = Map a6989586621681176070 a6989586621681176071 :: NonEmpty b

data IntersperseSym0 a6989586621681175998

data IntersperseSym1 a6989586621681175998 a6989586621681175999

type IntersperseSym2 (a6989586621681175998 :: a) (a6989586621681175999 :: NonEmpty a) = Intersperse a6989586621681175998 a6989586621681175999 :: NonEmpty a

data ScanlSym0 a6989586621681176040

data ScanlSym1 a6989586621681176040 a6989586621681176041

data ScanlSym2 a6989586621681176040 a6989586621681176041 a6989586621681176042

type ScanlSym3 (a6989586621681176040 :: (~>) b ((~>) a b)) (a6989586621681176041 :: b) (a6989586621681176042 :: [a]) = Scanl a6989586621681176040 a6989586621681176041 a6989586621681176042 :: NonEmpty b

data ScanrSym0 a6989586621681176028

data ScanrSym1 a6989586621681176028 a6989586621681176029

data ScanrSym2 a6989586621681176028 a6989586621681176029 a6989586621681176030

type ScanrSym3 (a6989586621681176028 :: (~>) a ((~>) b b)) (a6989586621681176029 :: b) (a6989586621681176030 :: [a]) = Scanr a6989586621681176028 a6989586621681176029 a6989586621681176030 :: NonEmpty b

data Scanl1Sym0 a6989586621681176017

data Scanl1Sym1 a6989586621681176017 a6989586621681176018

type Scanl1Sym2 (a6989586621681176017 :: (~>) a ((~>) a a)) (a6989586621681176018 :: NonEmpty a) = Scanl1 a6989586621681176017 a6989586621681176018 :: NonEmpty a

data Scanr1Sym0 a6989586621681176009

data Scanr1Sym1 a6989586621681176009 a6989586621681176010

type Scanr1Sym2 (a6989586621681176009 :: (~>) a ((~>) a a)) (a6989586621681176010 :: NonEmpty a) = Scanr1 a6989586621681176009 a6989586621681176010 :: NonEmpty a

data TransposeSym0 a6989586621681175707

type TransposeSym1 (a6989586621681175707 :: NonEmpty (NonEmpty a)) = Transpose a6989586621681175707 :: NonEmpty (NonEmpty a)

data SortBySym0 a6989586621681175699

data SortBySym1 a6989586621681175699 a6989586621681175700

type SortBySym2 (a6989586621681175699 :: (~>) a ((~>) a Ordering)) (a6989586621681175700 :: NonEmpty a) = SortBy a6989586621681175699 a6989586621681175700 :: NonEmpty a

data SortWithSym0 a6989586621681175690

data SortWithSym1 a6989586621681175690 a6989586621681175691

type SortWithSym2 (a6989586621681175690 :: (~>) a o) (a6989586621681175691 :: NonEmpty a) = SortWith a6989586621681175690 a6989586621681175691 :: NonEmpty a

data LengthSym0 a6989586621681176200

type LengthSym1 (a6989586621681176200 :: NonEmpty a) = Length a6989586621681176200 :: Nat

data HeadSym0 a6989586621681176135

type HeadSym1 (a6989586621681176135 :: NonEmpty a) = Head a6989586621681176135 :: a

data TailSym0 a6989586621681176131

type TailSym1 (a6989586621681176131 :: NonEmpty a) = Tail a6989586621681176131 :: [a]

data LastSym0 a6989586621681176126

type LastSym1 (a6989586621681176126 :: NonEmpty a) = Last a6989586621681176126 :: a

data InitSym0 a6989586621681176121

type InitSym1 (a6989586621681176121 :: NonEmpty a) = Init a6989586621681176121 :: [a]

data (<|@#@$) a6989586621681176114

data a6989586621681176114 <|@#@$$ a6989586621681176115

type (<|@#@$$$) (a6989586621681176114 :: a) (a6989586621681176115 :: NonEmpty a) = (<|) a6989586621681176114 a6989586621681176115 :: NonEmpty a

data ConsSym0 a6989586621681176107

data ConsSym1 a6989586621681176107 a6989586621681176108

type ConsSym2 (a6989586621681176107 :: a) (a6989586621681176108 :: NonEmpty a) = Cons a6989586621681176107 a6989586621681176108 :: NonEmpty a

data UnconsSym0 a6989586621681176164

type UnconsSym1 (a6989586621681176164 :: NonEmpty a) = Uncons a6989586621681176164 :: (a, Maybe (NonEmpty a))

data UnfoldrSym0 a6989586621681176140

data UnfoldrSym1 a6989586621681176140 a6989586621681176141

type UnfoldrSym2 (a6989586621681176140 :: (~>) a (b, Maybe a)) (a6989586621681176141 :: a) = Unfoldr a6989586621681176140 a6989586621681176141 :: NonEmpty b

data SortSym0 a6989586621681176098

type SortSym1 (a6989586621681176098 :: NonEmpty a) = Sort a6989586621681176098 :: NonEmpty a

data ReverseSym0 a6989586621681175993

type ReverseSym1 (a6989586621681175993 :: NonEmpty a) = Reverse a6989586621681175993 :: NonEmpty a

data InitsSym0 a6989586621681176065

type InitsSym1 (a6989586621681176065 :: [a]) = Inits a6989586621681176065 :: NonEmpty [a]

data TailsSym0 a6989586621681176059

type TailsSym1 (a6989586621681176059 :: [a]) = Tails a6989586621681176059 :: NonEmpty [a]

data UnfoldSym0 a6989586621681176175

data UnfoldSym1 a6989586621681176175 a6989586621681176176

data InsertSym0 a6989586621681176051

data InsertSym1 a6989586621681176051 a6989586621681176052

type InsertSym2 (a6989586621681176051 :: a) (a6989586621681176052 :: [a]) = Insert a6989586621681176051 a6989586621681176052 :: NonEmpty a

data TakeSym0 a6989586621681175985

data TakeSym1 a6989586621681175985 a6989586621681175986

type TakeSym2 (a6989586621681175985 :: Nat) (a6989586621681175986 :: NonEmpty a) = Take a6989586621681175985 a6989586621681175986 :: [a]

data DropSym0 a6989586621681175976

data DropSym1 a6989586621681175976 a6989586621681175977

type DropSym2 (a6989586621681175976 :: Nat) (a6989586621681175977 :: NonEmpty a) = Drop a6989586621681175976 a6989586621681175977 :: [a]

data SplitAtSym0 a6989586621681175967

data SplitAtSym1 a6989586621681175967 a6989586621681175968

type SplitAtSym2 (a6989586621681175967 :: Nat) (a6989586621681175968 :: NonEmpty a) = SplitAt a6989586621681175967 a6989586621681175968 :: ([a], [a])

data TakeWhileSym0 a6989586621681175958

data TakeWhileSym1 a6989586621681175958 a6989586621681175959

type TakeWhileSym2 (a6989586621681175958 :: (~>) a Bool) (a6989586621681175959 :: NonEmpty a) = TakeWhile a6989586621681175958 a6989586621681175959 :: [a]

data DropWhileSym0 a6989586621681175949

data DropWhileSym1 a6989586621681175949 a6989586621681175950

type DropWhileSym2 (a6989586621681175949 :: (~>) a Bool) (a6989586621681175950 :: NonEmpty a) = DropWhile a6989586621681175949 a6989586621681175950 :: [a]

data SpanSym0 a6989586621681175940

data SpanSym1 a6989586621681175940 a6989586621681175941

type SpanSym2 (a6989586621681175940 :: (~>) a Bool) (a6989586621681175941 :: NonEmpty a) = Span a6989586621681175940 a6989586621681175941 :: ([a], [a])

data BreakSym0 a6989586621681175931

data BreakSym1 a6989586621681175931 a6989586621681175932

type BreakSym2 (a6989586621681175931 :: (~>) a Bool) (a6989586621681175932 :: NonEmpty a) = Break a6989586621681175931 a6989586621681175932 :: ([a], [a])

data FilterSym0 a6989586621681175922

data FilterSym1 a6989586621681175922 a6989586621681175923

type FilterSym2 (a6989586621681175922 :: (~>) a Bool) (a6989586621681175923 :: NonEmpty a) = Filter a6989586621681175922 a6989586621681175923 :: [a]

data PartitionSym0 a6989586621681175913

data PartitionSym1 a6989586621681175913 a6989586621681175914

type PartitionSym2 (a6989586621681175913 :: (~>) a Bool) (a6989586621681175914 :: NonEmpty a) = Partition a6989586621681175913 a6989586621681175914 :: ([a], [a])

data GroupSym0 a6989586621681175906

type GroupSym1 (a6989586621681175906 :: [a]) = Group a6989586621681175906 :: [NonEmpty a]

data GroupBySym0 a6989586621681175873

data GroupBySym1 a6989586621681175873 a6989586621681175874

type GroupBySym2 (a6989586621681175873 :: (~>) a ((~>) a Bool)) (a6989586621681175874 :: [a]) = GroupBy a6989586621681175873 a6989586621681175874 :: [NonEmpty a]

data GroupWithSym0 a6989586621681175864

data GroupWithSym1 a6989586621681175864 a6989586621681175865

type GroupWithSym2 (a6989586621681175864 :: (~>) a b) (a6989586621681175865 :: [a]) = GroupWith a6989586621681175864 a6989586621681175865 :: [NonEmpty a]

data GroupAllWithSym0 a6989586621681175855

data GroupAllWithSym1 a6989586621681175855 a6989586621681175856

type GroupAllWithSym2 (a6989586621681175855 :: (~>) a b) (a6989586621681175856 :: [a]) = GroupAllWith a6989586621681175855 a6989586621681175856 :: [NonEmpty a]

data Group1Sym0 a6989586621681175848

type Group1Sym1 (a6989586621681175848 :: NonEmpty a) = Group1 a6989586621681175848 :: NonEmpty (NonEmpty a)

data GroupBy1Sym0 a6989586621681175821

data GroupBy1Sym1 a6989586621681175821 a6989586621681175822

type GroupBy1Sym2 (a6989586621681175821 :: (~>) a ((~>) a Bool)) (a6989586621681175822 :: NonEmpty a) = GroupBy1 a6989586621681175821 a6989586621681175822 :: NonEmpty (NonEmpty a)

data GroupWith1Sym0 a6989586621681175814

data GroupWith1Sym1 a6989586621681175814 a6989586621681175815

type GroupWith1Sym2 (a6989586621681175814 :: (~>) a b) (a6989586621681175815 :: NonEmpty a) = GroupWith1 a6989586621681175814 a6989586621681175815 :: NonEmpty (NonEmpty a)

data GroupAllWith1Sym0 a6989586621681175805

data GroupAllWith1Sym1 a6989586621681175805 a6989586621681175806

type GroupAllWith1Sym2 (a6989586621681175805 :: (~>) a b) (a6989586621681175806 :: NonEmpty a) = GroupAllWith1 a6989586621681175805 a6989586621681175806 :: NonEmpty (NonEmpty a)

data IsPrefixOfSym0 a6989586621681175794

data IsPrefixOfSym1 a6989586621681175794 a6989586621681175795

type IsPrefixOfSym2 (a6989586621681175794 :: [a]) (a6989586621681175795 :: NonEmpty a) = IsPrefixOf a6989586621681175794 a6989586621681175795 :: Bool

data NubSym0 a6989586621681175725

type NubSym1 (a6989586621681175725 :: NonEmpty a) = Nub a6989586621681175725 :: NonEmpty a

data NubBySym0 a6989586621681175712

data NubBySym1 a6989586621681175712 a6989586621681175713

type NubBySym2 (a6989586621681175712 :: (~>) a ((~>) a Bool)) (a6989586621681175713 :: NonEmpty a) = NubBy a6989586621681175712 a6989586621681175713 :: NonEmpty a

data (!!@#@$) a6989586621681175775

data a6989586621681175775 !!@#@$$ a6989586621681175776

type (!!@#@$$$) (a6989586621681175775 :: NonEmpty a) (a6989586621681175776 :: Nat) = (!!) a6989586621681175775 a6989586621681175776 :: a

data ZipSym0 a6989586621681175766

data ZipSym1 a6989586621681175766 a6989586621681175767

type ZipSym2 (a6989586621681175766 :: NonEmpty a) (a6989586621681175767 :: NonEmpty b) = Zip a6989586621681175766 a6989586621681175767 :: NonEmpty (a, b)

data ZipWithSym0 a6989586621681175755

data ZipWithSym1 a6989586621681175755 a6989586621681175756

data ZipWithSym2 a6989586621681175755 a6989586621681175756 a6989586621681175757

type ZipWithSym3 (a6989586621681175755 :: (~>) a ((~>) b c)) (a6989586621681175756 :: NonEmpty a) (a6989586621681175757 :: NonEmpty b) = ZipWith a6989586621681175755 a6989586621681175756 a6989586621681175757 :: NonEmpty c

data UnzipSym0 a6989586621681175729

type UnzipSym1 (a6989586621681175729 :: NonEmpty (a, b)) = Unzip a6989586621681175729 :: (NonEmpty a, NonEmpty b)

data FromListSym0 a6989586621681176091

type FromListSym1 (a6989586621681176091 :: [a]) = FromList a6989586621681176091 :: NonEmpty a

data ToListSym0 a6989586621681176086

type ToListSym1 (a6989586621681176086 :: NonEmpty a) = ToList a6989586621681176086 :: [a]

data NonEmpty_Sym0 a6989586621681176169

type NonEmpty_Sym1 (a6989586621681176169 :: [a]) = NonEmpty_ a6989586621681176169 :: Maybe (NonEmpty a)

data XorSym0 a6989586621681176189

type XorSym1 (a6989586621681176189 :: NonEmpty Bool) = Xor a6989586621681176189 :: Bool

Orphan instances

SMonadZip NonEmpty

Instance details

Methods

sMzip :: forall a b (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing (Apply (Apply MzipSym0 t) t)

sMzipWith :: forall a b c (t :: a ~> (b ~> c)) (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply MzipWithSym0 t) t) t)

sMunzip :: forall a b (t :: NonEmpty (a, b)). Sing t -> Sing (Apply MunzipSym0 t)

PMonadZip NonEmpty

Instance details

Associated Types

type Mzip arg arg :: m (a, b)

type MzipWith arg arg arg :: m c

type Munzip arg :: (m a, m b)