Data.Singletons.Prelude.Num

Description

Defines and exports promoted and singleton versions of definitions from GHC.Num.

Be warned that some of the associated type families in the PNum class ((+), (-), and (*)) clash with their counterparts for Nat in the GHC.TypeLits module.

Documentation

class PNum (a :: Type)

Associated Types

type (arg :: a) + (arg :: a) :: a infixl 6

type (arg :: a) - (arg :: a) :: a infixl 6

type (-) a a = Apply (Apply TFHelper_6989586621679530970Sym0 a) a

type (arg :: a) * (arg :: a) :: a infixl 7

type Negate (arg :: a) :: a

type Negate a = Apply Negate_6989586621679530981Sym0 a

type Abs (arg :: a) :: a

type Signum (arg :: a) :: a

type FromInteger (arg :: Nat) :: a

class SNum a where

Minimal complete definition

(%+), (%*), sAbs, sSignum, sFromInteger

Methods

(%+) :: forall (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply (+@#@$) t) t :: a) infixl 6

(%-) :: forall (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply (-@#@$) t) t :: a) infixl 6

default (%-) :: forall (t :: a) (t :: a). (Apply (Apply (-@#@$) t) t :: a) ~ Apply (Apply TFHelper_6989586621679530970Sym0 t) t => Sing t -> Sing t -> Sing (Apply (Apply (-@#@$) t) t :: a)

(%*) :: forall (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply (*@#@$) t) t :: a) infixl 7

sNegate :: forall (t :: a). Sing t -> Sing (Apply NegateSym0 t :: a)

default sNegate :: forall (t :: a). (Apply NegateSym0 t :: a) ~ Apply Negate_6989586621679530981Sym0 t => Sing t -> Sing (Apply NegateSym0 t :: a)

sAbs :: forall (t :: a). Sing t -> Sing (Apply AbsSym0 t :: a)

sSignum :: forall (t :: a). Sing t -> Sing (Apply SignumSym0 t :: a)

sFromInteger :: forall (t :: Nat). Sing t -> Sing (Apply FromIntegerSym0 t :: a)

type family Subtract (a :: a) (a :: a) :: a where ...

Equations

Subtract x y = Apply (Apply (-@#@$) y) x

sSubtract :: forall a (t :: a) (t :: a). SNum a => Sing t -> Sing t -> Sing (Apply (Apply SubtractSym0 t) t :: a)

Defunctionalization symbols

data (+@#@$) :: forall a6989586621679530929. (~>) a6989586621679530929 ((~>) a6989586621679530929 a6989586621679530929) infixl 6

data (+@#@$$) (arg6989586621679530948 :: a6989586621679530929) :: (~>) a6989586621679530929 a6989586621679530929 infixl 6

type (+@#@$$$) (arg6989586621679530948 :: a6989586621679530929) (arg6989586621679530949 :: a6989586621679530929) = (+) arg6989586621679530948 arg6989586621679530949

data (-@#@$) :: forall a6989586621679530929. (~>) a6989586621679530929 ((~>) a6989586621679530929 a6989586621679530929) infixl 6

data (-@#@$$) (arg6989586621679530952 :: a6989586621679530929) :: (~>) a6989586621679530929 a6989586621679530929 infixl 6

type (-@#@$$$) (arg6989586621679530952 :: a6989586621679530929) (arg6989586621679530953 :: a6989586621679530929) = (-) arg6989586621679530952 arg6989586621679530953

data (*@#@$) :: forall a6989586621679530929. (~>) a6989586621679530929 ((~>) a6989586621679530929 a6989586621679530929) infixl 7

data (*@#@$$) (arg6989586621679530956 :: a6989586621679530929) :: (~>) a6989586621679530929 a6989586621679530929 infixl 7

type (*@#@$$$) (arg6989586621679530956 :: a6989586621679530929) (arg6989586621679530957 :: a6989586621679530929) = * arg6989586621679530956 arg6989586621679530957

data NegateSym0 :: forall a6989586621679530929. (~>) a6989586621679530929 a6989586621679530929

type NegateSym1 (arg6989586621679530960 :: a6989586621679530929) = Negate arg6989586621679530960

data AbsSym0 :: forall a6989586621679530929. (~>) a6989586621679530929 a6989586621679530929

type AbsSym1 (arg6989586621679530962 :: a6989586621679530929) = Abs arg6989586621679530962

data SignumSym0 :: forall a6989586621679530929. (~>) a6989586621679530929 a6989586621679530929

type SignumSym1 (arg6989586621679530964 :: a6989586621679530929) = Signum arg6989586621679530964

data FromIntegerSym0 :: forall a6989586621679530929. (~>) Nat a6989586621679530929

type FromIntegerSym1 (arg6989586621679530966 :: Nat) = FromInteger arg6989586621679530966

data SubtractSym0 :: forall a6989586621679536633. (~>) a6989586621679536633 ((~>) a6989586621679536633 a6989586621679536633)

data SubtractSym1 (a6989586621679536637 :: a6989586621679536633) :: (~>) a6989586621679536633 a6989586621679536633

type SubtractSym2 (a6989586621679536637 :: a6989586621679536633) (a6989586621679536638 :: a6989586621679536633) = Subtract a6989586621679536637 a6989586621679536638