Data.Singletons.Prelude.Show

Description

Defines the SShow singleton version of the Show type class.

Documentation

class PShow (a :: Type)

Associated Types

type ShowsPrec (arg :: Nat) (arg :: a) (arg :: Symbol) :: Symbol

type ShowsPrec a a a = Apply (Apply (Apply ShowsPrec_6989586621680296660Sym0 a) a) a

type Show_ (arg :: a) :: Symbol

type Show_ a = Apply Show__6989586621680296674Sym0 a

type ShowList (arg :: [a]) (arg :: Symbol) :: Symbol

type ShowList a a = Apply (Apply ShowList_6989586621680296682Sym0 a) a

class SShow a where

Minimal complete definition

Nothing

Methods

sShowsPrec :: forall (t :: Nat) (t :: a) (t :: Symbol). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ShowsPrecSym0 t) t) t :: Symbol)

default sShowsPrec :: forall (t :: Nat) (t :: a) (t :: Symbol). (Apply (Apply (Apply ShowsPrecSym0 t) t) t :: Symbol) ~ Apply (Apply (Apply ShowsPrec_6989586621680296660Sym0 t) t) t => Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ShowsPrecSym0 t) t) t :: Symbol)

sShow_ :: forall (t :: a). Sing t -> Sing (Apply Show_Sym0 t :: Symbol)

default sShow_ :: forall (t :: a). (Apply Show_Sym0 t :: Symbol) ~ Apply Show__6989586621680296674Sym0 t => Sing t -> Sing (Apply Show_Sym0 t :: Symbol)

sShowList :: forall (t :: [a]) (t :: Symbol). Sing t -> Sing t -> Sing (Apply (Apply ShowListSym0 t) t :: Symbol)

default sShowList :: forall (t :: [a]) (t :: Symbol). (Apply (Apply ShowListSym0 t) t :: Symbol) ~ Apply (Apply ShowList_6989586621680296682Sym0 t) t => Sing t -> Sing t -> Sing (Apply (Apply ShowListSym0 t) t :: Symbol)

type SymbolS = Symbol -> Symbol

The shows functions return a function that prepends the output Symbol to an existing Symbol. This allows constant-time concatenation of results using function composition.

type SChar = Symbol

GHC currently has no notion of type-level Chars, so we fake them with single-character Symbols.

show_ :: Show a => a -> String

show, but with an extra underscore so that its promoted counterpart (Show_) will not clash with the Show class.

type family Shows (a :: a) (a :: Symbol) :: Symbol where ...

Equations

Shows s a_6989586621680296641 = Apply (Apply (Apply ShowsPrecSym0 (FromInteger 0)) s) a_6989586621680296641

sShows :: forall a (t :: a) (t :: Symbol). SShow a => Sing t -> Sing t -> Sing (Apply (Apply ShowsSym0 t) t :: Symbol)

type family ShowListWith (a :: (~>) a ((~>) Symbol Symbol)) (a :: [a]) (a :: Symbol) :: Symbol where ...

Equations

ShowListWith _ '[] s = Apply (Apply (<>@#@$) "[]") s

ShowListWith showx ('(:) x xs) s = Apply (Apply (<>@#@$) "[") (Apply (Apply showx x) (Apply (Let6989586621680296628ShowlSym4 showx x xs s) xs))

sShowListWith :: forall a (t :: (~>) a ((~>) Symbol Symbol)) (t :: [a]) (t :: Symbol). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ShowListWithSym0 t) t) t :: Symbol)

type family ShowChar (a :: Symbol) (a :: Symbol) :: Symbol where ...

Equations

ShowChar a_6989586621680296607 a_6989586621680296609 = Apply (Apply (<>@#@$) a_6989586621680296607) a_6989586621680296609

sShowChar :: forall (t :: Symbol) (t :: Symbol). Sing t -> Sing t -> Sing (Apply (Apply ShowCharSym0 t) t :: Symbol)

type family ShowString (a :: Symbol) (a :: Symbol) :: Symbol where ...

Equations

ShowString a_6989586621680296597 a_6989586621680296599 = Apply (Apply (<>@#@$) a_6989586621680296597) a_6989586621680296599

sShowString :: forall (t :: Symbol) (t :: Symbol). Sing t -> Sing t -> Sing (Apply (Apply ShowStringSym0 t) t :: Symbol)

type family ShowParen (a :: Bool) (a :: (~>) Symbol Symbol) (a :: Symbol) :: Symbol where ...

Equations

ShowParen b p a_6989586621680296589 = Apply (Case_6989586621680296594 b p a_6989586621680296589 b) a_6989586621680296589

sShowParen :: forall (t :: Bool) (t :: (~>) Symbol Symbol) (t :: Symbol). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ShowParenSym0 t) t) t :: Symbol)

type family ShowSpace (a :: Symbol) :: Symbol where ...

Equations

ShowSpace a_6989586621680296572 = Apply (Apply Lambda_6989586621680296577Sym0 a_6989586621680296572) a_6989586621680296572

sShowSpace :: forall (t :: Symbol). Sing t -> Sing (Apply ShowSpaceSym0 t :: Symbol)

type family ShowCommaSpace (a :: Symbol) :: Symbol where ...

Equations

ShowCommaSpace a_6989586621680296567 = Apply (Apply ShowStringSym0 ", ") a_6989586621680296567

sShowCommaSpace :: forall (t :: Symbol). Sing t -> Sing (Apply ShowCommaSpaceSym0 t :: Symbol)

type family AppPrec :: Nat where ...

Equations

AppPrec = FromInteger 10

sAppPrec :: Sing (AppPrecSym0 :: Nat)

type family AppPrec1 :: Nat where ...

Equations

AppPrec1 = FromInteger 11

sAppPrec1 :: Sing (AppPrec1Sym0 :: Nat)

Defunctionalization symbols

data ShowsPrecSym0 :: forall a6989586621680296207. (~>) Nat ((~>) a6989586621680296207 ((~>) Symbol Symbol))

data ShowsPrecSym1 (arg6989586621680296645 :: Nat) :: forall a6989586621680296207. (~>) a6989586621680296207 ((~>) Symbol Symbol)

data ShowsPrecSym2 (arg6989586621680296645 :: Nat) (arg6989586621680296646 :: a6989586621680296207) :: (~>) Symbol Symbol

type ShowsPrecSym3 (arg6989586621680296645 :: Nat) (arg6989586621680296646 :: a6989586621680296207) (arg6989586621680296647 :: Symbol) = ShowsPrec arg6989586621680296645 arg6989586621680296646 arg6989586621680296647

data Show_Sym0 :: forall a6989586621680296207. (~>) a6989586621680296207 Symbol

type Show_Sym1 (arg6989586621680296651 :: a6989586621680296207) = Show_ arg6989586621680296651

data ShowListSym0 :: forall a6989586621680296207. (~>) [a6989586621680296207] ((~>) Symbol Symbol)

data ShowListSym1 (arg6989586621680296653 :: [a6989586621680296207]) :: (~>) Symbol Symbol

type ShowListSym2 (arg6989586621680296653 :: [a6989586621680296207]) (arg6989586621680296654 :: Symbol) = ShowList arg6989586621680296653 arg6989586621680296654

data ShowsSym0 :: forall a6989586621680296192. (~>) a6989586621680296192 ((~>) Symbol Symbol)

data ShowsSym1 (a6989586621680296637 :: a6989586621680296192) :: (~>) Symbol Symbol

type ShowsSym2 (a6989586621680296637 :: a6989586621680296192) (a6989586621680296638 :: Symbol) = Shows a6989586621680296637 a6989586621680296638

data ShowListWithSym0 :: forall a6989586621680296191. (~>) ((~>) a6989586621680296191 ((~>) Symbol Symbol)) ((~>) [a6989586621680296191] ((~>) Symbol Symbol))

data ShowListWithSym1 (a6989586621680296617 :: (~>) a6989586621680296191 ((~>) Symbol Symbol)) :: (~>) [a6989586621680296191] ((~>) Symbol Symbol)

data ShowListWithSym2 (a6989586621680296617 :: (~>) a6989586621680296191 ((~>) Symbol Symbol)) (a6989586621680296618 :: [a6989586621680296191]) :: (~>) Symbol Symbol

type ShowListWithSym3 (a6989586621680296617 :: (~>) a6989586621680296191 ((~>) Symbol Symbol)) (a6989586621680296618 :: [a6989586621680296191]) (a6989586621680296619 :: Symbol) = ShowListWith a6989586621680296617 a6989586621680296618 a6989586621680296619

data ShowCharSym0 :: (~>) Symbol ((~>) Symbol Symbol)

data ShowCharSym1 (a6989586621680296611 :: Symbol) :: (~>) Symbol Symbol

type ShowCharSym2 (a6989586621680296611 :: Symbol) (a6989586621680296612 :: Symbol) = ShowChar a6989586621680296611 a6989586621680296612

data ShowStringSym0 :: (~>) Symbol ((~>) Symbol Symbol)

data ShowStringSym1 (a6989586621680296601 :: Symbol) :: (~>) Symbol Symbol

type ShowStringSym2 (a6989586621680296601 :: Symbol) (a6989586621680296602 :: Symbol) = ShowString a6989586621680296601 a6989586621680296602

data ShowParenSym0 :: (~>) Bool ((~>) ((~>) Symbol Symbol) ((~>) Symbol Symbol))

data ShowParenSym1 (a6989586621680296583 :: Bool) :: (~>) ((~>) Symbol Symbol) ((~>) Symbol Symbol)

data ShowParenSym2 (a6989586621680296583 :: Bool) (a6989586621680296584 :: (~>) Symbol Symbol) :: (~>) Symbol Symbol

data ShowSpaceSym0 :: (~>) Symbol Symbol

type ShowSpaceSym1 (a6989586621680296574 :: Symbol) = ShowSpace a6989586621680296574

data ShowCommaSpaceSym0 :: (~>) Symbol Symbol

type ShowCommaSpaceSym1 (a6989586621680296569 :: Symbol) = ShowCommaSpace a6989586621680296569

type AppPrecSym0 = AppPrec

type AppPrec1Sym0 = AppPrec1