Data.Singletons.Prelude.Tuple

Description

Defines functions and datatypes relating to the singleton for tuples, including a singletons version of all the definitions in Data.Tuple.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Tuple. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Singleton definitions

See Sing for more info.

data family Sing a

The singleton kind-indexed data family.

type STuple0 = (Sing :: () -> *)

type STuple2 = (Sing :: (a, b) -> *)

type STuple3 = (Sing :: (a, b, c) -> *)

type STuple4 = (Sing :: (a, b, c, d) -> *)

type STuple5 = (Sing :: (a, b, c, d, e) -> *)

type STuple6 = (Sing :: (a, b, c, d, e, f) -> *)

type STuple7 = (Sing :: (a, b, c, d, e, f, g) -> *)

Singletons from Data.Tuple

type family Fst a :: a

Equations

Fst `(x, _z_1627926696)` = x

sFst :: forall t. Sing t -> Sing (Apply FstSym0 t :: a)

type family Snd a :: b

Equations

Snd `(_z_1627926687, y)` = y

sSnd :: forall t. Sing t -> Sing (Apply SndSym0 t :: b)

type family Curry a a a :: c

Equations

Curry f x y = Apply f (Apply (Apply Tuple2Sym0 x) y)

sCurry :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply CurrySym0 t) t) t :: c)

type family Uncurry a a :: c

Equations

Uncurry f p = Apply (Apply f (Apply FstSym0 p)) (Apply SndSym0 p)

sUncurry :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply UncurrySym0 t) t :: c)

type family Swap a :: (b, a)

Equations

Swap `(a, b)` = Apply (Apply Tuple2Sym0 b) a

sSwap :: forall t. Sing t -> Sing (Apply SwapSym0 t :: (b, a))

Defunctionalization symbols

type Tuple0Sym0 = `()`

data Tuple2Sym0 l

data Tuple2Sym1 l l

type Tuple2Sym2 t t = `(t, t)`

data Tuple3Sym0 l

data Tuple3Sym1 l l

data Tuple3Sym2 l l l

type Tuple3Sym3 t t t = `(t, t, t)`

data Tuple4Sym0 l

data Tuple4Sym1 l l

data Tuple4Sym2 l l l

data Tuple4Sym3 l l l l

type Tuple4Sym4 t t t t = `(t, t, t, t)`

data Tuple5Sym0 l

data Tuple5Sym1 l l

data Tuple5Sym2 l l l

data Tuple5Sym3 l l l l

data Tuple5Sym4 l l l l l

type Tuple5Sym5 t t t t t = `(t, t, t, t, t)`

data Tuple6Sym0 l

data Tuple6Sym1 l l

data Tuple6Sym2 l l l

data Tuple6Sym3 l l l l

data Tuple6Sym4 l l l l l

data Tuple6Sym5 l l l l l l

type Tuple6Sym6 t t t t t t = `(t, t, t, t, t, t)`

data Tuple7Sym0 l

data Tuple7Sym1 l l

data Tuple7Sym2 l l l

data Tuple7Sym3 l l l l

data Tuple7Sym4 l l l l l

data Tuple7Sym5 l l l l l l

data Tuple7Sym6 l l l l l l l

type Tuple7Sym7 t t t t t t t = `(t, t, t, t, t, t, t)`

data FstSym0 l

type FstSym1 t = Fst t

data SndSym0 l

type SndSym1 t = Snd t

data CurrySym0 l

data CurrySym1 l l

data CurrySym2 l l l

type CurrySym3 t t t = Curry t t t

data UncurrySym0 l

data UncurrySym1 l l

type UncurrySym2 t t = Uncurry t t

data SwapSym0 l

type SwapSym1 t = Swap t