Data.Singletons.Prelude.Void

Description

Defines functions and datatypes relating to the singleton for Void, including a singleton version of all the definitions in Data.Void.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Void. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The Void singleton

type family Sing :: k -> Type

The singleton kind-indexed type family.

Just as Void has no constructors, the Sing instance above also has no constructors.

data SVoid :: Void -> Type

SVoid is a kind-restricted synonym for Sing: type SVoid (a :: Void) = Sing a

Singletons from Data.Void

type family Absurd (a :: Void) :: a where ...

Equations

Absurd a = Case_6989586621679371008 a a

sAbsurd :: forall a (t :: Void). Sing t -> Sing (Apply AbsurdSym0 t :: a)

Defunctionalization symbols

data AbsurdSym0 :: forall a6989586621679371002. (~>) Void a6989586621679371002

type AbsurdSym1 (a6989586621679371005 :: Void) = Absurd a6989586621679371005