Data.Singletons.TH

Description

This module contains everything you need to derive your own singletons via Template Haskell.

TURN ON -XScopedTypeVariables IN YOUR MODULE IF YOU WANT THIS TO WORK.

Primary Template Haskell generation functions

singletons :: Quasi q => q [Dec] -> q [Dec]

Make promoted and singleton versions of all declarations given, retaining the original declarations. See http://www.cis.upenn.edu/~eir/packages/singletons/README.html for further explanation.

singletonsOnly :: Quasi q => q [Dec] -> q [Dec]

Make promoted and singleton versions of all declarations given, discarding the original declarations.

genSingletons :: Quasi q => [Name] -> q [Dec]

Generate singleton definitions from a type that is already defined. For example, the singletons package itself uses

 $(genSingletons [''Bool, ''Maybe, ''Either, ''[]])

to generate singletons for Prelude types.

promote :: Quasi q => q [Dec] -> q [Dec]

Promote every declaration given to the type level, retaining the originals.

promoteOnly :: Quasi q => q [Dec] -> q [Dec]

Promote each declaration, discarding the originals.

Functions to generate equality instances

promoteEqInstances :: Quasi q => [Name] -> q [Dec]

Produce instances for '(:==)' (type-level equality) from the given types

promoteEqInstance :: Quasi q => Name -> q [Dec]

Produce an instance for '(:==)' (type-level equality) from the given type

singEqInstances :: Quasi q => [Name] -> q [Dec]

Create instances of SEq and type-level '(:==)' for each type in the list

singEqInstance :: Quasi q => Name -> q [Dec]

Create instance of SEq and type-level '(:==)' for the given type

singEqInstancesOnly :: Quasi q => [Name] -> q [Dec]

Create instances of SEq (only -- no instance for '(:==)', which SEq generally relies on) for each type in the list

singEqInstanceOnly :: Quasi q => Name -> q [Dec]

Create instances of SEq (only -- no instance for '(:==)', which SEq generally relies on) for the given type

singDecideInstances :: Quasi q => [Name] -> q [Dec]

Create instances of SDecide for each type in the list

singDecideInstance :: Quasi q => Name -> q [Dec]

Create instance of SDecide for the given type

Utility function

cases

Arguments

:: Quasi q

=> Name
The head of the type of the scrutinee. (Like ''Maybe or ''Bool.)

-> q Exp
The scrutinee, in a Template Haskell quote

-> q Exp
The body, in a Template Haskell quote

-> q Exp

The function cases generates a case expression where each right-hand side is identical. This may be useful if the type-checker requires knowledge of which constructor is used to satisfy equality or type-class constraints, but where each constructor is treated the same.

Basic singleton definitions

data family Sing a

The singleton kind-indexed data family.

class SingI a where

A SingI constraint is essentially an implicitly-passed singleton. If you need to satisfy this constraint with an explicit singleton, please see withSingI.

Methods

sing :: Sing a

Produce the singleton explicitly. You will likely need the ScopedTypeVariables extension to use this method the way you want.

class kparam ~ KProxy => SingKind kparam where

The SingKind class is essentially a kind class. It classifies all kinds for which singletons are defined. The class supports converting between a singleton type and the base (unrefined) type which it is built from.

Associated Types

type DemoteRep kparam :: *

Get a base type from a proxy for the promoted kind. For example, DemoteRep ('KProxy :: KProxy Bool) will be the type Bool.

Methods

fromSing :: Sing (a :: k) -> DemoteRep kparam

Convert a singleton to its unrefined version.

toSing :: DemoteRep kparam -> SomeSing kparam

Convert an unrefined type to an existentially-quantified singleton type.

type KindOf a = (KProxy :: KProxy k)

Convenient synonym to refer to the kind of a type variable: type KindOf (a :: k) = ('KProxy :: KProxy k)

type Demote a = DemoteRep (KProxy :: KProxy k)

Convenient abbreviation for DemoteRep: type Demote (a :: k) = DemoteRep ('KProxy :: KProxy k)

Auxiliary definitions

These definitions might be mentioned in code generated by Template Haskell, so they must be in scope.

type family a (==) b :: Bool

A type family to compute Boolean equality. Instances are provided only for open kinds, such as * and function kinds. Instances are also provided for datatypes exported from base. A poly-kinded instance is not provided, as a recursive definition for algebraic kinds is generally more useful.

type :== a b = a == b

A re-export of the type-level (==) that conforms to the singletons naming convention.

type family If cond tru fls :: k

Type-level If. If True a b ==> a; If False a b ==> b

sIf :: Sing a -> Sing b -> Sing c -> Sing (If a b c)

Conditional over singletons

type :&& a b = a && b

class kparam ~ KProxy => SEq kparam where

The singleton analogue of Eq. Unlike the definition for Eq, it is required that instances define a body for '(%:==)'. You may also supply a body for '(%:/=)'.

Methods

(%:==) :: forall a b. Sing a -> Sing b -> Sing (a :== b)

Boolean equality on singletons

(%:/=) :: forall a b. Sing a -> Sing b -> Sing (a :/= b)

Boolean disequality on singletons

data Any

The type constructor Any is type to which you can unsafely coerce any lifted type, and back.

It is lifted, and hence represented by a pointer

It does not claim to be a data type, and that's important for the code generator, because the code gen may enter a data value but never enters a function value.

It's also used to instantiate un-constrained type variables after type checking. For example, length has type

length :: forall a. [a] -> Int

and the list datacon for the empty list has type

[] :: forall a. [a]

In order to compose these two terms as length [] a type application is required, but there is no constraint on the choice. In this situation GHC uses Any:

length (Any *) ([] (Any *))

Note that Any is kind polymorphic, and takes a kind k as its first argument. The kind of Any is thus forall k. k -> k.

class kparam ~ KProxy => SDecide kparam where

Members of the SDecide kind class support decidable equality. Instances of this class are generated alongside singleton definitions for datatypes that derive an Eq instance.

Methods

(%~) :: forall a b. Sing a -> Sing b -> Decision (a :~: b)

Compute a proof or disproof of equality, given two singletons.

data a :~: b where

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: :~: k a1 a1

data Void

A logically uninhabited data type.

type Refuted a = a -> Void

Because we can never create a value of type Void, a function that type-checks at a -> Void shows that objects of type a can never exist. Thus, we say that a is Refuted

data Decision a

A Decision about a type a is either a proof of existence or a proof that a cannot exist.

Constructors

Proved a
Witness for a

Disproved (Refuted a)
Proof that no a exists

data KProxy t

A concrete, promotable proxy type, for use at the kind level There are no instances for this because it is intended at the kind level only

Constructors

KProxy

data SomeSing kproxy where

An existentially-quantified singleton. This type is useful when you want a singleton type, but there is no way of knowing, at compile-time, what the type index will be. To make use of this type, you will generally have to use a pattern-match:

 foo :: Bool -> ...
 foo b = case toSing b of
           SomeSing sb -> {- fancy dependently-typed code with sb -}

An example like the one above may be easier to write using withSomeSing.

Constructors

SomeSing :: Sing (a :: k) -> SomeSing (KProxy :: KProxy k)