Data.Singletons.TypeError

Description

Defines a drop-in replacement for TypeError (from GHC.TypeLits) that can be used at the value level as well. Since this is a drop-in replacement, it is not recommended to import all of GHC.TypeLits and Data.Singletons.TypeError at the same time, as many of the definitons in the latter deliberately clash with the former.

Documentation

type family TypeError (a :: PErrorMessage) :: b where ...

A drop-in replacement for TypeError. This also exists at the value-level as typeError.

Equations

TypeError a = TypeError (ConvertPErrorMessage a)

sTypeError :: HasCallStack => Sing err -> Sing (TypeError err)

The singleton for typeError.

Note that this is not quite as expressive as TypeError, as it is unable to handle ShowType constructors at all.

typeError :: HasCallStack => ErrorMessage -> a

The value-level counterpart to TypeError.

Note that this is not quite as expressive as TypeError, as it is unable to print the contents of ShowType constructors (it will simply print "<type>" in their place).

data ErrorMessage' s

A description of a custom type error.

This is a variation on ErrorMessage that is parameterized over what text type is used in the Text constructor. Instantiating it with Text gives you ErrorMessage, and instantiating it with Symbol gives you PErrorMessage.

Constructors

Text s
Show the text as is.

forall t. ShowType t
Pretty print the type. ShowType :: k -> ErrorMessage

(ErrorMessage' s) :<>: (ErrorMessage' s) infixl 6
Put two pieces of error message next to each other.

(ErrorMessage' s) :$$: (ErrorMessage' s) infixl 5
Stack two pieces of error message on top of each other.

type ErrorMessage = ErrorMessage' Text

A value-level ErrorMessage` which uses Text as its text type.

type PErrorMessage = ErrorMessage' Symbol

A type-level ErrorMessage` which uses Symbol as its text kind.

type family Sing :: k -> Type

The singleton kind-indexed type family.

data SErrorMessage :: PErrorMessage -> Type where

Constructors

SText :: Sing t -> SErrorMessage ('Text t)

SShowType :: Sing ty -> SErrorMessage ('ShowType ty)

(:%<>:) :: Sing e1 -> Sing e2 -> SErrorMessage (e1 :<>: e2) infixl 6

(:%$$:) :: Sing e1 -> Sing e2 -> SErrorMessage (e1 :$$: e2) infixl 5

type family ConvertPErrorMessage (a :: PErrorMessage) :: ErrorMessage where ...

Convert a PErrorMessage to a ErrorMessage from GHC.TypeLits.

Equations

ConvertPErrorMessage ('Text t) = 'Text t

ConvertPErrorMessage ('ShowType ty) = 'ShowType ty

ConvertPErrorMessage (e1 :<>: e2) = ConvertPErrorMessage e1 :<>: ConvertPErrorMessage e2

ConvertPErrorMessage (e1 :$$: e2) = ConvertPErrorMessage e1 :$$: ConvertPErrorMessage e2

showErrorMessage :: ErrorMessage -> String

Convert an ErrorMessage into a human-readable String.

Defunctionalization symbols

data TextSym0 :: forall (s6989586621681332717 :: Type). (~>) s6989586621681332717 (ErrorMessage' (s6989586621681332717 :: Type))

type TextSym1 (t6989586621681333461 :: s6989586621681332717) = 'Text t6989586621681333461

data ShowTypeSym0 :: forall t6989586621681332718 (s6989586621681332717 :: Type). (~>) t6989586621681332718 (ErrorMessage' (s6989586621681332717 :: Type))

type ShowTypeSym1 (t6989586621681333463 :: t6989586621681332718) = 'ShowType t6989586621681333463

data (:<>:@#@$) :: forall (s6989586621681332717 :: Type). (~>) (ErrorMessage' s6989586621681332717) ((~>) (ErrorMessage' s6989586621681332717) (ErrorMessage' (s6989586621681332717 :: Type))) infixl 6

data (:<>:@#@$$) (t6989586621681333465 :: ErrorMessage' (s6989586621681332717 :: Type)) :: (~>) (ErrorMessage' s6989586621681332717) (ErrorMessage' (s6989586621681332717 :: Type)) infixl 6

type (:<>:@#@$$$) (t6989586621681333465 :: ErrorMessage' s6989586621681332717) (t6989586621681333466 :: ErrorMessage' s6989586621681332717) = '(:<>:) t6989586621681333465 t6989586621681333466

data (:$$:@#@$) :: forall (s6989586621681332717 :: Type). (~>) (ErrorMessage' s6989586621681332717) ((~>) (ErrorMessage' s6989586621681332717) (ErrorMessage' (s6989586621681332717 :: Type))) infixl 5

data (:$$:@#@$$) (t6989586621681333469 :: ErrorMessage' (s6989586621681332717 :: Type)) :: (~>) (ErrorMessage' s6989586621681332717) (ErrorMessage' (s6989586621681332717 :: Type)) infixl 5

type (:$$:@#@$$$) (t6989586621681333469 :: ErrorMessage' s6989586621681332717) (t6989586621681333470 :: ErrorMessage' s6989586621681332717) = '(:$$:) t6989586621681333469 t6989586621681333470

data TypeErrorSym0 :: forall b6989586621681332701. (~>) PErrorMessage b6989586621681332701

type TypeErrorSym1 (a6989586621681332702 :: PErrorMessage) = TypeError a6989586621681332702