Data.Singletons.TypeRepTYPE

Description

This module defines singleton instances making TypeRep the singleton for the kind TYPE rep (for some RuntimeRep rep), an instantiation of which is the famous kind Type. The definitions don't fully line up with what is expected within the singletons library, so expect unusual results!

Documentation

type family Sing :: k -> Type

The singleton kind-indexed type family.

Here is the definition of the singleton for TYPE rep:

type instance Sing \@(TYPE rep) = TypeRep

Instances for SingI, SingKind, SEq, SDecide, and TestCoercion are also supplied.

data SomeTypeRepTYPE :: RuntimeRep -> Type where

A variant of SomeTypeRep whose underlying TypeRep is restricted to kind TYPE rep (for some RuntimeRep rep).

Constructors

SomeTypeRepTYPE :: forall (rep :: RuntimeRep) (a :: TYPE rep). !(TypeRep a) -> SomeTypeRepTYPE rep

Orphan instances

SingKind (TYPE rep)

Instance details

Associated Types

type Demote (TYPE rep) = (r :: Type)

Methods

fromSing :: forall (a :: TYPE rep). Sing a -> Demote (TYPE rep)

toSing :: Demote (TYPE rep) -> SomeSing (TYPE rep)

SDecide (TYPE rep)

Instance details

Methods

(%~) :: forall (a :: TYPE rep) (b :: TYPE rep). Sing a -> Sing b -> Decision (a :~: b)

PEq (TYPE rep)

Instance details

Associated Types

type x == y :: Bool

type x /= y :: Bool

SEq (TYPE rep)

Instance details

Methods

(%==) :: forall (a :: TYPE rep) (b :: TYPE rep). Sing a -> Sing b -> Sing (a == b)

(%/=) :: forall (a :: TYPE rep) (b :: TYPE rep). Sing a -> Sing b -> Sing (a /= b)

Typeable a => SingI (a :: TYPE rep)

Instance details

Methods

sing :: Sing a