Data.Singletons.Types

Description

Defines and exports types that are useful when working with singletons. Some of these are re-exports from Data.Type.Equality.

Documentation

data KProxy t

A concrete, promotable proxy type, for use at the kind level There are no instances for this because it is intended at the kind level only

Constructors

KProxy

data Proxy t

A concrete, poly-kinded proxy type

Constructors

Proxy

data a :~: b where

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: (:~:) k a1 a1

gcastWith :: (:~:) k a b -> ((~) k a b -> r) -> r

Generalized form of type-safe cast using propositional equality

class TestEquality f where

This class contains types where you can learn the equality of two types from information contained in terms. Typically, only singleton types should inhabit this class.

Methods

testEquality :: f a -> f b -> Maybe ((:~:) k a b)

Conditionally prove the equality of a and b.

type family Not a :: Bool

Type-level "not"

Equations

Not False = True

Not True = False

type family If cond tru fls :: k

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If k True tru fls = tru

If k False tru fls = fls

type family a == b :: Bool

A type family to compute Boolean equality. Instances are provided only for open kinds, such as * and function kinds. Instances are also provided for datatypes exported from base. A poly-kinded instance is not provided, as a recursive definition for algebraic kinds is generally more useful.

type (:==) a b = a == b

A re-export of the type-level (==) that conforms to the singletons naming convention.