Data.Singletons.Void

Description

This module is a reimplementation of Edward Kmett's void package. It is included within singletons to avoid depending on void and all the packages that depends on (including text). If this causes problems for you (that singletons has its own Void type), please let me (Richard Eisenberg) know at eir at cis.upenn.edu.

Documentation

data Void

A logically uninhabited data type.

absurd :: Void -> a

Since Void values logically don't exist, this witnesses the logical reasoning tool of "ex falso quodlibet".

vacuous :: Functor f => f Void -> f a

If Void is uninhabited then any Functor that holds only values of type Void is holding no values.

vacuousM :: Monad m => m Void -> m a

If Void is uninhabited then any Monad that holds values of type Void is holding no values.