Data.Promotion.Prelude

Description

Mimics the Haskell Prelude, but with promoted types.

Standard types, classes and related functions

Basic data types

type family If k (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If k True tru fls = tru

If k False tru fls = fls

type family Not (a :: Bool) :: Bool where ...

Equations

Not False = TrueSym0

Not True = FalseSym0

type family (a :: Bool) :&& (a :: Bool) :: Bool where ... infixr 3

Equations

False :&& _z_6989586621679545874 = FalseSym0

True :&& x = x

type family (a :: Bool) :|| (a :: Bool) :: Bool where ... infixr 2

Equations

False :|| x = x

True :|| _z_6989586621679545862 = TrueSym0

type family Otherwise :: Bool where ...

Equations

Otherwise = TrueSym0

maybe_ :: forall b a. b -> (a -> b) -> Maybe a -> b

type family Maybe_ (a :: b) (a :: TyFun a b -> Type) (a :: Maybe a) :: b where ...

Equations

Maybe_ n _z_6989586621679770024 Nothing = n

Maybe_ _z_6989586621679770027 f (Just x) = Apply f x

either_ :: forall a c b. (a -> c) -> (b -> c) -> Either a b -> c

type family Either_ (a :: TyFun a c -> Type) (a :: TyFun b c -> Type) (a :: Either a b) :: c where ...

Equations

Either_ f _z_6989586621679760570 (Left x) = Apply f x

Either_ _z_6989586621679760574 g (Right y) = Apply g y

data Symbol :: *

(Kind) This is the kind of type-level symbols. Declared here because class IP needs it

type family Fst (a :: (a, b)) :: a where ...

Equations

Fst '(x, _z_6989586621679781175) = x

type family Snd (a :: (a, b)) :: b where ...

Equations

Snd '(_z_6989586621679781166, y) = y

type family Curry (a :: TyFun (a, b) c -> Type) (a :: a) (a :: b) :: c where ...

Equations

Curry f x y = Apply f (Apply (Apply Tuple2Sym0 x) y)

type family Uncurry (a :: TyFun a (TyFun b c -> Type) -> Type) (a :: (a, b)) :: c where ...

Equations

Uncurry f p = Apply (Apply f (Apply FstSym0 p)) (Apply SndSym0 p)

Error reporting

type family Error (str :: k0) :: k

The promotion of error. This version is more poly-kinded for easier use.

data ErrorSym0 l

Promoted equality

module Data.Promotion.Prelude.Eq

Promoted comparisons

module Data.Promotion.Prelude.Ord

Promoted enumerations

As a matter of convenience, the promoted Prelude does not export promoted succ and pred, due to likely conflicts with unary numbers. Please import Enum directly if you want these.

module Data.Promotion.Prelude.Enum

Promoted numbers

module Data.Promotion.Prelude.Num

Miscellaneous functions

type family Id (a :: a) :: a where ...

Equations

Id x = x

type family Const (a :: a) (a :: b) :: a where ...

Equations

Const x _z_6989586621679549582 = x

type family ((a :: TyFun b c -> Type) :. (a :: TyFun a b -> Type)) (a :: a) :: c where ... infixr 9

Equations

(f :. g) a_6989586621679549545 = Apply (Apply (Apply (Apply Lambda_6989586621679549550Sym0 f) g) a_6989586621679549545) a_6989586621679549545

type family (f :: TyFun a b -> *) $ (x :: a) :: b infixr 0

type family (f :: TyFun a b -> *) $! (x :: a) :: b infixr 0

type family Flip (a :: TyFun a (TyFun b c -> Type) -> Type) (a :: b) (a :: a) :: c where ...

Equations

Flip f x y = Apply (Apply f y) x

type family AsTypeOf (a :: a) (a :: a) :: a where ...

Equations

AsTypeOf a_6989586621679549585 a_6989586621679549587 = Apply (Apply ConstSym0 a_6989586621679549585) a_6989586621679549587

type family Until (a :: TyFun a Bool -> Type) (a :: TyFun a a -> Type) (a :: a) :: a where ...

Equations

Until p f a_6989586621680072536 = Apply (Let6989586621680072541GoSym3 p f a_6989586621680072536) a_6989586621680072536

type family Seq (a :: a) (a :: b) :: b where ... infixr 0

Equations

Seq _z_6989586621679549508 x = x

List operations

type family Map (a :: TyFun a b -> Type) (a :: [a]) :: [b] where ...

Equations

Map _z_6989586621679549627 '[] = '[]

Map f ((:) x xs) = Apply (Apply (:$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

type family (a :: [a]) :++ (a :: [a]) :: [a] where ... infixr 5

Equations

'[] :++ ys = ys

((:) x xs) :++ ys = Apply (Apply (:$) x) (Apply (Apply (:++$) xs) ys)

type family Filter (a :: TyFun a Bool -> Type) (a :: [a]) :: [a] where ...

Equations

Filter _p '[] = '[]

Filter p ((:) x xs) = Case_6989586621679797784 p x xs (Let6989586621679797771Scrutinee_6989586621679796222Sym3 p x xs)

type family Head (a :: [a]) :: a where ...

Equations

Head ((:) a _z_6989586621679800389) = a

Head '[] = Apply ErrorSym0 "Data.Singletons.List.head: empty list"

type family Last (a :: [a]) :: a where ...

Equations

Last '[] = Apply ErrorSym0 "Data.Singletons.List.last: empty list"

Last '[x] = x

Last ((:) _z_6989586621679800379 ((:) x xs)) = Apply LastSym0 (Apply (Apply (:$) x) xs)

type family Tail (a :: [a]) :: [a] where ...

Equations

Tail ((:) _z_6989586621679800370 t) = t

Tail '[] = Apply ErrorSym0 "Data.Singletons.List.tail: empty list"

type family Init (a :: [a]) :: [a] where ...

Equations

Init '[] = Apply ErrorSym0 "Data.Singletons.List.init: empty list"

Init ((:) x xs) = Apply (Apply (Let6989586621679800339Init'Sym2 x xs) x) xs

type family Null (a :: [a]) :: Bool where ...

Equations

Null '[] = TrueSym0

Null ((:) _z_6989586621679800268 _z_6989586621679800271) = FalseSym0

type family Length (a :: [a]) :: Nat where ...

Equations

Length '[] = FromInteger 0

Length ((:) _z_6989586621679797183 xs) = Apply (Apply (:+$) (FromInteger 1)) (Apply LengthSym0 xs)

type family (a :: [a]) :!! (a :: Nat) :: a where ...

Equations

'[] :!! _z_6989586621679797135 = Apply ErrorSym0 "Data.Singletons.List.!!: index too large"

((:) x xs) :!! n = Case_6989586621679797154 x xs n (Let6989586621679797141Scrutinee_6989586621679796262Sym3 x xs n)

type family Reverse (a :: [a]) :: [a] where ...

Equations

Reverse l = Apply (Apply (Let6989586621679800234RevSym1 l) l) '[]

Reducing lists (folds)

type family Foldl (a :: TyFun b (TyFun a b -> Type) -> Type) (a :: b) (a :: [a]) :: b where ...

Equations

Foldl f z0 xs0 = Apply (Apply (Let6989586621679508526LgoSym3 f z0 xs0) z0) xs0

type family Foldl1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: a where ...

Equations

Foldl1 f ((:) x xs) = Apply (Apply (Apply FoldlSym0 f) x) xs

Foldl1 _z_6989586621679799478 '[] = Apply ErrorSym0 "Data.Singletons.List.foldl1: empty list"

type family Foldr (a :: TyFun a (TyFun b b -> Type) -> Type) (a :: b) (a :: [a]) :: b where ...

Equations

Foldr k z a_6989586621679549648 = Apply (Let6989586621679549653GoSym3 k z a_6989586621679549648) a_6989586621679549648

type family Foldr1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: a where ...

Equations

Foldr1 _z_6989586621679799436 '[x] = x

Foldr1 f ((:) x ((:) wild_6989586621679796168 wild_6989586621679796170)) = Apply (Apply f x) (Apply (Apply Foldr1Sym0 f) (Let6989586621679799444XsSym4 f x wild_6989586621679796168 wild_6989586621679796170))

Foldr1 _z_6989586621679799463 '[] = Apply ErrorSym0 "Data.Singletons.List.foldr1: empty list"

Special folds

type family And (a :: [Bool]) :: Bool where ...

Equations

And '[] = TrueSym0

And ((:) x xs) = Apply (Apply (:&&$) x) (Apply AndSym0 xs)

type family Or (a :: [Bool]) :: Bool where ...

Equations

Or '[] = FalseSym0

Or ((:) x xs) = Apply (Apply (:||$) x) (Apply OrSym0 xs)

any_ :: forall a. (a -> Bool) -> [a] -> Bool

type family Any_ (a :: TyFun a Bool -> Type) (a :: [a]) :: Bool where ...

Equations

Any_ _z_6989586621679785031 '[] = FalseSym0

Any_ p ((:) x xs) = Apply (Apply (:||$) (Apply p x)) (Apply (Apply Any_Sym0 p) xs)

type family All (a :: TyFun a Bool -> Type) (a :: [a]) :: Bool where ...

Equations

All _z_6989586621679799389 '[] = TrueSym0

All p ((:) x xs) = Apply (Apply (:&&$) (Apply p x)) (Apply (Apply AllSym0 p) xs)

type family Sum (a :: [a]) :: a where ...

Equations

Sum l = Apply (Apply (Let6989586621679797216Sum'Sym1 l) l) (FromInteger 0)

type family Product (a :: [a]) :: a where ...

Equations

Product l = Apply (Apply (Let6989586621679797192ProdSym1 l) l) (FromInteger 1)

type family Concat (a :: [[a]]) :: [a] where ...

Equations

Concat a_6989586621679799420 = Apply (Apply (Apply FoldrSym0 (:++$)) '[]) a_6989586621679799420

type family ConcatMap (a :: TyFun a [b] -> Type) (a :: [a]) :: [b] where ...

Equations

ConcatMap f a_6989586621679799416 = Apply (Apply (Apply FoldrSym0 (Apply (Apply (:.$) (:++$)) f)) '[]) a_6989586621679799416

type family Maximum (a :: [a]) :: a where ...

Equations

Maximum '[] = Apply ErrorSym0 "Data.Singletons.List.maximum: empty list"

Maximum ((:) wild_6989586621679796248 wild_6989586621679796250) = Apply (Apply Foldl1Sym0 MaxSym0) (Let6989586621679799661XsSym2 wild_6989586621679796248 wild_6989586621679796250)

type family Minimum (a :: [a]) :: a where ...

Equations

Minimum '[] = Apply ErrorSym0 "Data.Singletons.List.minimum: empty list"

Minimum ((:) wild_6989586621679796252 wild_6989586621679796254) = Apply (Apply Foldl1Sym0 MinSym0) (Let6989586621679799675XsSym2 wild_6989586621679796252 wild_6989586621679796254)

Building lists

Scans

type family Scanl (a :: TyFun b (TyFun a b -> Type) -> Type) (a :: b) (a :: [a]) :: [b] where ...

Equations

Scanl f q ls = Apply (Apply (:$) q) (Case_6989586621679799360 f q ls ls)

type family Scanl1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: [a] where ...

Equations

Scanl1 f ((:) x xs) = Apply (Apply (Apply ScanlSym0 f) x) xs

Scanl1 _z_6989586621679799377 '[] = '[]

type family Scanr (a :: TyFun a (TyFun b b -> Type) -> Type) (a :: b) (a :: [a]) :: [b] where ...

Equations

Scanr _z_6989586621679799310 q0 '[] = Apply (Apply (:$) q0) '[]

Scanr f q0 ((:) x xs) = Case_6989586621679799337 f q0 x xs (Let6989586621679799318Scrutinee_6989586621679796172Sym4 f q0 x xs)

type family Scanr1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: [a] where ...

Equations

Scanr1 _z_6989586621679799241 '[] = '[]

Scanr1 _z_6989586621679799244 '[x] = Apply (Apply (:$) x) '[]

Scanr1 f ((:) x ((:) wild_6989586621679796176 wild_6989586621679796178)) = Case_6989586621679799290 f x wild_6989586621679796176 wild_6989586621679796178 (Let6989586621679799271Scrutinee_6989586621679796174Sym4 f x wild_6989586621679796176 wild_6989586621679796178)

Infinite lists

type family Replicate (a :: Nat) (a :: a) :: [a] where ...

Equations

Replicate n x = Case_6989586621679797176 n x (Let6989586621679797168Scrutinee_6989586621679796260Sym2 n x)

Sublists

type family Take (a :: Nat) (a :: [a]) :: [a] where ...

Equations

Take _z_6989586621679797363 '[] = '[]

Take n ((:) x xs) = Case_6989586621679797382 n x xs (Let6989586621679797369Scrutinee_6989586621679796244Sym3 n x xs)

type family Drop (a :: Nat) (a :: [a]) :: [a] where ...

Equations

Drop _z_6989586621679797332 '[] = '[]

Drop n ((:) x xs) = Case_6989586621679797351 n x xs (Let6989586621679797338Scrutinee_6989586621679796246Sym3 n x xs)

type family SplitAt (a :: Nat) (a :: [a]) :: ([a], [a]) where ...

Equations

SplitAt n xs = Apply (Apply Tuple2Sym0 (Apply (Apply TakeSym0 n) xs)) (Apply (Apply DropSym0 n) xs)

type family TakeWhile (a :: TyFun a Bool -> Type) (a :: [a]) :: [a] where ...

Equations

TakeWhile _z_6989586621679797736 '[] = '[]

TakeWhile p ((:) x xs) = Case_6989586621679797755 p x xs (Let6989586621679797742Scrutinee_6989586621679796234Sym3 p x xs)

type family DropWhile (a :: TyFun a Bool -> Type) (a :: [a]) :: [a] where ...

Equations

DropWhile _z_6989586621679797692 '[] = '[]

DropWhile p ((:) x xs') = Case_6989586621679797724 p x xs' (Let6989586621679797711Scrutinee_6989586621679796236Sym3 p x xs')

type family Span (a :: TyFun a Bool -> Type) (a :: [a]) :: ([a], [a]) where ...

Equations

Span _z_6989586621679797510 '[] = Apply (Apply Tuple2Sym0 (Let6989586621679797513XsSym1 _z_6989586621679797510)) (Let6989586621679797513XsSym1 _z_6989586621679797510)

Span p ((:) x xs') = Case_6989586621679797546 p x xs' (Let6989586621679797533Scrutinee_6989586621679796240Sym3 p x xs')

type family Break (a :: TyFun a Bool -> Type) (a :: [a]) :: ([a], [a]) where ...

Equations

Break _z_6989586621679797405 '[] = Apply (Apply Tuple2Sym0 (Let6989586621679797408XsSym1 _z_6989586621679797405)) (Let6989586621679797408XsSym1 _z_6989586621679797405)

Break p ((:) x xs') = Case_6989586621679797441 p x xs' (Let6989586621679797428Scrutinee_6989586621679796242Sym3 p x xs')

Searching lists

type family Elem (a :: a) (a :: [a]) :: Bool where ...

Equations

Elem _z_6989586621679798752 '[] = FalseSym0

Elem x ((:) y ys) = Apply (Apply (:||$) (Apply (Apply (:==$) x) y)) (Apply (Apply ElemSym0 x) ys)

type family NotElem (a :: a) (a :: [a]) :: Bool where ...

Equations

NotElem _z_6989586621679798737 '[] = TrueSym0

NotElem x ((:) y ys) = Apply (Apply (:&&$) (Apply (Apply (:/=$) x) y)) (Apply (Apply NotElemSym0 x) ys)

type family Lookup (a :: a) (a :: [(a, b)]) :: Maybe b where ...

Equations

Lookup _key '[] = NothingSym0

Lookup key ((:) '(x, y) xys) = Case_6989586621679797320 key x y xys (Let6989586621679797301Scrutinee_6989586621679796256Sym4 key x y xys)

Zipping and unzipping lists

type family Zip (a :: [a]) (a :: [b]) :: [(a, b)] where ...

Equations

Zip ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply Tuple2Sym0 x) y)) (Apply (Apply ZipSym0 xs) ys)

Zip '[] '[] = '[]

Zip ((:) _z_6989586621679798614 _z_6989586621679798617) '[] = '[]

Zip '[] ((:) _z_6989586621679798620 _z_6989586621679798623) = '[]

type family Zip3 (a :: [a]) (a :: [b]) (a :: [c]) :: [(a, b, c)] where ...

Equations

Zip3 ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply Tuple3Sym0 a) b) c)) (Apply (Apply (Apply Zip3Sym0 as) bs) cs)

Zip3 '[] '[] '[] = '[]

Zip3 '[] '[] ((:) _z_6989586621679798547 _z_6989586621679798550) = '[]

Zip3 '[] ((:) _z_6989586621679798553 _z_6989586621679798556) '[] = '[]

Zip3 '[] ((:) _z_6989586621679798559 _z_6989586621679798562) ((:) _z_6989586621679798565 _z_6989586621679798568) = '[]

Zip3 ((:) _z_6989586621679798571 _z_6989586621679798574) '[] '[] = '[]

Zip3 ((:) _z_6989586621679798577 _z_6989586621679798580) '[] ((:) _z_6989586621679798583 _z_6989586621679798586) = '[]

Zip3 ((:) _z_6989586621679798589 _z_6989586621679798592) ((:) _z_6989586621679798595 _z_6989586621679798598) '[] = '[]

type family ZipWith (a :: TyFun a (TyFun b c -> Type) -> Type) (a :: [a]) (a :: [b]) :: [c] where ...

Equations

ZipWith f ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply f x) y)) (Apply (Apply (Apply ZipWithSym0 f) xs) ys)

ZipWith _z_6989586621679798505 '[] '[] = '[]

ZipWith _z_6989586621679798508 ((:) _z_6989586621679798511 _z_6989586621679798514) '[] = '[]

ZipWith _z_6989586621679798517 '[] ((:) _z_6989586621679798520 _z_6989586621679798523) = '[]

type family ZipWith3 (a :: TyFun a (TyFun b (TyFun c d -> Type) -> Type) -> Type) (a :: [a]) (a :: [b]) (a :: [c]) :: [d] where ...

Equations

ZipWith3 z ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply z a) b) c)) (Apply (Apply (Apply (Apply ZipWith3Sym0 z) as) bs) cs)

ZipWith3 _z_6989586621679798410 '[] '[] '[] = '[]

ZipWith3 _z_6989586621679798413 '[] '[] ((:) _z_6989586621679798416 _z_6989586621679798419) = '[]

ZipWith3 _z_6989586621679798422 '[] ((:) _z_6989586621679798425 _z_6989586621679798428) '[] = '[]

ZipWith3 _z_6989586621679798431 '[] ((:) _z_6989586621679798434 _z_6989586621679798437) ((:) _z_6989586621679798440 _z_6989586621679798443) = '[]

ZipWith3 _z_6989586621679798446 ((:) _z_6989586621679798449 _z_6989586621679798452) '[] '[] = '[]

ZipWith3 _z_6989586621679798455 ((:) _z_6989586621679798458 _z_6989586621679798461) '[] ((:) _z_6989586621679798464 _z_6989586621679798467) = '[]

ZipWith3 _z_6989586621679798470 ((:) _z_6989586621679798473 _z_6989586621679798476) ((:) _z_6989586621679798479 _z_6989586621679798482) '[] = '[]

type family Unzip (a :: [(a, b)]) :: ([a], [b]) where ...

Equations

Unzip xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798356Sym0 xs)) (Apply (Apply Tuple2Sym0 '[]) '[])) xs

type family Unzip3 (a :: [(a, b, c)]) :: ([a], [b], [c]) where ...

Equations

Unzip3 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798324Sym0 xs)) (Apply (Apply (Apply Tuple3Sym0 '[]) '[]) '[])) xs

Other datatypes

data Proxy k t :: forall k. k -> *

A concrete, poly-kinded proxy type

Constructors

Proxy

Defunctionalization symbols

type FalseSym0 = False

type TrueSym0 = True

data NotSym0 l

type NotSym1 t = Not t

data (:&&$) l

data l :&&$$ l

type (:&&$$$) t t = (:&&) t t

data (:||$) l

data l :||$$ l

type (:||$$$) t t = (:||) t t

type OtherwiseSym0 = Otherwise

type NothingSym0 = Nothing

data JustSym0 l

type JustSym1 t = Just t

data Maybe_Sym0 l

data Maybe_Sym1 l l

data Maybe_Sym2 l l l

type Maybe_Sym3 t t t = Maybe_ t t t

data LeftSym0 l

type LeftSym1 t = Left t

data RightSym0 l

type RightSym1 t = Right t

data Either_Sym0 l

data Either_Sym1 l l

data Either_Sym2 l l l

type Either_Sym3 t t t = Either_ t t t

type Tuple0Sym0 = '()

data Tuple2Sym0 l

data Tuple2Sym1 l l

type Tuple2Sym2 t t = '(t, t)

data Tuple3Sym0 l

data Tuple3Sym1 l l

data Tuple3Sym2 l l l

type Tuple3Sym3 t t t = '(t, t, t)

data Tuple4Sym0 l

data Tuple4Sym1 l l

data Tuple4Sym2 l l l

data Tuple4Sym3 l l l l

type Tuple4Sym4 t t t t = '(t, t, t, t)

data Tuple5Sym0 l

data Tuple5Sym1 l l

data Tuple5Sym2 l l l

data Tuple5Sym3 l l l l

data Tuple5Sym4 l l l l l

type Tuple5Sym5 t t t t t = '(t, t, t, t, t)

data Tuple6Sym0 l

data Tuple6Sym1 l l

data Tuple6Sym2 l l l

data Tuple6Sym3 l l l l

data Tuple6Sym4 l l l l l

data Tuple6Sym5 l l l l l l

type Tuple6Sym6 t t t t t t = '(t, t, t, t, t, t)

data Tuple7Sym0 l

data Tuple7Sym1 l l

data Tuple7Sym2 l l l

data Tuple7Sym3 l l l l

data Tuple7Sym4 l l l l l

data Tuple7Sym5 l l l l l l

data Tuple7Sym6 l l l l l l l

type Tuple7Sym7 t t t t t t t = '(t, t, t, t, t, t, t)

data FstSym0 l

type FstSym1 t = Fst t

data SndSym0 l

type SndSym1 t = Snd t

data CurrySym0 l

data CurrySym1 l l

data CurrySym2 l l l

type CurrySym3 t t t = Curry t t t

data UncurrySym0 l

data UncurrySym1 l l

type UncurrySym2 t t = Uncurry t t

data (:^$) l

data l :^$$ l

data IdSym0 l

type IdSym1 t = Id t

data ConstSym0 l

data ConstSym1 l l

type ConstSym2 t t = Const t t

data (:.$) l

data l :.$$ l

data (l :.$$$ l) l

data ($$) :: TyFun (TyFun a b -> *) (TyFun a b -> *) -> *

data ($$$) :: (TyFun a b -> *) -> TyFun a b -> *

type ($$$$) a b = ($) a b

data ($!$) :: TyFun (TyFun a b -> *) (TyFun a b -> *) -> *

data ($!$$) :: (TyFun a b -> *) -> TyFun a b -> *

type ($!$$$) a b = ($!) a b

data FlipSym0 l

data FlipSym1 l l

data FlipSym2 l l l

data AsTypeOfSym0 l

data AsTypeOfSym1 l l

type AsTypeOfSym2 t t = AsTypeOf t t

data SeqSym0 l

data SeqSym1 l l

type SeqSym2 t t = Seq t t

data (:$) l

data l :$$ l

type (:$$$) t t = (:) t t

type NilSym0 = '[]

data MapSym0 l

data MapSym1 l l

type MapSym2 t t = Map t t

data ReverseSym0 l

type ReverseSym1 t = Reverse t

data l :++$$ l

data (:++$) l

data HeadSym0 l

type HeadSym1 t = Head t

data LastSym0 l

type LastSym1 t = Last t

data TailSym0 l

type TailSym1 t = Tail t

data InitSym0 l

type InitSym1 t = Init t

data NullSym0 l

type NullSym1 t = Null t

data FoldlSym0 l

data FoldlSym1 l l

data FoldlSym2 l l l

type FoldlSym3 t t t = Foldl t t t

data Foldl1Sym0 l

data Foldl1Sym1 l l

type Foldl1Sym2 t t = Foldl1 t t

data FoldrSym0 l

data FoldrSym1 l l

data FoldrSym2 l l l

type FoldrSym3 t t t = Foldr t t t

data Foldr1Sym0 l

data Foldr1Sym1 l l

type Foldr1Sym2 t t = Foldr1 t t

data ConcatSym0 l

type ConcatSym1 t = Concat t

data ConcatMapSym0 l

data ConcatMapSym1 l l

type ConcatMapSym2 t t = ConcatMap t t

data MaximumBySym0 l

data MaximumBySym1 l l

type MaximumBySym2 t t = MaximumBy t t

data MinimumBySym0 l

data MinimumBySym1 l l

type MinimumBySym2 t t = MinimumBy t t

data AndSym0 l

type AndSym1 t = And t

data OrSym0 l

type OrSym1 t = Or t

data Any_Sym0 l

data Any_Sym1 l l

type Any_Sym2 t t = Any_ t t

data AllSym0 l

data AllSym1 l l

type AllSym2 t t = All t t

data ScanlSym0 l

data ScanlSym1 l l

data ScanlSym2 l l l

type ScanlSym3 t t t = Scanl t t t

data Scanl1Sym0 l

data Scanl1Sym1 l l

type Scanl1Sym2 t t = Scanl1 t t

data ScanrSym0 l

data ScanrSym1 l l

data ScanrSym2 l l l

type ScanrSym3 t t t = Scanr t t t

data Scanr1Sym0 l

data Scanr1Sym1 l l

type Scanr1Sym2 t t = Scanr1 t t

data ElemSym0 l

data ElemSym1 l l

type ElemSym2 t t = Elem t t

data NotElemSym0 l

data NotElemSym1 l l

type NotElemSym2 t t = NotElem t t

data ZipSym0 l

data ZipSym1 l l

type ZipSym2 t t = Zip t t

data Zip3Sym0 l

data Zip3Sym1 l l

data Zip3Sym2 l l l

type Zip3Sym3 t t t = Zip3 t t t

data ZipWithSym0 l

data ZipWithSym1 l l

data ZipWithSym2 l l l

type ZipWithSym3 t t t = ZipWith t t t

data ZipWith3Sym0 l

data ZipWith3Sym1 l l

data ZipWith3Sym2 l l l

data ZipWith3Sym3 l l l l

data UnzipSym0 l

type UnzipSym1 t = Unzip t

data UntilSym0 l

data UntilSym1 l l

data UntilSym2 l l l

type UntilSym3 t t t = Until t t t

data LengthSym0 l

type LengthSym1 t = Length t

data SumSym0 l

type SumSym1 t = Sum t

data ProductSym0 l

type ProductSym1 t = Product t

data ReplicateSym0 l

data ReplicateSym1 l l

type ReplicateSym2 t t = Replicate t t

data TakeSym0 l

data TakeSym1 l l

type TakeSym2 t t = Take t t

data DropSym0 l

data DropSym1 l l

type DropSym2 t t = Drop t t

data SplitAtSym0 l

data SplitAtSym1 l l

type SplitAtSym2 t t = SplitAt t t

data TakeWhileSym0 l

data TakeWhileSym1 l l

type TakeWhileSym2 t t = TakeWhile t t

data DropWhileSym0 l

data DropWhileSym1 l l

type DropWhileSym2 t t = DropWhile t t

data SpanSym0 l

data SpanSym1 l l

type SpanSym2 t t = Span t t

data BreakSym0 l

data BreakSym1 l l

type BreakSym2 t t = Break t t

data LookupSym0 l

data LookupSym1 l l

type LookupSym2 t t = Lookup t t

data FilterSym0 l

data FilterSym1 l l

type FilterSym2 t t = Filter t t

data (:!!$) l

data l :!!$$ l

type (:!!$$$) t t = (:!!) t t