Data.Promotion.Prelude.List

Description

Defines promoted functions and datatypes relating to List, including a promoted version of all the definitions in Data.List.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.List. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Basic functions

type family (a :: [a]) :++ (a :: [a]) :: [a] where ... infixr 5

Equations

'[] :++ ys = ys

((:) x xs) :++ ys = Apply (Apply (:$) x) (Apply (Apply (:++$) xs) ys)

type family Head (a :: [a]) :: a where ...

Equations

Head ((:) a _z_6989586621679800389) = a

Head '[] = Apply ErrorSym0 "Data.Singletons.List.head: empty list"

type family Last (a :: [a]) :: a where ...

Equations

Last '[] = Apply ErrorSym0 "Data.Singletons.List.last: empty list"

Last '[x] = x

Last ((:) _z_6989586621679800379 ((:) x xs)) = Apply LastSym0 (Apply (Apply (:$) x) xs)

type family Tail (a :: [a]) :: [a] where ...

Equations

Tail ((:) _z_6989586621679800370 t) = t

Tail '[] = Apply ErrorSym0 "Data.Singletons.List.tail: empty list"

type family Init (a :: [a]) :: [a] where ...

Equations

Init '[] = Apply ErrorSym0 "Data.Singletons.List.init: empty list"

Init ((:) x xs) = Apply (Apply (Let6989586621679800339Init'Sym2 x xs) x) xs

type family Null (a :: [a]) :: Bool where ...

Equations

Null '[] = TrueSym0

Null ((:) _z_6989586621679800268 _z_6989586621679800271) = FalseSym0

type family Length (a :: [a]) :: Nat where ...

Equations

Length '[] = FromInteger 0

Length ((:) _z_6989586621679797183 xs) = Apply (Apply (:+$) (FromInteger 1)) (Apply LengthSym0 xs)

List transformations

type family Map (a :: TyFun a b -> Type) (a :: [a]) :: [b] where ...

Equations

Map _z_6989586621679549627 '[] = '[]

Map f ((:) x xs) = Apply (Apply (:$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

type family Reverse (a :: [a]) :: [a] where ...

Equations

Reverse l = Apply (Apply (Let6989586621679800234RevSym1 l) l) '[]

type family Intersperse (a :: a) (a :: [a]) :: [a] where ...

Equations

Intersperse _z_6989586621679800212 '[] = '[]

Intersperse sep ((:) x xs) = Apply (Apply (:$) x) (Apply (Apply PrependToAllSym0 sep) xs)

type family Intercalate (a :: [a]) (a :: [[a]]) :: [a] where ...

Equations

Intercalate xs xss = Apply ConcatSym0 (Apply (Apply IntersperseSym0 xs) xss)

type family Transpose (a :: [[a]]) :: [[a]] where ...

Equations

Transpose '[] = '[]

Transpose ((:) '[] xss) = Apply TransposeSym0 xss

Transpose ((:) ((:) x xs) xss) = Apply (Apply (:$) (Apply (Apply (:$) x) (Apply (Apply MapSym0 HeadSym0) xss))) (Apply TransposeSym0 (Apply (Apply (:$) xs) (Apply (Apply MapSym0 TailSym0) xss)))

type family Subsequences (a :: [a]) :: [[a]] where ...

Equations

Subsequences xs = Apply (Apply (:$) '[]) (Apply NonEmptySubsequencesSym0 xs)

type family Permutations (a :: [a]) :: [[a]] where ...

Equations

Permutations xs0 = Apply (Apply (:$) xs0) (Apply (Apply (Let6989586621679799787PermsSym1 xs0) xs0) '[])

Reducing lists (folds)

type family Foldl (a :: TyFun b (TyFun a b -> Type) -> Type) (a :: b) (a :: [a]) :: b where ...

Equations

Foldl f z0 xs0 = Apply (Apply (Let6989586621679508526LgoSym3 f z0 xs0) z0) xs0

type family Foldl' (a :: TyFun b (TyFun a b -> Type) -> Type) (a :: b) (a :: [a]) :: b where ...

Equations

Foldl' f z0 xs0 = Apply (Apply (Let6989586621679799701LgoSym3 f z0 xs0) z0) xs0

type family Foldl1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: a where ...

Equations

Foldl1 f ((:) x xs) = Apply (Apply (Apply FoldlSym0 f) x) xs

Foldl1 _z_6989586621679799478 '[] = Apply ErrorSym0 "Data.Singletons.List.foldl1: empty list"

type family Foldl1' (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: a where ...

Equations

Foldl1' f ((:) x xs) = Apply (Apply (Apply Foldl'Sym0 f) x) xs

Foldl1' _z_6989586621679799779 '[] = Apply ErrorSym0 "Data.Singletons.List.foldl1': empty list"

type family Foldr (a :: TyFun a (TyFun b b -> Type) -> Type) (a :: b) (a :: [a]) :: b where ...

Equations

Foldr k z a_6989586621679549648 = Apply (Let6989586621679549653GoSym3 k z a_6989586621679549648) a_6989586621679549648

type family Foldr1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: a where ...

Equations

Foldr1 _z_6989586621679799436 '[x] = x

Foldr1 f ((:) x ((:) wild_6989586621679796168 wild_6989586621679796170)) = Apply (Apply f x) (Apply (Apply Foldr1Sym0 f) (Let6989586621679799444XsSym4 f x wild_6989586621679796168 wild_6989586621679796170))

Foldr1 _z_6989586621679799463 '[] = Apply ErrorSym0 "Data.Singletons.List.foldr1: empty list"

Special folds

type family Concat (a :: [[a]]) :: [a] where ...

Equations

Concat a_6989586621679799420 = Apply (Apply (Apply FoldrSym0 (:++$)) '[]) a_6989586621679799420

type family ConcatMap (a :: TyFun a [b] -> Type) (a :: [a]) :: [b] where ...

Equations

ConcatMap f a_6989586621679799416 = Apply (Apply (Apply FoldrSym0 (Apply (Apply (:.$) (:++$)) f)) '[]) a_6989586621679799416

type family And (a :: [Bool]) :: Bool where ...

Equations

And '[] = TrueSym0

And ((:) x xs) = Apply (Apply (:&&$) x) (Apply AndSym0 xs)

type family Or (a :: [Bool]) :: Bool where ...

Equations

Or '[] = FalseSym0

Or ((:) x xs) = Apply (Apply (:||$) x) (Apply OrSym0 xs)

type family Any_ (a :: TyFun a Bool -> Type) (a :: [a]) :: Bool where ...

Equations

Any_ _z_6989586621679785031 '[] = FalseSym0

Any_ p ((:) x xs) = Apply (Apply (:||$) (Apply p x)) (Apply (Apply Any_Sym0 p) xs)

type family All (a :: TyFun a Bool -> Type) (a :: [a]) :: Bool where ...

Equations

All _z_6989586621679799389 '[] = TrueSym0

All p ((:) x xs) = Apply (Apply (:&&$) (Apply p x)) (Apply (Apply AllSym0 p) xs)

type family Sum (a :: [a]) :: a where ...

Equations

Sum l = Apply (Apply (Let6989586621679797216Sum'Sym1 l) l) (FromInteger 0)

type family Product (a :: [a]) :: a where ...

Equations

Product l = Apply (Apply (Let6989586621679797192ProdSym1 l) l) (FromInteger 1)

type family Maximum (a :: [a]) :: a where ...

Equations

Maximum '[] = Apply ErrorSym0 "Data.Singletons.List.maximum: empty list"

Maximum ((:) wild_6989586621679796248 wild_6989586621679796250) = Apply (Apply Foldl1Sym0 MaxSym0) (Let6989586621679799661XsSym2 wild_6989586621679796248 wild_6989586621679796250)

type family Minimum (a :: [a]) :: a where ...

Equations

Minimum '[] = Apply ErrorSym0 "Data.Singletons.List.minimum: empty list"

Minimum ((:) wild_6989586621679796252 wild_6989586621679796254) = Apply (Apply Foldl1Sym0 MinSym0) (Let6989586621679799675XsSym2 wild_6989586621679796252 wild_6989586621679796254)

any_ :: forall a. (a -> Bool) -> [a] -> Bool

Building lists

Scans

type family Scanl (a :: TyFun b (TyFun a b -> Type) -> Type) (a :: b) (a :: [a]) :: [b] where ...

Equations

Scanl f q ls = Apply (Apply (:$) q) (Case_6989586621679799360 f q ls ls)

type family Scanl1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: [a] where ...

Equations

Scanl1 f ((:) x xs) = Apply (Apply (Apply ScanlSym0 f) x) xs

Scanl1 _z_6989586621679799377 '[] = '[]

type family Scanr (a :: TyFun a (TyFun b b -> Type) -> Type) (a :: b) (a :: [a]) :: [b] where ...

Equations

Scanr _z_6989586621679799310 q0 '[] = Apply (Apply (:$) q0) '[]

Scanr f q0 ((:) x xs) = Case_6989586621679799337 f q0 x xs (Let6989586621679799318Scrutinee_6989586621679796172Sym4 f q0 x xs)

type family Scanr1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: [a] where ...

Equations

Scanr1 _z_6989586621679799241 '[] = '[]

Scanr1 _z_6989586621679799244 '[x] = Apply (Apply (:$) x) '[]

Scanr1 f ((:) x ((:) wild_6989586621679796176 wild_6989586621679796178)) = Case_6989586621679799290 f x wild_6989586621679796176 wild_6989586621679796178 (Let6989586621679799271Scrutinee_6989586621679796174Sym4 f x wild_6989586621679796176 wild_6989586621679796178)

Accumulating maps

type family MapAccumL (a :: TyFun acc (TyFun x (acc, y) -> Type) -> Type) (a :: acc) (a :: [x]) :: (acc, [y]) where ...

Equations

MapAccumL _z_6989586621679799075 s '[] = Apply (Apply Tuple2Sym0 s) '[]

MapAccumL f s ((:) x xs) = Apply (Apply Tuple2Sym0 (Let6989586621679799083S''Sym4 f s x xs)) (Apply (Apply (:$) (Let6989586621679799083YSym4 f s x xs)) (Let6989586621679799083YsSym4 f s x xs))

type family MapAccumR (a :: TyFun acc (TyFun x (acc, y) -> Type) -> Type) (a :: acc) (a :: [x]) :: (acc, [y]) where ...

Equations

MapAccumR _z_6989586621679798903 s '[] = Apply (Apply Tuple2Sym0 s) '[]

MapAccumR f s ((:) x xs) = Apply (Apply Tuple2Sym0 (Let6989586621679798911S''Sym4 f s x xs)) (Apply (Apply (:$) (Let6989586621679798911YSym4 f s x xs)) (Let6989586621679798911YsSym4 f s x xs))

Infinite lists

type family Replicate (a :: Nat) (a :: a) :: [a] where ...

Equations

Replicate n x = Case_6989586621679797176 n x (Let6989586621679797168Scrutinee_6989586621679796260Sym2 n x)

Unfolding

type family Unfoldr (a :: TyFun b (Maybe (a, b)) -> Type) (a :: b) :: [a] where ...

Equations

Unfoldr f b = Case_6989586621679798883 f b (Let6989586621679798875Scrutinee_6989586621679796180Sym2 f b)

Sublists

Extracting sublists

type family Take (a :: Nat) (a :: [a]) :: [a] where ...

Equations

Take _z_6989586621679797363 '[] = '[]

Take n ((:) x xs) = Case_6989586621679797382 n x xs (Let6989586621679797369Scrutinee_6989586621679796244Sym3 n x xs)

type family Drop (a :: Nat) (a :: [a]) :: [a] where ...

Equations

Drop _z_6989586621679797332 '[] = '[]

Drop n ((:) x xs) = Case_6989586621679797351 n x xs (Let6989586621679797338Scrutinee_6989586621679796246Sym3 n x xs)

type family SplitAt (a :: Nat) (a :: [a]) :: ([a], [a]) where ...

Equations

SplitAt n xs = Apply (Apply Tuple2Sym0 (Apply (Apply TakeSym0 n) xs)) (Apply (Apply DropSym0 n) xs)

type family TakeWhile (a :: TyFun a Bool -> Type) (a :: [a]) :: [a] where ...

Equations

TakeWhile _z_6989586621679797736 '[] = '[]

TakeWhile p ((:) x xs) = Case_6989586621679797755 p x xs (Let6989586621679797742Scrutinee_6989586621679796234Sym3 p x xs)

type family DropWhile (a :: TyFun a Bool -> Type) (a :: [a]) :: [a] where ...

Equations

DropWhile _z_6989586621679797692 '[] = '[]

DropWhile p ((:) x xs') = Case_6989586621679797724 p x xs' (Let6989586621679797711Scrutinee_6989586621679796236Sym3 p x xs')

type family DropWhileEnd (a :: TyFun a Bool -> Type) (a :: [a]) :: [a] where ...

Equations

DropWhileEnd p a_6989586621679800283 = Apply (Apply (Apply FoldrSym0 (Apply (Apply Lambda_6989586621679800287Sym0 p) a_6989586621679800283)) '[]) a_6989586621679800283

type family Span (a :: TyFun a Bool -> Type) (a :: [a]) :: ([a], [a]) where ...

Equations

Span _z_6989586621679797510 '[] = Apply (Apply Tuple2Sym0 (Let6989586621679797513XsSym1 _z_6989586621679797510)) (Let6989586621679797513XsSym1 _z_6989586621679797510)

Span p ((:) x xs') = Case_6989586621679797546 p x xs' (Let6989586621679797533Scrutinee_6989586621679796240Sym3 p x xs')

type family Break (a :: TyFun a Bool -> Type) (a :: [a]) :: ([a], [a]) where ...

Equations

Break _z_6989586621679797405 '[] = Apply (Apply Tuple2Sym0 (Let6989586621679797408XsSym1 _z_6989586621679797405)) (Let6989586621679797408XsSym1 _z_6989586621679797405)

Break p ((:) x xs') = Case_6989586621679797441 p x xs' (Let6989586621679797428Scrutinee_6989586621679796242Sym3 p x xs')

type family StripPrefix (a :: [a]) (a :: [a]) :: Maybe [a] where ...

Equations

StripPrefix '[] ys = Apply JustSym0 ys

StripPrefix arg_6989586621680050931 arg_6989586621680050933 = Case_6989586621680051564 arg_6989586621680050931 arg_6989586621680050933 (Apply (Apply Tuple2Sym0 arg_6989586621680050931) arg_6989586621680050933)

type family Group (a :: [a]) :: [[a]] where ...

Equations

Group xs = Apply (Apply GroupBySym0 (:==$)) xs

type family Inits (a :: [a]) :: [[a]] where ...

Equations

Inits xs = Apply (Apply (:$) '[]) (Case_6989586621679798859 xs xs)

type family Tails (a :: [a]) :: [[a]] where ...

Equations

Tails xs = Apply (Apply (:$) xs) (Case_6989586621679798836 xs xs)

Predicates

type family IsPrefixOf (a :: [a]) (a :: [a]) :: Bool where ...

Equations

IsPrefixOf '[] '[] = TrueSym0

IsPrefixOf '[] ((:) _z_6989586621679798815 _z_6989586621679798818) = TrueSym0

IsPrefixOf ((:) _z_6989586621679798821 _z_6989586621679798824) '[] = FalseSym0

IsPrefixOf ((:) x xs) ((:) y ys) = Apply (Apply (:&&$) (Apply (Apply (:==$) x) y)) (Apply (Apply IsPrefixOfSym0 xs) ys)

type family IsSuffixOf (a :: [a]) (a :: [a]) :: Bool where ...

Equations

IsSuffixOf x y = Apply (Apply IsPrefixOfSym0 (Apply ReverseSym0 x)) (Apply ReverseSym0 y)

type family IsInfixOf (a :: [a]) (a :: [a]) :: Bool where ...

Equations

IsInfixOf needle haystack = Apply (Apply Any_Sym0 (Apply IsPrefixOfSym0 needle)) (Apply TailsSym0 haystack)

Searching lists

Searching by equality

type family Elem (a :: a) (a :: [a]) :: Bool where ...

Equations

Elem _z_6989586621679798752 '[] = FalseSym0

Elem x ((:) y ys) = Apply (Apply (:||$) (Apply (Apply (:==$) x) y)) (Apply (Apply ElemSym0 x) ys)

type family NotElem (a :: a) (a :: [a]) :: Bool where ...

Equations

NotElem _z_6989586621679798737 '[] = TrueSym0

NotElem x ((:) y ys) = Apply (Apply (:&&$) (Apply (Apply (:/=$) x) y)) (Apply (Apply NotElemSym0 x) ys)

type family Lookup (a :: a) (a :: [(a, b)]) :: Maybe b where ...

Equations

Lookup _key '[] = NothingSym0

Lookup key ((:) '(x, y) xys) = Case_6989586621679797320 key x y xys (Let6989586621679797301Scrutinee_6989586621679796256Sym4 key x y xys)

Searching with a predicate

type family Find (a :: TyFun a Bool -> Type) (a :: [a]) :: Maybe a where ...

Equations

Find p a_6989586621679797796 = Apply (Apply (Apply (:.$) ListToMaybeSym0) (Apply FilterSym0 p)) a_6989586621679797796

type family Filter (a :: TyFun a Bool -> Type) (a :: [a]) :: [a] where ...

Equations

Filter _p '[] = '[]

Filter p ((:) x xs) = Case_6989586621679797784 p x xs (Let6989586621679797771Scrutinee_6989586621679796222Sym3 p x xs)

type family Partition (a :: TyFun a Bool -> Type) (a :: [a]) :: ([a], [a]) where ...

Equations

Partition p xs = Apply (Apply (Apply FoldrSym0 (Apply SelectSym0 p)) (Apply (Apply Tuple2Sym0 '[]) '[])) xs

Indexing lists

type family (a :: [a]) :!! (a :: Nat) :: a where ...

Equations

'[] :!! _z_6989586621679797135 = Apply ErrorSym0 "Data.Singletons.List.!!: index too large"

((:) x xs) :!! n = Case_6989586621679797154 x xs n (Let6989586621679797141Scrutinee_6989586621679796262Sym3 x xs n)

type family ElemIndex (a :: a) (a :: [a]) :: Maybe Nat where ...

Equations

ElemIndex x a_6989586621679798724 = Apply (Apply FindIndexSym0 (Apply (:==$) x)) a_6989586621679798724

type family ElemIndices (a :: a) (a :: [a]) :: [Nat] where ...

Equations

ElemIndices x a_6989586621679798698 = Apply (Apply FindIndicesSym0 (Apply (:==$) x)) a_6989586621679798698

type family FindIndex (a :: TyFun a Bool -> Type) (a :: [a]) :: Maybe Nat where ...

Equations

FindIndex p a_6989586621679798711 = Apply (Apply (Apply (:.$) ListToMaybeSym0) (Apply FindIndicesSym0 p)) a_6989586621679798711

type family FindIndices (a :: TyFun a Bool -> Type) (a :: [a]) :: [Nat] where ...

Equations

FindIndices p xs = Apply (Apply MapSym0 SndSym0) (Apply (Apply FilterSym0 (Apply (Apply Lambda_6989586621679798666Sym0 p) xs)) (Apply (Apply ZipSym0 xs) (Apply (Apply (Let6989586621679798637BuildListSym2 p xs) (FromInteger 0)) xs)))

Zipping and unzipping lists

type family Zip (a :: [a]) (a :: [b]) :: [(a, b)] where ...

Equations

Zip ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply Tuple2Sym0 x) y)) (Apply (Apply ZipSym0 xs) ys)

Zip '[] '[] = '[]

Zip ((:) _z_6989586621679798614 _z_6989586621679798617) '[] = '[]

Zip '[] ((:) _z_6989586621679798620 _z_6989586621679798623) = '[]

type family Zip3 (a :: [a]) (a :: [b]) (a :: [c]) :: [(a, b, c)] where ...

Equations

Zip3 ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply Tuple3Sym0 a) b) c)) (Apply (Apply (Apply Zip3Sym0 as) bs) cs)

Zip3 '[] '[] '[] = '[]

Zip3 '[] '[] ((:) _z_6989586621679798547 _z_6989586621679798550) = '[]

Zip3 '[] ((:) _z_6989586621679798553 _z_6989586621679798556) '[] = '[]

Zip3 '[] ((:) _z_6989586621679798559 _z_6989586621679798562) ((:) _z_6989586621679798565 _z_6989586621679798568) = '[]

Zip3 ((:) _z_6989586621679798571 _z_6989586621679798574) '[] '[] = '[]

Zip3 ((:) _z_6989586621679798577 _z_6989586621679798580) '[] ((:) _z_6989586621679798583 _z_6989586621679798586) = '[]

Zip3 ((:) _z_6989586621679798589 _z_6989586621679798592) ((:) _z_6989586621679798595 _z_6989586621679798598) '[] = '[]

type family Zip4 (a :: [a]) (a :: [b]) (a :: [c]) (a :: [d]) :: [(a, b, c, d)] where ...

Equations

Zip4 a_6989586621680051518 a_6989586621680051520 a_6989586621680051522 a_6989586621680051524 = Apply (Apply (Apply (Apply (Apply ZipWith4Sym0 Tuple4Sym0) a_6989586621680051518) a_6989586621680051520) a_6989586621680051522) a_6989586621680051524

type family Zip5 (a :: [a]) (a :: [b]) (a :: [c]) (a :: [d]) (a :: [e]) :: [(a, b, c, d, e)] where ...

Equations

Zip5 a_6989586621680051473 a_6989586621680051475 a_6989586621680051477 a_6989586621680051479 a_6989586621680051481 = Apply (Apply (Apply (Apply (Apply (Apply ZipWith5Sym0 Tuple5Sym0) a_6989586621680051473) a_6989586621680051475) a_6989586621680051477) a_6989586621680051479) a_6989586621680051481

type family Zip6 (a :: [a]) (a :: [b]) (a :: [c]) (a :: [d]) (a :: [e]) (a :: [f]) :: [(a, b, c, d, e, f)] where ...

Equations

Zip6 a_6989586621680051416 a_6989586621680051418 a_6989586621680051420 a_6989586621680051422 a_6989586621680051424 a_6989586621680051426 = Apply (Apply (Apply (Apply (Apply (Apply (Apply ZipWith6Sym0 Tuple6Sym0) a_6989586621680051416) a_6989586621680051418) a_6989586621680051420) a_6989586621680051422) a_6989586621680051424) a_6989586621680051426

type family Zip7 (a :: [a]) (a :: [b]) (a :: [c]) (a :: [d]) (a :: [e]) (a :: [f]) (a :: [g]) :: [(a, b, c, d, e, f, g)] where ...

Equations

Zip7 a_6989586621680051346 a_6989586621680051348 a_6989586621680051350 a_6989586621680051352 a_6989586621680051354 a_6989586621680051356 a_6989586621680051358 = Apply (Apply (Apply (Apply (Apply (Apply (Apply (Apply ZipWith7Sym0 Tuple7Sym0) a_6989586621680051346) a_6989586621680051348) a_6989586621680051350) a_6989586621680051352) a_6989586621680051354) a_6989586621680051356) a_6989586621680051358

type family ZipWith (a :: TyFun a (TyFun b c -> Type) -> Type) (a :: [a]) (a :: [b]) :: [c] where ...

Equations

ZipWith f ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply f x) y)) (Apply (Apply (Apply ZipWithSym0 f) xs) ys)

ZipWith _z_6989586621679798505 '[] '[] = '[]

ZipWith _z_6989586621679798508 ((:) _z_6989586621679798511 _z_6989586621679798514) '[] = '[]

ZipWith _z_6989586621679798517 '[] ((:) _z_6989586621679798520 _z_6989586621679798523) = '[]

type family ZipWith3 (a :: TyFun a (TyFun b (TyFun c d -> Type) -> Type) -> Type) (a :: [a]) (a :: [b]) (a :: [c]) :: [d] where ...

Equations

ZipWith3 z ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply z a) b) c)) (Apply (Apply (Apply (Apply ZipWith3Sym0 z) as) bs) cs)

ZipWith3 _z_6989586621679798410 '[] '[] '[] = '[]

ZipWith3 _z_6989586621679798413 '[] '[] ((:) _z_6989586621679798416 _z_6989586621679798419) = '[]

ZipWith3 _z_6989586621679798422 '[] ((:) _z_6989586621679798425 _z_6989586621679798428) '[] = '[]

ZipWith3 _z_6989586621679798431 '[] ((:) _z_6989586621679798434 _z_6989586621679798437) ((:) _z_6989586621679798440 _z_6989586621679798443) = '[]

ZipWith3 _z_6989586621679798446 ((:) _z_6989586621679798449 _z_6989586621679798452) '[] '[] = '[]

ZipWith3 _z_6989586621679798455 ((:) _z_6989586621679798458 _z_6989586621679798461) '[] ((:) _z_6989586621679798464 _z_6989586621679798467) = '[]

ZipWith3 _z_6989586621679798470 ((:) _z_6989586621679798473 _z_6989586621679798476) ((:) _z_6989586621679798479 _z_6989586621679798482) '[] = '[]

type family ZipWith4 (a :: TyFun a (TyFun b (TyFun c (TyFun d e -> Type) -> Type) -> Type) -> Type) (a :: [a]) (a :: [b]) (a :: [c]) (a :: [d]) :: [e] where ...

Equations

ZipWith4 z ((:) a as) ((:) b bs) ((:) c cs) ((:) d ds) = Apply (Apply (:$) (Apply (Apply (Apply (Apply z a) b) c) d)) (Apply (Apply (Apply (Apply (Apply ZipWith4Sym0 z) as) bs) cs) ds)

ZipWith4 _z_6989586621680051331 _z_6989586621680051334 _z_6989586621680051337 _z_6989586621680051340 _z_6989586621680051343 = '[]

type family ZipWith5 (a :: TyFun a (TyFun b (TyFun c (TyFun d (TyFun e f -> Type) -> Type) -> Type) -> Type) -> Type) (a :: [a]) (a :: [b]) (a :: [c]) (a :: [d]) (a :: [e]) :: [f] where ...

Equations

ZipWith5 z ((:) a as) ((:) b bs) ((:) c cs) ((:) d ds) ((:) e es) = Apply (Apply (:$) (Apply (Apply (Apply (Apply (Apply z a) b) c) d) e)) (Apply (Apply (Apply (Apply (Apply (Apply ZipWith5Sym0 z) as) bs) cs) ds) es)

ZipWith5 _z_6989586621680051274 _z_6989586621680051277 _z_6989586621680051280 _z_6989586621680051283 _z_6989586621680051286 _z_6989586621680051289 = '[]

type family ZipWith6 (a :: TyFun a (TyFun b (TyFun c (TyFun d (TyFun e (TyFun f g -> Type) -> Type) -> Type) -> Type) -> Type) -> Type) (a :: [a]) (a :: [b]) (a :: [c]) (a :: [d]) (a :: [e]) (a :: [f]) :: [g] where ...

Equations

ZipWith6 z ((:) a as) ((:) b bs) ((:) c cs) ((:) d ds) ((:) e es) ((:) f fs) = Apply (Apply (:$) (Apply (Apply (Apply (Apply (Apply (Apply z a) b) c) d) e) f)) (Apply (Apply (Apply (Apply (Apply (Apply (Apply ZipWith6Sym0 z) as) bs) cs) ds) es) fs)

ZipWith6 _z_6989586621680051203 _z_6989586621680051206 _z_6989586621680051209 _z_6989586621680051212 _z_6989586621680051215 _z_6989586621680051218 _z_6989586621680051221 = '[]

type family ZipWith7 (a :: TyFun a (TyFun b (TyFun c (TyFun d (TyFun e (TyFun f (TyFun g h -> Type) -> Type) -> Type) -> Type) -> Type) -> Type) -> Type) (a :: [a]) (a :: [b]) (a :: [c]) (a :: [d]) (a :: [e]) (a :: [f]) (a :: [g]) :: [h] where ...

Equations

ZipWith7 z ((:) a as) ((:) b bs) ((:) c cs) ((:) d ds) ((:) e es) ((:) f fs) ((:) g gs) = Apply (Apply (:$) (Apply (Apply (Apply (Apply (Apply (Apply (Apply z a) b) c) d) e) f) g)) (Apply (Apply (Apply (Apply (Apply (Apply (Apply (Apply ZipWith7Sym0 z) as) bs) cs) ds) es) fs) gs)

ZipWith7 _z_6989586621680051117 _z_6989586621680051120 _z_6989586621680051123 _z_6989586621680051126 _z_6989586621680051129 _z_6989586621680051132 _z_6989586621680051135 _z_6989586621680051138 = '[]

type family Unzip (a :: [(a, b)]) :: ([a], [b]) where ...

Equations

Unzip xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798356Sym0 xs)) (Apply (Apply Tuple2Sym0 '[]) '[])) xs

type family Unzip3 (a :: [(a, b, c)]) :: ([a], [b], [c]) where ...

Equations

Unzip3 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798324Sym0 xs)) (Apply (Apply (Apply Tuple3Sym0 '[]) '[]) '[])) xs

type family Unzip4 (a :: [(a, b, c, d)]) :: ([a], [b], [c], [d]) where ...

Equations

Unzip4 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798290Sym0 xs)) (Apply (Apply (Apply (Apply Tuple4Sym0 '[]) '[]) '[]) '[])) xs

type family Unzip5 (a :: [(a, b, c, d, e)]) :: ([a], [b], [c], [d], [e]) where ...

Equations

Unzip5 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798254Sym0 xs)) (Apply (Apply (Apply (Apply (Apply Tuple5Sym0 '[]) '[]) '[]) '[]) '[])) xs

type family Unzip6 (a :: [(a, b, c, d, e, f)]) :: ([a], [b], [c], [d], [e], [f]) where ...

Equations

Unzip6 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798216Sym0 xs)) (Apply (Apply (Apply (Apply (Apply (Apply Tuple6Sym0 '[]) '[]) '[]) '[]) '[]) '[])) xs

type family Unzip7 (a :: [(a, b, c, d, e, f, g)]) :: ([a], [b], [c], [d], [e], [f], [g]) where ...

Equations

Unzip7 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798176Sym0 xs)) (Apply (Apply (Apply (Apply (Apply (Apply (Apply Tuple7Sym0 '[]) '[]) '[]) '[]) '[]) '[]) '[])) xs

Special lists

"Set" operations

type family Nub (a :: [a]) :: [a] where ...

Equations

Nub l = Apply (Apply (Let6989586621679798763Nub'Sym1 l) l) '[]

type family Delete (a :: a) (a :: [a]) :: [a] where ...

Equations

Delete a_6989586621679798141 a_6989586621679798143 = Apply (Apply (Apply DeleteBySym0 (:==$)) a_6989586621679798141) a_6989586621679798143

type family (a :: [a]) :\\ (a :: [a]) :: [a] where ... infix 5

Equations

a_6989586621679798156 :\\ a_6989586621679798158 = Apply (Apply (Apply FoldlSym0 (Apply FlipSym0 DeleteSym0)) a_6989586621679798156) a_6989586621679798158

type family Union (a :: [a]) (a :: [a]) :: [a] where ...

Equations

Union a_6989586621679798126 a_6989586621679798128 = Apply (Apply (Apply UnionBySym0 (:==$)) a_6989586621679798126) a_6989586621679798128

type family Intersect (a :: [a]) (a :: [a]) :: [a] where ...

Equations

Intersect a_6989586621679797929 a_6989586621679797931 = Apply (Apply (Apply IntersectBySym0 (:==$)) a_6989586621679797929) a_6989586621679797931

Ordered lists

type family Sort (a :: [a]) :: [a] where ...

Equations

Sort a_6989586621679798032 = Apply (Apply SortBySym0 CompareSym0) a_6989586621679798032

type family Insert (a :: a) (a :: [a]) :: [a] where ...

Equations

Insert e ls = Apply (Apply (Apply InsertBySym0 CompareSym0) e) ls

Generalized functions

The "By" operations

User-supplied equality (replacing an Eq context)

type family NubBy (a :: TyFun a (TyFun a Bool -> Type) -> Type) (a :: [a]) :: [a] where ...

Equations

NubBy eq l = Apply (Apply (Let6989586621679797070NubBy'Sym2 eq l) l) '[]

type family DeleteBy (a :: TyFun a (TyFun a Bool -> Type) -> Type) (a :: a) (a :: [a]) :: [a] where ...

Equations

DeleteBy _z_6989586621679798054 _z_6989586621679798057 '[] = '[]

DeleteBy eq x ((:) y ys) = Case_6989586621679798083 eq x y ys (Let6989586621679798064Scrutinee_6989586621679796206Sym4 eq x y ys)

type family DeleteFirstsBy (a :: TyFun a (TyFun a Bool -> Type) -> Type) (a :: [a]) (a :: [a]) :: [a] where ...

Equations

DeleteFirstsBy eq a_6989586621679798101 a_6989586621679798103 = Apply (Apply (Apply FoldlSym0 (Apply FlipSym0 (Apply DeleteBySym0 eq))) a_6989586621679798101) a_6989586621679798103

type family UnionBy (a :: TyFun a (TyFun a Bool -> Type) -> Type) (a :: [a]) (a :: [a]) :: [a] where ...

Equations

UnionBy eq xs ys = Apply (Apply (:++$) xs) (Apply (Apply (Apply FoldlSym0 (Apply FlipSym0 (Apply DeleteBySym0 eq))) (Apply (Apply NubBySym0 eq) ys)) xs)

type family GroupBy (a :: TyFun a (TyFun a Bool -> Type) -> Type) (a :: [a]) :: [[a]] where ...

Equations

GroupBy _z_6989586621679797615 '[] = '[]

GroupBy eq ((:) x xs) = Apply (Apply (:$) (Apply (Apply (:$) x) (Let6989586621679797621YsSym3 eq x xs))) (Apply (Apply GroupBySym0 eq) (Let6989586621679797621ZsSym3 eq x xs))

type family IntersectBy (a :: TyFun a (TyFun a Bool -> Type) -> Type) (a :: [a]) (a :: [a]) :: [a] where ...

Equations

IntersectBy _z_6989586621679797815 '[] '[] = '[]

IntersectBy _z_6989586621679797818 '[] ((:) _z_6989586621679797821 _z_6989586621679797824) = '[]

IntersectBy _z_6989586621679797827 ((:) _z_6989586621679797830 _z_6989586621679797833) '[] = '[]

IntersectBy eq ((:) wild_6989586621679796226 wild_6989586621679796228) ((:) wild_6989586621679796230 wild_6989586621679796232) = Apply (Apply FilterSym0 (Apply (Apply (Apply (Apply (Apply Lambda_6989586621679797892Sym0 eq) wild_6989586621679796226) wild_6989586621679796228) wild_6989586621679796230) wild_6989586621679796232)) (Let6989586621679797841XsSym5 eq wild_6989586621679796226 wild_6989586621679796228 wild_6989586621679796230 wild_6989586621679796232)

User-supplied comparison (replacing an Ord context)

type family SortBy (a :: TyFun a (TyFun a Ordering -> Type) -> Type) (a :: [a]) :: [a] where ...

Equations

SortBy cmp a_6989586621679798028 = Apply (Apply (Apply FoldrSym0 (Apply InsertBySym0 cmp)) '[]) a_6989586621679798028

type family InsertBy (a :: TyFun a (TyFun a Ordering -> Type) -> Type) (a :: a) (a :: [a]) :: [a] where ...

Equations

InsertBy _z_6989586621679797959 x '[] = Apply (Apply (:$) x) '[]

InsertBy cmp x ((:) y ys') = Case_6989586621679798005 cmp x y ys' (Let6989586621679797986Scrutinee_6989586621679796208Sym4 cmp x y ys')

type family MaximumBy (a :: TyFun a (TyFun a Ordering -> Type) -> Type) (a :: [a]) :: a where ...

Equations

MaximumBy _z_6989586621679799490 '[] = Apply ErrorSym0 "Data.Singletons.List.maximumBy: empty list"

MaximumBy cmp ((:) wild_6989586621679796212 wild_6989586621679796214) = Apply (Apply Foldl1Sym0 (Let6989586621679799509MaxBySym3 cmp wild_6989586621679796212 wild_6989586621679796214)) (Let6989586621679799496XsSym3 cmp wild_6989586621679796212 wild_6989586621679796214)

type family MinimumBy (a :: TyFun a (TyFun a Ordering -> Type) -> Type) (a :: [a]) :: a where ...

Equations

MinimumBy _z_6989586621679799577 '[] = Apply ErrorSym0 "Data.Singletons.List.minimumBy: empty list"

MinimumBy cmp ((:) wild_6989586621679796218 wild_6989586621679796220) = Apply (Apply Foldl1Sym0 (Let6989586621679799596MinBySym3 cmp wild_6989586621679796218 wild_6989586621679796220)) (Let6989586621679799583XsSym3 cmp wild_6989586621679796218 wild_6989586621679796220)

The "generic" operations

type family GenericLength (a :: [a]) :: i where ...

Equations

GenericLength '[] = FromInteger 0

GenericLength ((:) _z_6989586621679797030 xs) = Apply (Apply (:+$) (FromInteger 1)) (Apply GenericLengthSym0 xs)

type family GenericTake (a :: i) (a :: [a]) :: [a] where ...

Equations

GenericTake a_6989586621680051027 a_6989586621680051029 = Apply (Apply TakeSym0 a_6989586621680051027) a_6989586621680051029

type family GenericDrop (a :: i) (a :: [a]) :: [a] where ...

Equations

GenericDrop a_6989586621680051012 a_6989586621680051014 = Apply (Apply DropSym0 a_6989586621680051012) a_6989586621680051014

type family GenericSplitAt (a :: i) (a :: [a]) :: ([a], [a]) where ...

Equations

GenericSplitAt a_6989586621680050997 a_6989586621680050999 = Apply (Apply SplitAtSym0 a_6989586621680050997) a_6989586621680050999

type family GenericIndex (a :: [a]) (a :: i) :: a where ...

Equations

GenericIndex a_6989586621680050982 a_6989586621680050984 = Apply (Apply (:!!$) a_6989586621680050982) a_6989586621680050984

type family GenericReplicate (a :: i) (a :: a) :: [a] where ...

Equations

GenericReplicate a_6989586621680050967 a_6989586621680050969 = Apply (Apply ReplicateSym0 a_6989586621680050967) a_6989586621680050969

Defunctionalization symbols

type NilSym0 = '[]

data (:$) l

data l :$$ l

type (:$$$) t t = (:) t t

type (:++$$$) t t = (:++) t t

data l :++$$ l

data (:++$) l

data HeadSym0 l

type HeadSym1 t = Head t

data LastSym0 l

type LastSym1 t = Last t

data TailSym0 l

type TailSym1 t = Tail t

data InitSym0 l

type InitSym1 t = Init t

data NullSym0 l

type NullSym1 t = Null t

data MapSym0 l

data MapSym1 l l

type MapSym2 t t = Map t t

data ReverseSym0 l

type ReverseSym1 t = Reverse t

data IntersperseSym0 l

data IntersperseSym1 l l

type IntersperseSym2 t t = Intersperse t t

data IntercalateSym0 l

data IntercalateSym1 l l

type IntercalateSym2 t t = Intercalate t t

data SubsequencesSym0 l

type SubsequencesSym1 t = Subsequences t

data PermutationsSym0 l

type PermutationsSym1 t = Permutations t

data FoldlSym0 l

data FoldlSym1 l l

data FoldlSym2 l l l

type FoldlSym3 t t t = Foldl t t t

data Foldl'Sym0 l

data Foldl'Sym1 l l

data Foldl'Sym2 l l l

type Foldl'Sym3 t t t = Foldl' t t t

data Foldl1Sym0 l

data Foldl1Sym1 l l

type Foldl1Sym2 t t = Foldl1 t t

data Foldl1'Sym0 l

data Foldl1'Sym1 l l

type Foldl1'Sym2 t t = Foldl1' t t

data FoldrSym0 l

data FoldrSym1 l l

data FoldrSym2 l l l

type FoldrSym3 t t t = Foldr t t t

data Foldr1Sym0 l

data Foldr1Sym1 l l

type Foldr1Sym2 t t = Foldr1 t t

data ConcatSym0 l

type ConcatSym1 t = Concat t

data ConcatMapSym0 l

data ConcatMapSym1 l l

type ConcatMapSym2 t t = ConcatMap t t

data AndSym0 l

type AndSym1 t = And t

data OrSym0 l

type OrSym1 t = Or t

data Any_Sym0 l

data Any_Sym1 l l

type Any_Sym2 t t = Any_ t t

data AllSym0 l

data AllSym1 l l

type AllSym2 t t = All t t

data ScanlSym0 l

data ScanlSym1 l l

data ScanlSym2 l l l

type ScanlSym3 t t t = Scanl t t t

data Scanl1Sym0 l

data Scanl1Sym1 l l

type Scanl1Sym2 t t = Scanl1 t t

data ScanrSym0 l

data ScanrSym1 l l

data ScanrSym2 l l l

type ScanrSym3 t t t = Scanr t t t

data Scanr1Sym0 l

data Scanr1Sym1 l l

type Scanr1Sym2 t t = Scanr1 t t

data MapAccumLSym0 l

data MapAccumLSym1 l l

data MapAccumLSym2 l l l

type MapAccumLSym3 t t t = MapAccumL t t t

data MapAccumRSym0 l

data MapAccumRSym1 l l

data MapAccumRSym2 l l l

type MapAccumRSym3 t t t = MapAccumR t t t

data UnfoldrSym0 l

data UnfoldrSym1 l l

type UnfoldrSym2 t t = Unfoldr t t

data InitsSym0 l

type InitsSym1 t = Inits t

data TailsSym0 l

type TailsSym1 t = Tails t

data IsPrefixOfSym0 l

data IsPrefixOfSym1 l l

type IsPrefixOfSym2 t t = IsPrefixOf t t

data IsSuffixOfSym0 l

data IsSuffixOfSym1 l l

type IsSuffixOfSym2 t t = IsSuffixOf t t

data IsInfixOfSym0 l

data IsInfixOfSym1 l l

type IsInfixOfSym2 t t = IsInfixOf t t

data ElemSym0 l

data ElemSym1 l l

type ElemSym2 t t = Elem t t

data NotElemSym0 l

data NotElemSym1 l l

type NotElemSym2 t t = NotElem t t

data ZipSym0 l

data ZipSym1 l l

type ZipSym2 t t = Zip t t

data Zip3Sym0 l

data Zip3Sym1 l l

data Zip3Sym2 l l l

type Zip3Sym3 t t t = Zip3 t t t

data ZipWithSym0 l

data ZipWithSym1 l l

data ZipWithSym2 l l l

type ZipWithSym3 t t t = ZipWith t t t

data ZipWith3Sym0 l

data ZipWith3Sym1 l l

data ZipWith3Sym2 l l l

data ZipWith3Sym3 l l l l

type ZipWith3Sym4 t t t t = ZipWith3 t t t t

data UnzipSym0 l

type UnzipSym1 t = Unzip t

data Unzip3Sym0 l

type Unzip3Sym1 t = Unzip3 t

data Unzip4Sym0 l

type Unzip4Sym1 t = Unzip4 t

data Unzip5Sym0 l

type Unzip5Sym1 t = Unzip5 t

data Unzip6Sym0 l

type Unzip6Sym1 t = Unzip6 t

data Unzip7Sym0 l

type Unzip7Sym1 t = Unzip7 t

data DeleteSym0 l

data DeleteSym1 l l

type DeleteSym2 t t = Delete t t

data (:\\$) l

data l :\\$$ l

type (:\\$$$) t t = (:\\) t t

data IntersectSym0 l

data IntersectSym1 l l

type IntersectSym2 t t = Intersect t t

data InsertSym0 l

data InsertSym1 l l

type InsertSym2 t t = Insert t t

data SortSym0 l

type SortSym1 t = Sort t

data DeleteBySym0 l

data DeleteBySym1 l l

data DeleteBySym2 l l l

type DeleteBySym3 t t t = DeleteBy t t t

data DeleteFirstsBySym0 l

data DeleteFirstsBySym1 l l

data DeleteFirstsBySym2 l l l

type DeleteFirstsBySym3 t t t = DeleteFirstsBy t t t

data IntersectBySym0 l

data IntersectBySym1 l l

data IntersectBySym2 l l l

data SortBySym0 l

data SortBySym1 l l

type SortBySym2 t t = SortBy t t

data InsertBySym0 l

data InsertBySym1 l l

data InsertBySym2 l l l

type InsertBySym3 t t t = InsertBy t t t

data MaximumBySym0 l

data MaximumBySym1 l l

type MaximumBySym2 t t = MaximumBy t t

data MinimumBySym0 l

data MinimumBySym1 l l

type MinimumBySym2 t t = MinimumBy t t

data LengthSym0 l

type LengthSym1 t = Length t

data SumSym0 l

type SumSym1 t = Sum t

data ProductSym0 l

type ProductSym1 t = Product t

data ReplicateSym0 l

data ReplicateSym1 l l

type ReplicateSym2 t t = Replicate t t

data TransposeSym0 l

type TransposeSym1 t = Transpose t

data TakeSym0 l

data TakeSym1 l l

type TakeSym2 t t = Take t t

data DropSym0 l

data DropSym1 l l

type DropSym2 t t = Drop t t

data SplitAtSym0 l

data SplitAtSym1 l l

type SplitAtSym2 t t = SplitAt t t

data TakeWhileSym0 l

data TakeWhileSym1 l l

type TakeWhileSym2 t t = TakeWhile t t

data DropWhileSym0 l

data DropWhileSym1 l l

type DropWhileSym2 t t = DropWhile t t

data DropWhileEndSym0 l

data DropWhileEndSym1 l l

type DropWhileEndSym2 t t = DropWhileEnd t t

data SpanSym0 l

data SpanSym1 l l

type SpanSym2 t t = Span t t

data BreakSym0 l

data BreakSym1 l l

type BreakSym2 t t = Break t t

data StripPrefixSym0 l

data StripPrefixSym1 l l

type StripPrefixSym2 t t = StripPrefix t t

data MaximumSym0 l

type MaximumSym1 t = Maximum t

data MinimumSym0 l

type MinimumSym1 t = Minimum t

data GroupSym0 l

type GroupSym1 t = Group t

data GroupBySym0 l

data GroupBySym1 l l

type GroupBySym2 t t = GroupBy t t

data LookupSym0 l

data LookupSym1 l l

type LookupSym2 t t = Lookup t t

data FindSym0 l

data FindSym1 l l

type FindSym2 t t = Find t t

data FilterSym0 l

data FilterSym1 l l

type FilterSym2 t t = Filter t t

data PartitionSym0 l

data PartitionSym1 l l

type PartitionSym2 t t = Partition t t

data (:!!$) l

data l :!!$$ l

type (:!!$$$) t t = (:!!) t t

data ElemIndexSym0 l

data ElemIndexSym1 l l

type ElemIndexSym2 t t = ElemIndex t t

data ElemIndicesSym0 l

data ElemIndicesSym1 l l

type ElemIndicesSym2 t t = ElemIndices t t

data FindIndexSym0 l

data FindIndexSym1 l l

type FindIndexSym2 t t = FindIndex t t

data FindIndicesSym0 l

data FindIndicesSym1 l l

type FindIndicesSym2 t t = FindIndices t t

data Zip4Sym0 l

data Zip4Sym1 l l

data Zip4Sym2 l l l

data Zip4Sym3 l l l l

type Zip4Sym4 t t t t = Zip4 t t t t

data Zip5Sym0 l

data Zip5Sym1 l l

data Zip5Sym2 l l l

data Zip5Sym3 l l l l

data Zip5Sym4 l l l l l

type Zip5Sym5 t t t t t = Zip5 t t t t t

data Zip6Sym0 l

data Zip6Sym1 l l

data Zip6Sym2 l l l

data Zip6Sym3 l l l l

data Zip6Sym4 l l l l l

data Zip6Sym5 l l l l l l

type Zip6Sym6 t t t t t t = Zip6 t t t t t t

data Zip7Sym0 l

data Zip7Sym1 l l

data Zip7Sym2 l l l

data Zip7Sym3 l l l l

data Zip7Sym4 l l l l l

data Zip7Sym5 l l l l l l

data Zip7Sym6 l l l l l l l

type Zip7Sym7 t t t t t t t = Zip7 t t t t t t t

data ZipWith4Sym0 l

data ZipWith4Sym1 l l

data ZipWith4Sym2 l l l

data ZipWith4Sym3 l l l l

data ZipWith4Sym4 l l l l l

type ZipWith4Sym5 t t t t t = ZipWith4 t t t t t

data ZipWith5Sym0 l

data ZipWith5Sym1 l l

data ZipWith5Sym2 l l l

data ZipWith5Sym3 l l l l

data ZipWith5Sym4 l l l l l

data ZipWith5Sym5 l l l l l l

type ZipWith5Sym6 t t t t t t = ZipWith5 t t t t t t

data ZipWith6Sym0 l

data ZipWith6Sym1 l l

data ZipWith6Sym2 l l l

data ZipWith6Sym3 l l l l

data ZipWith6Sym4 l l l l l

data ZipWith6Sym5 l l l l l l

data ZipWith6Sym6 l l l l l l l

type ZipWith6Sym7 t t t t t t t = ZipWith6 t t t t t t t

data ZipWith7Sym0 l

data ZipWith7Sym1 l l

data ZipWith7Sym2 l l l

data ZipWith7Sym3 l l l l

data ZipWith7Sym4 l l l l l

data ZipWith7Sym5 l l l l l l

data ZipWith7Sym6 l l l l l l l

data ZipWith7Sym7 l l l l l l l l

type ZipWith7Sym8 t t t t t t t t = ZipWith7 t t t t t t t t

data NubSym0 l

type NubSym1 t = Nub t

data NubBySym0 l

data NubBySym1 l l

type NubBySym2 t t = NubBy t t

data UnionSym0 l

data UnionSym1 l l

type UnionSym2 t t = Union t t

data UnionBySym0 l

data UnionBySym1 l l

data UnionBySym2 l l l

type UnionBySym3 t t t = UnionBy t t t

data GenericLengthSym0 l

type GenericLengthSym1 t = GenericLength t

data GenericTakeSym0 l

data GenericTakeSym1 l l

type GenericTakeSym2 t t = GenericTake t t

data GenericDropSym0 l

data GenericDropSym1 l l

type GenericDropSym2 t t = GenericDrop t t

data GenericSplitAtSym0 l

data GenericSplitAtSym1 l l

type GenericSplitAtSym2 t t = GenericSplitAt t t

data GenericIndexSym0 l

data GenericIndexSym1 l l

type GenericIndexSym2 t t = GenericIndex t t

data GenericReplicateSym0 l

data GenericReplicateSym1 l l

type GenericReplicateSym2 t t = GenericReplicate t t