Data.Singletons

Description

This module exports the basic definitions to use singletons. For routine use, consider importing Prelude, which exports constructors for singletons based on types in the Prelude.

You may also want to read http://www.cis.upenn.edu/~eir/packages/singletons/README.html and the original paper presenting this library, available at http://www.cis.upenn.edu/~eir/papers/2012/singletons/paper.pdf.

Main singleton definitions

data family Sing a

The singleton kind-indexed data family.

See also Sing for exported constructors

class SingI a where

A SingI constraint is essentially an implicitly-passed singleton. If you need to satisfy this constraint with an explicit singleton, please see withSingI.

Methods

sing :: Sing a

Produce the singleton explicitly. You will likely need the ScopedTypeVariables extension to use this method the way you want.

class (kparam ~ KProxy) => SingKind kparam where

The SingKind class is essentially a kind class. It classifies all kinds for which singletons are defined. The class supports converting between a singleton type and the base (unrefined) type which it is built from.

Associated Types

type DemoteRep kparam :: *

Get a base type from a proxy for the promoted kind. For example, DemoteRep ('KProxy :: KProxy Bool) will be the type Bool.

Methods

fromSing :: Sing (a :: k) -> DemoteRep kparam

Convert a singleton to its unrefined version.

toSing :: DemoteRep kparam -> SomeSing kparam

Convert an unrefined type to an existentially-quantified singleton type.

Working with singletons

type KindOf a = (KProxy :: KProxy k)

Convenient synonym to refer to the kind of a type variable: type KindOf (a :: k) = ('KProxy :: KProxy k)

type Demote a = DemoteRep (KProxy :: KProxy k)

Convenient abbreviation for DemoteRep: type Demote (a :: k) = DemoteRep ('KProxy :: KProxy k)

data SingInstance a where

A SingInstance wraps up a SingI instance for explicit handling.

Constructors

SingInstance :: SingI a => SingInstance a

data SomeSing kproxy where

An existentially-quantified singleton. This type is useful when you want a singleton type, but there is no way of knowing, at compile-time, what the type index will be. To make use of this type, you will generally have to use a pattern-match:

foo :: Bool -> ...
foo b = case toSing b of
          SomeSing sb -> {- fancy dependently-typed code with sb -}

An example like the one above may be easier to write using withSomeSing.

Constructors

SomeSing :: Sing (a :: k) -> SomeSing (KProxy :: KProxy k)

singInstance :: forall a. Sing a -> SingInstance a

Get an implicit singleton (a SingI instance) from an explicit one.

withSingI :: Sing n -> (SingI n => r) -> r

Convenience function for creating a context with an implicit singleton available.

withSomeSing

Arguments

:: SingKind (KProxy :: KProxy k)

=> DemoteRep (KProxy :: KProxy k)
The original datatype

-> (forall a. Sing a -> r)
Function expecting a singleton

-> r

Convert a normal datatype (like Bool) to a singleton for that datatype, passing it into a continuation.

singByProxy :: SingI a => proxy a -> Sing a

Allows creation of a singleton when a proxy is at hand.

singByProxy# :: SingI a => Proxy# a -> Sing a

Allows creation of a singleton when a proxy# is at hand.

withSing :: SingI a => (Sing a -> b) -> b

A convenience function useful when we need to name a singleton value multiple times. Without this function, each use of sing could potentially refer to a different singleton, and one has to use type signatures (often with ScopedTypeVariables) to ensure that they are the same.

singThat :: forall a. (SingKind (KProxy :: KProxy k), SingI a) => (Demote a -> Bool) -> Maybe (Sing a)

A convenience function that names a singleton satisfying a certain property. If the singleton does not satisfy the property, then the function returns Nothing. The property is expressed in terms of the underlying representation of the singleton.

Defunctionalization

data TyFun :: * -> * -> *

Representation of the kind of a type-level function. The difference between term-level arrows and this type-level arrow is that at the term level applications can be unsaturated, whereas at the type level all applications have to be fully saturated.

data TyCon1 :: (k1 -> k2) -> TyFun k1 k2 -> *

Wrapper for converting the normal type-level arrow into a TyFun. For example, given:

data Nat = Zero | Succ Nat
type family Map (a :: TyFun a b -> *) (a :: [a]) :: [b]
  Map f '[] = '[]
  Map f (x ': xs) = Apply f x ': Map f xs

We can write:

Map (TyCon1 Succ) [Zero, Succ Zero]

data TyCon2 :: (k1 -> k2 -> k3) -> TyFun k1 (TyFun k2 k3 -> *) -> *

Similar to TyCon1, but for two-parameter type constructors.

data TyCon3 :: (k1 -> k2 -> k3 -> k4) -> TyFun k1 (TyFun k2 (TyFun k3 k4 -> *) -> *) -> *

data TyCon4 :: (k1 -> k2 -> k3 -> k4 -> k5) -> TyFun k1 (TyFun k2 (TyFun k3 (TyFun k4 k5 -> *) -> *) -> *) -> *

data TyCon5 :: (k1 -> k2 -> k3 -> k4 -> k5 -> k6) -> TyFun k1 (TyFun k2 (TyFun k3 (TyFun k4 (TyFun k5 k6 -> *) -> *) -> *) -> *) -> *

data TyCon6 :: (k1 -> k2 -> k3 -> k4 -> k5 -> k6 -> k7) -> TyFun k1 (TyFun k2 (TyFun k3 (TyFun k4 (TyFun k5 (TyFun k6 k7 -> *) -> *) -> *) -> *) -> *) -> *

data TyCon7 :: (k1 -> k2 -> k3 -> k4 -> k5 -> k6 -> k7 -> k8) -> TyFun k1 (TyFun k2 (TyFun k3 (TyFun k4 (TyFun k5 (TyFun k6 (TyFun k7 k8 -> *) -> *) -> *) -> *) -> *) -> *) -> *

data TyCon8 :: (k1 -> k2 -> k3 -> k4 -> k5 -> k6 -> k7 -> k8 -> k9) -> TyFun k1 (TyFun k2 (TyFun k3 (TyFun k4 (TyFun k5 (TyFun k6 (TyFun k7 (TyFun k8 k9 -> *) -> *) -> *) -> *) -> *) -> *) -> *) -> *

type family Apply f x :: k2

Type level function application

type (@@) a b = Apply a b infixl 9

An infix synonym for Apply

Defunctionalized singletons

When calling a higher-order singleton function, you need to use a singFun... function to wrap it. See singFun1.

singFun1 :: Proxy f -> SingFunction1 f -> Sing f

Use this function when passing a function on singletons as a higher-order function. You will often need an explicit type annotation to get this to work. For example:

falses = sMap (singFun1 (Proxy :: Proxy NotSym0) sNot)
              (STrue `SCons` STrue `SCons` SNil)

There are a family of singFun... functions, keyed by the number of parameters of the function.

singFun2 :: Proxy f -> SingFunction2 f -> Sing f

singFun3 :: Proxy f -> SingFunction3 f -> Sing f

singFun4 :: Proxy f -> SingFunction4 f -> Sing f

singFun5 :: Proxy f -> SingFunction5 f -> Sing f

singFun6 :: Proxy f -> SingFunction6 f -> Sing f

singFun7 :: Proxy f -> SingFunction7 f -> Sing f

singFun8 :: Proxy f -> SingFunction8 f -> Sing f

unSingFun1 :: Proxy f -> Sing f -> SingFunction1 f

This is the inverse of singFun1, and likewise for the other unSingFun... functions.

unSingFun2 :: Proxy f -> Sing f -> SingFunction2 f

unSingFun3 :: Proxy f -> Sing f -> SingFunction3 f

unSingFun4 :: Proxy f -> Sing f -> SingFunction4 f

unSingFun5 :: Proxy f -> Sing f -> SingFunction5 f

unSingFun6 :: Proxy f -> Sing f -> SingFunction6 f

unSingFun7 :: Proxy f -> Sing f -> SingFunction7 f

unSingFun8 :: Proxy f -> Sing f -> SingFunction8 f

These type synonyms are exported only to improve error messages; users should not have to mention them.

type SingFunction1 f = forall t. Sing t -> Sing (f @@ t)

type SingFunction2 f = forall t. Sing t -> SingFunction1 (f @@ t)

type SingFunction3 f = forall t. Sing t -> SingFunction2 (f @@ t)

type SingFunction4 f = forall t. Sing t -> SingFunction3 (f @@ t)

type SingFunction5 f = forall t. Sing t -> SingFunction4 (f @@ t)

type SingFunction6 f = forall t. Sing t -> SingFunction5 (f @@ t)

type SingFunction7 f = forall t. Sing t -> SingFunction6 (f @@ t)

type SingFunction8 f = forall t. Sing t -> SingFunction7 (f @@ t)

Auxiliary functions

bugInGHC :: forall a. a

GHC 7.8 sometimes warns about incomplete pattern matches when no such patterns are possible, due to GADT constraints. See the bug report at https://ghc.haskell.org/trac/ghc/ticket/3927. In such cases, it's useful to have a catch-all pattern that then has bugInGHC as its right-hand side.

data KProxy t :: * -> *

A concrete, promotable proxy type, for use at the kind level There are no instances for this because it is intended at the kind level only

Constructors

KProxy