Data.Singletons.Either

Description

Defines functions and datatypes relating to the singleton for Either, including a singletons version of all the definitions in Data.Either.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Either. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The Either singleton

data family Sing a

The singleton kind-indexed data family.

Though Haddock doesn't show it, the Sing instance above declares constructors

 SLeft  :: Sing a -> Sing (Left a)
 SRight :: Sing b -> Sing (Right b)

type SEither z = Sing z

SEither is a kind-restricted synonym for Sing: type SEither (a :: Either x y) = Sing a

Singletons from Data.Either

type family Either_ a a a :: c

sEither_ :: forall t t t. (forall t. Sing t -> Sing (t t)) -> (forall t. Sing t -> Sing (t t)) -> Sing t -> Sing (Either_ t t t)

The preceding two definitions are derived from the function either in Data.Either. The extra underscore is to avoid name clashes with the type Either.

type family Lefts a :: [a]

sLefts :: forall t. Sing t -> Sing (Lefts t)

type family Rights a :: [b]

sRights :: forall t. Sing t -> Sing (Rights t)

type family PartitionEithers a :: ([a], [b])

sPartitionEithers :: forall t. Sing t -> Sing (PartitionEithers t)

type family IsLeft a :: Bool

sIsLeft :: forall t. Sing t -> Sing (IsLeft t)

type family IsRight a :: Bool

sIsRight :: forall t. Sing t -> Sing (IsRight t)