Data.Singletons.Eq

Description

Defines the SEq singleton version of the Eq type class.

Documentation

class (kparam ~ KProxy) => SEq kparam where

The singleton analogue of Eq. Unlike the definition for Eq, it is required that instances define a body for '(%:==)'. You may also supply a body for '(%:/=)'.

Methods

(%:==) :: forall a b. Sing a -> Sing b -> Sing (a :== b)

Boolean equality on singletons

(%:/=) :: forall a b. Sing a -> Sing b -> Sing (a :/= b)

Boolean disequality on singletons

type family a == b :: Bool

A type family to compute Boolean equality. Instances are provided only for open kinds, such as * and function kinds. Instances are also provided for datatypes exported from base. A poly-kinded instance is not provided, as a recursive definition for algebraic kinds is generally more useful.

type (:==) a b = a == b

A re-export of the type-level (==) that conforms to the singletons naming convention.

type (:/=) a b = Not (a :== b)

A type synonym conforming to singletons naming conventions