Data.Singletons.Prelude

Description

Mimics the Haskell Prelude, but with singleton types. Includes the basic singleton definitions. Note: This is currently very incomplete!

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Basic singleton definitions

module Data.Singletons

data family Sing (a :: k)

The singleton kind-indexed data family.

Singleton type synonyms

These synonyms are all kind-restricted synonyms of Sing. For example SBool requires an argument of kind Bool.

type SBool = (Sing :: Bool -> Type)

type SList = (Sing :: [a] -> Type)

type SMaybe = (Sing :: Maybe a -> Type)

type SEither = (Sing :: Either a b -> Type)

type SOrdering = (Sing :: Ordering -> Type)

type STuple0 = (Sing :: () -> Type)

type STuple2 = (Sing :: (a, b) -> Type)

type STuple3 = (Sing :: (a, b, c) -> Type)

type STuple4 = (Sing :: (a, b, c, d) -> Type)

type STuple5 = (Sing :: (a, b, c, d, e) -> Type)

type STuple6 = (Sing :: (a, b, c, d, e, f) -> Type)

type STuple7 = (Sing :: (a, b, c, d, e, f, g) -> Type)

Functions working with Bool

type family If k (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If k True tru fls = tru

If k False tru fls = fls

sIf :: Sing a -> Sing b -> Sing c -> Sing (If a b c)

Conditional over singletons

type family Not (a :: Bool) :: Bool where ...

Equations

Not False = TrueSym0

Not True = FalseSym0

sNot :: forall t. Sing t -> Sing (Apply NotSym0 t :: Bool)

type family (a :: Bool) :&& (a :: Bool) :: Bool where ... infixr 3

Equations

False :&& _z_6989586621679545874 = FalseSym0

True :&& x = x

type family (a :: Bool) :|| (a :: Bool) :: Bool where ... infixr 2

Equations

False :|| x = x

True :|| _z_6989586621679545862 = TrueSym0

(%:&&) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:&&$) t) t :: Bool) infixr 3

(%:||) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:||$) t) t :: Bool) infixr 2

type family Otherwise :: Bool where ...

Equations

Otherwise = TrueSym0

sOtherwise :: Sing (OtherwiseSym0 :: Bool)

Error reporting

type family Error (str :: k0) :: k

The promotion of error. This version is more poly-kinded for easier use.

data ErrorSym0 l

sError :: Sing (str :: Symbol) -> a

The singleton for error

Singleton equality

module Data.Singletons.Prelude.Eq

Singleton comparisons

module Data.Singletons.Prelude.Ord

Singleton Enum and Bounded

As a matter of convenience, the singletons Prelude does not export promoted/singletonized succ and pred, due to likely conflicts with unary numbers. Please import Enum directly if you want these.

module Data.Singletons.Prelude.Enum

Singletons numbers

module Data.Singletons.Prelude.Num

Miscellaneous functions

type family Id (a :: a) :: a where ...

Equations

Id x = x

sId :: forall t. Sing t -> Sing (Apply IdSym0 t :: a)

type family Const (a :: a) (a :: b) :: a where ...

Equations

Const x _z_6989586621679549582 = x

sConst :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply ConstSym0 t) t :: a)

type family ((a :: TyFun b c -> Type) :. (a :: TyFun a b -> Type)) (a :: a) :: c where ... infixr 9

Equations

(f :. g) a_6989586621679549545 = Apply (Apply (Apply (Apply Lambda_6989586621679549550Sym0 f) g) a_6989586621679549545) a_6989586621679549545

(%:.) :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (:.$) t) t) t :: c) infixr 9

type family (f :: TyFun a b -> *) $ (x :: a) :: b infixr 0

(%$) :: forall f x. Sing f -> Sing x -> Sing ((($$) @@ f) @@ x) infixr 0

type family (f :: TyFun a b -> *) $! (x :: a) :: b infixr 0

(%$!) :: forall f x. Sing f -> Sing x -> Sing ((($!$) @@ f) @@ x) infixr 0

type family Flip (a :: TyFun a (TyFun b c -> Type) -> Type) (a :: b) (a :: a) :: c where ...

Equations

Flip f x y = Apply (Apply f y) x

sFlip :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FlipSym0 t) t) t :: c)

type family AsTypeOf (a :: a) (a :: a) :: a where ...

Equations

AsTypeOf a_6989586621679549585 a_6989586621679549587 = Apply (Apply ConstSym0 a_6989586621679549585) a_6989586621679549587

sAsTypeOf :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply AsTypeOfSym0 t) t :: a)

type family Seq (a :: a) (a :: b) :: b where ... infixr 0

Equations

Seq _z_6989586621679549508 x = x

sSeq :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply SeqSym0 t) t :: b) infixr 0

List operations

type family Map (a :: TyFun a b -> Type) (a :: [a]) :: [b] where ...

Equations

Map _z_6989586621679549627 '[] = '[]

Map f ((:) x xs) = Apply (Apply (:$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

sMap :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: [b])

type family (a :: [a]) :++ (a :: [a]) :: [a] where ... infixr 5

Equations

'[] :++ ys = ys

((:) x xs) :++ ys = Apply (Apply (:$) x) (Apply (Apply (:++$) xs) ys)

(%:++) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:++$) t) t :: [a]) infixr 5

type family Head (a :: [a]) :: a where ...

Equations

Head ((:) a _z_6989586621679800389) = a

Head '[] = Apply ErrorSym0 "Data.Singletons.List.head: empty list"

sHead :: forall t. Sing t -> Sing (Apply HeadSym0 t :: a)

type family Last (a :: [a]) :: a where ...

Equations

Last '[] = Apply ErrorSym0 "Data.Singletons.List.last: empty list"

Last '[x] = x

Last ((:) _z_6989586621679800379 ((:) x xs)) = Apply LastSym0 (Apply (Apply (:$) x) xs)

sLast :: forall t. Sing t -> Sing (Apply LastSym0 t :: a)

type family Tail (a :: [a]) :: [a] where ...

Equations

Tail ((:) _z_6989586621679800370 t) = t

Tail '[] = Apply ErrorSym0 "Data.Singletons.List.tail: empty list"

sTail :: forall t. Sing t -> Sing (Apply TailSym0 t :: [a])

type family Init (a :: [a]) :: [a] where ...

Equations

Init '[] = Apply ErrorSym0 "Data.Singletons.List.init: empty list"

Init ((:) x xs) = Apply (Apply (Let6989586621679800339Init'Sym2 x xs) x) xs

sInit :: forall t. Sing t -> Sing (Apply InitSym0 t :: [a])

type family Null (a :: [a]) :: Bool where ...

Equations

Null '[] = TrueSym0

Null ((:) _z_6989586621679800268 _z_6989586621679800271) = FalseSym0

sNull :: forall t. Sing t -> Sing (Apply NullSym0 t :: Bool)

type family Reverse (a :: [a]) :: [a] where ...

Equations

Reverse l = Apply (Apply (Let6989586621679800234RevSym1 l) l) '[]

sReverse :: forall t. Sing t -> Sing (Apply ReverseSym0 t :: [a])

Reducing lists (folds)

type family Foldl (a :: TyFun b (TyFun a b -> Type) -> Type) (a :: b) (a :: [a]) :: b where ...

Equations

Foldl f z0 xs0 = Apply (Apply (Let6989586621679508526LgoSym3 f z0 xs0) z0) xs0

sFoldl :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldlSym0 t) t) t :: b)

type family Foldl1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: a where ...

Equations

Foldl1 f ((:) x xs) = Apply (Apply (Apply FoldlSym0 f) x) xs

Foldl1 _z_6989586621679799478 '[] = Apply ErrorSym0 "Data.Singletons.List.foldl1: empty list"

sFoldl1 :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Foldl1Sym0 t) t :: a)

type family Foldr (a :: TyFun a (TyFun b b -> Type) -> Type) (a :: b) (a :: [a]) :: b where ...

Equations

Foldr k z a_6989586621679549648 = Apply (Let6989586621679549653GoSym3 k z a_6989586621679549648) a_6989586621679549648

sFoldr :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldrSym0 t) t) t :: b)

type family Foldr1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: a where ...

Equations

Foldr1 _z_6989586621679799436 '[x] = x

Foldr1 f ((:) x ((:) wild_6989586621679796168 wild_6989586621679796170)) = Apply (Apply f x) (Apply (Apply Foldr1Sym0 f) (Let6989586621679799444XsSym4 f x wild_6989586621679796168 wild_6989586621679796170))

Foldr1 _z_6989586621679799463 '[] = Apply ErrorSym0 "Data.Singletons.List.foldr1: empty list"

sFoldr1 :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Foldr1Sym0 t) t :: a)

Special folds

type family And (a :: [Bool]) :: Bool where ...

Equations

And '[] = TrueSym0

And ((:) x xs) = Apply (Apply (:&&$) x) (Apply AndSym0 xs)

sAnd :: forall t. Sing t -> Sing (Apply AndSym0 t :: Bool)

type family Or (a :: [Bool]) :: Bool where ...

Equations

Or '[] = FalseSym0

Or ((:) x xs) = Apply (Apply (:||$) x) (Apply OrSym0 xs)

sOr :: forall t. Sing t -> Sing (Apply OrSym0 t :: Bool)

type family Any_ (a :: TyFun a Bool -> Type) (a :: [a]) :: Bool where ...

Equations

Any_ _z_6989586621679785031 '[] = FalseSym0

Any_ p ((:) x xs) = Apply (Apply (:||$) (Apply p x)) (Apply (Apply Any_Sym0 p) xs)

sAny_ :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Any_Sym0 t) t :: Bool)

type family All (a :: TyFun a Bool -> Type) (a :: [a]) :: Bool where ...

Equations

All _z_6989586621679799389 '[] = TrueSym0

All p ((:) x xs) = Apply (Apply (:&&$) (Apply p x)) (Apply (Apply AllSym0 p) xs)

sAll :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply AllSym0 t) t :: Bool)

type family Concat (a :: [[a]]) :: [a] where ...

Equations

Concat a_6989586621679799420 = Apply (Apply (Apply FoldrSym0 (:++$)) '[]) a_6989586621679799420

sConcat :: forall t. Sing t -> Sing (Apply ConcatSym0 t :: [a])

type family ConcatMap (a :: TyFun a [b] -> Type) (a :: [a]) :: [b] where ...

Equations

ConcatMap f a_6989586621679799416 = Apply (Apply (Apply FoldrSym0 (Apply (Apply (:.$) (:++$)) f)) '[]) a_6989586621679799416

sConcatMap :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply ConcatMapSym0 t) t :: [b])

Scans

type family Scanl (a :: TyFun b (TyFun a b -> Type) -> Type) (a :: b) (a :: [a]) :: [b] where ...

Equations

Scanl f q ls = Apply (Apply (:$) q) (Case_6989586621679799360 f q ls ls)

sScanl :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanlSym0 t) t) t :: [b])

type family Scanl1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: [a] where ...

Equations

Scanl1 f ((:) x xs) = Apply (Apply (Apply ScanlSym0 f) x) xs

Scanl1 _z_6989586621679799377 '[] = '[]

sScanl1 :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Scanl1Sym0 t) t :: [a])

type family Scanr (a :: TyFun a (TyFun b b -> Type) -> Type) (a :: b) (a :: [a]) :: [b] where ...

Equations

Scanr _z_6989586621679799310 q0 '[] = Apply (Apply (:$) q0) '[]

Scanr f q0 ((:) x xs) = Case_6989586621679799337 f q0 x xs (Let6989586621679799318Scrutinee_6989586621679796172Sym4 f q0 x xs)

sScanr :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanrSym0 t) t) t :: [b])

type family Scanr1 (a :: TyFun a (TyFun a a -> Type) -> Type) (a :: [a]) :: [a] where ...

Equations

Scanr1 _z_6989586621679799241 '[] = '[]

Scanr1 _z_6989586621679799244 '[x] = Apply (Apply (:$) x) '[]

Scanr1 f ((:) x ((:) wild_6989586621679796176 wild_6989586621679796178)) = Case_6989586621679799290 f x wild_6989586621679796176 wild_6989586621679796178 (Let6989586621679799271Scrutinee_6989586621679796174Sym4 f x wild_6989586621679796176 wild_6989586621679796178)

sScanr1 :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply Scanr1Sym0 t) t :: [a])

Searching lists

type family Elem (a :: a) (a :: [a]) :: Bool where ...

Equations

Elem _z_6989586621679798752 '[] = FalseSym0

Elem x ((:) y ys) = Apply (Apply (:||$) (Apply (Apply (:==$) x) y)) (Apply (Apply ElemSym0 x) ys)

sElem :: forall t t. SEq a => Sing t -> Sing t -> Sing (Apply (Apply ElemSym0 t) t :: Bool)

type family NotElem (a :: a) (a :: [a]) :: Bool where ...

Equations

NotElem _z_6989586621679798737 '[] = TrueSym0

NotElem x ((:) y ys) = Apply (Apply (:&&$) (Apply (Apply (:/=$) x) y)) (Apply (Apply NotElemSym0 x) ys)

sNotElem :: forall t t. SEq a => Sing t -> Sing t -> Sing (Apply (Apply NotElemSym0 t) t :: Bool)

type family Lookup (a :: a) (a :: [(a, b)]) :: Maybe b where ...

Equations

Lookup _key '[] = NothingSym0

Lookup key ((:) '(x, y) xys) = Case_6989586621679797320 key x y xys (Let6989586621679797301Scrutinee_6989586621679796256Sym4 key x y xys)

sLookup :: forall t t. SEq a => Sing t -> Sing t -> Sing (Apply (Apply LookupSym0 t) t :: Maybe b)

Zipping and unzipping lists

type family Zip (a :: [a]) (a :: [b]) :: [(a, b)] where ...

Equations

Zip ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply Tuple2Sym0 x) y)) (Apply (Apply ZipSym0 xs) ys)

Zip '[] '[] = '[]

Zip ((:) _z_6989586621679798614 _z_6989586621679798617) '[] = '[]

Zip '[] ((:) _z_6989586621679798620 _z_6989586621679798623) = '[]

sZip :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply ZipSym0 t) t :: [(a, b)])

type family Zip3 (a :: [a]) (a :: [b]) (a :: [c]) :: [(a, b, c)] where ...

Equations

Zip3 ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply Tuple3Sym0 a) b) c)) (Apply (Apply (Apply Zip3Sym0 as) bs) cs)

Zip3 '[] '[] '[] = '[]

Zip3 '[] '[] ((:) _z_6989586621679798547 _z_6989586621679798550) = '[]

Zip3 '[] ((:) _z_6989586621679798553 _z_6989586621679798556) '[] = '[]

Zip3 '[] ((:) _z_6989586621679798559 _z_6989586621679798562) ((:) _z_6989586621679798565 _z_6989586621679798568) = '[]

Zip3 ((:) _z_6989586621679798571 _z_6989586621679798574) '[] '[] = '[]

Zip3 ((:) _z_6989586621679798577 _z_6989586621679798580) '[] ((:) _z_6989586621679798583 _z_6989586621679798586) = '[]

Zip3 ((:) _z_6989586621679798589 _z_6989586621679798592) ((:) _z_6989586621679798595 _z_6989586621679798598) '[] = '[]

sZip3 :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Zip3Sym0 t) t) t :: [(a, b, c)])

type family ZipWith (a :: TyFun a (TyFun b c -> Type) -> Type) (a :: [a]) (a :: [b]) :: [c] where ...

Equations

ZipWith f ((:) x xs) ((:) y ys) = Apply (Apply (:$) (Apply (Apply f x) y)) (Apply (Apply (Apply ZipWithSym0 f) xs) ys)

ZipWith _z_6989586621679798505 '[] '[] = '[]

ZipWith _z_6989586621679798508 ((:) _z_6989586621679798511 _z_6989586621679798514) '[] = '[]

ZipWith _z_6989586621679798517 '[] ((:) _z_6989586621679798520 _z_6989586621679798523) = '[]

sZipWith :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ZipWithSym0 t) t) t :: [c])

type family ZipWith3 (a :: TyFun a (TyFun b (TyFun c d -> Type) -> Type) -> Type) (a :: [a]) (a :: [b]) (a :: [c]) :: [d] where ...

Equations

ZipWith3 z ((:) a as) ((:) b bs) ((:) c cs) = Apply (Apply (:$) (Apply (Apply (Apply z a) b) c)) (Apply (Apply (Apply (Apply ZipWith3Sym0 z) as) bs) cs)

ZipWith3 _z_6989586621679798410 '[] '[] '[] = '[]

ZipWith3 _z_6989586621679798413 '[] '[] ((:) _z_6989586621679798416 _z_6989586621679798419) = '[]

ZipWith3 _z_6989586621679798422 '[] ((:) _z_6989586621679798425 _z_6989586621679798428) '[] = '[]

ZipWith3 _z_6989586621679798431 '[] ((:) _z_6989586621679798434 _z_6989586621679798437) ((:) _z_6989586621679798440 _z_6989586621679798443) = '[]

ZipWith3 _z_6989586621679798446 ((:) _z_6989586621679798449 _z_6989586621679798452) '[] '[] = '[]

ZipWith3 _z_6989586621679798455 ((:) _z_6989586621679798458 _z_6989586621679798461) '[] ((:) _z_6989586621679798464 _z_6989586621679798467) = '[]

ZipWith3 _z_6989586621679798470 ((:) _z_6989586621679798473 _z_6989586621679798476) ((:) _z_6989586621679798479 _z_6989586621679798482) '[] = '[]

sZipWith3 :: forall t t t t. Sing t -> Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (Apply ZipWith3Sym0 t) t) t) t :: [d])

type family Unzip (a :: [(a, b)]) :: ([a], [b]) where ...

Equations

Unzip xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798356Sym0 xs)) (Apply (Apply Tuple2Sym0 '[]) '[])) xs

sUnzip :: forall t. Sing t -> Sing (Apply UnzipSym0 t :: ([a], [b]))

type family Unzip3 (a :: [(a, b, c)]) :: ([a], [b], [c]) where ...

Equations

Unzip3 xs = Apply (Apply (Apply FoldrSym0 (Apply Lambda_6989586621679798324Sym0 xs)) (Apply (Apply (Apply Tuple3Sym0 '[]) '[]) '[])) xs

sUnzip3 :: forall t. Sing t -> Sing (Apply Unzip3Sym0 t :: ([a], [b], [c]))

Other datatypes

type family Maybe_ (a :: b) (a :: TyFun a b -> Type) (a :: Maybe a) :: b where ...

Equations

Maybe_ n _z_6989586621679770024 Nothing = n

Maybe_ _z_6989586621679770027 f (Just x) = Apply f x

sMaybe_ :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Maybe_Sym0 t) t) t :: b)

type family Either_ (a :: TyFun a c -> Type) (a :: TyFun b c -> Type) (a :: Either a b) :: c where ...

Equations

Either_ f _z_6989586621679760570 (Left x) = Apply f x

Either_ _z_6989586621679760574 g (Right y) = Apply g y

sEither_ :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Either_Sym0 t) t) t :: c)

type family Fst (a :: (a, b)) :: a where ...

Equations

Fst '(x, _z_6989586621679781175) = x

sFst :: forall t. Sing t -> Sing (Apply FstSym0 t :: a)

type family Snd (a :: (a, b)) :: b where ...

Equations

Snd '(_z_6989586621679781166, y) = y

sSnd :: forall t. Sing t -> Sing (Apply SndSym0 t :: b)

type family Curry (a :: TyFun (a, b) c -> Type) (a :: a) (a :: b) :: c where ...

Equations

Curry f x y = Apply f (Apply (Apply Tuple2Sym0 x) y)

sCurry :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply CurrySym0 t) t) t :: c)

type family Uncurry (a :: TyFun a (TyFun b c -> Type) -> Type) (a :: (a, b)) :: c where ...

Equations

Uncurry f p = Apply (Apply f (Apply FstSym0 p)) (Apply SndSym0 p)

sUncurry :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply UncurrySym0 t) t :: c)

data Symbol :: *

(Kind) This is the kind of type-level symbols. Declared here because class IP needs it

Other functions

either_ :: forall a c b. (a -> c) -> (b -> c) -> Either a b -> c

maybe_ :: forall b a. b -> (a -> b) -> Maybe a -> b

bool_ :: forall a. a -> a -> Bool -> a

any_ :: forall a. (a -> Bool) -> [a] -> Bool

Defunctionalization symbols

type FalseSym0 = False

type TrueSym0 = True

data NotSym0 l

type NotSym1 t = Not t

data (:&&$) l

data l :&&$$ l

type (:&&$$$) t t = (:&&) t t

data (:||$) l

data l :||$$ l

type (:||$$$) t t = (:||) t t

type OtherwiseSym0 = Otherwise

type NothingSym0 = Nothing

data JustSym0 l

type JustSym1 t = Just t

data Maybe_Sym0 l

data Maybe_Sym1 l l

data Maybe_Sym2 l l l

type Maybe_Sym3 t t t = Maybe_ t t t

data LeftSym0 l

type LeftSym1 t = Left t

data RightSym0 l

type RightSym1 t = Right t

data Either_Sym0 l

data Either_Sym1 l l

data Either_Sym2 l l l

type Either_Sym3 t t t = Either_ t t t

type Tuple0Sym0 = '()

data Tuple2Sym0 l

data Tuple2Sym1 l l

type Tuple2Sym2 t t = '(t, t)

data Tuple3Sym0 l

data Tuple3Sym1 l l

data Tuple3Sym2 l l l

type Tuple3Sym3 t t t = '(t, t, t)

data Tuple4Sym0 l

data Tuple4Sym1 l l

data Tuple4Sym2 l l l

data Tuple4Sym3 l l l l

type Tuple4Sym4 t t t t = '(t, t, t, t)

data Tuple5Sym0 l

data Tuple5Sym1 l l

data Tuple5Sym2 l l l

data Tuple5Sym3 l l l l

data Tuple5Sym4 l l l l l

type Tuple5Sym5 t t t t t = '(t, t, t, t, t)

data Tuple6Sym0 l

data Tuple6Sym1 l l

data Tuple6Sym2 l l l

data Tuple6Sym3 l l l l

data Tuple6Sym4 l l l l l

data Tuple6Sym5 l l l l l l

type Tuple6Sym6 t t t t t t = '(t, t, t, t, t, t)

data Tuple7Sym0 l

data Tuple7Sym1 l l

data Tuple7Sym2 l l l

data Tuple7Sym3 l l l l

data Tuple7Sym4 l l l l l

data Tuple7Sym5 l l l l l l

data Tuple7Sym6 l l l l l l l

type Tuple7Sym7 t t t t t t t = '(t, t, t, t, t, t, t)

data FstSym0 l

type FstSym1 t = Fst t

data SndSym0 l

type SndSym1 t = Snd t

data CurrySym0 l

data CurrySym1 l l

data CurrySym2 l l l

type CurrySym3 t t t = Curry t t t

data UncurrySym0 l

data UncurrySym1 l l

type UncurrySym2 t t = Uncurry t t

data IdSym0 l

type IdSym1 t = Id t

data ConstSym0 l

data ConstSym1 l l

type ConstSym2 t t = Const t t

data (:.$) l

data l :.$$ l

data (l :.$$$ l) l

data ($$) :: TyFun (TyFun a b -> *) (TyFun a b -> *) -> *

data ($$$) :: (TyFun a b -> *) -> TyFun a b -> *

type ($$$$) a b = ($) a b

data ($!$) :: TyFun (TyFun a b -> *) (TyFun a b -> *) -> *

data ($!$$) :: (TyFun a b -> *) -> TyFun a b -> *

type ($!$$$) a b = ($!) a b

data FlipSym0 l

data FlipSym1 l l

data FlipSym2 l l l

data AsTypeOfSym0 l

data AsTypeOfSym1 l l

type AsTypeOfSym2 t t = AsTypeOf t t

data SeqSym0 l

data SeqSym1 l l

type SeqSym2 t t = Seq t t

data (:$) l

data l :$$ l

type (:$$$) t t = (:) t t

type NilSym0 = '[]

data MapSym0 l

data MapSym1 l l

type MapSym2 t t = Map t t

data ReverseSym0 l

type ReverseSym1 t = Reverse t

data l :++$$ l

data (:++$) l

data HeadSym0 l

type HeadSym1 t = Head t

data LastSym0 l

type LastSym1 t = Last t

data TailSym0 l

type TailSym1 t = Tail t

data InitSym0 l

type InitSym1 t = Init t

data NullSym0 l

type NullSym1 t = Null t

data FoldlSym0 l

data FoldlSym1 l l

data FoldlSym2 l l l

type FoldlSym3 t t t = Foldl t t t

data Foldl1Sym0 l

data Foldl1Sym1 l l

type Foldl1Sym2 t t = Foldl1 t t

data FoldrSym0 l

data FoldrSym1 l l

data FoldrSym2 l l l

type FoldrSym3 t t t = Foldr t t t

data Foldr1Sym0 l

data Foldr1Sym1 l l

type Foldr1Sym2 t t = Foldr1 t t

data ConcatSym0 l

type ConcatSym1 t = Concat t

data ConcatMapSym0 l

data ConcatMapSym1 l l

type ConcatMapSym2 t t = ConcatMap t t

data AndSym0 l

type AndSym1 t = And t

data OrSym0 l

type OrSym1 t = Or t

data Any_Sym0 l

data Any_Sym1 l l

type Any_Sym2 t t = Any_ t t

data AllSym0 l

data AllSym1 l l

type AllSym2 t t = All t t

data ScanlSym0 l

data ScanlSym1 l l

data ScanlSym2 l l l

type ScanlSym3 t t t = Scanl t t t

data Scanl1Sym0 l

data Scanl1Sym1 l l

type Scanl1Sym2 t t = Scanl1 t t

data ScanrSym0 l

data ScanrSym1 l l

data ScanrSym2 l l l

type ScanrSym3 t t t = Scanr t t t

data Scanr1Sym0 l

data Scanr1Sym1 l l

type Scanr1Sym2 t t = Scanr1 t t

data ElemSym0 l

data ElemSym1 l l

type ElemSym2 t t = Elem t t

data NotElemSym0 l

data NotElemSym1 l l

type NotElemSym2 t t = NotElem t t

data ZipSym0 l

data ZipSym1 l l

type ZipSym2 t t = Zip t t

data Zip3Sym0 l

data Zip3Sym1 l l

data Zip3Sym2 l l l

type Zip3Sym3 t t t = Zip3 t t t

data ZipWithSym0 l

data ZipWithSym1 l l

data ZipWithSym2 l l l

type ZipWithSym3 t t t = ZipWith t t t

data ZipWith3Sym0 l

data ZipWith3Sym1 l l

data ZipWith3Sym2 l l l

data ZipWith3Sym3 l l l l

data UnzipSym0 l

type UnzipSym1 t = Unzip t