Data.Singletons.Prelude.Base

Description

Implements singletonized versions of functions from GHC.Base module.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Tuple. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Basic functions

type family Foldr (a :: TyFun a (TyFun b b -> Type) -> Type) (a :: b) (a :: [a]) :: b where ...

Equations

Foldr k z a_6989586621679549648 = Apply (Let6989586621679549653GoSym3 k z a_6989586621679549648) a_6989586621679549648

sFoldr :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldrSym0 t) t) t :: b)

type family Map (a :: TyFun a b -> Type) (a :: [a]) :: [b] where ...

Equations

Map _z_6989586621679549627 '[] = '[]

Map f ((:) x xs) = Apply (Apply (:$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

sMap :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: [b])

type family (a :: [a]) :++ (a :: [a]) :: [a] where ... infixr 5

Equations

'[] :++ ys = ys

((:) x xs) :++ ys = Apply (Apply (:$) x) (Apply (Apply (:++$) xs) ys)

(%:++) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:++$) t) t :: [a]) infixr 5

type family Otherwise :: Bool where ...

Equations

Otherwise = TrueSym0

sOtherwise :: Sing (OtherwiseSym0 :: Bool)

type family Id (a :: a) :: a where ...

Equations

Id x = x

sId :: forall t. Sing t -> Sing (Apply IdSym0 t :: a)

type family Const (a :: a) (a :: b) :: a where ...

Equations

Const x _z_6989586621679549582 = x

sConst :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply ConstSym0 t) t :: a)

type family ((a :: TyFun b c -> Type) :. (a :: TyFun a b -> Type)) (a :: a) :: c where ... infixr 9

Equations

(f :. g) a_6989586621679549545 = Apply (Apply (Apply (Apply Lambda_6989586621679549550Sym0 f) g) a_6989586621679549545) a_6989586621679549545

(%:.) :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (:.$) t) t) t :: c) infixr 9

type family (f :: TyFun a b -> *) $ (x :: a) :: b infixr 0

type family (f :: TyFun a b -> *) $! (x :: a) :: b infixr 0

(%$) :: forall f x. Sing f -> Sing x -> Sing ((($$) @@ f) @@ x) infixr 0

(%$!) :: forall f x. Sing f -> Sing x -> Sing ((($!$) @@ f) @@ x) infixr 0

type family Flip (a :: TyFun a (TyFun b c -> Type) -> Type) (a :: b) (a :: a) :: c where ...

Equations

Flip f x y = Apply (Apply f y) x

sFlip :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FlipSym0 t) t) t :: c)

type family AsTypeOf (a :: a) (a :: a) :: a where ...

Equations

AsTypeOf a_6989586621679549585 a_6989586621679549587 = Apply (Apply ConstSym0 a_6989586621679549585) a_6989586621679549587

sAsTypeOf :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply AsTypeOfSym0 t) t :: a)

type family Seq (a :: a) (a :: b) :: b where ... infixr 0

Equations

Seq _z_6989586621679549508 x = x

sSeq :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply SeqSym0 t) t :: b) infixr 0

Defunctionalization symbols

data FoldrSym0 l

data FoldrSym1 l l

data FoldrSym2 l l l

type FoldrSym3 t t t = Foldr t t t

data MapSym0 l

data MapSym1 l l

type MapSym2 t t = Map t t

data (:++$) l

data l :++$$ l

type (:++$$$) t t = (:++) t t

type OtherwiseSym0 = Otherwise

data IdSym0 l

type IdSym1 t = Id t

data ConstSym0 l

data ConstSym1 l l

type ConstSym2 t t = Const t t

data (:.$) l

data l :.$$ l

data (l :.$$$ l) l

type (:.$$$$) t t t = (:.) t t t

data ($$) :: TyFun (TyFun a b -> *) (TyFun a b -> *) -> *

data ($$$) :: (TyFun a b -> *) -> TyFun a b -> *

type ($$$$) a b = ($) a b

data ($!$) :: TyFun (TyFun a b -> *) (TyFun a b -> *) -> *

data ($!$$) :: (TyFun a b -> *) -> TyFun a b -> *

type ($!$$$) a b = ($!) a b

data FlipSym0 l

data FlipSym1 l l

data FlipSym2 l l l

type FlipSym3 t t t = Flip t t t

data AsTypeOfSym0 l

data AsTypeOfSym1 l l

type AsTypeOfSym2 t t = AsTypeOf t t

data SeqSym0 l

data SeqSym1 l l

type SeqSym2 t t = Seq t t