Data.Singletons.Prelude.Base

Description

Implements singletonized versions of functions from GHC.Base module.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Tuple. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Basic functions

type family Foldr (a :: (~>) a ((~>) b b)) (a :: b) (a :: [a]) :: b where ...

Equations

Foldr k z a_6989586621679391853 = Apply (Let6989586621679391858GoSym3 k z a_6989586621679391853) a_6989586621679391853

sFoldr :: forall a b (t :: (~>) a ((~>) b b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldrSym0 t) t) t :: b)

type family Map (a :: (~>) a b) (a :: [a]) :: [b] where ...

Equations

Map _ '[] = '[]

Map f ((:) x xs) = Apply (Apply (:@#@$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

sMap :: forall a b (t :: (~>) a b) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: [b])

type family (a :: [a]) ++ (a :: [a]) :: [a] where ... infixr 5

Equations

'[] ++ ys = ys

((:) x xs) ++ ys = Apply (Apply (:@#@$) x) (Apply (Apply (++@#@$) xs) ys)

(%++) :: forall a (t :: [a]) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply (++@#@$) t) t :: [a]) infixr 5

type family Otherwise :: Bool where ...

Equations

Otherwise = TrueSym0

sOtherwise :: Sing (OtherwiseSym0 :: Bool)

type family Id (a :: a) :: a where ...

Equations

Id x = x

sId :: forall a (t :: a). Sing t -> Sing (Apply IdSym0 t :: a)

type family Const (a :: a) (a :: b) :: a where ...

Equations

Const x _ = x

sConst :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply ConstSym0 t) t :: a)

type family ((a :: (~>) b c) :. (a :: (~>) a b)) (a :: a) :: c where ... infixr 9

Equations

(f :. g) a_6989586621679391801 = Apply (Apply (Apply (Apply Lambda_6989586621679391806Sym0 f) g) a_6989586621679391801) a_6989586621679391801

(%.) :: forall b c a (t :: (~>) b c) (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (.@#@$) t) t) t :: c) infixr 9

type family (a :: (~>) a b) $ (a :: a) :: b where ... infixr 0

Equations

f $ x = Apply f x

type family (a :: (~>) a b) $! (a :: a) :: b where ... infixr 0

Equations

f $! x = Apply f (Let6989586621679391777VxSym2 f x)

(%$) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($@#@$) t) t :: b) infixr 0

(%$!) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($!@#@$) t) t :: b) infixr 0

type family Until (a :: (~>) a Bool) (a :: (~>) a a) (a :: a) :: a where ...

Equations

Until p f a_6989586621679391751 = Apply (Let6989586621679391756GoSym3 p f a_6989586621679391751) a_6989586621679391751

sUntil :: forall a (t :: (~>) a Bool) (t :: (~>) a a) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply UntilSym0 t) t) t :: a)

type family Flip (a :: (~>) a ((~>) b c)) (a :: b) (a :: a) :: c where ...

Equations

Flip f x y = Apply (Apply f y) x

sFlip :: forall a b c (t :: (~>) a ((~>) b c)) (t :: b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FlipSym0 t) t) t :: c)

type family AsTypeOf (a :: a) (a :: a) :: a where ...

Equations

AsTypeOf a_6989586621679391819 a_6989586621679391821 = Apply (Apply ConstSym0 a_6989586621679391819) a_6989586621679391821

sAsTypeOf :: forall a (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply AsTypeOfSym0 t) t :: a)

type family Seq (a :: a) (a :: b) :: b where ... infixr 0

Equations

Seq _ x = x

sSeq :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply SeqSym0 t) t :: b) infixr 0

Defunctionalization symbols

data FoldrSym0 :: forall a6989586621679391638 b6989586621679391639. (~>) ((~>) a6989586621679391638 ((~>) b6989586621679391639 b6989586621679391639)) ((~>) b6989586621679391639 ((~>) [a6989586621679391638] b6989586621679391639))

data FoldrSym1 (a6989586621679391847 :: (~>) a6989586621679391638 ((~>) b6989586621679391639 b6989586621679391639)) :: (~>) b6989586621679391639 ((~>) [a6989586621679391638] b6989586621679391639)

data FoldrSym2 (a6989586621679391847 :: (~>) a6989586621679391638 ((~>) b6989586621679391639 b6989586621679391639)) (a6989586621679391848 :: b6989586621679391639) :: (~>) [a6989586621679391638] b6989586621679391639

type FoldrSym3 (a6989586621679391847 :: (~>) a6989586621679391638 ((~>) b6989586621679391639 b6989586621679391639)) (a6989586621679391848 :: b6989586621679391639) (a6989586621679391849 :: [a6989586621679391638]) = Foldr a6989586621679391847 a6989586621679391848 a6989586621679391849

data MapSym0 :: forall a6989586621679391636 b6989586621679391637. (~>) ((~>) a6989586621679391636 b6989586621679391637) ((~>) [a6989586621679391636] [b6989586621679391637])

data MapSym1 (a6989586621679391840 :: (~>) a6989586621679391636 b6989586621679391637) :: (~>) [a6989586621679391636] [b6989586621679391637]

type MapSym2 (a6989586621679391840 :: (~>) a6989586621679391636 b6989586621679391637) (a6989586621679391841 :: [a6989586621679391636]) = Map a6989586621679391840 a6989586621679391841

data (++@#@$) :: forall a6989586621679391635. (~>) [a6989586621679391635] ((~>) [a6989586621679391635] [a6989586621679391635]) infixr 5

data (++@#@$$) (a6989586621679391832 :: [a6989586621679391635]) :: (~>) [a6989586621679391635] [a6989586621679391635] infixr 5

type (++@#@$$$) (a6989586621679391832 :: [a6989586621679391635]) (a6989586621679391833 :: [a6989586621679391635]) = (++) a6989586621679391832 a6989586621679391833

type OtherwiseSym0 = Otherwise

data IdSym0 :: forall a6989586621679391634. (~>) a6989586621679391634 a6989586621679391634

type IdSym1 (a6989586621679391829 :: a6989586621679391634) = Id a6989586621679391829

data ConstSym0 :: forall a6989586621679391632 b6989586621679391633. (~>) a6989586621679391632 ((~>) b6989586621679391633 a6989586621679391632)

data ConstSym1 (a6989586621679391814 :: a6989586621679391632) :: forall b6989586621679391633. (~>) b6989586621679391633 a6989586621679391632

type ConstSym2 (a6989586621679391814 :: a6989586621679391632) (a6989586621679391815 :: b6989586621679391633) = Const a6989586621679391814 a6989586621679391815

data (.@#@$) :: forall a6989586621679391631 b6989586621679391629 c6989586621679391630. (~>) ((~>) b6989586621679391629 c6989586621679391630) ((~>) ((~>) a6989586621679391631 b6989586621679391629) ((~>) a6989586621679391631 c6989586621679391630)) infixr 9

data (.@#@$$) (a6989586621679391795 :: (~>) b6989586621679391629 c6989586621679391630) :: forall a6989586621679391631. (~>) ((~>) a6989586621679391631 b6989586621679391629) ((~>) a6989586621679391631 c6989586621679391630) infixr 9

data (a6989586621679391795 :: (~>) b6989586621679391629 c6989586621679391630) .@#@$$$ (a6989586621679391796 :: (~>) a6989586621679391631 b6989586621679391629) :: (~>) a6989586621679391631 c6989586621679391630 infixr 9

type (.@#@$$$$) (a6989586621679391795 :: (~>) b6989586621679391629 c6989586621679391630) (a6989586621679391796 :: (~>) a6989586621679391631 b6989586621679391629) (a6989586621679391797 :: a6989586621679391631) = (:.) a6989586621679391795 a6989586621679391796 a6989586621679391797

data ($@#@$) :: forall a6989586621679391623 b6989586621679391624. (~>) ((~>) a6989586621679391623 b6989586621679391624) ((~>) a6989586621679391623 b6989586621679391624) infixr 0

data ($@#@$$) (a6989586621679391780 :: (~>) a6989586621679391623 b6989586621679391624) :: (~>) a6989586621679391623 b6989586621679391624 infixr 0

type ($@#@$$$) (a6989586621679391780 :: (~>) a6989586621679391623 b6989586621679391624) (a6989586621679391781 :: a6989586621679391623) = ($) a6989586621679391780 a6989586621679391781

data ($!@#@$) :: forall a6989586621679391621 b6989586621679391622. (~>) ((~>) a6989586621679391621 b6989586621679391622) ((~>) a6989586621679391621 b6989586621679391622) infixr 0

data ($!@#@$$) (a6989586621679391771 :: (~>) a6989586621679391621 b6989586621679391622) :: (~>) a6989586621679391621 b6989586621679391622 infixr 0

type ($!@#@$$$) (a6989586621679391771 :: (~>) a6989586621679391621 b6989586621679391622) (a6989586621679391772 :: a6989586621679391621) = ($!) a6989586621679391771 a6989586621679391772

data UntilSym0 :: forall a6989586621679391620. (~>) ((~>) a6989586621679391620 Bool) ((~>) ((~>) a6989586621679391620 a6989586621679391620) ((~>) a6989586621679391620 a6989586621679391620))

data UntilSym1 (a6989586621679391745 :: (~>) a6989586621679391620 Bool) :: (~>) ((~>) a6989586621679391620 a6989586621679391620) ((~>) a6989586621679391620 a6989586621679391620)

data UntilSym2 (a6989586621679391745 :: (~>) a6989586621679391620 Bool) (a6989586621679391746 :: (~>) a6989586621679391620 a6989586621679391620) :: (~>) a6989586621679391620 a6989586621679391620

type UntilSym3 (a6989586621679391745 :: (~>) a6989586621679391620 Bool) (a6989586621679391746 :: (~>) a6989586621679391620 a6989586621679391620) (a6989586621679391747 :: a6989586621679391620) = Until a6989586621679391745 a6989586621679391746 a6989586621679391747

data FlipSym0 :: forall a6989586621679391626 b6989586621679391627 c6989586621679391628. (~>) ((~>) a6989586621679391626 ((~>) b6989586621679391627 c6989586621679391628)) ((~>) b6989586621679391627 ((~>) a6989586621679391626 c6989586621679391628))

data FlipSym1 (a6989586621679391786 :: (~>) a6989586621679391626 ((~>) b6989586621679391627 c6989586621679391628)) :: (~>) b6989586621679391627 ((~>) a6989586621679391626 c6989586621679391628)

data FlipSym2 (a6989586621679391786 :: (~>) a6989586621679391626 ((~>) b6989586621679391627 c6989586621679391628)) (a6989586621679391787 :: b6989586621679391627) :: (~>) a6989586621679391626 c6989586621679391628

type FlipSym3 (a6989586621679391786 :: (~>) a6989586621679391626 ((~>) b6989586621679391627 c6989586621679391628)) (a6989586621679391787 :: b6989586621679391627) (a6989586621679391788 :: a6989586621679391626) = Flip a6989586621679391786 a6989586621679391787 a6989586621679391788

data AsTypeOfSym0 :: forall a6989586621679391625. (~>) a6989586621679391625 ((~>) a6989586621679391625 a6989586621679391625)

data AsTypeOfSym1 (a6989586621679391823 :: a6989586621679391625) :: (~>) a6989586621679391625 a6989586621679391625

type AsTypeOfSym2 (a6989586621679391823 :: a6989586621679391625) (a6989586621679391824 :: a6989586621679391625) = AsTypeOf a6989586621679391823 a6989586621679391824

data SeqSym0 :: forall a6989586621679391618 b6989586621679391619. (~>) a6989586621679391618 ((~>) b6989586621679391619 b6989586621679391619) infixr 0

data SeqSym1 (a6989586621679391740 :: a6989586621679391618) :: forall b6989586621679391619. (~>) b6989586621679391619 b6989586621679391619 infixr 0

type SeqSym2 (a6989586621679391740 :: a6989586621679391618) (a6989586621679391741 :: b6989586621679391619) = Seq a6989586621679391740 a6989586621679391741