Data.Singletons.Prelude.Base

Description

Implements singletonized versions of functions from GHC.Base module.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Tuple. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Basic functions

type family Foldr (a :: (~>) a ((~>) b b)) (a :: b) (a :: [a]) :: b where ...

Equations

Foldr k z a_6989586621679547237 = Apply (Let6989586621679547242GoSym3 k z a_6989586621679547237) a_6989586621679547237

sFoldr :: forall a b (t :: (~>) a ((~>) b b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldrSym0 t) t) t :: b)

type family Map (a :: (~>) a b) (a :: [a]) :: [b] where ...

Equations

Map _ '[] = '[]

Map f ('(:) x xs) = Apply (Apply (:@#@$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

sMap :: forall a b (t :: (~>) a b) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: [b])

type family (a :: [a]) ++ (a :: [a]) :: [a] where ... infixr 5

Equations

'[] ++ ys = ys

('(:) x xs) ++ ys = Apply (Apply (:@#@$) x) (Apply (Apply (++@#@$) xs) ys)

(%++) :: forall a (t :: [a]) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply (++@#@$) t) t :: [a]) infixr 5

type family Otherwise :: Bool where ...

Equations

Otherwise = TrueSym0

sOtherwise :: Sing (OtherwiseSym0 :: Bool)

type family Id (a :: a) :: a where ...

Equations

Id x = x

sId :: forall a (t :: a). Sing t -> Sing (Apply IdSym0 t :: a)

type family Const (a :: a) (a :: b) :: a where ...

Equations

Const x _ = x

sConst :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply ConstSym0 t) t :: a)

type family ((a :: (~>) b c) . (a :: (~>) a b)) (a :: a) :: c where ... infixr 9

Equations

(f . g) a_6989586621679547195 = Apply (Apply (Apply (Apply Lambda_6989586621679547200Sym0 f) g) a_6989586621679547195) a_6989586621679547195

(%.) :: forall b c a (t :: (~>) b c) (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (.@#@$) t) t) t :: c) infixr 9

type family (a :: (~>) a b) $ (a :: a) :: b where ... infixr 0

Equations

f $ x = Apply f x

type family (a :: (~>) a b) $! (a :: a) :: b where ... infixr 0

Equations

f $! x = Apply f (Let6989586621679547161VxSym2 f x)

(%$) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($@#@$) t) t :: b) infixr 0

(%$!) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($!@#@$) t) t :: b) infixr 0

type family Until (a :: (~>) a Bool) (a :: (~>) a a) (a :: a) :: a where ...

Equations

Until p f a_6989586621679547135 = Apply (Let6989586621679547140GoSym3 p f a_6989586621679547135) a_6989586621679547135

sUntil :: forall a (t :: (~>) a Bool) (t :: (~>) a a) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply UntilSym0 t) t) t :: a)

type family Flip (a :: (~>) a ((~>) b c)) (a :: b) (a :: a) :: c where ...

Equations

Flip f x y = Apply (Apply f y) x

sFlip :: forall a b c (t :: (~>) a ((~>) b c)) (t :: b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FlipSym0 t) t) t :: c)

type family AsTypeOf (a :: a) (a :: a) :: a where ...

Equations

AsTypeOf a_6989586621679547170 a_6989586621679547172 = Apply (Apply ConstSym0 a_6989586621679547170) a_6989586621679547172

sAsTypeOf :: forall a (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply AsTypeOfSym0 t) t :: a)

type family Seq (a :: a) (a :: b) :: b where ... infixr 0

Equations

Seq _ x = x

sSeq :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply SeqSym0 t) t :: b) infixr 0

Defunctionalization symbols

data FoldrSym0 :: forall a6989586621679547022 b6989586621679547023. (~>) ((~>) a6989586621679547022 ((~>) b6989586621679547023 b6989586621679547023)) ((~>) b6989586621679547023 ((~>) [a6989586621679547022] b6989586621679547023))

data FoldrSym1 (a6989586621679547231 :: (~>) a6989586621679547022 ((~>) b6989586621679547023 b6989586621679547023)) :: (~>) b6989586621679547023 ((~>) [a6989586621679547022] b6989586621679547023)

data FoldrSym2 (a6989586621679547231 :: (~>) a6989586621679547022 ((~>) b6989586621679547023 b6989586621679547023)) (a6989586621679547232 :: b6989586621679547023) :: (~>) [a6989586621679547022] b6989586621679547023

type FoldrSym3 (a6989586621679547231 :: (~>) a6989586621679547022 ((~>) b6989586621679547023 b6989586621679547023)) (a6989586621679547232 :: b6989586621679547023) (a6989586621679547233 :: [a6989586621679547022]) = Foldr a6989586621679547231 a6989586621679547232 a6989586621679547233

data MapSym0 :: forall a6989586621679547020 b6989586621679547021. (~>) ((~>) a6989586621679547020 b6989586621679547021) ((~>) [a6989586621679547020] [b6989586621679547021])

data MapSym1 (a6989586621679547224 :: (~>) a6989586621679547020 b6989586621679547021) :: (~>) [a6989586621679547020] [b6989586621679547021]

type MapSym2 (a6989586621679547224 :: (~>) a6989586621679547020 b6989586621679547021) (a6989586621679547225 :: [a6989586621679547020]) = Map a6989586621679547224 a6989586621679547225

data (++@#@$) :: forall a6989586621679547019. (~>) [a6989586621679547019] ((~>) [a6989586621679547019] [a6989586621679547019]) infixr 5

data (++@#@$$) (a6989586621679547216 :: [a6989586621679547019]) :: (~>) [a6989586621679547019] [a6989586621679547019] infixr 5

type (++@#@$$$) (a6989586621679547216 :: [a6989586621679547019]) (a6989586621679547217 :: [a6989586621679547019]) = (++) a6989586621679547216 a6989586621679547217

type OtherwiseSym0 = Otherwise

data IdSym0 :: forall a6989586621679547018. (~>) a6989586621679547018 a6989586621679547018

type IdSym1 (a6989586621679547213 :: a6989586621679547018) = Id a6989586621679547213

data ConstSym0 :: forall a6989586621679547016 b6989586621679547017. (~>) a6989586621679547016 ((~>) b6989586621679547017 a6989586621679547016)

data ConstSym1 (a6989586621679547208 :: a6989586621679547016) :: forall b6989586621679547017. (~>) b6989586621679547017 a6989586621679547016

type ConstSym2 (a6989586621679547208 :: a6989586621679547016) (a6989586621679547209 :: b6989586621679547017) = Const a6989586621679547208 a6989586621679547209

data (.@#@$) :: forall b6989586621679547013 c6989586621679547014 a6989586621679547015. (~>) ((~>) b6989586621679547013 c6989586621679547014) ((~>) ((~>) a6989586621679547015 b6989586621679547013) ((~>) a6989586621679547015 c6989586621679547014)) infixr 9

data (.@#@$$) (a6989586621679547189 :: (~>) b6989586621679547013 c6989586621679547014) :: forall a6989586621679547015. (~>) ((~>) a6989586621679547015 b6989586621679547013) ((~>) a6989586621679547015 c6989586621679547014) infixr 9

data (a6989586621679547189 :: (~>) b6989586621679547013 c6989586621679547014) .@#@$$$ (a6989586621679547190 :: (~>) a6989586621679547015 b6989586621679547013) :: (~>) a6989586621679547015 c6989586621679547014 infixr 9

type (.@#@$$$$) (a6989586621679547189 :: (~>) b6989586621679547013 c6989586621679547014) (a6989586621679547190 :: (~>) a6989586621679547015 b6989586621679547013) (a6989586621679547191 :: a6989586621679547015) = (.) a6989586621679547189 a6989586621679547190 a6989586621679547191

data ($@#@$) :: forall a6989586621679547007 b6989586621679547008. (~>) ((~>) a6989586621679547007 b6989586621679547008) ((~>) a6989586621679547007 b6989586621679547008) infixr 0

data ($@#@$$) (a6989586621679547164 :: (~>) a6989586621679547007 b6989586621679547008) :: (~>) a6989586621679547007 b6989586621679547008 infixr 0

type ($@#@$$$) (a6989586621679547164 :: (~>) a6989586621679547007 b6989586621679547008) (a6989586621679547165 :: a6989586621679547007) = ($) a6989586621679547164 a6989586621679547165

data ($!@#@$) :: forall a6989586621679547005 b6989586621679547006. (~>) ((~>) a6989586621679547005 b6989586621679547006) ((~>) a6989586621679547005 b6989586621679547006) infixr 0

data ($!@#@$$) (a6989586621679547155 :: (~>) a6989586621679547005 b6989586621679547006) :: (~>) a6989586621679547005 b6989586621679547006 infixr 0

type ($!@#@$$$) (a6989586621679547155 :: (~>) a6989586621679547005 b6989586621679547006) (a6989586621679547156 :: a6989586621679547005) = ($!) a6989586621679547155 a6989586621679547156

data UntilSym0 :: forall a6989586621679547004. (~>) ((~>) a6989586621679547004 Bool) ((~>) ((~>) a6989586621679547004 a6989586621679547004) ((~>) a6989586621679547004 a6989586621679547004))

data UntilSym1 (a6989586621679547129 :: (~>) a6989586621679547004 Bool) :: (~>) ((~>) a6989586621679547004 a6989586621679547004) ((~>) a6989586621679547004 a6989586621679547004)

data UntilSym2 (a6989586621679547129 :: (~>) a6989586621679547004 Bool) (a6989586621679547130 :: (~>) a6989586621679547004 a6989586621679547004) :: (~>) a6989586621679547004 a6989586621679547004

type UntilSym3 (a6989586621679547129 :: (~>) a6989586621679547004 Bool) (a6989586621679547130 :: (~>) a6989586621679547004 a6989586621679547004) (a6989586621679547131 :: a6989586621679547004) = Until a6989586621679547129 a6989586621679547130 a6989586621679547131

data FlipSym0 :: forall a6989586621679547010 b6989586621679547011 c6989586621679547012. (~>) ((~>) a6989586621679547010 ((~>) b6989586621679547011 c6989586621679547012)) ((~>) b6989586621679547011 ((~>) a6989586621679547010 c6989586621679547012))

data FlipSym1 (a6989586621679547180 :: (~>) a6989586621679547010 ((~>) b6989586621679547011 c6989586621679547012)) :: (~>) b6989586621679547011 ((~>) a6989586621679547010 c6989586621679547012)

data FlipSym2 (a6989586621679547180 :: (~>) a6989586621679547010 ((~>) b6989586621679547011 c6989586621679547012)) (a6989586621679547181 :: b6989586621679547011) :: (~>) a6989586621679547010 c6989586621679547012

type FlipSym3 (a6989586621679547180 :: (~>) a6989586621679547010 ((~>) b6989586621679547011 c6989586621679547012)) (a6989586621679547181 :: b6989586621679547011) (a6989586621679547182 :: a6989586621679547010) = Flip a6989586621679547180 a6989586621679547181 a6989586621679547182

data AsTypeOfSym0 :: forall a6989586621679547009. (~>) a6989586621679547009 ((~>) a6989586621679547009 a6989586621679547009)

data AsTypeOfSym1 (a6989586621679547174 :: a6989586621679547009) :: (~>) a6989586621679547009 a6989586621679547009

type AsTypeOfSym2 (a6989586621679547174 :: a6989586621679547009) (a6989586621679547175 :: a6989586621679547009) = AsTypeOf a6989586621679547174 a6989586621679547175

data SeqSym0 :: forall a6989586621679547002 b6989586621679547003. (~>) a6989586621679547002 ((~>) b6989586621679547003 b6989586621679547003) infixr 0

data SeqSym1 (a6989586621679547124 :: a6989586621679547002) :: forall b6989586621679547003. (~>) b6989586621679547003 b6989586621679547003 infixr 0

type SeqSym2 (a6989586621679547124 :: a6989586621679547002) (a6989586621679547125 :: b6989586621679547003) = Seq a6989586621679547124 a6989586621679547125