Data.Singletons.Prelude.Base

Description

Implements singletonized versions of functions from GHC.Base module.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Tuple. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Basic functions

type family Foldr (a :: (~>) a ((~>) b b)) (a :: b) (a :: [a]) :: b where ...

Equations

Foldr k z a_6989586621679548280 = Apply (Let6989586621679548285GoSym3 k z a_6989586621679548280) a_6989586621679548280

sFoldr :: forall a b (t :: (~>) a ((~>) b b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldrSym0 t) t) t :: b)

type family Map (a :: (~>) a b) (a :: [a]) :: [b] where ...

Equations

Map _ '[] = '[]

Map f ((:) x xs) = Apply (Apply (:@#@$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

sMap :: forall a b (t :: (~>) a b) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: [b])

type family (a :: [a]) ++ (a :: [a]) :: [a] where ... infixr 5

Equations

'[] ++ ys = ys

((:) x xs) ++ ys = Apply (Apply (:@#@$) x) (Apply (Apply (++@#@$) xs) ys)

(%++) :: forall a (t :: [a]) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply (++@#@$) t) t :: [a]) infixr 5

type family Otherwise :: Bool where ...

Equations

Otherwise = TrueSym0

sOtherwise :: Sing (OtherwiseSym0 :: Bool)

type family Id (a :: a) :: a where ...

Equations

Id x = x

sId :: forall a (t :: a). Sing t -> Sing (Apply IdSym0 t :: a)

type family Const (a :: a) (a :: b) :: a where ...

Equations

Const x _ = x

sConst :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply ConstSym0 t) t :: a)

type family ((a :: (~>) b c) :. (a :: (~>) a b)) (a :: a) :: c where ... infixr 9

Equations

(f :. g) a_6989586621679548228 = Apply (Apply (Apply (Apply Lambda_6989586621679548233Sym0 f) g) a_6989586621679548228) a_6989586621679548228

(%.) :: forall b c a (t :: (~>) b c) (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (.@#@$) t) t) t :: c) infixr 9

type family (a :: (~>) a b) $ (a :: a) :: b where ... infixr 0

Equations

f $ x = Apply f x

type family (a :: (~>) a b) $! (a :: a) :: b where ... infixr 0

Equations

f $! x = Apply f (Let6989586621679548204VxSym2 f x)

(%$) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($@#@$) t) t :: b) infixr 0

(%$!) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($!@#@$) t) t :: b) infixr 0

type family Until (a :: (~>) a Bool) (a :: (~>) a a) (a :: a) :: a where ...

Equations

Until p f a_6989586621679548178 = Apply (Let6989586621679548183GoSym3 p f a_6989586621679548178) a_6989586621679548178

sUntil :: forall a (t :: (~>) a Bool) (t :: (~>) a a) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply UntilSym0 t) t) t :: a)

type family Flip (a :: (~>) a ((~>) b c)) (a :: b) (a :: a) :: c where ...

Equations

Flip f x y = Apply (Apply f y) x

sFlip :: forall a b c (t :: (~>) a ((~>) b c)) (t :: b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FlipSym0 t) t) t :: c)

type family AsTypeOf (a :: a) (a :: a) :: a where ...

Equations

AsTypeOf a_6989586621679548246 a_6989586621679548248 = Apply (Apply ConstSym0 a_6989586621679548246) a_6989586621679548248

sAsTypeOf :: forall a (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply AsTypeOfSym0 t) t :: a)

type family Seq (a :: a) (a :: b) :: b where ... infixr 0

Equations

Seq _ x = x

sSeq :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply SeqSym0 t) t :: b) infixr 0

Defunctionalization symbols

data FoldrSym0 :: forall a6989586621679548065 b6989586621679548066. (~>) ((~>) a6989586621679548065 ((~>) b6989586621679548066 b6989586621679548066)) ((~>) b6989586621679548066 ((~>) [a6989586621679548065] b6989586621679548066))

data FoldrSym1 (a6989586621679548274 :: (~>) a6989586621679548065 ((~>) b6989586621679548066 b6989586621679548066)) :: (~>) b6989586621679548066 ((~>) [a6989586621679548065] b6989586621679548066)

data FoldrSym2 (a6989586621679548274 :: (~>) a6989586621679548065 ((~>) b6989586621679548066 b6989586621679548066)) (a6989586621679548275 :: b6989586621679548066) :: (~>) [a6989586621679548065] b6989586621679548066

type FoldrSym3 (a6989586621679548274 :: (~>) a6989586621679548065 ((~>) b6989586621679548066 b6989586621679548066)) (a6989586621679548275 :: b6989586621679548066) (a6989586621679548276 :: [a6989586621679548065]) = Foldr a6989586621679548274 a6989586621679548275 a6989586621679548276

data MapSym0 :: forall a6989586621679548063 b6989586621679548064. (~>) ((~>) a6989586621679548063 b6989586621679548064) ((~>) [a6989586621679548063] [b6989586621679548064])

data MapSym1 (a6989586621679548267 :: (~>) a6989586621679548063 b6989586621679548064) :: (~>) [a6989586621679548063] [b6989586621679548064]

type MapSym2 (a6989586621679548267 :: (~>) a6989586621679548063 b6989586621679548064) (a6989586621679548268 :: [a6989586621679548063]) = Map a6989586621679548267 a6989586621679548268

data (++@#@$) :: forall a6989586621679548062. (~>) [a6989586621679548062] ((~>) [a6989586621679548062] [a6989586621679548062]) infixr 5

data (++@#@$$) (a6989586621679548259 :: [a6989586621679548062]) :: (~>) [a6989586621679548062] [a6989586621679548062] infixr 5

type (++@#@$$$) (a6989586621679548259 :: [a6989586621679548062]) (a6989586621679548260 :: [a6989586621679548062]) = (++) a6989586621679548259 a6989586621679548260

type OtherwiseSym0 = Otherwise

data IdSym0 :: forall a6989586621679548061. (~>) a6989586621679548061 a6989586621679548061

type IdSym1 (a6989586621679548256 :: a6989586621679548061) = Id a6989586621679548256

data ConstSym0 :: forall a6989586621679548059 b6989586621679548060. (~>) a6989586621679548059 ((~>) b6989586621679548060 a6989586621679548059)

data ConstSym1 (a6989586621679548241 :: a6989586621679548059) :: forall b6989586621679548060. (~>) b6989586621679548060 a6989586621679548059

type ConstSym2 (a6989586621679548241 :: a6989586621679548059) (a6989586621679548242 :: b6989586621679548060) = Const a6989586621679548241 a6989586621679548242

data (.@#@$) :: forall a6989586621679548058 b6989586621679548056 c6989586621679548057. (~>) ((~>) b6989586621679548056 c6989586621679548057) ((~>) ((~>) a6989586621679548058 b6989586621679548056) ((~>) a6989586621679548058 c6989586621679548057)) infixr 9

data (.@#@$$) (a6989586621679548222 :: (~>) b6989586621679548056 c6989586621679548057) :: forall a6989586621679548058. (~>) ((~>) a6989586621679548058 b6989586621679548056) ((~>) a6989586621679548058 c6989586621679548057) infixr 9

data (a6989586621679548222 :: (~>) b6989586621679548056 c6989586621679548057) .@#@$$$ (a6989586621679548223 :: (~>) a6989586621679548058 b6989586621679548056) :: (~>) a6989586621679548058 c6989586621679548057 infixr 9

type (.@#@$$$$) (a6989586621679548222 :: (~>) b6989586621679548056 c6989586621679548057) (a6989586621679548223 :: (~>) a6989586621679548058 b6989586621679548056) (a6989586621679548224 :: a6989586621679548058) = (:.) a6989586621679548222 a6989586621679548223 a6989586621679548224

data ($@#@$) :: forall a6989586621679548050 b6989586621679548051. (~>) ((~>) a6989586621679548050 b6989586621679548051) ((~>) a6989586621679548050 b6989586621679548051) infixr 0

data ($@#@$$) (a6989586621679548207 :: (~>) a6989586621679548050 b6989586621679548051) :: (~>) a6989586621679548050 b6989586621679548051 infixr 0

type ($@#@$$$) (a6989586621679548207 :: (~>) a6989586621679548050 b6989586621679548051) (a6989586621679548208 :: a6989586621679548050) = ($) a6989586621679548207 a6989586621679548208

data ($!@#@$) :: forall a6989586621679548048 b6989586621679548049. (~>) ((~>) a6989586621679548048 b6989586621679548049) ((~>) a6989586621679548048 b6989586621679548049) infixr 0

data ($!@#@$$) (a6989586621679548198 :: (~>) a6989586621679548048 b6989586621679548049) :: (~>) a6989586621679548048 b6989586621679548049 infixr 0

type ($!@#@$$$) (a6989586621679548198 :: (~>) a6989586621679548048 b6989586621679548049) (a6989586621679548199 :: a6989586621679548048) = ($!) a6989586621679548198 a6989586621679548199

data UntilSym0 :: forall a6989586621679548047. (~>) ((~>) a6989586621679548047 Bool) ((~>) ((~>) a6989586621679548047 a6989586621679548047) ((~>) a6989586621679548047 a6989586621679548047))

data UntilSym1 (a6989586621679548172 :: (~>) a6989586621679548047 Bool) :: (~>) ((~>) a6989586621679548047 a6989586621679548047) ((~>) a6989586621679548047 a6989586621679548047)

data UntilSym2 (a6989586621679548172 :: (~>) a6989586621679548047 Bool) (a6989586621679548173 :: (~>) a6989586621679548047 a6989586621679548047) :: (~>) a6989586621679548047 a6989586621679548047

type UntilSym3 (a6989586621679548172 :: (~>) a6989586621679548047 Bool) (a6989586621679548173 :: (~>) a6989586621679548047 a6989586621679548047) (a6989586621679548174 :: a6989586621679548047) = Until a6989586621679548172 a6989586621679548173 a6989586621679548174

data FlipSym0 :: forall a6989586621679548053 b6989586621679548054 c6989586621679548055. (~>) ((~>) a6989586621679548053 ((~>) b6989586621679548054 c6989586621679548055)) ((~>) b6989586621679548054 ((~>) a6989586621679548053 c6989586621679548055))

data FlipSym1 (a6989586621679548213 :: (~>) a6989586621679548053 ((~>) b6989586621679548054 c6989586621679548055)) :: (~>) b6989586621679548054 ((~>) a6989586621679548053 c6989586621679548055)

data FlipSym2 (a6989586621679548213 :: (~>) a6989586621679548053 ((~>) b6989586621679548054 c6989586621679548055)) (a6989586621679548214 :: b6989586621679548054) :: (~>) a6989586621679548053 c6989586621679548055

type FlipSym3 (a6989586621679548213 :: (~>) a6989586621679548053 ((~>) b6989586621679548054 c6989586621679548055)) (a6989586621679548214 :: b6989586621679548054) (a6989586621679548215 :: a6989586621679548053) = Flip a6989586621679548213 a6989586621679548214 a6989586621679548215

data AsTypeOfSym0 :: forall a6989586621679548052. (~>) a6989586621679548052 ((~>) a6989586621679548052 a6989586621679548052)

data AsTypeOfSym1 (a6989586621679548250 :: a6989586621679548052) :: (~>) a6989586621679548052 a6989586621679548052

type AsTypeOfSym2 (a6989586621679548250 :: a6989586621679548052) (a6989586621679548251 :: a6989586621679548052) = AsTypeOf a6989586621679548250 a6989586621679548251

data SeqSym0 :: forall a6989586621679548045 b6989586621679548046. (~>) a6989586621679548045 ((~>) b6989586621679548046 b6989586621679548046) infixr 0

data SeqSym1 (a6989586621679548167 :: a6989586621679548045) :: forall b6989586621679548046. (~>) b6989586621679548046 b6989586621679548046 infixr 0

type SeqSym2 (a6989586621679548167 :: a6989586621679548045) (a6989586621679548168 :: b6989586621679548046) = Seq a6989586621679548167 a6989586621679548168