Data.Singletons.Prelude.Base

Description

Implements singletonized versions of functions from GHC.Base module.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Tuple. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Basic functions

type family Foldr (a :: (~>) a ((~>) b b)) (a :: b) (a :: [a]) :: b where ...

Equations

Foldr k z a_6989586621679548224 = Apply (Let6989586621679548229GoSym3 k z a_6989586621679548224) a_6989586621679548224

sFoldr :: forall a b (t :: (~>) a ((~>) b b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldrSym0 t) t) t :: b)

type family Map (a :: (~>) a b) (a :: [a]) :: [b] where ...

Equations

Map _ '[] = '[]

Map f ((:) x xs) = Apply (Apply (:@#@$) (Apply f x)) (Apply (Apply MapSym0 f) xs)

sMap :: forall a b (t :: (~>) a b) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: [b])

type family (a :: [a]) ++ (a :: [a]) :: [a] where ... infixr 5

Equations

'[] ++ ys = ys

((:) x xs) ++ ys = Apply (Apply (:@#@$) x) (Apply (Apply (++@#@$) xs) ys)

(%++) :: forall a (t :: [a]) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply (++@#@$) t) t :: [a]) infixr 5

type family Otherwise :: Bool where ...

Equations

Otherwise = TrueSym0

sOtherwise :: Sing (OtherwiseSym0 :: Bool)

type family Id (a :: a) :: a where ...

Equations

Id x = x

sId :: forall a (t :: a). Sing t -> Sing (Apply IdSym0 t :: a)

type family Const (a :: a) (a :: b) :: a where ...

Equations

Const x _ = x

sConst :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply ConstSym0 t) t :: a)

type family ((a :: (~>) b c) :. (a :: (~>) a b)) (a :: a) :: c where ... infixr 9

Equations

(f :. g) a_6989586621679548172 = Apply (Apply (Apply (Apply Lambda_6989586621679548177Sym0 f) g) a_6989586621679548172) a_6989586621679548172

(%.) :: forall b c a (t :: (~>) b c) (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (.@#@$) t) t) t :: c) infixr 9

type family (a :: (~>) a b) $ (a :: a) :: b where ... infixr 0

Equations

f $ x = Apply f x

type family (a :: (~>) a b) $! (a :: a) :: b where ... infixr 0

Equations

f $! x = Apply f (Let6989586621679548148VxSym2 f x)

(%$) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($@#@$) t) t :: b) infixr 0

(%$!) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($!@#@$) t) t :: b) infixr 0

type family Until (a :: (~>) a Bool) (a :: (~>) a a) (a :: a) :: a where ...

Equations

Until p f a_6989586621679548122 = Apply (Let6989586621679548127GoSym3 p f a_6989586621679548122) a_6989586621679548122

sUntil :: forall a (t :: (~>) a Bool) (t :: (~>) a a) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply UntilSym0 t) t) t :: a)

type family Flip (a :: (~>) a ((~>) b c)) (a :: b) (a :: a) :: c where ...

Equations

Flip f x y = Apply (Apply f y) x

sFlip :: forall a b c (t :: (~>) a ((~>) b c)) (t :: b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FlipSym0 t) t) t :: c)

type family AsTypeOf (a :: a) (a :: a) :: a where ...

Equations

AsTypeOf a_6989586621679548190 a_6989586621679548192 = Apply (Apply ConstSym0 a_6989586621679548190) a_6989586621679548192

sAsTypeOf :: forall a (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply AsTypeOfSym0 t) t :: a)

type family Seq (a :: a) (a :: b) :: b where ... infixr 0

Equations

Seq _ x = x

sSeq :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply SeqSym0 t) t :: b) infixr 0

Defunctionalization symbols

data FoldrSym0 :: forall a6989586621679548009 b6989586621679548010. (~>) ((~>) a6989586621679548009 ((~>) b6989586621679548010 b6989586621679548010)) ((~>) b6989586621679548010 ((~>) [a6989586621679548009] b6989586621679548010))

data FoldrSym1 (a6989586621679548218 :: (~>) a6989586621679548009 ((~>) b6989586621679548010 b6989586621679548010)) :: (~>) b6989586621679548010 ((~>) [a6989586621679548009] b6989586621679548010)

data FoldrSym2 (a6989586621679548218 :: (~>) a6989586621679548009 ((~>) b6989586621679548010 b6989586621679548010)) (a6989586621679548219 :: b6989586621679548010) :: (~>) [a6989586621679548009] b6989586621679548010

type FoldrSym3 (a6989586621679548218 :: (~>) a6989586621679548009 ((~>) b6989586621679548010 b6989586621679548010)) (a6989586621679548219 :: b6989586621679548010) (a6989586621679548220 :: [a6989586621679548009]) = Foldr a6989586621679548218 a6989586621679548219 a6989586621679548220

data MapSym0 :: forall a6989586621679548007 b6989586621679548008. (~>) ((~>) a6989586621679548007 b6989586621679548008) ((~>) [a6989586621679548007] [b6989586621679548008])

data MapSym1 (a6989586621679548211 :: (~>) a6989586621679548007 b6989586621679548008) :: (~>) [a6989586621679548007] [b6989586621679548008]

type MapSym2 (a6989586621679548211 :: (~>) a6989586621679548007 b6989586621679548008) (a6989586621679548212 :: [a6989586621679548007]) = Map a6989586621679548211 a6989586621679548212

data (++@#@$) :: forall a6989586621679548006. (~>) [a6989586621679548006] ((~>) [a6989586621679548006] [a6989586621679548006]) infixr 5

data (++@#@$$) (a6989586621679548203 :: [a6989586621679548006]) :: (~>) [a6989586621679548006] [a6989586621679548006] infixr 5

type (++@#@$$$) (a6989586621679548203 :: [a6989586621679548006]) (a6989586621679548204 :: [a6989586621679548006]) = (++) a6989586621679548203 a6989586621679548204

type OtherwiseSym0 = Otherwise

data IdSym0 :: forall a6989586621679548005. (~>) a6989586621679548005 a6989586621679548005

type IdSym1 (a6989586621679548200 :: a6989586621679548005) = Id a6989586621679548200

data ConstSym0 :: forall a6989586621679548003 b6989586621679548004. (~>) a6989586621679548003 ((~>) b6989586621679548004 a6989586621679548003)

data ConstSym1 (a6989586621679548185 :: a6989586621679548003) :: forall b6989586621679548004. (~>) b6989586621679548004 a6989586621679548003

type ConstSym2 (a6989586621679548185 :: a6989586621679548003) (a6989586621679548186 :: b6989586621679548004) = Const a6989586621679548185 a6989586621679548186

data (.@#@$) :: forall a6989586621679548002 b6989586621679548000 c6989586621679548001. (~>) ((~>) b6989586621679548000 c6989586621679548001) ((~>) ((~>) a6989586621679548002 b6989586621679548000) ((~>) a6989586621679548002 c6989586621679548001)) infixr 9

data (.@#@$$) (a6989586621679548166 :: (~>) b6989586621679548000 c6989586621679548001) :: forall a6989586621679548002. (~>) ((~>) a6989586621679548002 b6989586621679548000) ((~>) a6989586621679548002 c6989586621679548001) infixr 9

data (a6989586621679548166 :: (~>) b6989586621679548000 c6989586621679548001) .@#@$$$ (a6989586621679548167 :: (~>) a6989586621679548002 b6989586621679548000) :: (~>) a6989586621679548002 c6989586621679548001 infixr 9

type (.@#@$$$$) (a6989586621679548166 :: (~>) b6989586621679548000 c6989586621679548001) (a6989586621679548167 :: (~>) a6989586621679548002 b6989586621679548000) (a6989586621679548168 :: a6989586621679548002) = (:.) a6989586621679548166 a6989586621679548167 a6989586621679548168

data ($@#@$) :: forall a6989586621679547994 b6989586621679547995. (~>) ((~>) a6989586621679547994 b6989586621679547995) ((~>) a6989586621679547994 b6989586621679547995) infixr 0

data ($@#@$$) (a6989586621679548151 :: (~>) a6989586621679547994 b6989586621679547995) :: (~>) a6989586621679547994 b6989586621679547995 infixr 0

type ($@#@$$$) (a6989586621679548151 :: (~>) a6989586621679547994 b6989586621679547995) (a6989586621679548152 :: a6989586621679547994) = ($) a6989586621679548151 a6989586621679548152

data ($!@#@$) :: forall a6989586621679547992 b6989586621679547993. (~>) ((~>) a6989586621679547992 b6989586621679547993) ((~>) a6989586621679547992 b6989586621679547993) infixr 0

data ($!@#@$$) (a6989586621679548142 :: (~>) a6989586621679547992 b6989586621679547993) :: (~>) a6989586621679547992 b6989586621679547993 infixr 0

type ($!@#@$$$) (a6989586621679548142 :: (~>) a6989586621679547992 b6989586621679547993) (a6989586621679548143 :: a6989586621679547992) = ($!) a6989586621679548142 a6989586621679548143

data UntilSym0 :: forall a6989586621679547991. (~>) ((~>) a6989586621679547991 Bool) ((~>) ((~>) a6989586621679547991 a6989586621679547991) ((~>) a6989586621679547991 a6989586621679547991))

data UntilSym1 (a6989586621679548116 :: (~>) a6989586621679547991 Bool) :: (~>) ((~>) a6989586621679547991 a6989586621679547991) ((~>) a6989586621679547991 a6989586621679547991)

data UntilSym2 (a6989586621679548116 :: (~>) a6989586621679547991 Bool) (a6989586621679548117 :: (~>) a6989586621679547991 a6989586621679547991) :: (~>) a6989586621679547991 a6989586621679547991

type UntilSym3 (a6989586621679548116 :: (~>) a6989586621679547991 Bool) (a6989586621679548117 :: (~>) a6989586621679547991 a6989586621679547991) (a6989586621679548118 :: a6989586621679547991) = Until a6989586621679548116 a6989586621679548117 a6989586621679548118

data FlipSym0 :: forall a6989586621679547997 b6989586621679547998 c6989586621679547999. (~>) ((~>) a6989586621679547997 ((~>) b6989586621679547998 c6989586621679547999)) ((~>) b6989586621679547998 ((~>) a6989586621679547997 c6989586621679547999))

data FlipSym1 (a6989586621679548157 :: (~>) a6989586621679547997 ((~>) b6989586621679547998 c6989586621679547999)) :: (~>) b6989586621679547998 ((~>) a6989586621679547997 c6989586621679547999)

data FlipSym2 (a6989586621679548157 :: (~>) a6989586621679547997 ((~>) b6989586621679547998 c6989586621679547999)) (a6989586621679548158 :: b6989586621679547998) :: (~>) a6989586621679547997 c6989586621679547999

type FlipSym3 (a6989586621679548157 :: (~>) a6989586621679547997 ((~>) b6989586621679547998 c6989586621679547999)) (a6989586621679548158 :: b6989586621679547998) (a6989586621679548159 :: a6989586621679547997) = Flip a6989586621679548157 a6989586621679548158 a6989586621679548159

data AsTypeOfSym0 :: forall a6989586621679547996. (~>) a6989586621679547996 ((~>) a6989586621679547996 a6989586621679547996)

data AsTypeOfSym1 (a6989586621679548194 :: a6989586621679547996) :: (~>) a6989586621679547996 a6989586621679547996

type AsTypeOfSym2 (a6989586621679548194 :: a6989586621679547996) (a6989586621679548195 :: a6989586621679547996) = AsTypeOf a6989586621679548194 a6989586621679548195

data SeqSym0 :: forall a6989586621679547989 b6989586621679547990. (~>) a6989586621679547989 ((~>) b6989586621679547990 b6989586621679547990) infixr 0

data SeqSym1 (a6989586621679548111 :: a6989586621679547989) :: forall b6989586621679547990. (~>) b6989586621679547990 b6989586621679547990 infixr 0

type SeqSym2 (a6989586621679548111 :: a6989586621679547989) (a6989586621679548112 :: b6989586621679547990) = Seq a6989586621679548111 a6989586621679548112