Data.Singletons.Prelude.Bool

Description

Defines functions and datatypes relating to the singleton for Bool, including a singletons version of all the definitions in Data.Bool.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Bool. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The Bool singleton

data family Sing (a :: k)

The singleton kind-indexed data family.

Though Haddock doesn't show it, the Sing instance above declares constructors

SFalse :: Sing False
STrue  :: Sing True

type SBool = (Sing :: Bool -> Type)

SBool is a kind-restricted synonym for Sing: type SBool (a :: Bool) = Sing a

Conditionals

type family If k (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If k True tru fls = tru

If k False tru fls = fls

sIf :: Sing a -> Sing b -> Sing c -> Sing (If a b c)

Conditional over singletons

Singletons from Data.Bool

type family Not (a :: Bool) :: Bool where ...

Equations

Not False = TrueSym0

Not True = FalseSym0

sNot :: forall t. Sing t -> Sing (Apply NotSym0 t :: Bool)

type family (a :: Bool) :&& (a :: Bool) :: Bool where ... infixr 3

Equations

False :&& _z_6989586621679545874 = FalseSym0

True :&& x = x

type family (a :: Bool) :|| (a :: Bool) :: Bool where ... infixr 2

Equations

False :|| x = x

True :|| _z_6989586621679545862 = TrueSym0

(%:&&) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:&&$) t) t :: Bool) infixr 3

(%:||) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:||$) t) t :: Bool) infixr 2

The following are derived from the function bool in Data.Bool. The extra underscore is to avoid name clashes with the type Bool.

bool_ :: forall a. a -> a -> Bool -> a

type family Bool_ (a :: a) (a :: a) (a :: Bool) :: a where ...

Equations

Bool_ fls _tru False = fls

Bool_ _fls tru True = tru

sBool_ :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Bool_Sym0 t) t) t :: a)

type family Otherwise :: Bool where ...

Equations

Otherwise = TrueSym0

sOtherwise :: Sing (OtherwiseSym0 :: Bool)

Defunctionalization symbols

type TrueSym0 = True

type FalseSym0 = False

data NotSym0 l

type NotSym1 t = Not t

data (:&&$) l

data l :&&$$ l

type (:&&$$$) t t = (:&&) t t

data (:||$) l

data l :||$$ l

type (:||$$$) t t = (:||) t t

data Bool_Sym0 l

data Bool_Sym1 l l

data Bool_Sym2 l l l

type Bool_Sym3 t t t = Bool_ t t t

type OtherwiseSym0 = Otherwise