Data.Singletons.Prelude.Bool

Description

Defines functions and datatypes relating to the singleton for Bool, including a singletons version of all the definitions in Data.Bool.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Bool. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The Bool singleton

type family Sing :: k -> Type

The singleton kind-indexed type family.

data SBool :: Bool -> Type where

Constructors

SFalse :: SBool 'False

STrue :: SBool 'True

Conditionals

type family If (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If 'True (tru :: k) (fls :: k) = tru

If 'False (tru :: k) (fls :: k) = fls

sIf :: Sing a -> Sing b -> Sing c -> Sing (If a b c)

Conditional over singletons

Singletons from Data.Bool

type family Not (a :: Bool) = (res :: Bool) | res -> a where ...

Type-level "not". An injective type family since 4.10.0.0.

Since: base-4.7.0.0

Equations

Not 'False = 'True

Not 'True = 'False

sNot :: Sing a -> Sing (Not a)

Negation of a singleton

type family (a :: Bool) && (b :: Bool) :: Bool where ... infixr 3

Type-level "and"

Equations

'False && a = 'False

'True && a = a

a && 'False = 'False

a && 'True = a

a && a = a

type family (a :: Bool) || (b :: Bool) :: Bool where ... infixr 2

Type-level "or"

Equations

'False || a = a

'True || a = 'True

a || 'False = a

a || 'True = 'True

a || a = a

(%&&) :: Sing a -> Sing b -> Sing (a && b) infixr 3

Conjunction of singletons

(%||) :: Sing a -> Sing b -> Sing (a || b) infixr 2

Disjunction of singletons

The following are derived from the function bool in Data.Bool. The extra underscore is to avoid name clashes with the type Bool.

bool_ :: a -> a -> Bool -> a

type family Bool_ (a :: a) (a :: a) (a :: Bool) :: a where ...

Equations

Bool_ fls _tru 'False = fls

Bool_ _fls tru 'True = tru

sBool_ :: forall a (t :: a) (t :: a) (t :: Bool). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Bool_Sym0 t) t) t :: a)

type family Otherwise :: Bool where ...

Equations

Otherwise = TrueSym0

sOtherwise :: Sing (OtherwiseSym0 :: Bool)

Defunctionalization symbols

type TrueSym0 = 'True

type FalseSym0 = 'False

data NotSym0 :: (~>) Bool Bool

type NotSym1 (a6989586621679378735 :: Bool) = Not a6989586621679378735

data (&&@#@$) :: (~>) Bool ((~>) Bool Bool) infixr 3

data (&&@#@$$) (a6989586621679378203 :: Bool) :: (~>) Bool Bool infixr 3

type (&&@#@$$$) (a6989586621679378203 :: Bool) (b6989586621679378204 :: Bool) = (&&) a6989586621679378203 b6989586621679378204

data (||@#@$) :: (~>) Bool ((~>) Bool Bool) infixr 2

data (||@#@$$) (a6989586621679378441 :: Bool) :: (~>) Bool Bool infixr 2

type (||@#@$$$) (a6989586621679378441 :: Bool) (b6989586621679378442 :: Bool) = (||) a6989586621679378441 b6989586621679378442

data Bool_Sym0 :: forall a6989586621679377441. (~>) a6989586621679377441 ((~>) a6989586621679377441 ((~>) Bool a6989586621679377441))

data Bool_Sym1 (a6989586621679377447 :: a6989586621679377441) :: (~>) a6989586621679377441 ((~>) Bool a6989586621679377441)

data Bool_Sym2 (a6989586621679377447 :: a6989586621679377441) (a6989586621679377448 :: a6989586621679377441) :: (~>) Bool a6989586621679377441

type Bool_Sym3 (a6989586621679377447 :: a6989586621679377441) (a6989586621679377448 :: a6989586621679377441) (a6989586621679377449 :: Bool) = Bool_ a6989586621679377447 a6989586621679377448 a6989586621679377449

type OtherwiseSym0 = Otherwise