Data.Singletons.Prelude.Either

Description

Defines functions and datatypes relating to the singleton for Either, including a singletons version of all the definitions in Data.Either.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Either. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The Either singleton

type family Sing :: k -> Type

The singleton kind-indexed type family.

data SEither :: forall a b. Either a b -> Type where

Constructors

SLeft :: forall a (n :: a). (Sing (n :: a)) -> SEither ('Left n)

SRight :: forall b (n :: b). (Sing (n :: b)) -> SEither ('Right n)

Singletons from Data.Either

either_ :: (a -> c) -> (b -> c) -> Either a b -> c

type family Either_ (a :: (~>) a c) (a :: (~>) b c) (a :: Either a b) :: c where ...

Equations

Either_ f _ ('Left x) = Apply f x

Either_ _ g ('Right y) = Apply g y

sEither_ :: forall a c b (t :: (~>) a c) (t :: (~>) b c) (t :: Either a b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Either_Sym0 t) t) t :: c)

The preceding two definitions are derived from the function either in Data.Either. The extra underscore is to avoid name clashes with the type Either.

type family Lefts (a :: [Either a b]) :: [a] where ...

Equations

Lefts '[] = '[]

Lefts ('(:) ('Left x) xs) = Apply (Apply (:@#@$) x) (Apply LeftsSym0 xs)

Lefts ('(:) ('Right _) xs) = Apply LeftsSym0 xs

sLefts :: forall a b (t :: [Either a b]). Sing t -> Sing (Apply LeftsSym0 t :: [a])

type family Rights (a :: [Either a b]) :: [b] where ...

Equations

Rights '[] = '[]

Rights ('(:) ('Left _) xs) = Apply RightsSym0 xs

Rights ('(:) ('Right x) xs) = Apply (Apply (:@#@$) x) (Apply RightsSym0 xs)

sRights :: forall a b (t :: [Either a b]). Sing t -> Sing (Apply RightsSym0 t :: [b])

type family PartitionEithers (a :: [Either a b]) :: ([a], [b]) where ...

Equations

PartitionEithers a_6989586621680473348 = Apply (Apply (Apply FoldrSym0 (Apply (Apply Either_Sym0 (Let6989586621680473353LeftSym1 a_6989586621680473348)) (Let6989586621680473353RightSym1 a_6989586621680473348))) (Apply (Apply Tuple2Sym0 '[]) '[])) a_6989586621680473348

sPartitionEithers :: forall a b (t :: [Either a b]). Sing t -> Sing (Apply PartitionEithersSym0 t :: ([a], [b]))

type family IsLeft (a :: Either a b) :: Bool where ...

Equations

IsLeft ('Left _) = TrueSym0

IsLeft ('Right _) = FalseSym0

sIsLeft :: forall a b (t :: Either a b). Sing t -> Sing (Apply IsLeftSym0 t :: Bool)

type family IsRight (a :: Either a b) :: Bool where ...

Equations

IsRight ('Left _) = FalseSym0

IsRight ('Right _) = TrueSym0

sIsRight :: forall a b (t :: Either a b). Sing t -> Sing (Apply IsRightSym0 t :: Bool)

Defunctionalization symbols

data LeftSym0 :: forall (a6989586621679095180 :: Type) (b6989586621679095181 :: Type). (~>) a6989586621679095180 (Either (a6989586621679095180 :: Type) (b6989586621679095181 :: Type))

type LeftSym1 (t6989586621679316758 :: a6989586621679095180) = 'Left t6989586621679316758

data RightSym0 :: forall (b6989586621679095181 :: Type) (a6989586621679095180 :: Type). (~>) b6989586621679095181 (Either (a6989586621679095180 :: Type) (b6989586621679095181 :: Type))

type RightSym1 (t6989586621679316760 :: b6989586621679095181) = 'Right t6989586621679316760

data Either_Sym0 :: forall a6989586621680471621 c6989586621680471622 b6989586621680471623. (~>) ((~>) a6989586621680471621 c6989586621680471622) ((~>) ((~>) b6989586621680471623 c6989586621680471622) ((~>) (Either a6989586621680471621 b6989586621680471623) c6989586621680471622))

data Either_Sym1 (a6989586621680471657 :: (~>) a6989586621680471621 c6989586621680471622) :: forall b6989586621680471623. (~>) ((~>) b6989586621680471623 c6989586621680471622) ((~>) (Either a6989586621680471621 b6989586621680471623) c6989586621680471622)

data Either_Sym2 (a6989586621680471657 :: (~>) a6989586621680471621 c6989586621680471622) (a6989586621680471658 :: (~>) b6989586621680471623 c6989586621680471622) :: (~>) (Either a6989586621680471621 b6989586621680471623) c6989586621680471622

type Either_Sym3 (a6989586621680471657 :: (~>) a6989586621680471621 c6989586621680471622) (a6989586621680471658 :: (~>) b6989586621680471623 c6989586621680471622) (a6989586621680471659 :: Either a6989586621680471621 b6989586621680471623) = Either_ a6989586621680471657 a6989586621680471658 a6989586621680471659

data LeftsSym0 :: forall a6989586621680473103 b6989586621680473104. (~>) [Either a6989586621680473103 b6989586621680473104] [a6989586621680473103]

type LeftsSym1 (a6989586621680473375 :: [Either a6989586621680473103 b6989586621680473104]) = Lefts a6989586621680473375

data RightsSym0 :: forall a6989586621680473101 b6989586621680473102. (~>) [Either a6989586621680473101 b6989586621680473102] [b6989586621680473102]

type RightsSym1 (a6989586621680473370 :: [Either a6989586621680473101 b6989586621680473102]) = Rights a6989586621680473370

data IsLeftSym0 :: forall a6989586621680473097 b6989586621680473098. (~>) (Either a6989586621680473097 b6989586621680473098) Bool

type IsLeftSym1 (a6989586621680473346 :: Either a6989586621680473097 b6989586621680473098) = IsLeft a6989586621680473346

data IsRightSym0 :: forall a6989586621680473095 b6989586621680473096. (~>) (Either a6989586621680473095 b6989586621680473096) Bool

type IsRightSym1 (a6989586621680473344 :: Either a6989586621680473095 b6989586621680473096) = IsRight a6989586621680473344