Data.Singletons.Prelude.Function

Description

Defines singleton versions of the definitions in Data.Function.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Function. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Prelude re-exports

type family Id (a :: a) :: a where ...

Equations

Id x = x

sId :: forall a (t :: a). Sing t -> Sing (Apply IdSym0 t :: a)

type family Const (a :: a) (a :: b) :: a where ...

Equations

Const x _ = x

sConst :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply ConstSym0 t) t :: a)

type family ((a :: (~>) b c) :. (a :: (~>) a b)) (a :: a) :: c where ... infixr 9

Equations

(f :. g) a_6989586621679391801 = Apply (Apply (Apply (Apply Lambda_6989586621679391806Sym0 f) g) a_6989586621679391801) a_6989586621679391801

(%.) :: forall b c a (t :: (~>) b c) (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (.@#@$) t) t) t :: c) infixr 9

type family Flip (a :: (~>) a ((~>) b c)) (a :: b) (a :: a) :: c where ...

Equations

Flip f x y = Apply (Apply f y) x

sFlip :: forall a b c (t :: (~>) a ((~>) b c)) (t :: b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FlipSym0 t) t) t :: c)

type family (a :: (~>) a b) $ (a :: a) :: b where ... infixr 0

Equations

f $ x = Apply f x

(%$) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($@#@$) t) t :: b) infixr 0

Other combinators

type family (a :: a) & (a :: (~>) a b) :: b where ... infixl 1

Equations

x & f = Apply f x

(%&) :: forall a b (t :: a) (t :: (~>) a b). Sing t -> Sing t -> Sing (Apply (Apply (&@#@$) t) t :: b) infixl 1

type family On (a :: (~>) b ((~>) b c)) (a :: (~>) a b) (a :: a) (a :: a) :: c where ... infixl 0

Equations

On ty f a_6989586621679560468 a_6989586621679560470 = Apply (Apply (Apply (Apply (Apply (Apply Lambda_6989586621679560476Sym0 ty) f) a_6989586621679560468) a_6989586621679560470) a_6989586621679560468) a_6989586621679560470

sOn :: forall b c a (t :: (~>) b ((~>) b c)) (t :: (~>) a b) (t :: a) (t :: a). Sing t -> Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (Apply OnSym0 t) t) t) t :: c) infixl 0

Defunctionalization symbols

data IdSym0 :: forall a6989586621679391634. (~>) a6989586621679391634 a6989586621679391634

type IdSym1 (a6989586621679391829 :: a6989586621679391634) = Id a6989586621679391829

data ConstSym0 :: forall a6989586621679391632 b6989586621679391633. (~>) a6989586621679391632 ((~>) b6989586621679391633 a6989586621679391632)

data ConstSym1 (a6989586621679391814 :: a6989586621679391632) :: forall b6989586621679391633. (~>) b6989586621679391633 a6989586621679391632

type ConstSym2 (a6989586621679391814 :: a6989586621679391632) (a6989586621679391815 :: b6989586621679391633) = Const a6989586621679391814 a6989586621679391815

data (.@#@$) :: forall a6989586621679391631 b6989586621679391629 c6989586621679391630. (~>) ((~>) b6989586621679391629 c6989586621679391630) ((~>) ((~>) a6989586621679391631 b6989586621679391629) ((~>) a6989586621679391631 c6989586621679391630)) infixr 9

data (.@#@$$) (a6989586621679391795 :: (~>) b6989586621679391629 c6989586621679391630) :: forall a6989586621679391631. (~>) ((~>) a6989586621679391631 b6989586621679391629) ((~>) a6989586621679391631 c6989586621679391630) infixr 9

data (a6989586621679391795 :: (~>) b6989586621679391629 c6989586621679391630) .@#@$$$ (a6989586621679391796 :: (~>) a6989586621679391631 b6989586621679391629) :: (~>) a6989586621679391631 c6989586621679391630 infixr 9

type (.@#@$$$$) (a6989586621679391795 :: (~>) b6989586621679391629 c6989586621679391630) (a6989586621679391796 :: (~>) a6989586621679391631 b6989586621679391629) (a6989586621679391797 :: a6989586621679391631) = (:.) a6989586621679391795 a6989586621679391796 a6989586621679391797

data FlipSym0 :: forall a6989586621679391626 b6989586621679391627 c6989586621679391628. (~>) ((~>) a6989586621679391626 ((~>) b6989586621679391627 c6989586621679391628)) ((~>) b6989586621679391627 ((~>) a6989586621679391626 c6989586621679391628))

data FlipSym1 (a6989586621679391786 :: (~>) a6989586621679391626 ((~>) b6989586621679391627 c6989586621679391628)) :: (~>) b6989586621679391627 ((~>) a6989586621679391626 c6989586621679391628)

data FlipSym2 (a6989586621679391786 :: (~>) a6989586621679391626 ((~>) b6989586621679391627 c6989586621679391628)) (a6989586621679391787 :: b6989586621679391627) :: (~>) a6989586621679391626 c6989586621679391628

type FlipSym3 (a6989586621679391786 :: (~>) a6989586621679391626 ((~>) b6989586621679391627 c6989586621679391628)) (a6989586621679391787 :: b6989586621679391627) (a6989586621679391788 :: a6989586621679391626) = Flip a6989586621679391786 a6989586621679391787 a6989586621679391788

data ($@#@$) :: forall a6989586621679391623 b6989586621679391624. (~>) ((~>) a6989586621679391623 b6989586621679391624) ((~>) a6989586621679391623 b6989586621679391624) infixr 0

data ($@#@$$) (a6989586621679391780 :: (~>) a6989586621679391623 b6989586621679391624) :: (~>) a6989586621679391623 b6989586621679391624 infixr 0

type ($@#@$$$) (a6989586621679391780 :: (~>) a6989586621679391623 b6989586621679391624) (a6989586621679391781 :: a6989586621679391623) = ($) a6989586621679391780 a6989586621679391781

data (&@#@$) :: forall a6989586621679560441 b6989586621679560442. (~>) a6989586621679560441 ((~>) ((~>) a6989586621679560441 b6989586621679560442) b6989586621679560442) infixl 1

data (&@#@$$) (a6989586621679560454 :: a6989586621679560441) :: forall b6989586621679560442. (~>) ((~>) a6989586621679560441 b6989586621679560442) b6989586621679560442 infixl 1

type (&@#@$$$) (a6989586621679560454 :: a6989586621679560441) (a6989586621679560455 :: (~>) a6989586621679560441 b6989586621679560442) = (&) a6989586621679560454 a6989586621679560455

data OnSym0 :: forall a6989586621679560445 b6989586621679560443 c6989586621679560444. (~>) ((~>) b6989586621679560443 ((~>) b6989586621679560443 c6989586621679560444)) ((~>) ((~>) a6989586621679560445 b6989586621679560443) ((~>) a6989586621679560445 ((~>) a6989586621679560445 c6989586621679560444))) infixl 0

data OnSym1 (a6989586621679560460 :: (~>) b6989586621679560443 ((~>) b6989586621679560443 c6989586621679560444)) :: forall a6989586621679560445. (~>) ((~>) a6989586621679560445 b6989586621679560443) ((~>) a6989586621679560445 ((~>) a6989586621679560445 c6989586621679560444)) infixl 0

data OnSym2 (a6989586621679560460 :: (~>) b6989586621679560443 ((~>) b6989586621679560443 c6989586621679560444)) (a6989586621679560461 :: (~>) a6989586621679560445 b6989586621679560443) :: (~>) a6989586621679560445 ((~>) a6989586621679560445 c6989586621679560444) infixl 0

data OnSym3 (a6989586621679560460 :: (~>) b6989586621679560443 ((~>) b6989586621679560443 c6989586621679560444)) (a6989586621679560461 :: (~>) a6989586621679560445 b6989586621679560443) (a6989586621679560462 :: a6989586621679560445) :: (~>) a6989586621679560445 c6989586621679560444 infixl 0

type OnSym4 (a6989586621679560460 :: (~>) b6989586621679560443 ((~>) b6989586621679560443 c6989586621679560444)) (a6989586621679560461 :: (~>) a6989586621679560445 b6989586621679560443) (a6989586621679560462 :: a6989586621679560445) (a6989586621679560463 :: a6989586621679560445) = On a6989586621679560460 a6989586621679560461 a6989586621679560462 a6989586621679560463