Data.Singletons.Prelude.Function

Description

Defines singleton versions of the definitions in Data.Function.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Function. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Prelude re-exports

type family Id (a :: a) :: a where ...

Equations

Id x = x

sId :: forall a (t :: a). Sing t -> Sing (Apply IdSym0 t :: a)

type family Const (a :: a) (a :: b) :: a where ...

Equations

Const x _ = x

sConst :: forall a b (t :: a) (t :: b). Sing t -> Sing t -> Sing (Apply (Apply ConstSym0 t) t :: a)

type family ((a :: (~>) b c) . (a :: (~>) a b)) (a :: a) :: c where ... infixr 9

Equations

(f . g) a_6989586621679547195 = Apply (Apply (Apply (Apply Lambda_6989586621679547200Sym0 f) g) a_6989586621679547195) a_6989586621679547195

(%.) :: forall b c a (t :: (~>) b c) (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (.@#@$) t) t) t :: c) infixr 9

type family Flip (a :: (~>) a ((~>) b c)) (a :: b) (a :: a) :: c where ...

Equations

Flip f x y = Apply (Apply f y) x

sFlip :: forall a b c (t :: (~>) a ((~>) b c)) (t :: b) (t :: a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FlipSym0 t) t) t :: c)

type family (a :: (~>) a b) $ (a :: a) :: b where ... infixr 0

Equations

f $ x = Apply f x

(%$) :: forall a b (t :: (~>) a b) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply ($@#@$) t) t :: b) infixr 0

Other combinators

type family (a :: a) & (a :: (~>) a b) :: b where ... infixl 1

Equations

x & f = Apply f x

(%&) :: forall a b (t :: a) (t :: (~>) a b). Sing t -> Sing t -> Sing (Apply (Apply (&@#@$) t) t :: b) infixl 1

type family On (a :: (~>) b ((~>) b c)) (a :: (~>) a b) (a :: a) (a :: a) :: c where ... infixl 0

Equations

On ty f a_6989586621679758168 a_6989586621679758170 = Apply (Apply (Apply (Apply (Apply (Apply Lambda_6989586621679758176Sym0 ty) f) a_6989586621679758168) a_6989586621679758170) a_6989586621679758168) a_6989586621679758170

sOn :: forall b c a (t :: (~>) b ((~>) b c)) (t :: (~>) a b) (t :: a) (t :: a). Sing t -> Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (Apply OnSym0 t) t) t) t :: c) infixl 0

Defunctionalization symbols

data IdSym0 :: forall a6989586621679547018. (~>) a6989586621679547018 a6989586621679547018

type IdSym1 (a6989586621679547213 :: a6989586621679547018) = Id a6989586621679547213

data ConstSym0 :: forall a6989586621679547016 b6989586621679547017. (~>) a6989586621679547016 ((~>) b6989586621679547017 a6989586621679547016)

data ConstSym1 (a6989586621679547208 :: a6989586621679547016) :: forall b6989586621679547017. (~>) b6989586621679547017 a6989586621679547016

type ConstSym2 (a6989586621679547208 :: a6989586621679547016) (a6989586621679547209 :: b6989586621679547017) = Const a6989586621679547208 a6989586621679547209

data (.@#@$) :: forall b6989586621679547013 c6989586621679547014 a6989586621679547015. (~>) ((~>) b6989586621679547013 c6989586621679547014) ((~>) ((~>) a6989586621679547015 b6989586621679547013) ((~>) a6989586621679547015 c6989586621679547014)) infixr 9

data (.@#@$$) (a6989586621679547189 :: (~>) b6989586621679547013 c6989586621679547014) :: forall a6989586621679547015. (~>) ((~>) a6989586621679547015 b6989586621679547013) ((~>) a6989586621679547015 c6989586621679547014) infixr 9

data (a6989586621679547189 :: (~>) b6989586621679547013 c6989586621679547014) .@#@$$$ (a6989586621679547190 :: (~>) a6989586621679547015 b6989586621679547013) :: (~>) a6989586621679547015 c6989586621679547014 infixr 9

type (.@#@$$$$) (a6989586621679547189 :: (~>) b6989586621679547013 c6989586621679547014) (a6989586621679547190 :: (~>) a6989586621679547015 b6989586621679547013) (a6989586621679547191 :: a6989586621679547015) = (.) a6989586621679547189 a6989586621679547190 a6989586621679547191

data FlipSym0 :: forall a6989586621679547010 b6989586621679547011 c6989586621679547012. (~>) ((~>) a6989586621679547010 ((~>) b6989586621679547011 c6989586621679547012)) ((~>) b6989586621679547011 ((~>) a6989586621679547010 c6989586621679547012))

data FlipSym1 (a6989586621679547180 :: (~>) a6989586621679547010 ((~>) b6989586621679547011 c6989586621679547012)) :: (~>) b6989586621679547011 ((~>) a6989586621679547010 c6989586621679547012)

data FlipSym2 (a6989586621679547180 :: (~>) a6989586621679547010 ((~>) b6989586621679547011 c6989586621679547012)) (a6989586621679547181 :: b6989586621679547011) :: (~>) a6989586621679547010 c6989586621679547012

type FlipSym3 (a6989586621679547180 :: (~>) a6989586621679547010 ((~>) b6989586621679547011 c6989586621679547012)) (a6989586621679547181 :: b6989586621679547011) (a6989586621679547182 :: a6989586621679547010) = Flip a6989586621679547180 a6989586621679547181 a6989586621679547182

data ($@#@$) :: forall a6989586621679547007 b6989586621679547008. (~>) ((~>) a6989586621679547007 b6989586621679547008) ((~>) a6989586621679547007 b6989586621679547008) infixr 0

data ($@#@$$) (a6989586621679547164 :: (~>) a6989586621679547007 b6989586621679547008) :: (~>) a6989586621679547007 b6989586621679547008 infixr 0

type ($@#@$$$) (a6989586621679547164 :: (~>) a6989586621679547007 b6989586621679547008) (a6989586621679547165 :: a6989586621679547007) = ($) a6989586621679547164 a6989586621679547165

data (&@#@$) :: forall a6989586621679758141 b6989586621679758142. (~>) a6989586621679758141 ((~>) ((~>) a6989586621679758141 b6989586621679758142) b6989586621679758142) infixl 1

data (&@#@$$) (a6989586621679758154 :: a6989586621679758141) :: forall b6989586621679758142. (~>) ((~>) a6989586621679758141 b6989586621679758142) b6989586621679758142 infixl 1

type (&@#@$$$) (a6989586621679758154 :: a6989586621679758141) (a6989586621679758155 :: (~>) a6989586621679758141 b6989586621679758142) = (&) a6989586621679758154 a6989586621679758155

data OnSym0 :: forall b6989586621679758143 c6989586621679758144 a6989586621679758145. (~>) ((~>) b6989586621679758143 ((~>) b6989586621679758143 c6989586621679758144)) ((~>) ((~>) a6989586621679758145 b6989586621679758143) ((~>) a6989586621679758145 ((~>) a6989586621679758145 c6989586621679758144))) infixl 0

data OnSym1 (a6989586621679758160 :: (~>) b6989586621679758143 ((~>) b6989586621679758143 c6989586621679758144)) :: forall a6989586621679758145. (~>) ((~>) a6989586621679758145 b6989586621679758143) ((~>) a6989586621679758145 ((~>) a6989586621679758145 c6989586621679758144)) infixl 0

data OnSym2 (a6989586621679758160 :: (~>) b6989586621679758143 ((~>) b6989586621679758143 c6989586621679758144)) (a6989586621679758161 :: (~>) a6989586621679758145 b6989586621679758143) :: (~>) a6989586621679758145 ((~>) a6989586621679758145 c6989586621679758144) infixl 0

data OnSym3 (a6989586621679758160 :: (~>) b6989586621679758143 ((~>) b6989586621679758143 c6989586621679758144)) (a6989586621679758161 :: (~>) a6989586621679758145 b6989586621679758143) (a6989586621679758162 :: a6989586621679758145) :: (~>) a6989586621679758145 c6989586621679758144 infixl 0

type OnSym4 (a6989586621679758160 :: (~>) b6989586621679758143 ((~>) b6989586621679758143 c6989586621679758144)) (a6989586621679758161 :: (~>) a6989586621679758145 b6989586621679758143) (a6989586621679758162 :: a6989586621679758145) (a6989586621679758163 :: a6989586621679758145) = On a6989586621679758160 a6989586621679758161 a6989586621679758162 a6989586621679758163