Data.Singletons.Prelude.List.NonEmpty

Description

Defines functions and datatypes relating to the singleton for NonEmpty, including a singletons version of all the definitions in Data.List.NonEmpty.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.List.NonEmpty. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The NonEmpty singleton

data family Sing :: k -> Type infixr 5

The singleton kind-indexed data family.

Though Haddock doesn't show it, the Sing instance above declares constructor

(:%|) :: Sing h -> Sing t -> Sing (h :| t)

type SNonEmpty = (Sing :: NonEmpty a -> Type)

SNonEmpty is a kind-restricted synonym for Sing: type SNonEmpty (a :: NonEmpty) = Sing a

Non-empty stream transformations

type family Map (a :: (~>) a b) (a :: NonEmpty a) :: NonEmpty b where ...

Equations

Map f ((:|) a as) = Apply (Apply (:|@#@$) (Apply f a)) (Apply (Apply ListmapSym0 f) as)

sMap :: forall a b (t :: (~>) a b) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: NonEmpty b)

type family Intersperse (a :: a) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Intersperse a ((:|) b bs) = Apply (Apply (:|@#@$) b) (Case_6989586621680632652 a b bs bs)

sIntersperse :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply IntersperseSym0 t) t :: NonEmpty a)

type family Scanl (a :: (~>) b ((~>) a b)) (a :: b) (a :: [a]) :: NonEmpty b where ...

Equations

Scanl f z a_6989586621680632805 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply ListscanlSym0 f) z)) a_6989586621680632805

sScanl :: forall b a (t :: (~>) b ((~>) a b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanlSym0 t) t) t :: NonEmpty b)

type family Scanr (a :: (~>) a ((~>) b b)) (a :: b) (a :: [a]) :: NonEmpty b where ...

Equations

Scanr f z a_6989586621680632816 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply ListscanrSym0 f) z)) a_6989586621680632816

sScanr :: forall a b (t :: (~>) a ((~>) b b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanrSym0 t) t) t :: NonEmpty b)

type family Scanl1 (a :: (~>) a ((~>) a a)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Scanl1 f ((:|) a as) = Apply FromListSym0 (Apply (Apply (Apply ListscanlSym0 f) a) as)

sScanl1 :: forall a (t :: (~>) a ((~>) a a)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply Scanl1Sym0 t) t :: NonEmpty a)

type family Scanr1 (a :: (~>) a ((~>) a a)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Scanr1 f ((:|) a as) = Apply FromListSym0 (Apply (Apply Listscanr1Sym0 f) (Apply (Apply (:@#@$) a) as))

sScanr1 :: forall a (t :: (~>) a ((~>) a a)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply Scanr1Sym0 t) t :: NonEmpty a)

type family Transpose (a :: NonEmpty (NonEmpty a)) :: NonEmpty (NonEmpty a) where ...

Equations

Transpose a_6989586621680632946 = Apply (Apply (Apply (.@#@$) (Apply FmapSym0 FromListSym0)) (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply (.@#@$) ListtransposeSym0) (Apply (Apply (.@#@$) ToListSym0) (Apply FmapSym0 ToListSym0))))) a_6989586621680632946

sTranspose :: forall a (t :: NonEmpty (NonEmpty a)). Sing t -> Sing (Apply TransposeSym0 t :: NonEmpty (NonEmpty a))

type family SortBy (a :: (~>) a ((~>) a Ordering)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

SortBy f a_6989586621680632759 = Apply (Apply LiftSym0 (Apply ListsortBySym0 f)) a_6989586621680632759

sSortBy :: forall a (t :: (~>) a ((~>) a Ordering)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SortBySym0 t) t :: NonEmpty a)

type family SortWith (a :: (~>) a o) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

SortWith a_6989586621680632763 a_6989586621680632765 = Apply (Apply (Apply (Apply (.@#@$) SortBySym0) ComparingSym0) a_6989586621680632763) a_6989586621680632765

sSortWith :: forall o a (t :: (~>) a o) (t :: NonEmpty a). SOrd o => Sing t -> Sing t -> Sing (Apply (Apply SortWithSym0 t) t :: NonEmpty a)

type family Length (a :: NonEmpty a) :: Nat where ...

Equations

Length ((:|) _ xs) = Apply (Apply (+@#@$) (FromInteger 1)) (Apply ListlengthSym0 xs)

sLength :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply LengthSym0 t :: Nat)

type family Head (a :: NonEmpty a) :: a where ...

Equations

Head ((:|) a _) = a

sHead :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply HeadSym0 t :: a)

type family Tail (a :: NonEmpty a) :: [a] where ...

Equations

Tail ((:|) _ as) = as

sTail :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply TailSym0 t :: [a])

type family Last (a :: NonEmpty a) :: a where ...

Equations

Last ((:|) a as) = Apply ListlastSym0 (Apply (Apply (:@#@$) a) as)

sLast :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply LastSym0 t :: a)

type family Init (a :: NonEmpty a) :: [a] where ...

Equations

Init ((:|) a as) = Apply ListinitSym0 (Apply (Apply (:@#@$) a) as)

sInit :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply InitSym0 t :: [a])

type family (a :: a) <| (a :: NonEmpty a) :: NonEmpty a where ...

Equations

a <| ((:|) b bs) = Apply (Apply (:|@#@$) a) (Apply (Apply (:@#@$) b) bs)

(%<|) :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply (<|@#@$) t) t :: NonEmpty a)

type family Cons (a :: a) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Cons a_6989586621680632847 a_6989586621680632849 = Apply (Apply (<|@#@$) a_6989586621680632847) a_6989586621680632849

sCons :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply ConsSym0 t) t :: NonEmpty a)

type family Uncons (a :: NonEmpty a) :: (a, Maybe (NonEmpty a)) where ...

Equations

Uncons ((:|) a as) = Apply (Apply Tuple2Sym0 a) (Apply NonEmpty_Sym0 as)

sUncons :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply UnconsSym0 t :: (a, Maybe (NonEmpty a)))

type family Unfoldr (a :: (~>) a (b, Maybe a)) (a :: a) :: NonEmpty b where ...

Equations

Unfoldr f a = Case_6989586621680632895 f a (Let6989586621680632892Scrutinee_6989586621680631228Sym2 f a)

sUnfoldr :: forall a b (t :: (~>) a (b, Maybe a)) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply UnfoldrSym0 t) t :: NonEmpty b)

type family Sort (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Sort a_6989586621680632835 = Apply (Apply LiftSym0 ListsortSym0) a_6989586621680632835

sSort :: forall a (t :: NonEmpty a). SOrd a => Sing t -> Sing (Apply SortSym0 t :: NonEmpty a)

type family Reverse (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Reverse a_6989586621680632750 = Apply (Apply LiftSym0 ListreverseSym0) a_6989586621680632750

sReverse :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply ReverseSym0 t :: NonEmpty a)

type family Inits (a :: [a]) :: NonEmpty [a] where ...

Equations

Inits a_6989586621680632781 = Apply (Apply (Apply (.@#@$) FromListSym0) ListinitsSym0) a_6989586621680632781

sInits :: forall a (t :: [a]). Sing t -> Sing (Apply InitsSym0 t :: NonEmpty [a])

type family Tails (a :: [a]) :: NonEmpty [a] where ...

Equations

Tails a_6989586621680632786 = Apply (Apply (Apply (.@#@$) FromListSym0) ListtailsSym0) a_6989586621680632786

sTails :: forall a (t :: [a]). Sing t -> Sing (Apply TailsSym0 t :: NonEmpty [a])

type family Unfold (a :: (~>) a (b, Maybe a)) (a :: a) :: NonEmpty b where ...

Equations

Unfold f a = Case_6989586621680632917 f a (Let6989586621680632914Scrutinee_6989586621680631218Sym2 f a)

sUnfold :: forall a b (t :: (~>) a (b, Maybe a)) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply UnfoldSym0 t) t :: NonEmpty b)

type family Insert (a :: a) (a :: [a]) :: NonEmpty a where ...

Equations

Insert a a_6989586621680632795 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply ListinsertSym0 a)) a_6989586621680632795

sInsert :: forall a (t :: a) (t :: [a]). SOrd a => Sing t -> Sing t -> Sing (Apply (Apply InsertSym0 t) t :: NonEmpty a)

type family Take (a :: Nat) (a :: NonEmpty a) :: [a] where ...

Equations

Take n a_6989586621680632670 = Apply (Apply (Apply (.@#@$) (Apply ListtakeSym0 n)) ToListSym0) a_6989586621680632670

sTake :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply TakeSym0 t) t :: [a])

type family Drop (a :: Nat) (a :: NonEmpty a) :: [a] where ...

Equations

Drop n a_6989586621680632678 = Apply (Apply (Apply (.@#@$) (Apply ListdropSym0 n)) ToListSym0) a_6989586621680632678

sDrop :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply DropSym0 t) t :: [a])

type family SplitAt (a :: Nat) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

SplitAt n a_6989586621680632686 = Apply (Apply (Apply (.@#@$) (Apply ListsplitAtSym0 n)) ToListSym0) a_6989586621680632686

sSplitAt :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SplitAtSym0 t) t :: ([a], [a]))

type family TakeWhile (a :: (~>) a Bool) (a :: NonEmpty a) :: [a] where ...

Equations

TakeWhile p a_6989586621680632694 = Apply (Apply (Apply (.@#@$) (Apply ListtakeWhileSym0 p)) ToListSym0) a_6989586621680632694

sTakeWhile :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply TakeWhileSym0 t) t :: [a])

type family DropWhile (a :: (~>) a Bool) (a :: NonEmpty a) :: [a] where ...

Equations

DropWhile p a_6989586621680632702 = Apply (Apply (Apply (.@#@$) (Apply ListdropWhileSym0 p)) ToListSym0) a_6989586621680632702

sDropWhile :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply DropWhileSym0 t) t :: [a])

type family Span (a :: (~>) a Bool) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

Span p a_6989586621680632710 = Apply (Apply (Apply (.@#@$) (Apply ListspanSym0 p)) ToListSym0) a_6989586621680632710

sSpan :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SpanSym0 t) t :: ([a], [a]))

type family Break (a :: (~>) a Bool) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

Break p a_6989586621680632718 = Apply (Apply SpanSym0 (Apply (Apply (.@#@$) NotSym0) p)) a_6989586621680632718

sBreak :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply BreakSym0 t) t :: ([a], [a]))

type family Filter (a :: (~>) a Bool) (a :: NonEmpty a) :: [a] where ...

Equations

Filter p a_6989586621680632726 = Apply (Apply (Apply (.@#@$) (Apply ListfilterSym0 p)) ToListSym0) a_6989586621680632726

sFilter :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply FilterSym0 t) t :: [a])

type family Partition (a :: (~>) a Bool) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

Partition p a_6989586621680632734 = Apply (Apply (Apply (.@#@$) (Apply ListpartitionSym0 p)) ToListSym0) a_6989586621680632734

sPartition :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply PartitionSym0 t) t :: ([a], [a]))

type family Group (a :: [a]) :: [NonEmpty a] where ...

Equations

Group a_6989586621680632640 = Apply (Apply GroupBySym0 (==@#@$)) a_6989586621680632640

sGroup :: forall a (t :: [a]). SEq a => Sing t -> Sing (Apply GroupSym0 t :: [NonEmpty a])

type family GroupBy (a :: (~>) a ((~>) a Bool)) (a :: [a]) :: [NonEmpty a] where ...

Equations

GroupBy eq0 a_6989586621680632536 = Apply (Apply (Let6989586621680632540GoSym2 eq0 a_6989586621680632536) eq0) a_6989586621680632536

sGroupBy :: forall a (t :: (~>) a ((~>) a Bool)) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply GroupBySym0 t) t :: [NonEmpty a])

type family GroupWith (a :: (~>) a b) (a :: [a]) :: [NonEmpty a] where ...

Equations

GroupWith f a_6989586621680632584 = Apply (Apply GroupBySym0 (Apply (Apply OnSym0 (==@#@$)) f)) a_6989586621680632584

sGroupWith :: forall b a (t :: (~>) a b) (t :: [a]). SEq b => Sing t -> Sing t -> Sing (Apply (Apply GroupWithSym0 t) t :: [NonEmpty a])

type family GroupAllWith (a :: (~>) a b) (a :: [a]) :: [NonEmpty a] where ...

Equations

GroupAllWith f a_6989586621680632592 = Apply (Apply (Apply (.@#@$) (Apply GroupWithSym0 f)) (Apply ListsortBySym0 (Apply (Apply OnSym0 CompareSym0) f))) a_6989586621680632592

sGroupAllWith :: forall b a (t :: (~>) a b) (t :: [a]). SOrd b => Sing t -> Sing t -> Sing (Apply (Apply GroupAllWithSym0 t) t :: [NonEmpty a])

type family Group1 (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

Group1 a_6989586621680632627 = Apply (Apply GroupBy1Sym0 (==@#@$)) a_6989586621680632627

sGroup1 :: forall a (t :: NonEmpty a). SEq a => Sing t -> Sing (Apply Group1Sym0 t :: NonEmpty (NonEmpty a))

type family GroupBy1 (a :: (~>) a ((~>) a Bool)) (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

GroupBy1 eq ((:|) x xs) = Apply (Apply (:|@#@$) (Apply (Apply (:|@#@$) x) (Let6989586621680632603YsSym3 eq x xs))) (Apply (Apply GroupBySym0 eq) (Let6989586621680632603ZsSym3 eq x xs))

sGroupBy1 :: forall a (t :: (~>) a ((~>) a Bool)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply GroupBy1Sym0 t) t :: NonEmpty (NonEmpty a))

type family GroupWith1 (a :: (~>) a b) (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

GroupWith1 f a_6989586621680632636 = Apply (Apply GroupBy1Sym0 (Apply (Apply OnSym0 (==@#@$)) f)) a_6989586621680632636

sGroupWith1 :: forall b a (t :: (~>) a b) (t :: NonEmpty a). SEq b => Sing t -> Sing t -> Sing (Apply (Apply GroupWith1Sym0 t) t :: NonEmpty (NonEmpty a))

type family GroupAllWith1 (a :: (~>) a b) (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

GroupAllWith1 f a_6989586621680632777 = Apply (Apply (Apply (.@#@$) (Apply GroupWith1Sym0 f)) (Apply SortWithSym0 f)) a_6989586621680632777

sGroupAllWith1 :: forall b a (t :: (~>) a b) (t :: NonEmpty a). SOrd b => Sing t -> Sing t -> Sing (Apply (Apply GroupAllWith1Sym0 t) t :: NonEmpty (NonEmpty a))

type family IsPrefixOf (a :: [a]) (a :: NonEmpty a) :: Bool where ...

Equations

IsPrefixOf '[] _ = TrueSym0

IsPrefixOf ((:) y ys) ((:|) x xs) = Apply (Apply (&&@#@$) (Apply (Apply (==@#@$) y) x)) (Apply (Apply ListisPrefixOfSym0 ys) xs)

sIsPrefixOf :: forall a (t :: [a]) (t :: NonEmpty a). SEq a => Sing t -> Sing t -> Sing (Apply (Apply IsPrefixOfSym0 t) t :: Bool)

type family Nub (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Nub a_6989586621680632453 = Apply (Apply NubBySym0 (==@#@$)) a_6989586621680632453

sNub :: forall a (t :: NonEmpty a). SEq a => Sing t -> Sing (Apply NubSym0 t :: NonEmpty a)

type family NubBy (a :: (~>) a ((~>) a Bool)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

NubBy eq ((:|) a as) = Apply (Apply (:|@#@$) a) (Apply (Apply ListnubBySym0 eq) (Apply (Apply ListfilterSym0 (Apply (Apply (Apply Lambda_6989586621680632445Sym0 eq) a) as)) as))

sNubBy :: forall a (t :: (~>) a ((~>) a Bool)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply NubBySym0 t) t :: NonEmpty a)

type family (a :: NonEmpty a) !! (a :: Nat) :: a where ...

Equations

arg_6989586621680631240 !! arg_6989586621680631242 = Case_6989586621680632512 arg_6989586621680631240 arg_6989586621680631242 (Apply (Apply Tuple2Sym0 arg_6989586621680631240) arg_6989586621680631242)

(%!!) :: forall a (t :: NonEmpty a) (t :: Nat). Sing t -> Sing t -> Sing (Apply (Apply (!!@#@$) t) t :: a)

type family Zip (a :: NonEmpty a) (a :: NonEmpty b) :: NonEmpty (a, b) where ...

Equations

Zip ((:|) x xs) ((:|) y ys) = Apply (Apply (:|@#@$) (Apply (Apply Tuple2Sym0 x) y)) (Apply (Apply ListzipSym0 xs) ys)

sZip :: forall a b (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing (Apply (Apply ZipSym0 t) t :: NonEmpty (a, b))

type family ZipWith (a :: (~>) a ((~>) b c)) (a :: NonEmpty a) (a :: NonEmpty b) :: NonEmpty c where ...

Equations

ZipWith f ((:|) x xs) ((:|) y ys) = Apply (Apply (:|@#@$) (Apply (Apply f x) y)) (Apply (Apply (Apply ListzipWithSym0 f) xs) ys)

sZipWith :: forall a b c (t :: (~>) a ((~>) b c)) (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ZipWithSym0 t) t) t :: NonEmpty c)

type family Unzip (a :: NonEmpty (a, b)) :: (NonEmpty a, NonEmpty b) where ...

Equations

Unzip ((:|) '(a, b) asbs) = Apply (Apply Tuple2Sym0 (Apply (Apply (:|@#@$) a) (Let6989586621680632463AsSym3 a b asbs))) (Apply (Apply (:|@#@$) b) (Let6989586621680632463BsSym3 a b asbs))

sUnzip :: forall a b (t :: NonEmpty (a, b)). Sing t -> Sing (Apply UnzipSym0 t :: (NonEmpty a, NonEmpty b))

type family FromList (a :: [a]) :: NonEmpty a where ...

Equations

FromList ((:) a as) = Apply (Apply (:|@#@$) a) as

FromList '[] = Apply ErrorSym0 "NonEmpty.fromList: empty list"

sFromList :: forall a (t :: [a]). Sing t -> Sing (Apply FromListSym0 t :: NonEmpty a)

type family ToList (a :: NonEmpty a) :: [a] where ...

Equations

ToList ((:|) a as) = Apply (Apply (:@#@$) a) as

sToList :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply ToListSym0 t :: [a])

type family NonEmpty_ (a :: [a]) :: Maybe (NonEmpty a) where ...

Equations

NonEmpty_ '[] = NothingSym0

NonEmpty_ ((:) a as) = Apply JustSym0 (Apply (Apply (:|@#@$) a) as)

sNonEmpty_ :: forall a (t :: [a]). Sing t -> Sing (Apply NonEmpty_Sym0 t :: Maybe (NonEmpty a))

type family Xor (a :: NonEmpty Bool) :: Bool where ...

Equations

Xor ((:|) x xs) = Apply (Apply (Apply FoldrSym0 (Let6989586621680632927Xor'Sym2 x xs)) x) xs

sXor :: forall (t :: NonEmpty Bool). Sing t -> Sing (Apply XorSym0 t :: Bool)

Defunctionalization symbols

data (:|@#@$) :: forall (a6989586621679066696 :: Type). (~>) a6989586621679066696 ((~>) [a6989586621679066696] (NonEmpty (a6989586621679066696 :: Type))) infixr 5

data (:|@#@$$) (t6989586621679227992 :: (a6989586621679066696 :: Type)) :: (~>) [a6989586621679066696] (NonEmpty (a6989586621679066696 :: Type)) infixr 5

type (:|@#@$$$) (t6989586621679227992 :: a6989586621679066696) (t6989586621679227993 :: [a6989586621679066696]) = (:|) t6989586621679227992 t6989586621679227993

data MapSym0 :: forall a6989586621680631034 b6989586621680631035. (~>) ((~>) a6989586621680631034 b6989586621680631035) ((~>) (NonEmpty a6989586621680631034) (NonEmpty b6989586621680631035))

data MapSym1 (a6989586621680632655 :: (~>) a6989586621680631034 b6989586621680631035) :: (~>) (NonEmpty a6989586621680631034) (NonEmpty b6989586621680631035)

type MapSym2 (a6989586621680632655 :: (~>) a6989586621680631034 b6989586621680631035) (a6989586621680632656 :: NonEmpty a6989586621680631034) = Map a6989586621680632655 a6989586621680632656

data IntersperseSym0 :: forall a6989586621680631024. (~>) a6989586621680631024 ((~>) (NonEmpty a6989586621680631024) (NonEmpty a6989586621680631024))

data IntersperseSym1 (a6989586621680632645 :: a6989586621680631024) :: (~>) (NonEmpty a6989586621680631024) (NonEmpty a6989586621680631024)

type IntersperseSym2 (a6989586621680632645 :: a6989586621680631024) (a6989586621680632646 :: NonEmpty a6989586621680631024) = Intersperse a6989586621680632645 a6989586621680632646

data ScanlSym0 :: forall a6989586621680631030 b6989586621680631029. (~>) ((~>) b6989586621680631029 ((~>) a6989586621680631030 b6989586621680631029)) ((~>) b6989586621680631029 ((~>) [a6989586621680631030] (NonEmpty b6989586621680631029)))

data ScanlSym1 (a6989586621680632799 :: (~>) b6989586621680631029 ((~>) a6989586621680631030 b6989586621680631029)) :: (~>) b6989586621680631029 ((~>) [a6989586621680631030] (NonEmpty b6989586621680631029))

data ScanlSym2 (a6989586621680632799 :: (~>) b6989586621680631029 ((~>) a6989586621680631030 b6989586621680631029)) (a6989586621680632800 :: b6989586621680631029) :: (~>) [a6989586621680631030] (NonEmpty b6989586621680631029)

type ScanlSym3 (a6989586621680632799 :: (~>) b6989586621680631029 ((~>) a6989586621680631030 b6989586621680631029)) (a6989586621680632800 :: b6989586621680631029) (a6989586621680632801 :: [a6989586621680631030]) = Scanl a6989586621680632799 a6989586621680632800 a6989586621680632801

data ScanrSym0 :: forall a6989586621680631027 b6989586621680631028. (~>) ((~>) a6989586621680631027 ((~>) b6989586621680631028 b6989586621680631028)) ((~>) b6989586621680631028 ((~>) [a6989586621680631027] (NonEmpty b6989586621680631028)))

data ScanrSym1 (a6989586621680632810 :: (~>) a6989586621680631027 ((~>) b6989586621680631028 b6989586621680631028)) :: (~>) b6989586621680631028 ((~>) [a6989586621680631027] (NonEmpty b6989586621680631028))

data ScanrSym2 (a6989586621680632810 :: (~>) a6989586621680631027 ((~>) b6989586621680631028 b6989586621680631028)) (a6989586621680632811 :: b6989586621680631028) :: (~>) [a6989586621680631027] (NonEmpty b6989586621680631028)

type ScanrSym3 (a6989586621680632810 :: (~>) a6989586621680631027 ((~>) b6989586621680631028 b6989586621680631028)) (a6989586621680632811 :: b6989586621680631028) (a6989586621680632812 :: [a6989586621680631027]) = Scanr a6989586621680632810 a6989586621680632811 a6989586621680632812

data Scanl1Sym0 :: forall a6989586621680631026. (~>) ((~>) a6989586621680631026 ((~>) a6989586621680631026 a6989586621680631026)) ((~>) (NonEmpty a6989586621680631026) (NonEmpty a6989586621680631026))

data Scanl1Sym1 (a6989586621680632821 :: (~>) a6989586621680631026 ((~>) a6989586621680631026 a6989586621680631026)) :: (~>) (NonEmpty a6989586621680631026) (NonEmpty a6989586621680631026)

type Scanl1Sym2 (a6989586621680632821 :: (~>) a6989586621680631026 ((~>) a6989586621680631026 a6989586621680631026)) (a6989586621680632822 :: NonEmpty a6989586621680631026) = Scanl1 a6989586621680632821 a6989586621680632822

data Scanr1Sym0 :: forall a6989586621680631025. (~>) ((~>) a6989586621680631025 ((~>) a6989586621680631025 a6989586621680631025)) ((~>) (NonEmpty a6989586621680631025) (NonEmpty a6989586621680631025))

data Scanr1Sym1 (a6989586621680632828 :: (~>) a6989586621680631025 ((~>) a6989586621680631025 a6989586621680631025)) :: (~>) (NonEmpty a6989586621680631025) (NonEmpty a6989586621680631025)

type Scanr1Sym2 (a6989586621680632828 :: (~>) a6989586621680631025 ((~>) a6989586621680631025 a6989586621680631025)) (a6989586621680632829 :: NonEmpty a6989586621680631025) = Scanr1 a6989586621680632828 a6989586621680632829

data TransposeSym0 :: forall a6989586621680630990. (~>) (NonEmpty (NonEmpty a6989586621680630990)) (NonEmpty (NonEmpty a6989586621680630990))

type TransposeSym1 (a6989586621680632948 :: NonEmpty (NonEmpty a6989586621680630990)) = Transpose a6989586621680632948

data SortBySym0 :: forall a6989586621680630989. (~>) ((~>) a6989586621680630989 ((~>) a6989586621680630989 Ordering)) ((~>) (NonEmpty a6989586621680630989) (NonEmpty a6989586621680630989))

data SortBySym1 (a6989586621680632755 :: (~>) a6989586621680630989 ((~>) a6989586621680630989 Ordering)) :: (~>) (NonEmpty a6989586621680630989) (NonEmpty a6989586621680630989)

type SortBySym2 (a6989586621680632755 :: (~>) a6989586621680630989 ((~>) a6989586621680630989 Ordering)) (a6989586621680632756 :: NonEmpty a6989586621680630989) = SortBy a6989586621680632755 a6989586621680632756

data SortWithSym0 :: forall a6989586621680630988 o6989586621680630987. (~>) ((~>) a6989586621680630988 o6989586621680630987) ((~>) (NonEmpty a6989586621680630988) (NonEmpty a6989586621680630988))

data SortWithSym1 (a6989586621680632767 :: (~>) a6989586621680630988 o6989586621680630987) :: (~>) (NonEmpty a6989586621680630988) (NonEmpty a6989586621680630988)

type SortWithSym2 (a6989586621680632767 :: (~>) a6989586621680630988 o6989586621680630987) (a6989586621680632768 :: NonEmpty a6989586621680630988) = SortWith a6989586621680632767 a6989586621680632768

data LengthSym0 :: forall a6989586621680631053. (~>) (NonEmpty a6989586621680631053) Nat

type LengthSym1 (a6989586621680632936 :: NonEmpty a6989586621680631053) = Length a6989586621680632936

data HeadSym0 :: forall a6989586621680631046. (~>) (NonEmpty a6989586621680631046) a6989586621680631046

type HeadSym1 (a6989586621680632868 :: NonEmpty a6989586621680631046) = Head a6989586621680632868

data TailSym0 :: forall a6989586621680631045. (~>) (NonEmpty a6989586621680631045) [a6989586621680631045]

type TailSym1 (a6989586621680632865 :: NonEmpty a6989586621680631045) = Tail a6989586621680632865

data LastSym0 :: forall a6989586621680631044. (~>) (NonEmpty a6989586621680631044) a6989586621680631044

type LastSym1 (a6989586621680632861 :: NonEmpty a6989586621680631044) = Last a6989586621680632861

data InitSym0 :: forall a6989586621680631043. (~>) (NonEmpty a6989586621680631043) [a6989586621680631043]

type InitSym1 (a6989586621680632857 :: NonEmpty a6989586621680631043) = Init a6989586621680632857

data (<|@#@$) :: forall a6989586621680631042. (~>) a6989586621680631042 ((~>) (NonEmpty a6989586621680631042) (NonEmpty a6989586621680631042))

data (<|@#@$$) (a6989586621680632840 :: a6989586621680631042) :: (~>) (NonEmpty a6989586621680631042) (NonEmpty a6989586621680631042)

type (<|@#@$$$) (a6989586621680632840 :: a6989586621680631042) (a6989586621680632841 :: NonEmpty a6989586621680631042) = (<|) a6989586621680632840 a6989586621680632841

data ConsSym0 :: forall a6989586621680631041. (~>) a6989586621680631041 ((~>) (NonEmpty a6989586621680631041) (NonEmpty a6989586621680631041))

data ConsSym1 (a6989586621680632851 :: a6989586621680631041) :: (~>) (NonEmpty a6989586621680631041) (NonEmpty a6989586621680631041)

type ConsSym2 (a6989586621680632851 :: a6989586621680631041) (a6989586621680632852 :: NonEmpty a6989586621680631041) = Cons a6989586621680632851 a6989586621680632852

data UnconsSym0 :: forall a6989586621680631049. (~>) (NonEmpty a6989586621680631049) (a6989586621680631049, Maybe (NonEmpty a6989586621680631049))

type UnconsSym1 (a6989586621680632904 :: NonEmpty a6989586621680631049) = Uncons a6989586621680632904

data UnfoldrSym0 :: forall a6989586621680631047 b6989586621680631048. (~>) ((~>) a6989586621680631047 (b6989586621680631048, Maybe a6989586621680631047)) ((~>) a6989586621680631047 (NonEmpty b6989586621680631048))

data UnfoldrSym1 (a6989586621680632871 :: (~>) a6989586621680631047 (b6989586621680631048, Maybe a6989586621680631047)) :: (~>) a6989586621680631047 (NonEmpty b6989586621680631048)

type UnfoldrSym2 (a6989586621680632871 :: (~>) a6989586621680631047 (b6989586621680631048, Maybe a6989586621680631047)) (a6989586621680632872 :: a6989586621680631047) = Unfoldr a6989586621680632871 a6989586621680632872

data SortSym0 :: forall a6989586621680631040. (~>) (NonEmpty a6989586621680631040) (NonEmpty a6989586621680631040)

type SortSym1 (a6989586621680632837 :: NonEmpty a6989586621680631040) = Sort a6989586621680632837

data ReverseSym0 :: forall a6989586621680631023. (~>) (NonEmpty a6989586621680631023) (NonEmpty a6989586621680631023)

type ReverseSym1 (a6989586621680632752 :: NonEmpty a6989586621680631023) = Reverse a6989586621680632752

data InitsSym0 :: forall a6989586621680631033. (~>) [a6989586621680631033] (NonEmpty [a6989586621680631033])

type InitsSym1 (a6989586621680632783 :: [a6989586621680631033]) = Inits a6989586621680632783

data TailsSym0 :: forall a6989586621680631032. (~>) [a6989586621680631032] (NonEmpty [a6989586621680631032])

type TailsSym1 (a6989586621680632788 :: [a6989586621680631032]) = Tails a6989586621680632788

data UnfoldSym0 :: forall a6989586621680631051 b6989586621680631052. (~>) ((~>) a6989586621680631051 (b6989586621680631052, Maybe a6989586621680631051)) ((~>) a6989586621680631051 (NonEmpty b6989586621680631052))

data UnfoldSym1 (a6989586621680632908 :: (~>) a6989586621680631051 (b6989586621680631052, Maybe a6989586621680631051)) :: (~>) a6989586621680631051 (NonEmpty b6989586621680631052)

data InsertSym0 :: forall a6989586621680631031. (~>) a6989586621680631031 ((~>) [a6989586621680631031] (NonEmpty a6989586621680631031))

data InsertSym1 (a6989586621680632791 :: a6989586621680631031) :: (~>) [a6989586621680631031] (NonEmpty a6989586621680631031)

type InsertSym2 (a6989586621680632791 :: a6989586621680631031) (a6989586621680632792 :: [a6989586621680631031]) = Insert a6989586621680632791 a6989586621680632792

data TakeSym0 :: forall a6989586621680631022. (~>) Nat ((~>) (NonEmpty a6989586621680631022) [a6989586621680631022])

data TakeSym1 (a6989586621680632666 :: Nat) :: forall a6989586621680631022. (~>) (NonEmpty a6989586621680631022) [a6989586621680631022]

type TakeSym2 (a6989586621680632666 :: Nat) (a6989586621680632667 :: NonEmpty a6989586621680631022) = Take a6989586621680632666 a6989586621680632667

data DropSym0 :: forall a6989586621680631021. (~>) Nat ((~>) (NonEmpty a6989586621680631021) [a6989586621680631021])

data DropSym1 (a6989586621680632674 :: Nat) :: forall a6989586621680631021. (~>) (NonEmpty a6989586621680631021) [a6989586621680631021]

type DropSym2 (a6989586621680632674 :: Nat) (a6989586621680632675 :: NonEmpty a6989586621680631021) = Drop a6989586621680632674 a6989586621680632675

data SplitAtSym0 :: forall a6989586621680631020. (~>) Nat ((~>) (NonEmpty a6989586621680631020) ([a6989586621680631020], [a6989586621680631020]))

data SplitAtSym1 (a6989586621680632682 :: Nat) :: forall a6989586621680631020. (~>) (NonEmpty a6989586621680631020) ([a6989586621680631020], [a6989586621680631020])

type SplitAtSym2 (a6989586621680632682 :: Nat) (a6989586621680632683 :: NonEmpty a6989586621680631020) = SplitAt a6989586621680632682 a6989586621680632683

data TakeWhileSym0 :: forall a6989586621680631019. (~>) ((~>) a6989586621680631019 Bool) ((~>) (NonEmpty a6989586621680631019) [a6989586621680631019])

data TakeWhileSym1 (a6989586621680632690 :: (~>) a6989586621680631019 Bool) :: (~>) (NonEmpty a6989586621680631019) [a6989586621680631019]

type TakeWhileSym2 (a6989586621680632690 :: (~>) a6989586621680631019 Bool) (a6989586621680632691 :: NonEmpty a6989586621680631019) = TakeWhile a6989586621680632690 a6989586621680632691

data DropWhileSym0 :: forall a6989586621680631018. (~>) ((~>) a6989586621680631018 Bool) ((~>) (NonEmpty a6989586621680631018) [a6989586621680631018])

data DropWhileSym1 (a6989586621680632698 :: (~>) a6989586621680631018 Bool) :: (~>) (NonEmpty a6989586621680631018) [a6989586621680631018]

type DropWhileSym2 (a6989586621680632698 :: (~>) a6989586621680631018 Bool) (a6989586621680632699 :: NonEmpty a6989586621680631018) = DropWhile a6989586621680632698 a6989586621680632699

data SpanSym0 :: forall a6989586621680631017. (~>) ((~>) a6989586621680631017 Bool) ((~>) (NonEmpty a6989586621680631017) ([a6989586621680631017], [a6989586621680631017]))

data SpanSym1 (a6989586621680632706 :: (~>) a6989586621680631017 Bool) :: (~>) (NonEmpty a6989586621680631017) ([a6989586621680631017], [a6989586621680631017])

type SpanSym2 (a6989586621680632706 :: (~>) a6989586621680631017 Bool) (a6989586621680632707 :: NonEmpty a6989586621680631017) = Span a6989586621680632706 a6989586621680632707

data BreakSym0 :: forall a6989586621680631016. (~>) ((~>) a6989586621680631016 Bool) ((~>) (NonEmpty a6989586621680631016) ([a6989586621680631016], [a6989586621680631016]))

data BreakSym1 (a6989586621680632714 :: (~>) a6989586621680631016 Bool) :: (~>) (NonEmpty a6989586621680631016) ([a6989586621680631016], [a6989586621680631016])

type BreakSym2 (a6989586621680632714 :: (~>) a6989586621680631016 Bool) (a6989586621680632715 :: NonEmpty a6989586621680631016) = Break a6989586621680632714 a6989586621680632715

data FilterSym0 :: forall a6989586621680631015. (~>) ((~>) a6989586621680631015 Bool) ((~>) (NonEmpty a6989586621680631015) [a6989586621680631015])

data FilterSym1 (a6989586621680632722 :: (~>) a6989586621680631015 Bool) :: (~>) (NonEmpty a6989586621680631015) [a6989586621680631015]

type FilterSym2 (a6989586621680632722 :: (~>) a6989586621680631015 Bool) (a6989586621680632723 :: NonEmpty a6989586621680631015) = Filter a6989586621680632722 a6989586621680632723

data PartitionSym0 :: forall a6989586621680631014. (~>) ((~>) a6989586621680631014 Bool) ((~>) (NonEmpty a6989586621680631014) ([a6989586621680631014], [a6989586621680631014]))

data PartitionSym1 (a6989586621680632730 :: (~>) a6989586621680631014 Bool) :: (~>) (NonEmpty a6989586621680631014) ([a6989586621680631014], [a6989586621680631014])

type PartitionSym2 (a6989586621680632730 :: (~>) a6989586621680631014 Bool) (a6989586621680632731 :: NonEmpty a6989586621680631014) = Partition a6989586621680632730 a6989586621680632731

data GroupSym0 :: forall a6989586621680631013. (~>) [a6989586621680631013] [NonEmpty a6989586621680631013]

type GroupSym1 (a6989586621680632642 :: [a6989586621680631013]) = Group a6989586621680632642

data GroupBySym0 :: forall a6989586621680631012. (~>) ((~>) a6989586621680631012 ((~>) a6989586621680631012 Bool)) ((~>) [a6989586621680631012] [NonEmpty a6989586621680631012])

data GroupBySym1 (a6989586621680632532 :: (~>) a6989586621680631012 ((~>) a6989586621680631012 Bool)) :: (~>) [a6989586621680631012] [NonEmpty a6989586621680631012]

type GroupBySym2 (a6989586621680632532 :: (~>) a6989586621680631012 ((~>) a6989586621680631012 Bool)) (a6989586621680632533 :: [a6989586621680631012]) = GroupBy a6989586621680632532 a6989586621680632533

data GroupWithSym0 :: forall a6989586621680631011 b6989586621680631010. (~>) ((~>) a6989586621680631011 b6989586621680631010) ((~>) [a6989586621680631011] [NonEmpty a6989586621680631011])

data GroupWithSym1 (a6989586621680632580 :: (~>) a6989586621680631011 b6989586621680631010) :: (~>) [a6989586621680631011] [NonEmpty a6989586621680631011]

type GroupWithSym2 (a6989586621680632580 :: (~>) a6989586621680631011 b6989586621680631010) (a6989586621680632581 :: [a6989586621680631011]) = GroupWith a6989586621680632580 a6989586621680632581

data GroupAllWithSym0 :: forall a6989586621680631009 b6989586621680631008. (~>) ((~>) a6989586621680631009 b6989586621680631008) ((~>) [a6989586621680631009] [NonEmpty a6989586621680631009])

data GroupAllWithSym1 (a6989586621680632588 :: (~>) a6989586621680631009 b6989586621680631008) :: (~>) [a6989586621680631009] [NonEmpty a6989586621680631009]

type GroupAllWithSym2 (a6989586621680632588 :: (~>) a6989586621680631009 b6989586621680631008) (a6989586621680632589 :: [a6989586621680631009]) = GroupAllWith a6989586621680632588 a6989586621680632589

data Group1Sym0 :: forall a6989586621680631007. (~>) (NonEmpty a6989586621680631007) (NonEmpty (NonEmpty a6989586621680631007))

type Group1Sym1 (a6989586621680632629 :: NonEmpty a6989586621680631007) = Group1 a6989586621680632629

data GroupBy1Sym0 :: forall a6989586621680631006. (~>) ((~>) a6989586621680631006 ((~>) a6989586621680631006 Bool)) ((~>) (NonEmpty a6989586621680631006) (NonEmpty (NonEmpty a6989586621680631006)))

data GroupBy1Sym1 (a6989586621680632596 :: (~>) a6989586621680631006 ((~>) a6989586621680631006 Bool)) :: (~>) (NonEmpty a6989586621680631006) (NonEmpty (NonEmpty a6989586621680631006))

type GroupBy1Sym2 (a6989586621680632596 :: (~>) a6989586621680631006 ((~>) a6989586621680631006 Bool)) (a6989586621680632597 :: NonEmpty a6989586621680631006) = GroupBy1 a6989586621680632596 a6989586621680632597

data GroupWith1Sym0 :: forall a6989586621680631005 b6989586621680631004. (~>) ((~>) a6989586621680631005 b6989586621680631004) ((~>) (NonEmpty a6989586621680631005) (NonEmpty (NonEmpty a6989586621680631005)))

data GroupWith1Sym1 (a6989586621680632632 :: (~>) a6989586621680631005 b6989586621680631004) :: (~>) (NonEmpty a6989586621680631005) (NonEmpty (NonEmpty a6989586621680631005))

type GroupWith1Sym2 (a6989586621680632632 :: (~>) a6989586621680631005 b6989586621680631004) (a6989586621680632633 :: NonEmpty a6989586621680631005) = GroupWith1 a6989586621680632632 a6989586621680632633

data GroupAllWith1Sym0 :: forall a6989586621680631003 b6989586621680631002. (~>) ((~>) a6989586621680631003 b6989586621680631002) ((~>) (NonEmpty a6989586621680631003) (NonEmpty (NonEmpty a6989586621680631003)))

data GroupAllWith1Sym1 (a6989586621680632773 :: (~>) a6989586621680631003 b6989586621680631002) :: (~>) (NonEmpty a6989586621680631003) (NonEmpty (NonEmpty a6989586621680631003))

type GroupAllWith1Sym2 (a6989586621680632773 :: (~>) a6989586621680631003 b6989586621680631002) (a6989586621680632774 :: NonEmpty a6989586621680631003) = GroupAllWith1 a6989586621680632773 a6989586621680632774

data IsPrefixOfSym0 :: forall a6989586621680631001. (~>) [a6989586621680631001] ((~>) (NonEmpty a6989586621680631001) Bool)

data IsPrefixOfSym1 (a6989586621680632524 :: [a6989586621680631001]) :: (~>) (NonEmpty a6989586621680631001) Bool

type IsPrefixOfSym2 (a6989586621680632524 :: [a6989586621680631001]) (a6989586621680632525 :: NonEmpty a6989586621680631001) = IsPrefixOf a6989586621680632524 a6989586621680632525

data NubSym0 :: forall a6989586621680630992. (~>) (NonEmpty a6989586621680630992) (NonEmpty a6989586621680630992)

type NubSym1 (a6989586621680632455 :: NonEmpty a6989586621680630992) = Nub a6989586621680632455

data NubBySym0 :: forall a6989586621680630991. (~>) ((~>) a6989586621680630991 ((~>) a6989586621680630991 Bool)) ((~>) (NonEmpty a6989586621680630991) (NonEmpty a6989586621680630991))

data NubBySym1 (a6989586621680632438 :: (~>) a6989586621680630991 ((~>) a6989586621680630991 Bool)) :: (~>) (NonEmpty a6989586621680630991) (NonEmpty a6989586621680630991)

type NubBySym2 (a6989586621680632438 :: (~>) a6989586621680630991 ((~>) a6989586621680630991 Bool)) (a6989586621680632439 :: NonEmpty a6989586621680630991) = NubBy a6989586621680632438 a6989586621680632439

data (!!@#@$) :: forall a6989586621680631000. (~>) (NonEmpty a6989586621680631000) ((~>) Nat a6989586621680631000)

data (!!@#@$$) (a6989586621680632506 :: NonEmpty a6989586621680631000) :: (~>) Nat a6989586621680631000

type (!!@#@$$$) (a6989586621680632506 :: NonEmpty a6989586621680631000) (a6989586621680632507 :: Nat) = (!!) a6989586621680632506 a6989586621680632507

data ZipSym0 :: forall a6989586621680630998 b6989586621680630999. (~>) (NonEmpty a6989586621680630998) ((~>) (NonEmpty b6989586621680630999) (NonEmpty (a6989586621680630998, b6989586621680630999)))

data ZipSym1 (a6989586621680632498 :: NonEmpty a6989586621680630998) :: forall b6989586621680630999. (~>) (NonEmpty b6989586621680630999) (NonEmpty (a6989586621680630998, b6989586621680630999))

type ZipSym2 (a6989586621680632498 :: NonEmpty a6989586621680630998) (a6989586621680632499 :: NonEmpty b6989586621680630999) = Zip a6989586621680632498 a6989586621680632499

data ZipWithSym0 :: forall a6989586621680630995 b6989586621680630996 c6989586621680630997. (~>) ((~>) a6989586621680630995 ((~>) b6989586621680630996 c6989586621680630997)) ((~>) (NonEmpty a6989586621680630995) ((~>) (NonEmpty b6989586621680630996) (NonEmpty c6989586621680630997)))

data ZipWithSym1 (a6989586621680632487 :: (~>) a6989586621680630995 ((~>) b6989586621680630996 c6989586621680630997)) :: (~>) (NonEmpty a6989586621680630995) ((~>) (NonEmpty b6989586621680630996) (NonEmpty c6989586621680630997))

data ZipWithSym2 (a6989586621680632487 :: (~>) a6989586621680630995 ((~>) b6989586621680630996 c6989586621680630997)) (a6989586621680632488 :: NonEmpty a6989586621680630995) :: (~>) (NonEmpty b6989586621680630996) (NonEmpty c6989586621680630997)

type ZipWithSym3 (a6989586621680632487 :: (~>) a6989586621680630995 ((~>) b6989586621680630996 c6989586621680630997)) (a6989586621680632488 :: NonEmpty a6989586621680630995) (a6989586621680632489 :: NonEmpty b6989586621680630996) = ZipWith a6989586621680632487 a6989586621680632488 a6989586621680632489

data UnzipSym0 :: forall a6989586621680630993 b6989586621680630994. (~>) (NonEmpty (a6989586621680630993, b6989586621680630994)) (NonEmpty a6989586621680630993, NonEmpty b6989586621680630994)

type UnzipSym1 (a6989586621680632458 :: NonEmpty (a6989586621680630993, b6989586621680630994)) = Unzip a6989586621680632458

data FromListSym0 :: forall a6989586621680631039. (~>) [a6989586621680631039] (NonEmpty a6989586621680631039)

type FromListSym1 (a6989586621680632738 :: [a6989586621680631039]) = FromList a6989586621680632738

data ToListSym0 :: forall a6989586621680631038. (~>) (NonEmpty a6989586621680631038) [a6989586621680631038]

type ToListSym1 (a6989586621680632662 :: NonEmpty a6989586621680631038) = ToList a6989586621680632662

data NonEmpty_Sym0 :: forall a6989586621680631050. (~>) [a6989586621680631050] (Maybe (NonEmpty a6989586621680631050))

type NonEmpty_Sym1 (a6989586621680632900 :: [a6989586621680631050]) = NonEmpty_ a6989586621680632900

data XorSym0 :: (~>) (NonEmpty Bool) Bool

type XorSym1 (a6989586621680632923 :: NonEmpty Bool) = Xor a6989586621680632923

Orphan instances

SMonadZip NonEmpty

Instance details

Methods

sMzip :: Sing t -> Sing t -> Sing (Apply (Apply MzipSym0 t) t)

sMzipWith :: Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply MzipWithSym0 t) t) t)

sMunzip :: Sing t -> Sing (Apply MunzipSym0 t)

PMonadZip NonEmpty

Instance details

Associated Types

type Mzip arg arg :: m (a, b)

type MzipWith arg arg arg :: m c

type Munzip arg :: (m a, m b)