Data.Singletons.Prelude.List.NonEmpty

Description

Defines functions and datatypes relating to the singleton for NonEmpty, including a singletons version of all the definitions in Data.List.NonEmpty.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.List.NonEmpty. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The NonEmpty singleton

type family Sing :: k -> Type

The singleton kind-indexed type family.

data SNonEmpty :: forall a. NonEmpty a -> Type where

Constructors

(:%|) :: forall a (n :: a) (n :: [a]). (Sing (n :: a)) -> (Sing (n :: [a])) -> SNonEmpty ('(:|) n n) infixr 5

Non-empty stream transformations

type family Map (a :: (~>) a b) (a :: NonEmpty a) :: NonEmpty b where ...

Equations

Map f ('(:|) a as) = Apply (Apply (:|@#@$) (Apply f a)) (Apply (Apply ListmapSym0 f) as)

sMap :: forall a b (t :: (~>) a b) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: NonEmpty b)

type family Intersperse (a :: a) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Intersperse a ('(:|) b bs) = Apply (Apply (:|@#@$) b) (Case_6989586621681166670 a b bs bs)

sIntersperse :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply IntersperseSym0 t) t :: NonEmpty a)

type family Scanl (a :: (~>) b ((~>) a b)) (a :: b) (a :: [a]) :: NonEmpty b where ...

Equations

Scanl f z a_6989586621681166704 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply ListscanlSym0 f) z)) a_6989586621681166704

sScanl :: forall b a (t :: (~>) b ((~>) a b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanlSym0 t) t) t :: NonEmpty b)

type family Scanr (a :: (~>) a ((~>) b b)) (a :: b) (a :: [a]) :: NonEmpty b where ...

Equations

Scanr f z a_6989586621681166693 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply ListscanrSym0 f) z)) a_6989586621681166693

sScanr :: forall a b (t :: (~>) a ((~>) b b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanrSym0 t) t) t :: NonEmpty b)

type family Scanl1 (a :: (~>) a ((~>) a a)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Scanl1 f ('(:|) a as) = Apply FromListSym0 (Apply (Apply (Apply ListscanlSym0 f) a) as)

sScanl1 :: forall a (t :: (~>) a ((~>) a a)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply Scanl1Sym0 t) t :: NonEmpty a)

type family Scanr1 (a :: (~>) a ((~>) a a)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Scanr1 f ('(:|) a as) = Apply FromListSym0 (Apply (Apply Listscanr1Sym0 f) (Apply (Apply (:@#@$) a) as))

sScanr1 :: forall a (t :: (~>) a ((~>) a a)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply Scanr1Sym0 t) t :: NonEmpty a)

type family Transpose (a :: NonEmpty (NonEmpty a)) :: NonEmpty (NonEmpty a) where ...

Equations

Transpose a_6989586621681166366 = Apply (Apply (Apply (.@#@$) (Apply FmapSym0 FromListSym0)) (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply (.@#@$) ListtransposeSym0) (Apply (Apply (.@#@$) ToListSym0) (Apply FmapSym0 ToListSym0))))) a_6989586621681166366

sTranspose :: forall a (t :: NonEmpty (NonEmpty a)). Sing t -> Sing (Apply TransposeSym0 t :: NonEmpty (NonEmpty a))

type family SortBy (a :: (~>) a ((~>) a Ordering)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

SortBy f a_6989586621681166362 = Apply (Apply LiftSym0 (Apply ListsortBySym0 f)) a_6989586621681166362

sSortBy :: forall a (t :: (~>) a ((~>) a Ordering)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SortBySym0 t) t :: NonEmpty a)

type family SortWith (a :: (~>) a o) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

SortWith a_6989586621681166348 a_6989586621681166350 = Apply (Apply (Apply (Apply (.@#@$) SortBySym0) ComparingSym0) a_6989586621681166348) a_6989586621681166350

sSortWith :: forall a o (t :: (~>) a o) (t :: NonEmpty a). SOrd o => Sing t -> Sing t -> Sing (Apply (Apply SortWithSym0 t) t :: NonEmpty a)

type family Length (a :: NonEmpty a) :: Nat where ...

Equations

Length ('(:|) _ xs) = Apply (Apply (+@#@$) (FromInteger 1)) (Apply ListlengthSym0 xs)

sLength :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply LengthSym0 t :: Nat)

type family Head (a :: NonEmpty a) :: a where ...

Equations

Head ('(:|) a _) = a

sHead :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply HeadSym0 t :: a)

type family Tail (a :: NonEmpty a) :: [a] where ...

Equations

Tail ('(:|) _ as) = as

sTail :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply TailSym0 t :: [a])

type family Last (a :: NonEmpty a) :: a where ...

Equations

Last ('(:|) a as) = Apply ListlastSym0 (Apply (Apply (:@#@$) a) as)

sLast :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply LastSym0 t :: a)

type family Init (a :: NonEmpty a) :: [a] where ...

Equations

Init ('(:|) a as) = Apply ListinitSym0 (Apply (Apply (:@#@$) a) as)

sInit :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply InitSym0 t :: [a])

type family (a :: a) <| (a :: NonEmpty a) :: NonEmpty a where ...

Equations

a <| ('(:|) b bs) = Apply (Apply (:|@#@$) a) (Apply (Apply (:@#@$) b) bs)

(%<|) :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply (<|@#@$) t) t :: NonEmpty a)

type family Cons (a :: a) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Cons a_6989586621681166755 a_6989586621681166757 = Apply (Apply (<|@#@$) a_6989586621681166755) a_6989586621681166757

sCons :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply ConsSym0 t) t :: NonEmpty a)

type family Uncons (a :: NonEmpty a) :: (a, Maybe (NonEmpty a)) where ...

Equations

Uncons ('(:|) a as) = Apply (Apply Tuple2Sym0 a) (Apply NonEmpty_Sym0 as)

sUncons :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply UnconsSym0 t :: (a, Maybe (NonEmpty a)))

type family Unfoldr (a :: (~>) a (b, Maybe a)) (a :: a) :: NonEmpty b where ...

Equations

Unfoldr f a = Case_6989586621681166810 f a (Let6989586621681166807Scrutinee_6989586621681165307Sym2 f a)

sUnfoldr :: forall a b (t :: (~>) a (b, Maybe a)) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply UnfoldrSym0 t) t :: NonEmpty b)

type family Sort (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Sort a_6989586621681166750 = Apply (Apply LiftSym0 ListsortSym0) a_6989586621681166750

sSort :: forall a (t :: NonEmpty a). SOrd a => Sing t -> Sing (Apply SortSym0 t :: NonEmpty a)

type family Reverse (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Reverse a_6989586621681166658 = Apply (Apply LiftSym0 ListreverseSym0) a_6989586621681166658

sReverse :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply ReverseSym0 t :: NonEmpty a)

type family Inits (a :: [a]) :: NonEmpty [a] where ...

Equations

Inits a_6989586621681166722 = Apply (Apply (Apply (.@#@$) FromListSym0) ListinitsSym0) a_6989586621681166722

sInits :: forall a (t :: [a]). Sing t -> Sing (Apply InitsSym0 t :: NonEmpty [a])

type family Tails (a :: [a]) :: NonEmpty [a] where ...

Equations

Tails a_6989586621681166717 = Apply (Apply (Apply (.@#@$) FromListSym0) ListtailsSym0) a_6989586621681166717

sTails :: forall a (t :: [a]). Sing t -> Sing (Apply TailsSym0 t :: NonEmpty [a])

type family Unfold (a :: (~>) a (b, Maybe a)) (a :: a) :: NonEmpty b where ...

Equations

Unfold f a = Case_6989586621681166832 f a (Let6989586621681166829Scrutinee_6989586621681165297Sym2 f a)

sUnfold :: forall a b (t :: (~>) a (b, Maybe a)) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply UnfoldSym0 t) t :: NonEmpty b)

type family Insert (a :: a) (a :: [a]) :: NonEmpty a where ...

Equations

Insert a a_6989586621681166713 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply ListinsertSym0 a)) a_6989586621681166713

sInsert :: forall a (t :: a) (t :: [a]). SOrd a => Sing t -> Sing t -> Sing (Apply (Apply InsertSym0 t) t :: NonEmpty a)

type family Take (a :: Nat) (a :: NonEmpty a) :: [a] where ...

Equations

Take n a_6989586621681166654 = Apply (Apply (Apply (.@#@$) (Apply ListtakeSym0 n)) ToListSym0) a_6989586621681166654

sTake :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply TakeSym0 t) t :: [a])

type family Drop (a :: Nat) (a :: NonEmpty a) :: [a] where ...

Equations

Drop n a_6989586621681166646 = Apply (Apply (Apply (.@#@$) (Apply ListdropSym0 n)) ToListSym0) a_6989586621681166646

sDrop :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply DropSym0 t) t :: [a])

type family SplitAt (a :: Nat) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

SplitAt n a_6989586621681166638 = Apply (Apply (Apply (.@#@$) (Apply ListsplitAtSym0 n)) ToListSym0) a_6989586621681166638

sSplitAt :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SplitAtSym0 t) t :: ([a], [a]))

type family TakeWhile (a :: (~>) a Bool) (a :: NonEmpty a) :: [a] where ...

Equations

TakeWhile p a_6989586621681166630 = Apply (Apply (Apply (.@#@$) (Apply ListtakeWhileSym0 p)) ToListSym0) a_6989586621681166630

sTakeWhile :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply TakeWhileSym0 t) t :: [a])

type family DropWhile (a :: (~>) a Bool) (a :: NonEmpty a) :: [a] where ...

Equations

DropWhile p a_6989586621681166622 = Apply (Apply (Apply (.@#@$) (Apply ListdropWhileSym0 p)) ToListSym0) a_6989586621681166622

sDropWhile :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply DropWhileSym0 t) t :: [a])

type family Span (a :: (~>) a Bool) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

Span p a_6989586621681166614 = Apply (Apply (Apply (.@#@$) (Apply ListspanSym0 p)) ToListSym0) a_6989586621681166614

sSpan :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SpanSym0 t) t :: ([a], [a]))

type family Break (a :: (~>) a Bool) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

Break p a_6989586621681166606 = Apply (Apply SpanSym0 (Apply (Apply (.@#@$) NotSym0) p)) a_6989586621681166606

sBreak :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply BreakSym0 t) t :: ([a], [a]))

type family Filter (a :: (~>) a Bool) (a :: NonEmpty a) :: [a] where ...

Equations

Filter p a_6989586621681166598 = Apply (Apply (Apply (.@#@$) (Apply ListfilterSym0 p)) ToListSym0) a_6989586621681166598

sFilter :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply FilterSym0 t) t :: [a])

type family Partition (a :: (~>) a Bool) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

Partition p a_6989586621681166590 = Apply (Apply (Apply (.@#@$) (Apply ListpartitionSym0 p)) ToListSym0) a_6989586621681166590

sPartition :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply PartitionSym0 t) t :: ([a], [a]))

type family Group (a :: [a]) :: [NonEmpty a] where ...

Equations

Group a_6989586621681166581 = Apply (Apply GroupBySym0 (==@#@$)) a_6989586621681166581

sGroup :: forall a (t :: [a]). SEq a => Sing t -> Sing (Apply GroupSym0 t :: [NonEmpty a])

type family GroupBy (a :: (~>) a ((~>) a Bool)) (a :: [a]) :: [NonEmpty a] where ...

Equations

GroupBy eq0 a_6989586621681166537 = Apply (Apply (Let6989586621681166541GoSym2 eq0 a_6989586621681166537) eq0) a_6989586621681166537

sGroupBy :: forall a (t :: (~>) a ((~>) a Bool)) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply GroupBySym0 t) t :: [NonEmpty a])

type family GroupWith (a :: (~>) a b) (a :: [a]) :: [NonEmpty a] where ...

Equations

GroupWith f a_6989586621681166529 = Apply (Apply GroupBySym0 (Apply (Apply OnSym0 (==@#@$)) f)) a_6989586621681166529

sGroupWith :: forall a b (t :: (~>) a b) (t :: [a]). SEq b => Sing t -> Sing t -> Sing (Apply (Apply GroupWithSym0 t) t :: [NonEmpty a])

type family GroupAllWith (a :: (~>) a b) (a :: [a]) :: [NonEmpty a] where ...

Equations

GroupAllWith f a_6989586621681166521 = Apply (Apply (Apply (.@#@$) (Apply GroupWithSym0 f)) (Apply ListsortBySym0 (Apply (Apply OnSym0 CompareSym0) f))) a_6989586621681166521

sGroupAllWith :: forall a b (t :: (~>) a b) (t :: [a]). SOrd b => Sing t -> Sing t -> Sing (Apply (Apply GroupAllWithSym0 t) t :: [NonEmpty a])

type family Group1 (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

Group1 a_6989586621681166512 = Apply (Apply GroupBy1Sym0 (==@#@$)) a_6989586621681166512

sGroup1 :: forall a (t :: NonEmpty a). SEq a => Sing t -> Sing (Apply Group1Sym0 t :: NonEmpty (NonEmpty a))

type family GroupBy1 (a :: (~>) a ((~>) a Bool)) (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

GroupBy1 eq ('(:|) x xs) = Apply (Apply (:|@#@$) (Apply (Apply (:|@#@$) x) (Let6989586621681166488YsSym3 eq x xs))) (Apply (Apply GroupBySym0 eq) (Let6989586621681166488ZsSym3 eq x xs))

sGroupBy1 :: forall a (t :: (~>) a ((~>) a Bool)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply GroupBy1Sym0 t) t :: NonEmpty (NonEmpty a))

type family GroupWith1 (a :: (~>) a b) (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

GroupWith1 f a_6989586621681166477 = Apply (Apply GroupBy1Sym0 (Apply (Apply OnSym0 (==@#@$)) f)) a_6989586621681166477

sGroupWith1 :: forall a b (t :: (~>) a b) (t :: NonEmpty a). SEq b => Sing t -> Sing t -> Sing (Apply (Apply GroupWith1Sym0 t) t :: NonEmpty (NonEmpty a))

type family GroupAllWith1 (a :: (~>) a b) (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

GroupAllWith1 f a_6989586621681166469 = Apply (Apply (Apply (.@#@$) (Apply GroupWith1Sym0 f)) (Apply SortWithSym0 f)) a_6989586621681166469

sGroupAllWith1 :: forall a b (t :: (~>) a b) (t :: NonEmpty a). SOrd b => Sing t -> Sing t -> Sing (Apply (Apply GroupAllWith1Sym0 t) t :: NonEmpty (NonEmpty a))

type family IsPrefixOf (a :: [a]) (a :: NonEmpty a) :: Bool where ...

Equations

IsPrefixOf '[] _ = TrueSym0

IsPrefixOf ('(:) y ys) ('(:|) x xs) = Apply (Apply (&&@#@$) (Apply (Apply (==@#@$) y) x)) (Apply (Apply ListisPrefixOfSym0 ys) xs)

sIsPrefixOf :: forall a (t :: [a]) (t :: NonEmpty a). SEq a => Sing t -> Sing t -> Sing (Apply (Apply IsPrefixOfSym0 t) t :: Bool)

type family Nub (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Nub a_6989586621681166386 = Apply (Apply NubBySym0 (==@#@$)) a_6989586621681166386

sNub :: forall a (t :: NonEmpty a). SEq a => Sing t -> Sing (Apply NubSym0 t :: NonEmpty a)

type family NubBy (a :: (~>) a ((~>) a Bool)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

NubBy eq ('(:|) a as) = Apply (Apply (:|@#@$) a) (Apply (Apply ListnubBySym0 eq) (Apply (Apply ListfilterSym0 (Apply (Apply (Apply Lambda_6989586621681166378Sym0 eq) a) as)) as))

sNubBy :: forall a (t :: (~>) a ((~>) a Bool)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply NubBySym0 t) t :: NonEmpty a)

type family (a :: NonEmpty a) !! (a :: Nat) :: a where ...

Equations

arg_6989586621681165319 !! arg_6989586621681165321 = Case_6989586621681166445 arg_6989586621681165319 arg_6989586621681165321 (Apply (Apply Tuple2Sym0 arg_6989586621681165319) arg_6989586621681165321)

(%!!) :: forall a (t :: NonEmpty a) (t :: Nat). Sing t -> Sing t -> Sing (Apply (Apply (!!@#@$) t) t :: a)

type family Zip (a :: NonEmpty a) (a :: NonEmpty b) :: NonEmpty (a, b) where ...

Equations

Zip ('(:|) x xs) ('(:|) y ys) = Apply (Apply (:|@#@$) (Apply (Apply Tuple2Sym0 x) y)) (Apply (Apply ListzipSym0 xs) ys)

sZip :: forall a b (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing (Apply (Apply ZipSym0 t) t :: NonEmpty (a, b))

type family ZipWith (a :: (~>) a ((~>) b c)) (a :: NonEmpty a) (a :: NonEmpty b) :: NonEmpty c where ...

Equations

ZipWith f ('(:|) x xs) ('(:|) y ys) = Apply (Apply (:|@#@$) (Apply (Apply f x) y)) (Apply (Apply (Apply ListzipWithSym0 f) xs) ys)

sZipWith :: forall a b c (t :: (~>) a ((~>) b c)) (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ZipWithSym0 t) t) t :: NonEmpty c)

type family Unzip (a :: NonEmpty (a, b)) :: (NonEmpty a, NonEmpty b) where ...

Equations

Unzip ('(:|) '(a, b) asbs) = Apply (Apply Tuple2Sym0 (Apply (Apply (:|@#@$) a) (Let6989586621681166396AsSym3 a b asbs))) (Apply (Apply (:|@#@$) b) (Let6989586621681166396BsSym3 a b asbs))

sUnzip :: forall a b (t :: NonEmpty (a, b)). Sing t -> Sing (Apply UnzipSym0 t :: (NonEmpty a, NonEmpty b))

type family FromList (a :: [a]) :: NonEmpty a where ...

Equations

FromList ('(:) a as) = Apply (Apply (:|@#@$) a) as

FromList '[] = Apply ErrorSym0 "NonEmpty.fromList: empty list"

sFromList :: forall a (t :: [a]). Sing t -> Sing (Apply FromListSym0 t :: NonEmpty a)

type family ToList (a :: NonEmpty a) :: [a] where ...

Equations

ToList ('(:|) a as) = Apply (Apply (:@#@$) a) as

sToList :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply ToListSym0 t :: [a])

type family NonEmpty_ (a :: [a]) :: Maybe (NonEmpty a) where ...

Equations

NonEmpty_ '[] = NothingSym0

NonEmpty_ ('(:) a as) = Apply JustSym0 (Apply (Apply (:|@#@$) a) as)

sNonEmpty_ :: forall a (t :: [a]). Sing t -> Sing (Apply NonEmpty_Sym0 t :: Maybe (NonEmpty a))

type family Xor (a :: NonEmpty Bool) :: Bool where ...

Equations

Xor ('(:|) x xs) = Apply (Apply (Apply FoldrSym0 (Let6989586621681166842Xor'Sym2 x xs)) x) xs

sXor :: forall (t :: NonEmpty Bool). Sing t -> Sing (Apply XorSym0 t :: Bool)

Defunctionalization symbols

data (:|@#@$) :: forall (a6989586621679069839 :: Type). (~>) a6989586621679069839 ((~>) [a6989586621679069839] (NonEmpty (a6989586621679069839 :: Type))) infixr 5

data (:|@#@$$) (t6989586621679316782 :: a6989586621679069839 :: Type) :: (~>) [a6989586621679069839] (NonEmpty (a6989586621679069839 :: Type)) infixr 5

type (:|@#@$$$) (t6989586621679316782 :: a6989586621679069839) (t6989586621679316783 :: [a6989586621679069839]) = '(:|) t6989586621679316782 t6989586621679316783

data MapSym0 :: forall a6989586621681165113 b6989586621681165114. (~>) ((~>) a6989586621681165113 b6989586621681165114) ((~>) (NonEmpty a6989586621681165113) (NonEmpty b6989586621681165114))

data MapSym1 (a6989586621681166727 :: (~>) a6989586621681165113 b6989586621681165114) :: (~>) (NonEmpty a6989586621681165113) (NonEmpty b6989586621681165114)

type MapSym2 (a6989586621681166727 :: (~>) a6989586621681165113 b6989586621681165114) (a6989586621681166728 :: NonEmpty a6989586621681165113) = Map a6989586621681166727 a6989586621681166728

data IntersperseSym0 :: forall a6989586621681165103. (~>) a6989586621681165103 ((~>) (NonEmpty a6989586621681165103) (NonEmpty a6989586621681165103))

data IntersperseSym1 (a6989586621681166663 :: a6989586621681165103) :: (~>) (NonEmpty a6989586621681165103) (NonEmpty a6989586621681165103)

type IntersperseSym2 (a6989586621681166663 :: a6989586621681165103) (a6989586621681166664 :: NonEmpty a6989586621681165103) = Intersperse a6989586621681166663 a6989586621681166664

data ScanlSym0 :: forall b6989586621681165108 a6989586621681165109. (~>) ((~>) b6989586621681165108 ((~>) a6989586621681165109 b6989586621681165108)) ((~>) b6989586621681165108 ((~>) [a6989586621681165109] (NonEmpty b6989586621681165108)))

data ScanlSym1 (a6989586621681166698 :: (~>) b6989586621681165108 ((~>) a6989586621681165109 b6989586621681165108)) :: (~>) b6989586621681165108 ((~>) [a6989586621681165109] (NonEmpty b6989586621681165108))

data ScanlSym2 (a6989586621681166698 :: (~>) b6989586621681165108 ((~>) a6989586621681165109 b6989586621681165108)) (a6989586621681166699 :: b6989586621681165108) :: (~>) [a6989586621681165109] (NonEmpty b6989586621681165108)

type ScanlSym3 (a6989586621681166698 :: (~>) b6989586621681165108 ((~>) a6989586621681165109 b6989586621681165108)) (a6989586621681166699 :: b6989586621681165108) (a6989586621681166700 :: [a6989586621681165109]) = Scanl a6989586621681166698 a6989586621681166699 a6989586621681166700

data ScanrSym0 :: forall a6989586621681165106 b6989586621681165107. (~>) ((~>) a6989586621681165106 ((~>) b6989586621681165107 b6989586621681165107)) ((~>) b6989586621681165107 ((~>) [a6989586621681165106] (NonEmpty b6989586621681165107)))

data ScanrSym1 (a6989586621681166687 :: (~>) a6989586621681165106 ((~>) b6989586621681165107 b6989586621681165107)) :: (~>) b6989586621681165107 ((~>) [a6989586621681165106] (NonEmpty b6989586621681165107))

data ScanrSym2 (a6989586621681166687 :: (~>) a6989586621681165106 ((~>) b6989586621681165107 b6989586621681165107)) (a6989586621681166688 :: b6989586621681165107) :: (~>) [a6989586621681165106] (NonEmpty b6989586621681165107)

type ScanrSym3 (a6989586621681166687 :: (~>) a6989586621681165106 ((~>) b6989586621681165107 b6989586621681165107)) (a6989586621681166688 :: b6989586621681165107) (a6989586621681166689 :: [a6989586621681165106]) = Scanr a6989586621681166687 a6989586621681166688 a6989586621681166689

data Scanl1Sym0 :: forall a6989586621681165105. (~>) ((~>) a6989586621681165105 ((~>) a6989586621681165105 a6989586621681165105)) ((~>) (NonEmpty a6989586621681165105) (NonEmpty a6989586621681165105))

data Scanl1Sym1 (a6989586621681166680 :: (~>) a6989586621681165105 ((~>) a6989586621681165105 a6989586621681165105)) :: (~>) (NonEmpty a6989586621681165105) (NonEmpty a6989586621681165105)

type Scanl1Sym2 (a6989586621681166680 :: (~>) a6989586621681165105 ((~>) a6989586621681165105 a6989586621681165105)) (a6989586621681166681 :: NonEmpty a6989586621681165105) = Scanl1 a6989586621681166680 a6989586621681166681

data Scanr1Sym0 :: forall a6989586621681165104. (~>) ((~>) a6989586621681165104 ((~>) a6989586621681165104 a6989586621681165104)) ((~>) (NonEmpty a6989586621681165104) (NonEmpty a6989586621681165104))

data Scanr1Sym1 (a6989586621681166673 :: (~>) a6989586621681165104 ((~>) a6989586621681165104 a6989586621681165104)) :: (~>) (NonEmpty a6989586621681165104) (NonEmpty a6989586621681165104)

type Scanr1Sym2 (a6989586621681166673 :: (~>) a6989586621681165104 ((~>) a6989586621681165104 a6989586621681165104)) (a6989586621681166674 :: NonEmpty a6989586621681165104) = Scanr1 a6989586621681166673 a6989586621681166674

data TransposeSym0 :: forall a6989586621681165069. (~>) (NonEmpty (NonEmpty a6989586621681165069)) (NonEmpty (NonEmpty a6989586621681165069))

type TransposeSym1 (a6989586621681166368 :: NonEmpty (NonEmpty a6989586621681165069)) = Transpose a6989586621681166368

data SortBySym0 :: forall a6989586621681165068. (~>) ((~>) a6989586621681165068 ((~>) a6989586621681165068 Ordering)) ((~>) (NonEmpty a6989586621681165068) (NonEmpty a6989586621681165068))

data SortBySym1 (a6989586621681166358 :: (~>) a6989586621681165068 ((~>) a6989586621681165068 Ordering)) :: (~>) (NonEmpty a6989586621681165068) (NonEmpty a6989586621681165068)

type SortBySym2 (a6989586621681166358 :: (~>) a6989586621681165068 ((~>) a6989586621681165068 Ordering)) (a6989586621681166359 :: NonEmpty a6989586621681165068) = SortBy a6989586621681166358 a6989586621681166359

data SortWithSym0 :: forall a6989586621681165067 o6989586621681165066. (~>) ((~>) a6989586621681165067 o6989586621681165066) ((~>) (NonEmpty a6989586621681165067) (NonEmpty a6989586621681165067))

data SortWithSym1 (a6989586621681166352 :: (~>) a6989586621681165067 o6989586621681165066) :: (~>) (NonEmpty a6989586621681165067) (NonEmpty a6989586621681165067)

type SortWithSym2 (a6989586621681166352 :: (~>) a6989586621681165067 o6989586621681165066) (a6989586621681166353 :: NonEmpty a6989586621681165067) = SortWith a6989586621681166352 a6989586621681166353

data LengthSym0 :: forall a6989586621681165132. (~>) (NonEmpty a6989586621681165132) Nat

type LengthSym1 (a6989586621681166851 :: NonEmpty a6989586621681165132) = Length a6989586621681166851

data HeadSym0 :: forall a6989586621681165125. (~>) (NonEmpty a6989586621681165125) a6989586621681165125

type HeadSym1 (a6989586621681166783 :: NonEmpty a6989586621681165125) = Head a6989586621681166783

data TailSym0 :: forall a6989586621681165124. (~>) (NonEmpty a6989586621681165124) [a6989586621681165124]

type TailSym1 (a6989586621681166780 :: NonEmpty a6989586621681165124) = Tail a6989586621681166780

data LastSym0 :: forall a6989586621681165123. (~>) (NonEmpty a6989586621681165123) a6989586621681165123

type LastSym1 (a6989586621681166776 :: NonEmpty a6989586621681165123) = Last a6989586621681166776

data InitSym0 :: forall a6989586621681165122. (~>) (NonEmpty a6989586621681165122) [a6989586621681165122]

type InitSym1 (a6989586621681166772 :: NonEmpty a6989586621681165122) = Init a6989586621681166772

data (<|@#@$) :: forall a6989586621681165121. (~>) a6989586621681165121 ((~>) (NonEmpty a6989586621681165121) (NonEmpty a6989586621681165121))

data (<|@#@$$) (a6989586621681166765 :: a6989586621681165121) :: (~>) (NonEmpty a6989586621681165121) (NonEmpty a6989586621681165121)

type (<|@#@$$$) (a6989586621681166765 :: a6989586621681165121) (a6989586621681166766 :: NonEmpty a6989586621681165121) = (<|) a6989586621681166765 a6989586621681166766

data ConsSym0 :: forall a6989586621681165120. (~>) a6989586621681165120 ((~>) (NonEmpty a6989586621681165120) (NonEmpty a6989586621681165120))

data ConsSym1 (a6989586621681166759 :: a6989586621681165120) :: (~>) (NonEmpty a6989586621681165120) (NonEmpty a6989586621681165120)

type ConsSym2 (a6989586621681166759 :: a6989586621681165120) (a6989586621681166760 :: NonEmpty a6989586621681165120) = Cons a6989586621681166759 a6989586621681166760

data UnconsSym0 :: forall a6989586621681165128. (~>) (NonEmpty a6989586621681165128) (a6989586621681165128, Maybe (NonEmpty a6989586621681165128))

type UnconsSym1 (a6989586621681166815 :: NonEmpty a6989586621681165128) = Uncons a6989586621681166815

data UnfoldrSym0 :: forall a6989586621681165126 b6989586621681165127. (~>) ((~>) a6989586621681165126 (b6989586621681165127, Maybe a6989586621681165126)) ((~>) a6989586621681165126 (NonEmpty b6989586621681165127))

data UnfoldrSym1 (a6989586621681166786 :: (~>) a6989586621681165126 (b6989586621681165127, Maybe a6989586621681165126)) :: (~>) a6989586621681165126 (NonEmpty b6989586621681165127)

type UnfoldrSym2 (a6989586621681166786 :: (~>) a6989586621681165126 (b6989586621681165127, Maybe a6989586621681165126)) (a6989586621681166787 :: a6989586621681165126) = Unfoldr a6989586621681166786 a6989586621681166787

data SortSym0 :: forall a6989586621681165119. (~>) (NonEmpty a6989586621681165119) (NonEmpty a6989586621681165119)

type SortSym1 (a6989586621681166752 :: NonEmpty a6989586621681165119) = Sort a6989586621681166752

data ReverseSym0 :: forall a6989586621681165102. (~>) (NonEmpty a6989586621681165102) (NonEmpty a6989586621681165102)

type ReverseSym1 (a6989586621681166660 :: NonEmpty a6989586621681165102) = Reverse a6989586621681166660

data InitsSym0 :: forall a6989586621681165112. (~>) [a6989586621681165112] (NonEmpty [a6989586621681165112])

type InitsSym1 (a6989586621681166724 :: [a6989586621681165112]) = Inits a6989586621681166724

data TailsSym0 :: forall a6989586621681165111. (~>) [a6989586621681165111] (NonEmpty [a6989586621681165111])

type TailsSym1 (a6989586621681166719 :: [a6989586621681165111]) = Tails a6989586621681166719

data UnfoldSym0 :: forall a6989586621681165130 b6989586621681165131. (~>) ((~>) a6989586621681165130 (b6989586621681165131, Maybe a6989586621681165130)) ((~>) a6989586621681165130 (NonEmpty b6989586621681165131))

data UnfoldSym1 (a6989586621681166823 :: (~>) a6989586621681165130 (b6989586621681165131, Maybe a6989586621681165130)) :: (~>) a6989586621681165130 (NonEmpty b6989586621681165131)

data InsertSym0 :: forall a6989586621681165110. (~>) a6989586621681165110 ((~>) [a6989586621681165110] (NonEmpty a6989586621681165110))

data InsertSym1 (a6989586621681166709 :: a6989586621681165110) :: (~>) [a6989586621681165110] (NonEmpty a6989586621681165110)

type InsertSym2 (a6989586621681166709 :: a6989586621681165110) (a6989586621681166710 :: [a6989586621681165110]) = Insert a6989586621681166709 a6989586621681166710

data TakeSym0 :: forall a6989586621681165101. (~>) Nat ((~>) (NonEmpty a6989586621681165101) [a6989586621681165101])

data TakeSym1 (a6989586621681166650 :: Nat) :: forall a6989586621681165101. (~>) (NonEmpty a6989586621681165101) [a6989586621681165101]

type TakeSym2 (a6989586621681166650 :: Nat) (a6989586621681166651 :: NonEmpty a6989586621681165101) = Take a6989586621681166650 a6989586621681166651

data DropSym0 :: forall a6989586621681165100. (~>) Nat ((~>) (NonEmpty a6989586621681165100) [a6989586621681165100])

data DropSym1 (a6989586621681166642 :: Nat) :: forall a6989586621681165100. (~>) (NonEmpty a6989586621681165100) [a6989586621681165100]

type DropSym2 (a6989586621681166642 :: Nat) (a6989586621681166643 :: NonEmpty a6989586621681165100) = Drop a6989586621681166642 a6989586621681166643

data SplitAtSym0 :: forall a6989586621681165099. (~>) Nat ((~>) (NonEmpty a6989586621681165099) ([a6989586621681165099], [a6989586621681165099]))

data SplitAtSym1 (a6989586621681166634 :: Nat) :: forall a6989586621681165099. (~>) (NonEmpty a6989586621681165099) ([a6989586621681165099], [a6989586621681165099])

type SplitAtSym2 (a6989586621681166634 :: Nat) (a6989586621681166635 :: NonEmpty a6989586621681165099) = SplitAt a6989586621681166634 a6989586621681166635

data TakeWhileSym0 :: forall a6989586621681165098. (~>) ((~>) a6989586621681165098 Bool) ((~>) (NonEmpty a6989586621681165098) [a6989586621681165098])

data TakeWhileSym1 (a6989586621681166626 :: (~>) a6989586621681165098 Bool) :: (~>) (NonEmpty a6989586621681165098) [a6989586621681165098]

type TakeWhileSym2 (a6989586621681166626 :: (~>) a6989586621681165098 Bool) (a6989586621681166627 :: NonEmpty a6989586621681165098) = TakeWhile a6989586621681166626 a6989586621681166627

data DropWhileSym0 :: forall a6989586621681165097. (~>) ((~>) a6989586621681165097 Bool) ((~>) (NonEmpty a6989586621681165097) [a6989586621681165097])

data DropWhileSym1 (a6989586621681166618 :: (~>) a6989586621681165097 Bool) :: (~>) (NonEmpty a6989586621681165097) [a6989586621681165097]

type DropWhileSym2 (a6989586621681166618 :: (~>) a6989586621681165097 Bool) (a6989586621681166619 :: NonEmpty a6989586621681165097) = DropWhile a6989586621681166618 a6989586621681166619

data SpanSym0 :: forall a6989586621681165096. (~>) ((~>) a6989586621681165096 Bool) ((~>) (NonEmpty a6989586621681165096) ([a6989586621681165096], [a6989586621681165096]))

data SpanSym1 (a6989586621681166610 :: (~>) a6989586621681165096 Bool) :: (~>) (NonEmpty a6989586621681165096) ([a6989586621681165096], [a6989586621681165096])

type SpanSym2 (a6989586621681166610 :: (~>) a6989586621681165096 Bool) (a6989586621681166611 :: NonEmpty a6989586621681165096) = Span a6989586621681166610 a6989586621681166611

data BreakSym0 :: forall a6989586621681165095. (~>) ((~>) a6989586621681165095 Bool) ((~>) (NonEmpty a6989586621681165095) ([a6989586621681165095], [a6989586621681165095]))

data BreakSym1 (a6989586621681166602 :: (~>) a6989586621681165095 Bool) :: (~>) (NonEmpty a6989586621681165095) ([a6989586621681165095], [a6989586621681165095])

type BreakSym2 (a6989586621681166602 :: (~>) a6989586621681165095 Bool) (a6989586621681166603 :: NonEmpty a6989586621681165095) = Break a6989586621681166602 a6989586621681166603

data FilterSym0 :: forall a6989586621681165094. (~>) ((~>) a6989586621681165094 Bool) ((~>) (NonEmpty a6989586621681165094) [a6989586621681165094])

data FilterSym1 (a6989586621681166594 :: (~>) a6989586621681165094 Bool) :: (~>) (NonEmpty a6989586621681165094) [a6989586621681165094]

type FilterSym2 (a6989586621681166594 :: (~>) a6989586621681165094 Bool) (a6989586621681166595 :: NonEmpty a6989586621681165094) = Filter a6989586621681166594 a6989586621681166595

data PartitionSym0 :: forall a6989586621681165093. (~>) ((~>) a6989586621681165093 Bool) ((~>) (NonEmpty a6989586621681165093) ([a6989586621681165093], [a6989586621681165093]))

data PartitionSym1 (a6989586621681166586 :: (~>) a6989586621681165093 Bool) :: (~>) (NonEmpty a6989586621681165093) ([a6989586621681165093], [a6989586621681165093])

type PartitionSym2 (a6989586621681166586 :: (~>) a6989586621681165093 Bool) (a6989586621681166587 :: NonEmpty a6989586621681165093) = Partition a6989586621681166586 a6989586621681166587

data GroupSym0 :: forall a6989586621681165092. (~>) [a6989586621681165092] [NonEmpty a6989586621681165092]

type GroupSym1 (a6989586621681166583 :: [a6989586621681165092]) = Group a6989586621681166583

data GroupBySym0 :: forall a6989586621681165091. (~>) ((~>) a6989586621681165091 ((~>) a6989586621681165091 Bool)) ((~>) [a6989586621681165091] [NonEmpty a6989586621681165091])

data GroupBySym1 (a6989586621681166533 :: (~>) a6989586621681165091 ((~>) a6989586621681165091 Bool)) :: (~>) [a6989586621681165091] [NonEmpty a6989586621681165091]

type GroupBySym2 (a6989586621681166533 :: (~>) a6989586621681165091 ((~>) a6989586621681165091 Bool)) (a6989586621681166534 :: [a6989586621681165091]) = GroupBy a6989586621681166533 a6989586621681166534

data GroupWithSym0 :: forall a6989586621681165090 b6989586621681165089. (~>) ((~>) a6989586621681165090 b6989586621681165089) ((~>) [a6989586621681165090] [NonEmpty a6989586621681165090])

data GroupWithSym1 (a6989586621681166525 :: (~>) a6989586621681165090 b6989586621681165089) :: (~>) [a6989586621681165090] [NonEmpty a6989586621681165090]

type GroupWithSym2 (a6989586621681166525 :: (~>) a6989586621681165090 b6989586621681165089) (a6989586621681166526 :: [a6989586621681165090]) = GroupWith a6989586621681166525 a6989586621681166526

data GroupAllWithSym0 :: forall a6989586621681165088 b6989586621681165087. (~>) ((~>) a6989586621681165088 b6989586621681165087) ((~>) [a6989586621681165088] [NonEmpty a6989586621681165088])

data GroupAllWithSym1 (a6989586621681166517 :: (~>) a6989586621681165088 b6989586621681165087) :: (~>) [a6989586621681165088] [NonEmpty a6989586621681165088]

type GroupAllWithSym2 (a6989586621681166517 :: (~>) a6989586621681165088 b6989586621681165087) (a6989586621681166518 :: [a6989586621681165088]) = GroupAllWith a6989586621681166517 a6989586621681166518

data Group1Sym0 :: forall a6989586621681165086. (~>) (NonEmpty a6989586621681165086) (NonEmpty (NonEmpty a6989586621681165086))

type Group1Sym1 (a6989586621681166514 :: NonEmpty a6989586621681165086) = Group1 a6989586621681166514

data GroupBy1Sym0 :: forall a6989586621681165085. (~>) ((~>) a6989586621681165085 ((~>) a6989586621681165085 Bool)) ((~>) (NonEmpty a6989586621681165085) (NonEmpty (NonEmpty a6989586621681165085)))

data GroupBy1Sym1 (a6989586621681166481 :: (~>) a6989586621681165085 ((~>) a6989586621681165085 Bool)) :: (~>) (NonEmpty a6989586621681165085) (NonEmpty (NonEmpty a6989586621681165085))

type GroupBy1Sym2 (a6989586621681166481 :: (~>) a6989586621681165085 ((~>) a6989586621681165085 Bool)) (a6989586621681166482 :: NonEmpty a6989586621681165085) = GroupBy1 a6989586621681166481 a6989586621681166482

data GroupWith1Sym0 :: forall a6989586621681165084 b6989586621681165083. (~>) ((~>) a6989586621681165084 b6989586621681165083) ((~>) (NonEmpty a6989586621681165084) (NonEmpty (NonEmpty a6989586621681165084)))

data GroupWith1Sym1 (a6989586621681166473 :: (~>) a6989586621681165084 b6989586621681165083) :: (~>) (NonEmpty a6989586621681165084) (NonEmpty (NonEmpty a6989586621681165084))

type GroupWith1Sym2 (a6989586621681166473 :: (~>) a6989586621681165084 b6989586621681165083) (a6989586621681166474 :: NonEmpty a6989586621681165084) = GroupWith1 a6989586621681166473 a6989586621681166474

data GroupAllWith1Sym0 :: forall a6989586621681165082 b6989586621681165081. (~>) ((~>) a6989586621681165082 b6989586621681165081) ((~>) (NonEmpty a6989586621681165082) (NonEmpty (NonEmpty a6989586621681165082)))

data GroupAllWith1Sym1 (a6989586621681166465 :: (~>) a6989586621681165082 b6989586621681165081) :: (~>) (NonEmpty a6989586621681165082) (NonEmpty (NonEmpty a6989586621681165082))

type GroupAllWith1Sym2 (a6989586621681166465 :: (~>) a6989586621681165082 b6989586621681165081) (a6989586621681166466 :: NonEmpty a6989586621681165082) = GroupAllWith1 a6989586621681166465 a6989586621681166466

data IsPrefixOfSym0 :: forall a6989586621681165080. (~>) [a6989586621681165080] ((~>) (NonEmpty a6989586621681165080) Bool)

data IsPrefixOfSym1 (a6989586621681166457 :: [a6989586621681165080]) :: (~>) (NonEmpty a6989586621681165080) Bool

type IsPrefixOfSym2 (a6989586621681166457 :: [a6989586621681165080]) (a6989586621681166458 :: NonEmpty a6989586621681165080) = IsPrefixOf a6989586621681166457 a6989586621681166458

data NubSym0 :: forall a6989586621681165071. (~>) (NonEmpty a6989586621681165071) (NonEmpty a6989586621681165071)

type NubSym1 (a6989586621681166388 :: NonEmpty a6989586621681165071) = Nub a6989586621681166388

data NubBySym0 :: forall a6989586621681165070. (~>) ((~>) a6989586621681165070 ((~>) a6989586621681165070 Bool)) ((~>) (NonEmpty a6989586621681165070) (NonEmpty a6989586621681165070))

data NubBySym1 (a6989586621681166371 :: (~>) a6989586621681165070 ((~>) a6989586621681165070 Bool)) :: (~>) (NonEmpty a6989586621681165070) (NonEmpty a6989586621681165070)

type NubBySym2 (a6989586621681166371 :: (~>) a6989586621681165070 ((~>) a6989586621681165070 Bool)) (a6989586621681166372 :: NonEmpty a6989586621681165070) = NubBy a6989586621681166371 a6989586621681166372

data (!!@#@$) :: forall a6989586621681165079. (~>) (NonEmpty a6989586621681165079) ((~>) Nat a6989586621681165079)

data (!!@#@$$) (a6989586621681166439 :: NonEmpty a6989586621681165079) :: (~>) Nat a6989586621681165079

type (!!@#@$$$) (a6989586621681166439 :: NonEmpty a6989586621681165079) (a6989586621681166440 :: Nat) = (!!) a6989586621681166439 a6989586621681166440

data ZipSym0 :: forall a6989586621681165077 b6989586621681165078. (~>) (NonEmpty a6989586621681165077) ((~>) (NonEmpty b6989586621681165078) (NonEmpty (a6989586621681165077, b6989586621681165078)))

data ZipSym1 (a6989586621681166431 :: NonEmpty a6989586621681165077) :: forall b6989586621681165078. (~>) (NonEmpty b6989586621681165078) (NonEmpty (a6989586621681165077, b6989586621681165078))

type ZipSym2 (a6989586621681166431 :: NonEmpty a6989586621681165077) (a6989586621681166432 :: NonEmpty b6989586621681165078) = Zip a6989586621681166431 a6989586621681166432

data ZipWithSym0 :: forall a6989586621681165074 b6989586621681165075 c6989586621681165076. (~>) ((~>) a6989586621681165074 ((~>) b6989586621681165075 c6989586621681165076)) ((~>) (NonEmpty a6989586621681165074) ((~>) (NonEmpty b6989586621681165075) (NonEmpty c6989586621681165076)))

data ZipWithSym1 (a6989586621681166420 :: (~>) a6989586621681165074 ((~>) b6989586621681165075 c6989586621681165076)) :: (~>) (NonEmpty a6989586621681165074) ((~>) (NonEmpty b6989586621681165075) (NonEmpty c6989586621681165076))

data ZipWithSym2 (a6989586621681166420 :: (~>) a6989586621681165074 ((~>) b6989586621681165075 c6989586621681165076)) (a6989586621681166421 :: NonEmpty a6989586621681165074) :: (~>) (NonEmpty b6989586621681165075) (NonEmpty c6989586621681165076)

type ZipWithSym3 (a6989586621681166420 :: (~>) a6989586621681165074 ((~>) b6989586621681165075 c6989586621681165076)) (a6989586621681166421 :: NonEmpty a6989586621681165074) (a6989586621681166422 :: NonEmpty b6989586621681165075) = ZipWith a6989586621681166420 a6989586621681166421 a6989586621681166422

data UnzipSym0 :: forall a6989586621681165072 b6989586621681165073. (~>) (NonEmpty (a6989586621681165072, b6989586621681165073)) (NonEmpty a6989586621681165072, NonEmpty b6989586621681165073)

type UnzipSym1 (a6989586621681166391 :: NonEmpty (a6989586621681165072, b6989586621681165073)) = Unzip a6989586621681166391

data FromListSym0 :: forall a6989586621681165118. (~>) [a6989586621681165118] (NonEmpty a6989586621681165118)

type FromListSym1 (a6989586621681166746 :: [a6989586621681165118]) = FromList a6989586621681166746

data ToListSym0 :: forall a6989586621681165117. (~>) (NonEmpty a6989586621681165117) [a6989586621681165117]

type ToListSym1 (a6989586621681166742 :: NonEmpty a6989586621681165117) = ToList a6989586621681166742

data NonEmpty_Sym0 :: forall a6989586621681165129. (~>) [a6989586621681165129] (Maybe (NonEmpty a6989586621681165129))

type NonEmpty_Sym1 (a6989586621681166819 :: [a6989586621681165129]) = NonEmpty_ a6989586621681166819

data XorSym0 :: (~>) (NonEmpty Bool) Bool

type XorSym1 (a6989586621681166838 :: NonEmpty Bool) = Xor a6989586621681166838

Orphan instances

SMonadZip NonEmpty

Instance details

Methods

sMzip :: forall a b (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing (Apply (Apply MzipSym0 t) t)

sMzipWith :: forall a b c (t :: a ~> (b ~> c)) (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply MzipWithSym0 t) t) t)

sMunzip :: forall a b (t :: NonEmpty (a, b)). Sing t -> Sing (Apply MunzipSym0 t)

PMonadZip NonEmpty

Instance details

Associated Types

type Mzip arg arg :: m (a, b)

type MzipWith arg arg arg :: m c

type Munzip arg :: (m a, m b)