Data.Singletons.Prelude.List.NonEmpty

Description

Defines functions and datatypes relating to the singleton for NonEmpty, including a singletons version of all the definitions in Data.List.NonEmpty.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.List.NonEmpty. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

The NonEmpty singleton

data family Sing :: k -> Type infixr 5

The singleton kind-indexed data family.

Though Haddock doesn't show it, the Sing instance above declares constructor

(:%|) :: Sing h -> Sing t -> Sing (h :| t)

type SNonEmpty = (Sing :: NonEmpty a -> Type)

SNonEmpty is a kind-restricted synonym for Sing: type SNonEmpty (a :: NonEmpty) = Sing a

Non-empty stream transformations

type family Map (a :: (~>) a b) (a :: NonEmpty a) :: NonEmpty b where ...

Equations

Map f ((:|) a as) = Apply (Apply (:|@#@$) (Apply f a)) (Apply (Apply ListmapSym0 f) as)

sMap :: forall a b (t :: (~>) a b) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply MapSym0 t) t :: NonEmpty b)

type family Intersperse (a :: a) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Intersperse a ((:|) b bs) = Apply (Apply (:|@#@$) b) (Case_6989586621681195483 a b bs bs)

sIntersperse :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply IntersperseSym0 t) t :: NonEmpty a)

type family Scanl (a :: (~>) b ((~>) a b)) (a :: b) (a :: [a]) :: NonEmpty b where ...

Equations

Scanl f z a_6989586621681195636 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply ListscanlSym0 f) z)) a_6989586621681195636

sScanl :: forall b a (t :: (~>) b ((~>) a b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanlSym0 t) t) t :: NonEmpty b)

type family Scanr (a :: (~>) a ((~>) b b)) (a :: b) (a :: [a]) :: NonEmpty b where ...

Equations

Scanr f z a_6989586621681195647 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply ListscanrSym0 f) z)) a_6989586621681195647

sScanr :: forall a b (t :: (~>) a ((~>) b b)) (t :: b) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ScanrSym0 t) t) t :: NonEmpty b)

type family Scanl1 (a :: (~>) a ((~>) a a)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Scanl1 f ((:|) a as) = Apply FromListSym0 (Apply (Apply (Apply ListscanlSym0 f) a) as)

sScanl1 :: forall a (t :: (~>) a ((~>) a a)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply Scanl1Sym0 t) t :: NonEmpty a)

type family Scanr1 (a :: (~>) a ((~>) a a)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Scanr1 f ((:|) a as) = Apply FromListSym0 (Apply (Apply Listscanr1Sym0 f) (Apply (Apply (:@#@$) a) as))

sScanr1 :: forall a (t :: (~>) a ((~>) a a)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply Scanr1Sym0 t) t :: NonEmpty a)

type family Transpose (a :: NonEmpty (NonEmpty a)) :: NonEmpty (NonEmpty a) where ...

Equations

Transpose a_6989586621681195777 = Apply (Apply (Apply (.@#@$) (Apply FmapSym0 FromListSym0)) (Apply (Apply (.@#@$) FromListSym0) (Apply (Apply (.@#@$) ListtransposeSym0) (Apply (Apply (.@#@$) ToListSym0) (Apply FmapSym0 ToListSym0))))) a_6989586621681195777

sTranspose :: forall a (t :: NonEmpty (NonEmpty a)). Sing t -> Sing (Apply TransposeSym0 t :: NonEmpty (NonEmpty a))

type family SortBy (a :: (~>) a ((~>) a Ordering)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

SortBy f a_6989586621681195590 = Apply (Apply LiftSym0 (Apply ListsortBySym0 f)) a_6989586621681195590

sSortBy :: forall a (t :: (~>) a ((~>) a Ordering)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SortBySym0 t) t :: NonEmpty a)

type family SortWith (a :: (~>) a o) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

SortWith a_6989586621681195594 a_6989586621681195596 = Apply (Apply (Apply (Apply (.@#@$) SortBySym0) ComparingSym0) a_6989586621681195594) a_6989586621681195596

sSortWith :: forall o a (t :: (~>) a o) (t :: NonEmpty a). SOrd o => Sing t -> Sing t -> Sing (Apply (Apply SortWithSym0 t) t :: NonEmpty a)

type family Length (a :: NonEmpty a) :: Nat where ...

Equations

Length ((:|) _ xs) = Apply (Apply (+@#@$) (FromInteger 1)) (Apply ListlengthSym0 xs)

sLength :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply LengthSym0 t :: Nat)

type family Head (a :: NonEmpty a) :: a where ...

Equations

Head ((:|) a _) = a

sHead :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply HeadSym0 t :: a)

type family Tail (a :: NonEmpty a) :: [a] where ...

Equations

Tail ((:|) _ as) = as

sTail :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply TailSym0 t :: [a])

type family Last (a :: NonEmpty a) :: a where ...

Equations

Last ((:|) a as) = Apply ListlastSym0 (Apply (Apply (:@#@$) a) as)

sLast :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply LastSym0 t :: a)

type family Init (a :: NonEmpty a) :: [a] where ...

Equations

Init ((:|) a as) = Apply ListinitSym0 (Apply (Apply (:@#@$) a) as)

sInit :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply InitSym0 t :: [a])

type family (a :: a) <| (a :: NonEmpty a) :: NonEmpty a where ...

Equations

a <| ((:|) b bs) = Apply (Apply (:|@#@$) a) (Apply (Apply (:@#@$) b) bs)

(%<|) :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply (<|@#@$) t) t :: NonEmpty a)

type family Cons (a :: a) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Cons a_6989586621681195678 a_6989586621681195680 = Apply (Apply (<|@#@$) a_6989586621681195678) a_6989586621681195680

sCons :: forall a (t :: a) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply ConsSym0 t) t :: NonEmpty a)

type family Uncons (a :: NonEmpty a) :: (a, Maybe (NonEmpty a)) where ...

Equations

Uncons ((:|) a as) = Apply (Apply Tuple2Sym0 a) (Apply NonEmpty_Sym0 as)

sUncons :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply UnconsSym0 t :: (a, Maybe (NonEmpty a)))

type family Unfoldr (a :: (~>) a (b, Maybe a)) (a :: a) :: NonEmpty b where ...

Equations

Unfoldr f a = Case_6989586621681195726 f a (Let6989586621681195723Scrutinee_6989586621681194059Sym2 f a)

sUnfoldr :: forall a b (t :: (~>) a (b, Maybe a)) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply UnfoldrSym0 t) t :: NonEmpty b)

type family Sort (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Sort a_6989586621681195666 = Apply (Apply LiftSym0 ListsortSym0) a_6989586621681195666

sSort :: forall a (t :: NonEmpty a). SOrd a => Sing t -> Sing (Apply SortSym0 t :: NonEmpty a)

type family Reverse (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Reverse a_6989586621681195581 = Apply (Apply LiftSym0 ListreverseSym0) a_6989586621681195581

sReverse :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply ReverseSym0 t :: NonEmpty a)

type family Inits (a :: [a]) :: NonEmpty [a] where ...

Equations

Inits a_6989586621681195612 = Apply (Apply (Apply (.@#@$) FromListSym0) ListinitsSym0) a_6989586621681195612

sInits :: forall a (t :: [a]). Sing t -> Sing (Apply InitsSym0 t :: NonEmpty [a])

type family Tails (a :: [a]) :: NonEmpty [a] where ...

Equations

Tails a_6989586621681195617 = Apply (Apply (Apply (.@#@$) FromListSym0) ListtailsSym0) a_6989586621681195617

sTails :: forall a (t :: [a]). Sing t -> Sing (Apply TailsSym0 t :: NonEmpty [a])

type family Unfold (a :: (~>) a (b, Maybe a)) (a :: a) :: NonEmpty b where ...

Equations

Unfold f a = Case_6989586621681195748 f a (Let6989586621681195745Scrutinee_6989586621681194049Sym2 f a)

sUnfold :: forall a b (t :: (~>) a (b, Maybe a)) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply UnfoldSym0 t) t :: NonEmpty b)

type family Insert (a :: a) (a :: [a]) :: NonEmpty a where ...

Equations

Insert a a_6989586621681195626 = Apply (Apply (Apply (.@#@$) FromListSym0) (Apply ListinsertSym0 a)) a_6989586621681195626

sInsert :: forall a (t :: a) (t :: [a]). SOrd a => Sing t -> Sing t -> Sing (Apply (Apply InsertSym0 t) t :: NonEmpty a)

type family Take (a :: Nat) (a :: NonEmpty a) :: [a] where ...

Equations

Take n a_6989586621681195501 = Apply (Apply (Apply (.@#@$) (Apply ListtakeSym0 n)) ToListSym0) a_6989586621681195501

sTake :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply TakeSym0 t) t :: [a])

type family Drop (a :: Nat) (a :: NonEmpty a) :: [a] where ...

Equations

Drop n a_6989586621681195509 = Apply (Apply (Apply (.@#@$) (Apply ListdropSym0 n)) ToListSym0) a_6989586621681195509

sDrop :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply DropSym0 t) t :: [a])

type family SplitAt (a :: Nat) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

SplitAt n a_6989586621681195517 = Apply (Apply (Apply (.@#@$) (Apply ListsplitAtSym0 n)) ToListSym0) a_6989586621681195517

sSplitAt :: forall a (t :: Nat) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SplitAtSym0 t) t :: ([a], [a]))

type family TakeWhile (a :: (~>) a Bool) (a :: NonEmpty a) :: [a] where ...

Equations

TakeWhile p a_6989586621681195525 = Apply (Apply (Apply (.@#@$) (Apply ListtakeWhileSym0 p)) ToListSym0) a_6989586621681195525

sTakeWhile :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply TakeWhileSym0 t) t :: [a])

type family DropWhile (a :: (~>) a Bool) (a :: NonEmpty a) :: [a] where ...

Equations

DropWhile p a_6989586621681195533 = Apply (Apply (Apply (.@#@$) (Apply ListdropWhileSym0 p)) ToListSym0) a_6989586621681195533

sDropWhile :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply DropWhileSym0 t) t :: [a])

type family Span (a :: (~>) a Bool) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

Span p a_6989586621681195541 = Apply (Apply (Apply (.@#@$) (Apply ListspanSym0 p)) ToListSym0) a_6989586621681195541

sSpan :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply SpanSym0 t) t :: ([a], [a]))

type family Break (a :: (~>) a Bool) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

Break p a_6989586621681195549 = Apply (Apply SpanSym0 (Apply (Apply (.@#@$) NotSym0) p)) a_6989586621681195549

sBreak :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply BreakSym0 t) t :: ([a], [a]))

type family Filter (a :: (~>) a Bool) (a :: NonEmpty a) :: [a] where ...

Equations

Filter p a_6989586621681195557 = Apply (Apply (Apply (.@#@$) (Apply ListfilterSym0 p)) ToListSym0) a_6989586621681195557

sFilter :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply FilterSym0 t) t :: [a])

type family Partition (a :: (~>) a Bool) (a :: NonEmpty a) :: ([a], [a]) where ...

Equations

Partition p a_6989586621681195565 = Apply (Apply (Apply (.@#@$) (Apply ListpartitionSym0 p)) ToListSym0) a_6989586621681195565

sPartition :: forall a (t :: (~>) a Bool) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply PartitionSym0 t) t :: ([a], [a]))

type family Group (a :: [a]) :: [NonEmpty a] where ...

Equations

Group a_6989586621681195471 = Apply (Apply GroupBySym0 (==@#@$)) a_6989586621681195471

sGroup :: forall a (t :: [a]). SEq a => Sing t -> Sing (Apply GroupSym0 t :: [NonEmpty a])

type family GroupBy (a :: (~>) a ((~>) a Bool)) (a :: [a]) :: [NonEmpty a] where ...

Equations

GroupBy eq0 a_6989586621681195367 = Apply (Apply (Let6989586621681195371GoSym2 eq0 a_6989586621681195367) eq0) a_6989586621681195367

sGroupBy :: forall a (t :: (~>) a ((~>) a Bool)) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply GroupBySym0 t) t :: [NonEmpty a])

type family GroupWith (a :: (~>) a b) (a :: [a]) :: [NonEmpty a] where ...

Equations

GroupWith f a_6989586621681195415 = Apply (Apply GroupBySym0 (Apply (Apply OnSym0 (==@#@$)) f)) a_6989586621681195415

sGroupWith :: forall b a (t :: (~>) a b) (t :: [a]). SEq b => Sing t -> Sing t -> Sing (Apply (Apply GroupWithSym0 t) t :: [NonEmpty a])

type family GroupAllWith (a :: (~>) a b) (a :: [a]) :: [NonEmpty a] where ...

Equations

GroupAllWith f a_6989586621681195423 = Apply (Apply (Apply (.@#@$) (Apply GroupWithSym0 f)) (Apply ListsortBySym0 (Apply (Apply OnSym0 CompareSym0) f))) a_6989586621681195423

sGroupAllWith :: forall b a (t :: (~>) a b) (t :: [a]). SOrd b => Sing t -> Sing t -> Sing (Apply (Apply GroupAllWithSym0 t) t :: [NonEmpty a])

type family Group1 (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

Group1 a_6989586621681195458 = Apply (Apply GroupBy1Sym0 (==@#@$)) a_6989586621681195458

sGroup1 :: forall a (t :: NonEmpty a). SEq a => Sing t -> Sing (Apply Group1Sym0 t :: NonEmpty (NonEmpty a))

type family GroupBy1 (a :: (~>) a ((~>) a Bool)) (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

GroupBy1 eq ((:|) x xs) = Apply (Apply (:|@#@$) (Apply (Apply (:|@#@$) x) (Let6989586621681195434YsSym3 eq x xs))) (Apply (Apply GroupBySym0 eq) (Let6989586621681195434ZsSym3 eq x xs))

sGroupBy1 :: forall a (t :: (~>) a ((~>) a Bool)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply GroupBy1Sym0 t) t :: NonEmpty (NonEmpty a))

type family GroupWith1 (a :: (~>) a b) (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

GroupWith1 f a_6989586621681195467 = Apply (Apply GroupBy1Sym0 (Apply (Apply OnSym0 (==@#@$)) f)) a_6989586621681195467

sGroupWith1 :: forall b a (t :: (~>) a b) (t :: NonEmpty a). SEq b => Sing t -> Sing t -> Sing (Apply (Apply GroupWith1Sym0 t) t :: NonEmpty (NonEmpty a))

type family GroupAllWith1 (a :: (~>) a b) (a :: NonEmpty a) :: NonEmpty (NonEmpty a) where ...

Equations

GroupAllWith1 f a_6989586621681195608 = Apply (Apply (Apply (.@#@$) (Apply GroupWith1Sym0 f)) (Apply SortWithSym0 f)) a_6989586621681195608

sGroupAllWith1 :: forall b a (t :: (~>) a b) (t :: NonEmpty a). SOrd b => Sing t -> Sing t -> Sing (Apply (Apply GroupAllWith1Sym0 t) t :: NonEmpty (NonEmpty a))

type family IsPrefixOf (a :: [a]) (a :: NonEmpty a) :: Bool where ...

Equations

IsPrefixOf '[] _ = TrueSym0

IsPrefixOf ((:) y ys) ((:|) x xs) = Apply (Apply (&&@#@$) (Apply (Apply (==@#@$) y) x)) (Apply (Apply ListisPrefixOfSym0 ys) xs)

sIsPrefixOf :: forall a (t :: [a]) (t :: NonEmpty a). SEq a => Sing t -> Sing t -> Sing (Apply (Apply IsPrefixOfSym0 t) t :: Bool)

type family Nub (a :: NonEmpty a) :: NonEmpty a where ...

Equations

Nub a_6989586621681195284 = Apply (Apply NubBySym0 (==@#@$)) a_6989586621681195284

sNub :: forall a (t :: NonEmpty a). SEq a => Sing t -> Sing (Apply NubSym0 t :: NonEmpty a)

type family NubBy (a :: (~>) a ((~>) a Bool)) (a :: NonEmpty a) :: NonEmpty a where ...

Equations

NubBy eq ((:|) a as) = Apply (Apply (:|@#@$) a) (Apply (Apply ListnubBySym0 eq) (Apply (Apply ListfilterSym0 (Apply (Apply (Apply Lambda_6989586621681195276Sym0 eq) a) as)) as))

sNubBy :: forall a (t :: (~>) a ((~>) a Bool)) (t :: NonEmpty a). Sing t -> Sing t -> Sing (Apply (Apply NubBySym0 t) t :: NonEmpty a)

type family (a :: NonEmpty a) !! (a :: Nat) :: a where ...

Equations

arg_6989586621681194071 !! arg_6989586621681194073 = Case_6989586621681195343 arg_6989586621681194071 arg_6989586621681194073 (Apply (Apply Tuple2Sym0 arg_6989586621681194071) arg_6989586621681194073)

(%!!) :: forall a (t :: NonEmpty a) (t :: Nat). Sing t -> Sing t -> Sing (Apply (Apply (!!@#@$) t) t :: a)

type family Zip (a :: NonEmpty a) (a :: NonEmpty b) :: NonEmpty (a, b) where ...

Equations

Zip ((:|) x xs) ((:|) y ys) = Apply (Apply (:|@#@$) (Apply (Apply Tuple2Sym0 x) y)) (Apply (Apply ListzipSym0 xs) ys)

sZip :: forall a b (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing (Apply (Apply ZipSym0 t) t :: NonEmpty (a, b))

type family ZipWith (a :: (~>) a ((~>) b c)) (a :: NonEmpty a) (a :: NonEmpty b) :: NonEmpty c where ...

Equations

ZipWith f ((:|) x xs) ((:|) y ys) = Apply (Apply (:|@#@$) (Apply (Apply f x) y)) (Apply (Apply (Apply ListzipWithSym0 f) xs) ys)

sZipWith :: forall a b c (t :: (~>) a ((~>) b c)) (t :: NonEmpty a) (t :: NonEmpty b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ZipWithSym0 t) t) t :: NonEmpty c)

type family Unzip (a :: NonEmpty (a, b)) :: (NonEmpty a, NonEmpty b) where ...

Equations

Unzip ((:|) '(a, b) asbs) = Apply (Apply Tuple2Sym0 (Apply (Apply (:|@#@$) a) (Let6989586621681195294AsSym3 a b asbs))) (Apply (Apply (:|@#@$) b) (Let6989586621681195294BsSym3 a b asbs))

sUnzip :: forall a b (t :: NonEmpty (a, b)). Sing t -> Sing (Apply UnzipSym0 t :: (NonEmpty a, NonEmpty b))

type family FromList (a :: [a]) :: NonEmpty a where ...

Equations

FromList ((:) a as) = Apply (Apply (:|@#@$) a) as

FromList '[] = Apply ErrorSym0 "NonEmpty.fromList: empty list"

sFromList :: forall a (t :: [a]). Sing t -> Sing (Apply FromListSym0 t :: NonEmpty a)

type family ToList (a :: NonEmpty a) :: [a] where ...

Equations

ToList ((:|) a as) = Apply (Apply (:@#@$) a) as

sToList :: forall a (t :: NonEmpty a). Sing t -> Sing (Apply ToListSym0 t :: [a])

type family NonEmpty_ (a :: [a]) :: Maybe (NonEmpty a) where ...

Equations

NonEmpty_ '[] = NothingSym0

NonEmpty_ ((:) a as) = Apply JustSym0 (Apply (Apply (:|@#@$) a) as)

sNonEmpty_ :: forall a (t :: [a]). Sing t -> Sing (Apply NonEmpty_Sym0 t :: Maybe (NonEmpty a))

type family Xor (a :: NonEmpty Bool) :: Bool where ...

Equations

Xor ((:|) x xs) = Apply (Apply (Apply FoldrSym0 (Let6989586621681195758Xor'Sym2 x xs)) x) xs

sXor :: forall (t :: NonEmpty Bool). Sing t -> Sing (Apply XorSym0 t :: Bool)

Defunctionalization symbols

data (:|@#@$) :: forall (a6989586621679070532 :: Type). (~>) a6989586621679070532 ((~>) [a6989586621679070532] (NonEmpty (a6989586621679070532 :: Type))) infixr 5

data (:|@#@$$) (t6989586621679317665 :: (a6989586621679070532 :: Type)) :: (~>) [a6989586621679070532] (NonEmpty (a6989586621679070532 :: Type)) infixr 5

type (:|@#@$$$) (t6989586621679317665 :: a6989586621679070532) (t6989586621679317666 :: [a6989586621679070532]) = (:|) t6989586621679317665 t6989586621679317666

data MapSym0 :: forall a6989586621681193865 b6989586621681193866. (~>) ((~>) a6989586621681193865 b6989586621681193866) ((~>) (NonEmpty a6989586621681193865) (NonEmpty b6989586621681193866))

data MapSym1 (a6989586621681195486 :: (~>) a6989586621681193865 b6989586621681193866) :: (~>) (NonEmpty a6989586621681193865) (NonEmpty b6989586621681193866)

type MapSym2 (a6989586621681195486 :: (~>) a6989586621681193865 b6989586621681193866) (a6989586621681195487 :: NonEmpty a6989586621681193865) = Map a6989586621681195486 a6989586621681195487

data IntersperseSym0 :: forall a6989586621681193855. (~>) a6989586621681193855 ((~>) (NonEmpty a6989586621681193855) (NonEmpty a6989586621681193855))

data IntersperseSym1 (a6989586621681195476 :: a6989586621681193855) :: (~>) (NonEmpty a6989586621681193855) (NonEmpty a6989586621681193855)

type IntersperseSym2 (a6989586621681195476 :: a6989586621681193855) (a6989586621681195477 :: NonEmpty a6989586621681193855) = Intersperse a6989586621681195476 a6989586621681195477

data ScanlSym0 :: forall a6989586621681193861 b6989586621681193860. (~>) ((~>) b6989586621681193860 ((~>) a6989586621681193861 b6989586621681193860)) ((~>) b6989586621681193860 ((~>) [a6989586621681193861] (NonEmpty b6989586621681193860)))

data ScanlSym1 (a6989586621681195630 :: (~>) b6989586621681193860 ((~>) a6989586621681193861 b6989586621681193860)) :: (~>) b6989586621681193860 ((~>) [a6989586621681193861] (NonEmpty b6989586621681193860))

data ScanlSym2 (a6989586621681195630 :: (~>) b6989586621681193860 ((~>) a6989586621681193861 b6989586621681193860)) (a6989586621681195631 :: b6989586621681193860) :: (~>) [a6989586621681193861] (NonEmpty b6989586621681193860)

type ScanlSym3 (a6989586621681195630 :: (~>) b6989586621681193860 ((~>) a6989586621681193861 b6989586621681193860)) (a6989586621681195631 :: b6989586621681193860) (a6989586621681195632 :: [a6989586621681193861]) = Scanl a6989586621681195630 a6989586621681195631 a6989586621681195632

data ScanrSym0 :: forall a6989586621681193858 b6989586621681193859. (~>) ((~>) a6989586621681193858 ((~>) b6989586621681193859 b6989586621681193859)) ((~>) b6989586621681193859 ((~>) [a6989586621681193858] (NonEmpty b6989586621681193859)))

data ScanrSym1 (a6989586621681195641 :: (~>) a6989586621681193858 ((~>) b6989586621681193859 b6989586621681193859)) :: (~>) b6989586621681193859 ((~>) [a6989586621681193858] (NonEmpty b6989586621681193859))

data ScanrSym2 (a6989586621681195641 :: (~>) a6989586621681193858 ((~>) b6989586621681193859 b6989586621681193859)) (a6989586621681195642 :: b6989586621681193859) :: (~>) [a6989586621681193858] (NonEmpty b6989586621681193859)

type ScanrSym3 (a6989586621681195641 :: (~>) a6989586621681193858 ((~>) b6989586621681193859 b6989586621681193859)) (a6989586621681195642 :: b6989586621681193859) (a6989586621681195643 :: [a6989586621681193858]) = Scanr a6989586621681195641 a6989586621681195642 a6989586621681195643

data Scanl1Sym0 :: forall a6989586621681193857. (~>) ((~>) a6989586621681193857 ((~>) a6989586621681193857 a6989586621681193857)) ((~>) (NonEmpty a6989586621681193857) (NonEmpty a6989586621681193857))

data Scanl1Sym1 (a6989586621681195652 :: (~>) a6989586621681193857 ((~>) a6989586621681193857 a6989586621681193857)) :: (~>) (NonEmpty a6989586621681193857) (NonEmpty a6989586621681193857)

type Scanl1Sym2 (a6989586621681195652 :: (~>) a6989586621681193857 ((~>) a6989586621681193857 a6989586621681193857)) (a6989586621681195653 :: NonEmpty a6989586621681193857) = Scanl1 a6989586621681195652 a6989586621681195653

data Scanr1Sym0 :: forall a6989586621681193856. (~>) ((~>) a6989586621681193856 ((~>) a6989586621681193856 a6989586621681193856)) ((~>) (NonEmpty a6989586621681193856) (NonEmpty a6989586621681193856))

data Scanr1Sym1 (a6989586621681195659 :: (~>) a6989586621681193856 ((~>) a6989586621681193856 a6989586621681193856)) :: (~>) (NonEmpty a6989586621681193856) (NonEmpty a6989586621681193856)

type Scanr1Sym2 (a6989586621681195659 :: (~>) a6989586621681193856 ((~>) a6989586621681193856 a6989586621681193856)) (a6989586621681195660 :: NonEmpty a6989586621681193856) = Scanr1 a6989586621681195659 a6989586621681195660

data TransposeSym0 :: forall a6989586621681193821. (~>) (NonEmpty (NonEmpty a6989586621681193821)) (NonEmpty (NonEmpty a6989586621681193821))

type TransposeSym1 (a6989586621681195779 :: NonEmpty (NonEmpty a6989586621681193821)) = Transpose a6989586621681195779

data SortBySym0 :: forall a6989586621681193820. (~>) ((~>) a6989586621681193820 ((~>) a6989586621681193820 Ordering)) ((~>) (NonEmpty a6989586621681193820) (NonEmpty a6989586621681193820))

data SortBySym1 (a6989586621681195586 :: (~>) a6989586621681193820 ((~>) a6989586621681193820 Ordering)) :: (~>) (NonEmpty a6989586621681193820) (NonEmpty a6989586621681193820)

type SortBySym2 (a6989586621681195586 :: (~>) a6989586621681193820 ((~>) a6989586621681193820 Ordering)) (a6989586621681195587 :: NonEmpty a6989586621681193820) = SortBy a6989586621681195586 a6989586621681195587

data SortWithSym0 :: forall a6989586621681193819 o6989586621681193818. (~>) ((~>) a6989586621681193819 o6989586621681193818) ((~>) (NonEmpty a6989586621681193819) (NonEmpty a6989586621681193819))

data SortWithSym1 (a6989586621681195598 :: (~>) a6989586621681193819 o6989586621681193818) :: (~>) (NonEmpty a6989586621681193819) (NonEmpty a6989586621681193819)

type SortWithSym2 (a6989586621681195598 :: (~>) a6989586621681193819 o6989586621681193818) (a6989586621681195599 :: NonEmpty a6989586621681193819) = SortWith a6989586621681195598 a6989586621681195599

data LengthSym0 :: forall a6989586621681193884. (~>) (NonEmpty a6989586621681193884) Nat

type LengthSym1 (a6989586621681195767 :: NonEmpty a6989586621681193884) = Length a6989586621681195767

data HeadSym0 :: forall a6989586621681193877. (~>) (NonEmpty a6989586621681193877) a6989586621681193877

type HeadSym1 (a6989586621681195699 :: NonEmpty a6989586621681193877) = Head a6989586621681195699

data TailSym0 :: forall a6989586621681193876. (~>) (NonEmpty a6989586621681193876) [a6989586621681193876]

type TailSym1 (a6989586621681195696 :: NonEmpty a6989586621681193876) = Tail a6989586621681195696

data LastSym0 :: forall a6989586621681193875. (~>) (NonEmpty a6989586621681193875) a6989586621681193875

type LastSym1 (a6989586621681195692 :: NonEmpty a6989586621681193875) = Last a6989586621681195692

data InitSym0 :: forall a6989586621681193874. (~>) (NonEmpty a6989586621681193874) [a6989586621681193874]

type InitSym1 (a6989586621681195688 :: NonEmpty a6989586621681193874) = Init a6989586621681195688

data (<|@#@$) :: forall a6989586621681193873. (~>) a6989586621681193873 ((~>) (NonEmpty a6989586621681193873) (NonEmpty a6989586621681193873))

data (<|@#@$$) (a6989586621681195671 :: a6989586621681193873) :: (~>) (NonEmpty a6989586621681193873) (NonEmpty a6989586621681193873)

type (<|@#@$$$) (a6989586621681195671 :: a6989586621681193873) (a6989586621681195672 :: NonEmpty a6989586621681193873) = (<|) a6989586621681195671 a6989586621681195672

data ConsSym0 :: forall a6989586621681193872. (~>) a6989586621681193872 ((~>) (NonEmpty a6989586621681193872) (NonEmpty a6989586621681193872))

data ConsSym1 (a6989586621681195682 :: a6989586621681193872) :: (~>) (NonEmpty a6989586621681193872) (NonEmpty a6989586621681193872)

type ConsSym2 (a6989586621681195682 :: a6989586621681193872) (a6989586621681195683 :: NonEmpty a6989586621681193872) = Cons a6989586621681195682 a6989586621681195683

data UnconsSym0 :: forall a6989586621681193880. (~>) (NonEmpty a6989586621681193880) (a6989586621681193880, Maybe (NonEmpty a6989586621681193880))

type UnconsSym1 (a6989586621681195735 :: NonEmpty a6989586621681193880) = Uncons a6989586621681195735

data UnfoldrSym0 :: forall a6989586621681193878 b6989586621681193879. (~>) ((~>) a6989586621681193878 (b6989586621681193879, Maybe a6989586621681193878)) ((~>) a6989586621681193878 (NonEmpty b6989586621681193879))

data UnfoldrSym1 (a6989586621681195702 :: (~>) a6989586621681193878 (b6989586621681193879, Maybe a6989586621681193878)) :: (~>) a6989586621681193878 (NonEmpty b6989586621681193879)

type UnfoldrSym2 (a6989586621681195702 :: (~>) a6989586621681193878 (b6989586621681193879, Maybe a6989586621681193878)) (a6989586621681195703 :: a6989586621681193878) = Unfoldr a6989586621681195702 a6989586621681195703

data SortSym0 :: forall a6989586621681193871. (~>) (NonEmpty a6989586621681193871) (NonEmpty a6989586621681193871)

type SortSym1 (a6989586621681195668 :: NonEmpty a6989586621681193871) = Sort a6989586621681195668

data ReverseSym0 :: forall a6989586621681193854. (~>) (NonEmpty a6989586621681193854) (NonEmpty a6989586621681193854)

type ReverseSym1 (a6989586621681195583 :: NonEmpty a6989586621681193854) = Reverse a6989586621681195583

data InitsSym0 :: forall a6989586621681193864. (~>) [a6989586621681193864] (NonEmpty [a6989586621681193864])

type InitsSym1 (a6989586621681195614 :: [a6989586621681193864]) = Inits a6989586621681195614

data TailsSym0 :: forall a6989586621681193863. (~>) [a6989586621681193863] (NonEmpty [a6989586621681193863])

type TailsSym1 (a6989586621681195619 :: [a6989586621681193863]) = Tails a6989586621681195619

data UnfoldSym0 :: forall a6989586621681193882 b6989586621681193883. (~>) ((~>) a6989586621681193882 (b6989586621681193883, Maybe a6989586621681193882)) ((~>) a6989586621681193882 (NonEmpty b6989586621681193883))

data UnfoldSym1 (a6989586621681195739 :: (~>) a6989586621681193882 (b6989586621681193883, Maybe a6989586621681193882)) :: (~>) a6989586621681193882 (NonEmpty b6989586621681193883)

data InsertSym0 :: forall a6989586621681193862. (~>) a6989586621681193862 ((~>) [a6989586621681193862] (NonEmpty a6989586621681193862))

data InsertSym1 (a6989586621681195622 :: a6989586621681193862) :: (~>) [a6989586621681193862] (NonEmpty a6989586621681193862)

type InsertSym2 (a6989586621681195622 :: a6989586621681193862) (a6989586621681195623 :: [a6989586621681193862]) = Insert a6989586621681195622 a6989586621681195623

data TakeSym0 :: forall a6989586621681193853. (~>) Nat ((~>) (NonEmpty a6989586621681193853) [a6989586621681193853])

data TakeSym1 (a6989586621681195497 :: Nat) :: forall a6989586621681193853. (~>) (NonEmpty a6989586621681193853) [a6989586621681193853]

type TakeSym2 (a6989586621681195497 :: Nat) (a6989586621681195498 :: NonEmpty a6989586621681193853) = Take a6989586621681195497 a6989586621681195498

data DropSym0 :: forall a6989586621681193852. (~>) Nat ((~>) (NonEmpty a6989586621681193852) [a6989586621681193852])

data DropSym1 (a6989586621681195505 :: Nat) :: forall a6989586621681193852. (~>) (NonEmpty a6989586621681193852) [a6989586621681193852]

type DropSym2 (a6989586621681195505 :: Nat) (a6989586621681195506 :: NonEmpty a6989586621681193852) = Drop a6989586621681195505 a6989586621681195506

data SplitAtSym0 :: forall a6989586621681193851. (~>) Nat ((~>) (NonEmpty a6989586621681193851) ([a6989586621681193851], [a6989586621681193851]))

data SplitAtSym1 (a6989586621681195513 :: Nat) :: forall a6989586621681193851. (~>) (NonEmpty a6989586621681193851) ([a6989586621681193851], [a6989586621681193851])

type SplitAtSym2 (a6989586621681195513 :: Nat) (a6989586621681195514 :: NonEmpty a6989586621681193851) = SplitAt a6989586621681195513 a6989586621681195514

data TakeWhileSym0 :: forall a6989586621681193850. (~>) ((~>) a6989586621681193850 Bool) ((~>) (NonEmpty a6989586621681193850) [a6989586621681193850])

data TakeWhileSym1 (a6989586621681195521 :: (~>) a6989586621681193850 Bool) :: (~>) (NonEmpty a6989586621681193850) [a6989586621681193850]

type TakeWhileSym2 (a6989586621681195521 :: (~>) a6989586621681193850 Bool) (a6989586621681195522 :: NonEmpty a6989586621681193850) = TakeWhile a6989586621681195521 a6989586621681195522

data DropWhileSym0 :: forall a6989586621681193849. (~>) ((~>) a6989586621681193849 Bool) ((~>) (NonEmpty a6989586621681193849) [a6989586621681193849])

data DropWhileSym1 (a6989586621681195529 :: (~>) a6989586621681193849 Bool) :: (~>) (NonEmpty a6989586621681193849) [a6989586621681193849]

type DropWhileSym2 (a6989586621681195529 :: (~>) a6989586621681193849 Bool) (a6989586621681195530 :: NonEmpty a6989586621681193849) = DropWhile a6989586621681195529 a6989586621681195530

data SpanSym0 :: forall a6989586621681193848. (~>) ((~>) a6989586621681193848 Bool) ((~>) (NonEmpty a6989586621681193848) ([a6989586621681193848], [a6989586621681193848]))

data SpanSym1 (a6989586621681195537 :: (~>) a6989586621681193848 Bool) :: (~>) (NonEmpty a6989586621681193848) ([a6989586621681193848], [a6989586621681193848])

type SpanSym2 (a6989586621681195537 :: (~>) a6989586621681193848 Bool) (a6989586621681195538 :: NonEmpty a6989586621681193848) = Span a6989586621681195537 a6989586621681195538

data BreakSym0 :: forall a6989586621681193847. (~>) ((~>) a6989586621681193847 Bool) ((~>) (NonEmpty a6989586621681193847) ([a6989586621681193847], [a6989586621681193847]))

data BreakSym1 (a6989586621681195545 :: (~>) a6989586621681193847 Bool) :: (~>) (NonEmpty a6989586621681193847) ([a6989586621681193847], [a6989586621681193847])

type BreakSym2 (a6989586621681195545 :: (~>) a6989586621681193847 Bool) (a6989586621681195546 :: NonEmpty a6989586621681193847) = Break a6989586621681195545 a6989586621681195546

data FilterSym0 :: forall a6989586621681193846. (~>) ((~>) a6989586621681193846 Bool) ((~>) (NonEmpty a6989586621681193846) [a6989586621681193846])

data FilterSym1 (a6989586621681195553 :: (~>) a6989586621681193846 Bool) :: (~>) (NonEmpty a6989586621681193846) [a6989586621681193846]

type FilterSym2 (a6989586621681195553 :: (~>) a6989586621681193846 Bool) (a6989586621681195554 :: NonEmpty a6989586621681193846) = Filter a6989586621681195553 a6989586621681195554

data PartitionSym0 :: forall a6989586621681193845. (~>) ((~>) a6989586621681193845 Bool) ((~>) (NonEmpty a6989586621681193845) ([a6989586621681193845], [a6989586621681193845]))

data PartitionSym1 (a6989586621681195561 :: (~>) a6989586621681193845 Bool) :: (~>) (NonEmpty a6989586621681193845) ([a6989586621681193845], [a6989586621681193845])

type PartitionSym2 (a6989586621681195561 :: (~>) a6989586621681193845 Bool) (a6989586621681195562 :: NonEmpty a6989586621681193845) = Partition a6989586621681195561 a6989586621681195562

data GroupSym0 :: forall a6989586621681193844. (~>) [a6989586621681193844] [NonEmpty a6989586621681193844]

type GroupSym1 (a6989586621681195473 :: [a6989586621681193844]) = Group a6989586621681195473

data GroupBySym0 :: forall a6989586621681193843. (~>) ((~>) a6989586621681193843 ((~>) a6989586621681193843 Bool)) ((~>) [a6989586621681193843] [NonEmpty a6989586621681193843])

data GroupBySym1 (a6989586621681195363 :: (~>) a6989586621681193843 ((~>) a6989586621681193843 Bool)) :: (~>) [a6989586621681193843] [NonEmpty a6989586621681193843]

type GroupBySym2 (a6989586621681195363 :: (~>) a6989586621681193843 ((~>) a6989586621681193843 Bool)) (a6989586621681195364 :: [a6989586621681193843]) = GroupBy a6989586621681195363 a6989586621681195364

data GroupWithSym0 :: forall a6989586621681193842 b6989586621681193841. (~>) ((~>) a6989586621681193842 b6989586621681193841) ((~>) [a6989586621681193842] [NonEmpty a6989586621681193842])

data GroupWithSym1 (a6989586621681195411 :: (~>) a6989586621681193842 b6989586621681193841) :: (~>) [a6989586621681193842] [NonEmpty a6989586621681193842]

type GroupWithSym2 (a6989586621681195411 :: (~>) a6989586621681193842 b6989586621681193841) (a6989586621681195412 :: [a6989586621681193842]) = GroupWith a6989586621681195411 a6989586621681195412

data GroupAllWithSym0 :: forall a6989586621681193840 b6989586621681193839. (~>) ((~>) a6989586621681193840 b6989586621681193839) ((~>) [a6989586621681193840] [NonEmpty a6989586621681193840])

data GroupAllWithSym1 (a6989586621681195419 :: (~>) a6989586621681193840 b6989586621681193839) :: (~>) [a6989586621681193840] [NonEmpty a6989586621681193840]

type GroupAllWithSym2 (a6989586621681195419 :: (~>) a6989586621681193840 b6989586621681193839) (a6989586621681195420 :: [a6989586621681193840]) = GroupAllWith a6989586621681195419 a6989586621681195420

data Group1Sym0 :: forall a6989586621681193838. (~>) (NonEmpty a6989586621681193838) (NonEmpty (NonEmpty a6989586621681193838))

type Group1Sym1 (a6989586621681195460 :: NonEmpty a6989586621681193838) = Group1 a6989586621681195460

data GroupBy1Sym0 :: forall a6989586621681193837. (~>) ((~>) a6989586621681193837 ((~>) a6989586621681193837 Bool)) ((~>) (NonEmpty a6989586621681193837) (NonEmpty (NonEmpty a6989586621681193837)))

data GroupBy1Sym1 (a6989586621681195427 :: (~>) a6989586621681193837 ((~>) a6989586621681193837 Bool)) :: (~>) (NonEmpty a6989586621681193837) (NonEmpty (NonEmpty a6989586621681193837))

type GroupBy1Sym2 (a6989586621681195427 :: (~>) a6989586621681193837 ((~>) a6989586621681193837 Bool)) (a6989586621681195428 :: NonEmpty a6989586621681193837) = GroupBy1 a6989586621681195427 a6989586621681195428

data GroupWith1Sym0 :: forall a6989586621681193836 b6989586621681193835. (~>) ((~>) a6989586621681193836 b6989586621681193835) ((~>) (NonEmpty a6989586621681193836) (NonEmpty (NonEmpty a6989586621681193836)))

data GroupWith1Sym1 (a6989586621681195463 :: (~>) a6989586621681193836 b6989586621681193835) :: (~>) (NonEmpty a6989586621681193836) (NonEmpty (NonEmpty a6989586621681193836))

type GroupWith1Sym2 (a6989586621681195463 :: (~>) a6989586621681193836 b6989586621681193835) (a6989586621681195464 :: NonEmpty a6989586621681193836) = GroupWith1 a6989586621681195463 a6989586621681195464

data GroupAllWith1Sym0 :: forall a6989586621681193834 b6989586621681193833. (~>) ((~>) a6989586621681193834 b6989586621681193833) ((~>) (NonEmpty a6989586621681193834) (NonEmpty (NonEmpty a6989586621681193834)))

data GroupAllWith1Sym1 (a6989586621681195604 :: (~>) a6989586621681193834 b6989586621681193833) :: (~>) (NonEmpty a6989586621681193834) (NonEmpty (NonEmpty a6989586621681193834))

type GroupAllWith1Sym2 (a6989586621681195604 :: (~>) a6989586621681193834 b6989586621681193833) (a6989586621681195605 :: NonEmpty a6989586621681193834) = GroupAllWith1 a6989586621681195604 a6989586621681195605

data IsPrefixOfSym0 :: forall a6989586621681193832. (~>) [a6989586621681193832] ((~>) (NonEmpty a6989586621681193832) Bool)

data IsPrefixOfSym1 (a6989586621681195355 :: [a6989586621681193832]) :: (~>) (NonEmpty a6989586621681193832) Bool

type IsPrefixOfSym2 (a6989586621681195355 :: [a6989586621681193832]) (a6989586621681195356 :: NonEmpty a6989586621681193832) = IsPrefixOf a6989586621681195355 a6989586621681195356

data NubSym0 :: forall a6989586621681193823. (~>) (NonEmpty a6989586621681193823) (NonEmpty a6989586621681193823)

type NubSym1 (a6989586621681195286 :: NonEmpty a6989586621681193823) = Nub a6989586621681195286

data NubBySym0 :: forall a6989586621681193822. (~>) ((~>) a6989586621681193822 ((~>) a6989586621681193822 Bool)) ((~>) (NonEmpty a6989586621681193822) (NonEmpty a6989586621681193822))

data NubBySym1 (a6989586621681195269 :: (~>) a6989586621681193822 ((~>) a6989586621681193822 Bool)) :: (~>) (NonEmpty a6989586621681193822) (NonEmpty a6989586621681193822)

type NubBySym2 (a6989586621681195269 :: (~>) a6989586621681193822 ((~>) a6989586621681193822 Bool)) (a6989586621681195270 :: NonEmpty a6989586621681193822) = NubBy a6989586621681195269 a6989586621681195270

data (!!@#@$) :: forall a6989586621681193831. (~>) (NonEmpty a6989586621681193831) ((~>) Nat a6989586621681193831)

data (!!@#@$$) (a6989586621681195337 :: NonEmpty a6989586621681193831) :: (~>) Nat a6989586621681193831

type (!!@#@$$$) (a6989586621681195337 :: NonEmpty a6989586621681193831) (a6989586621681195338 :: Nat) = (!!) a6989586621681195337 a6989586621681195338

data ZipSym0 :: forall a6989586621681193829 b6989586621681193830. (~>) (NonEmpty a6989586621681193829) ((~>) (NonEmpty b6989586621681193830) (NonEmpty (a6989586621681193829, b6989586621681193830)))

data ZipSym1 (a6989586621681195329 :: NonEmpty a6989586621681193829) :: forall b6989586621681193830. (~>) (NonEmpty b6989586621681193830) (NonEmpty (a6989586621681193829, b6989586621681193830))

type ZipSym2 (a6989586621681195329 :: NonEmpty a6989586621681193829) (a6989586621681195330 :: NonEmpty b6989586621681193830) = Zip a6989586621681195329 a6989586621681195330

data ZipWithSym0 :: forall a6989586621681193826 b6989586621681193827 c6989586621681193828. (~>) ((~>) a6989586621681193826 ((~>) b6989586621681193827 c6989586621681193828)) ((~>) (NonEmpty a6989586621681193826) ((~>) (NonEmpty b6989586621681193827) (NonEmpty c6989586621681193828)))

data ZipWithSym1 (a6989586621681195318 :: (~>) a6989586621681193826 ((~>) b6989586621681193827 c6989586621681193828)) :: (~>) (NonEmpty a6989586621681193826) ((~>) (NonEmpty b6989586621681193827) (NonEmpty c6989586621681193828))

data ZipWithSym2 (a6989586621681195318 :: (~>) a6989586621681193826 ((~>) b6989586621681193827 c6989586621681193828)) (a6989586621681195319 :: NonEmpty a6989586621681193826) :: (~>) (NonEmpty b6989586621681193827) (NonEmpty c6989586621681193828)

type ZipWithSym3 (a6989586621681195318 :: (~>) a6989586621681193826 ((~>) b6989586621681193827 c6989586621681193828)) (a6989586621681195319 :: NonEmpty a6989586621681193826) (a6989586621681195320 :: NonEmpty b6989586621681193827) = ZipWith a6989586621681195318 a6989586621681195319 a6989586621681195320

data UnzipSym0 :: forall a6989586621681193824 b6989586621681193825. (~>) (NonEmpty (a6989586621681193824, b6989586621681193825)) (NonEmpty a6989586621681193824, NonEmpty b6989586621681193825)

type UnzipSym1 (a6989586621681195289 :: NonEmpty (a6989586621681193824, b6989586621681193825)) = Unzip a6989586621681195289

data FromListSym0 :: forall a6989586621681193870. (~>) [a6989586621681193870] (NonEmpty a6989586621681193870)

type FromListSym1 (a6989586621681195569 :: [a6989586621681193870]) = FromList a6989586621681195569

data ToListSym0 :: forall a6989586621681193869. (~>) (NonEmpty a6989586621681193869) [a6989586621681193869]

type ToListSym1 (a6989586621681195493 :: NonEmpty a6989586621681193869) = ToList a6989586621681195493

data NonEmpty_Sym0 :: forall a6989586621681193881. (~>) [a6989586621681193881] (Maybe (NonEmpty a6989586621681193881))

type NonEmpty_Sym1 (a6989586621681195731 :: [a6989586621681193881]) = NonEmpty_ a6989586621681195731

data XorSym0 :: (~>) (NonEmpty Bool) Bool

type XorSym1 (a6989586621681195754 :: NonEmpty Bool) = Xor a6989586621681195754

Orphan instances

SMonadZip NonEmpty

Instance details

Methods

sMzip :: Sing t -> Sing t -> Sing (Apply (Apply MzipSym0 t) t)

sMzipWith :: Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply MzipWithSym0 t) t) t)

sMunzip :: Sing t -> Sing (Apply MunzipSym0 t)

PMonadZip NonEmpty

Instance details

Associated Types

type Mzip arg arg :: m (a, b)

type MzipWith arg arg arg :: m c

type Munzip arg :: (m a, m b)