Data.Singletons.Prelude.Maybe

Description

Defines functions and datatypes relating to the singleton for Maybe, including a singletons version of all the definitions in Data.Maybe.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Maybe. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Documentation

data family Sing (a :: k)

The singleton kind-indexed data family.

Though Haddock doesn't show it, the Sing instance above declares constructors

SNothing :: Sing Nothing
SJust    :: Sing a -> Sing (Just a)

type SMaybe = (Sing :: Maybe a -> Type)

SBool is a kind-restricted synonym for Sing: type SMaybe (a :: Maybe k) = Sing a

Singletons from Data.Maybe

maybe_ :: forall b a. b -> (a -> b) -> Maybe a -> b

type family Maybe_ (a :: b) (a :: TyFun a b -> Type) (a :: Maybe a) :: b where ...

Equations

Maybe_ n _z_6989586621679770024 Nothing = n

Maybe_ _z_6989586621679770027 f (Just x) = Apply f x

sMaybe_ :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Maybe_Sym0 t) t) t :: b)

The preceding two definitions are derived from the function maybe in Data.Maybe. The extra underscore is to avoid name clashes with the type Maybe.

type family IsJust (a :: Maybe a) :: Bool where ...

Equations

IsJust Nothing = FalseSym0

IsJust (Just _z_6989586621679771377) = TrueSym0

sIsJust :: forall t. Sing t -> Sing (Apply IsJustSym0 t :: Bool)

type family IsNothing (a :: Maybe a) :: Bool where ...

Equations

IsNothing Nothing = TrueSym0

IsNothing (Just _z_6989586621679771370) = FalseSym0

sIsNothing :: forall t. Sing t -> Sing (Apply IsNothingSym0 t :: Bool)

type family FromJust (a :: Maybe a) :: a where ...

Equations

FromJust Nothing = Apply ErrorSym0 "Maybe.fromJust: Nothing"

FromJust (Just x) = x

sFromJust :: forall t. Sing t -> Sing (Apply FromJustSym0 t :: a)

type family FromMaybe (a :: a) (a :: Maybe a) :: a where ...

Equations

FromMaybe d x = Case_6989586621679771357 d x x

sFromMaybe :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply FromMaybeSym0 t) t :: a)

type family ListToMaybe (a :: [a]) :: Maybe a where ...

Equations

ListToMaybe '[] = NothingSym0

ListToMaybe ((:) a _z_6989586621679771338) = Apply JustSym0 a

sListToMaybe :: forall t. Sing t -> Sing (Apply ListToMaybeSym0 t :: Maybe a)

type family MaybeToList (a :: Maybe a) :: [a] where ...

Equations

MaybeToList Nothing = '[]

MaybeToList (Just x) = Apply (Apply (:$) x) '[]

sMaybeToList :: forall t. Sing t -> Sing (Apply MaybeToListSym0 t :: [a])

type family CatMaybes (a :: [Maybe a]) :: [a] where ...

Equations

CatMaybes '[] = '[]

CatMaybes ((:) (Just x) xs) = Apply (Apply (:$) x) (Apply CatMaybesSym0 xs)

CatMaybes ((:) Nothing xs) = Apply CatMaybesSym0 xs

sCatMaybes :: forall t. Sing t -> Sing (Apply CatMaybesSym0 t :: [a])

type family MapMaybe (a :: TyFun a (Maybe b) -> Type) (a :: [a]) :: [b] where ...

Equations

MapMaybe _z_6989586621679771290 '[] = '[]

MapMaybe f ((:) x xs) = Case_6989586621679771322 f x xs (Let6989586621679771309Scrutinee_6989586621679771279Sym3 f x xs)

sMapMaybe :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply MapMaybeSym0 t) t :: [b])

Defunctionalization symbols

type NothingSym0 = Nothing

data JustSym0 l

type JustSym1 t = Just t

data Maybe_Sym0 l

data Maybe_Sym1 l l

data Maybe_Sym2 l l l

type Maybe_Sym3 t t t = Maybe_ t t t

data IsJustSym0 l

type IsJustSym1 t = IsJust t

data IsNothingSym0 l

type IsNothingSym1 t = IsNothing t

data FromJustSym0 l

type FromJustSym1 t = FromJust t

data FromMaybeSym0 l

data FromMaybeSym1 l l

type FromMaybeSym2 t t = FromMaybe t t

data ListToMaybeSym0 l

type ListToMaybeSym1 t = ListToMaybe t

data MaybeToListSym0 l

type MaybeToListSym1 t = MaybeToList t

data CatMaybesSym0 l

type CatMaybesSym1 t = CatMaybes t

data MapMaybeSym0 l

data MapMaybeSym1 l l

type MapMaybeSym2 t t = MapMaybe t t