Data.Singletons.Prelude.Maybe

Description

Defines functions and datatypes relating to the singleton for Maybe, including a singletons version of all the definitions in Data.Maybe.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Maybe. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Documentation

type family Sing :: k -> Type

The singleton kind-indexed type family.

data SMaybe :: forall a. Maybe a -> Type where

Constructors

SNothing :: SMaybe 'Nothing

SJust :: forall a (n :: a). (Sing (n :: a)) -> SMaybe ('Just n)

Singletons from Data.Maybe

maybe_ :: b -> (a -> b) -> Maybe a -> b

type family Maybe_ (a :: b) (a :: (~>) a b) (a :: Maybe a) :: b where ...

Equations

Maybe_ n _ 'Nothing = n

Maybe_ _ f ('Just x) = Apply f x

sMaybe_ :: forall b a (t :: b) (t :: (~>) a b) (t :: Maybe a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Maybe_Sym0 t) t) t :: b)

The preceding two definitions are derived from the function maybe in Data.Maybe. The extra underscore is to avoid name clashes with the type Maybe.

type family IsJust (a :: Maybe a) :: Bool where ...

Equations

IsJust 'Nothing = FalseSym0

IsJust ('Just _) = TrueSym0

sIsJust :: forall a (t :: Maybe a). Sing t -> Sing (Apply IsJustSym0 t :: Bool)

type family IsNothing (a :: Maybe a) :: Bool where ...

Equations

IsNothing 'Nothing = TrueSym0

IsNothing ('Just _) = FalseSym0

sIsNothing :: forall a (t :: Maybe a). Sing t -> Sing (Apply IsNothingSym0 t :: Bool)

type family FromJust (a :: Maybe a) :: a where ...

Equations

FromJust 'Nothing = Apply ErrorSym0 "Maybe.fromJust: Nothing"

FromJust ('Just x) = x

sFromJust :: forall a (t :: Maybe a). Sing t -> Sing (Apply FromJustSym0 t :: a)

type family FromMaybe (a :: a) (a :: Maybe a) :: a where ...

Equations

FromMaybe d x = Case_6989586621679518070 d x x

sFromMaybe :: forall a (t :: a) (t :: Maybe a). Sing t -> Sing t -> Sing (Apply (Apply FromMaybeSym0 t) t :: a)

type family ListToMaybe (a :: [a]) :: Maybe a where ...

Equations

ListToMaybe '[] = NothingSym0

ListToMaybe ('(:) a _) = Apply JustSym0 a

sListToMaybe :: forall a (t :: [a]). Sing t -> Sing (Apply ListToMaybeSym0 t :: Maybe a)

type family MaybeToList (a :: Maybe a) :: [a] where ...

Equations

MaybeToList 'Nothing = '[]

MaybeToList ('Just x) = Apply (Apply (:@#@$) x) '[]

sMaybeToList :: forall a (t :: Maybe a). Sing t -> Sing (Apply MaybeToListSym0 t :: [a])

type family CatMaybes (a :: [Maybe a]) :: [a] where ...

Equations

CatMaybes '[] = '[]

CatMaybes ('(:) ('Just x) xs) = Apply (Apply (:@#@$) x) (Apply CatMaybesSym0 xs)

CatMaybes ('(:) 'Nothing xs) = Apply CatMaybesSym0 xs

sCatMaybes :: forall a (t :: [Maybe a]). Sing t -> Sing (Apply CatMaybesSym0 t :: [a])

type family MapMaybe (a :: (~>) a (Maybe b)) (a :: [a]) :: [b] where ...

Equations

MapMaybe _ '[] = '[]

MapMaybe f ('(:) x xs) = Case_6989586621679518049 f x xs (Let6989586621679518045Scrutinee_6989586621679517908Sym3 f x xs)

sMapMaybe :: forall a b (t :: (~>) a (Maybe b)) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply MapMaybeSym0 t) t :: [b])

Defunctionalization symbols

type NothingSym0 = 'Nothing

data JustSym0 :: forall (a3530822107858468865 :: Type). (~>) a3530822107858468865 (Maybe (a3530822107858468865 :: Type))

type JustSym1 (t6989586621679316691 :: a3530822107858468865) = 'Just t6989586621679316691

data Maybe_Sym0 :: forall b6989586621679516451 a6989586621679516452. (~>) b6989586621679516451 ((~>) ((~>) a6989586621679516452 b6989586621679516451) ((~>) (Maybe a6989586621679516452) b6989586621679516451))

data Maybe_Sym1 (a6989586621679516469 :: b6989586621679516451) :: forall a6989586621679516452. (~>) ((~>) a6989586621679516452 b6989586621679516451) ((~>) (Maybe a6989586621679516452) b6989586621679516451)

data Maybe_Sym2 (a6989586621679516469 :: b6989586621679516451) (a6989586621679516470 :: (~>) a6989586621679516452 b6989586621679516451) :: (~>) (Maybe a6989586621679516452) b6989586621679516451

type Maybe_Sym3 (a6989586621679516469 :: b6989586621679516451) (a6989586621679516470 :: (~>) a6989586621679516452 b6989586621679516451) (a6989586621679516471 :: Maybe a6989586621679516452) = Maybe_ a6989586621679516469 a6989586621679516470 a6989586621679516471

data IsJustSym0 :: forall a6989586621679517881. (~>) (Maybe a6989586621679517881) Bool

type IsJustSym1 (a6989586621679518079 :: Maybe a6989586621679517881) = IsJust a6989586621679518079

data IsNothingSym0 :: forall a6989586621679517880. (~>) (Maybe a6989586621679517880) Bool

type IsNothingSym1 (a6989586621679518077 :: Maybe a6989586621679517880) = IsNothing a6989586621679518077

data FromJustSym0 :: forall a6989586621679517879. (~>) (Maybe a6989586621679517879) a6989586621679517879

type FromJustSym1 (a6989586621679518074 :: Maybe a6989586621679517879) = FromJust a6989586621679518074

data FromMaybeSym0 :: forall a6989586621679517878. (~>) a6989586621679517878 ((~>) (Maybe a6989586621679517878) a6989586621679517878)

data FromMaybeSym1 (a6989586621679518064 :: a6989586621679517878) :: (~>) (Maybe a6989586621679517878) a6989586621679517878

type FromMaybeSym2 (a6989586621679518064 :: a6989586621679517878) (a6989586621679518065 :: Maybe a6989586621679517878) = FromMaybe a6989586621679518064 a6989586621679518065

data ListToMaybeSym0 :: forall a6989586621679517876. (~>) [a6989586621679517876] (Maybe a6989586621679517876)

type ListToMaybeSym1 (a6989586621679518058 :: [a6989586621679517876]) = ListToMaybe a6989586621679518058

data MaybeToListSym0 :: forall a6989586621679517877. (~>) (Maybe a6989586621679517877) [a6989586621679517877]

type MaybeToListSym1 (a6989586621679518061 :: Maybe a6989586621679517877) = MaybeToList a6989586621679518061

data CatMaybesSym0 :: forall a6989586621679517875. (~>) [Maybe a6989586621679517875] [a6989586621679517875]

type CatMaybesSym1 (a6989586621679518053 :: [Maybe a6989586621679517875]) = CatMaybes a6989586621679518053

data MapMaybeSym0 :: forall a6989586621679517873 b6989586621679517874. (~>) ((~>) a6989586621679517873 (Maybe b6989586621679517874)) ((~>) [a6989586621679517873] [b6989586621679517874])

data MapMaybeSym1 (a6989586621679518034 :: (~>) a6989586621679517873 (Maybe b6989586621679517874)) :: (~>) [a6989586621679517873] [b6989586621679517874]

type MapMaybeSym2 (a6989586621679518034 :: (~>) a6989586621679517873 (Maybe b6989586621679517874)) (a6989586621679518035 :: [a6989586621679517873]) = MapMaybe a6989586621679518034 a6989586621679518035