Data.Singletons.Prelude.Maybe

Description

Defines functions and datatypes relating to the singleton for Maybe, including a singletons version of all the definitions in Data.Maybe.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Maybe. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Documentation

data family Sing :: k -> Type

The singleton kind-indexed data family.

Though Haddock doesn't show it, the Sing instance above declares constructors

SNothing :: Sing Nothing
SJust    :: Sing a -> Sing (Just a)

type SMaybe = (Sing :: Maybe a -> Type)

SBool is a kind-restricted synonym for Sing: type SMaybe (a :: Maybe k) = Sing a

Singletons from Data.Maybe

maybe_ :: b -> (a -> b) -> Maybe a -> b

type family Maybe_ (a :: b) (a :: (~>) a b) (a :: Maybe a) :: b where ...

Equations

Maybe_ n _ Nothing = n

Maybe_ _ f (Just x) = Apply f x

sMaybe_ :: forall b a (t :: b) (t :: (~>) a b) (t :: Maybe a). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply Maybe_Sym0 t) t) t :: b)

The preceding two definitions are derived from the function maybe in Data.Maybe. The extra underscore is to avoid name clashes with the type Maybe.

type family IsJust (a :: Maybe a) :: Bool where ...

Equations

IsJust Nothing = FalseSym0

IsJust (Just _) = TrueSym0

sIsJust :: forall a (t :: Maybe a). Sing t -> Sing (Apply IsJustSym0 t :: Bool)

type family IsNothing (a :: Maybe a) :: Bool where ...

Equations

IsNothing Nothing = TrueSym0

IsNothing (Just _) = FalseSym0

sIsNothing :: forall a (t :: Maybe a). Sing t -> Sing (Apply IsNothingSym0 t :: Bool)

type family FromJust (a :: Maybe a) :: a where ...

Equations

FromJust Nothing = Apply ErrorSym0 "Maybe.fromJust: Nothing"

FromJust (Just x) = x

sFromJust :: forall a (t :: Maybe a). Sing t -> Sing (Apply FromJustSym0 t :: a)

type family FromMaybe (a :: a) (a :: Maybe a) :: a where ...

Equations

FromMaybe d x = Case_6989586621679522122 d x x

sFromMaybe :: forall a (t :: a) (t :: Maybe a). Sing t -> Sing t -> Sing (Apply (Apply FromMaybeSym0 t) t :: a)

type family ListToMaybe (a :: [a]) :: Maybe a where ...

Equations

ListToMaybe '[] = NothingSym0

ListToMaybe ((:) a _) = Apply JustSym0 a

sListToMaybe :: forall a (t :: [a]). Sing t -> Sing (Apply ListToMaybeSym0 t :: Maybe a)

type family MaybeToList (a :: Maybe a) :: [a] where ...

Equations

MaybeToList Nothing = '[]

MaybeToList (Just x) = Apply (Apply (:@#@$) x) '[]

sMaybeToList :: forall a (t :: Maybe a). Sing t -> Sing (Apply MaybeToListSym0 t :: [a])

type family CatMaybes (a :: [Maybe a]) :: [a] where ...

Equations

CatMaybes '[] = '[]

CatMaybes ((:) (Just x) xs) = Apply (Apply (:@#@$) x) (Apply CatMaybesSym0 xs)

CatMaybes ((:) Nothing xs) = Apply CatMaybesSym0 xs

sCatMaybes :: forall a (t :: [Maybe a]). Sing t -> Sing (Apply CatMaybesSym0 t :: [a])

type family MapMaybe (a :: (~>) a (Maybe b)) (a :: [a]) :: [b] where ...

Equations

MapMaybe _ '[] = '[]

MapMaybe f ((:) x xs) = Case_6989586621679522101 f x xs (Let6989586621679522097Scrutinee_6989586621679521954Sym3 f x xs)

sMapMaybe :: forall a b (t :: (~>) a (Maybe b)) (t :: [a]). Sing t -> Sing t -> Sing (Apply (Apply MapMaybeSym0 t) t :: [b])

Defunctionalization symbols

type NothingSym0 = Nothing

data JustSym0 :: forall (a3530822107858468865 :: Type). (~>) a3530822107858468865 (Maybe (a3530822107858468865 :: Type))

type JustSym1 (t6989586621679317575 :: a3530822107858468865) = Just t6989586621679317575

data Maybe_Sym0 :: forall a6989586621679520498 b6989586621679520497. (~>) b6989586621679520497 ((~>) ((~>) a6989586621679520498 b6989586621679520497) ((~>) (Maybe a6989586621679520498) b6989586621679520497))

data Maybe_Sym1 (a6989586621679520515 :: b6989586621679520497) :: forall a6989586621679520498. (~>) ((~>) a6989586621679520498 b6989586621679520497) ((~>) (Maybe a6989586621679520498) b6989586621679520497)

data Maybe_Sym2 (a6989586621679520515 :: b6989586621679520497) (a6989586621679520516 :: (~>) a6989586621679520498 b6989586621679520497) :: (~>) (Maybe a6989586621679520498) b6989586621679520497

type Maybe_Sym3 (a6989586621679520515 :: b6989586621679520497) (a6989586621679520516 :: (~>) a6989586621679520498 b6989586621679520497) (a6989586621679520517 :: Maybe a6989586621679520498) = Maybe_ a6989586621679520515 a6989586621679520516 a6989586621679520517

data IsJustSym0 :: forall a6989586621679521927. (~>) (Maybe a6989586621679521927) Bool

type IsJustSym1 (a6989586621679522131 :: Maybe a6989586621679521927) = IsJust a6989586621679522131

data IsNothingSym0 :: forall a6989586621679521926. (~>) (Maybe a6989586621679521926) Bool

type IsNothingSym1 (a6989586621679522129 :: Maybe a6989586621679521926) = IsNothing a6989586621679522129

data FromJustSym0 :: forall a6989586621679521925. (~>) (Maybe a6989586621679521925) a6989586621679521925

type FromJustSym1 (a6989586621679522126 :: Maybe a6989586621679521925) = FromJust a6989586621679522126

data FromMaybeSym0 :: forall a6989586621679521924. (~>) a6989586621679521924 ((~>) (Maybe a6989586621679521924) a6989586621679521924)

data FromMaybeSym1 (a6989586621679522116 :: a6989586621679521924) :: (~>) (Maybe a6989586621679521924) a6989586621679521924

type FromMaybeSym2 (a6989586621679522116 :: a6989586621679521924) (a6989586621679522117 :: Maybe a6989586621679521924) = FromMaybe a6989586621679522116 a6989586621679522117

data ListToMaybeSym0 :: forall a6989586621679521922. (~>) [a6989586621679521922] (Maybe a6989586621679521922)

type ListToMaybeSym1 (a6989586621679522110 :: [a6989586621679521922]) = ListToMaybe a6989586621679522110

data MaybeToListSym0 :: forall a6989586621679521923. (~>) (Maybe a6989586621679521923) [a6989586621679521923]

type MaybeToListSym1 (a6989586621679522113 :: Maybe a6989586621679521923) = MaybeToList a6989586621679522113

data CatMaybesSym0 :: forall a6989586621679521921. (~>) [Maybe a6989586621679521921] [a6989586621679521921]

type CatMaybesSym1 (a6989586621679522105 :: [Maybe a6989586621679521921]) = CatMaybes a6989586621679522105

data MapMaybeSym0 :: forall a6989586621679521919 b6989586621679521920. (~>) ((~>) a6989586621679521919 (Maybe b6989586621679521920)) ((~>) [a6989586621679521919] [b6989586621679521920])

data MapMaybeSym1 (a6989586621679522086 :: (~>) a6989586621679521919 (Maybe b6989586621679521920)) :: (~>) [a6989586621679521919] [b6989586621679521920]

type MapMaybeSym2 (a6989586621679522086 :: (~>) a6989586621679521919 (Maybe b6989586621679521920)) (a6989586621679522087 :: [a6989586621679521919]) = MapMaybe a6989586621679522086 a6989586621679522087