Data.Singletons.Prelude.Num

Description

Defines and exports promoted and singleton versions of definitions from GHC.Num.

Be warned that some of the associated type families in the PNum class ((+), (-), and (*)) clash with their counterparts for Nat in the GHC.TypeLits module.

Documentation

class PNum (a :: Type)

Associated Types

type (arg :: a) + (arg :: a) :: a infixl 6

type (arg :: a) - (arg :: a) :: a infixl 6

type (arg :: a) * (arg :: a) :: a infixl 7

type Negate (arg :: a) :: a

type Abs (arg :: a) :: a

type Signum (arg :: a) :: a

type FromInteger (arg :: Nat) :: a

class SNum a where

Minimal complete definition

(%+), (%*), sAbs, sSignum, sFromInteger

Methods

(%+) :: forall (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply (+@#@$) t) t :: a) infixl 6

(%-) :: forall (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply (-@#@$) t) t :: a) infixl 6

(%*) :: forall (t :: a) (t :: a). Sing t -> Sing t -> Sing (Apply (Apply (*@#@$) t) t :: a) infixl 7

sNegate :: forall (t :: a). Sing t -> Sing (Apply NegateSym0 t :: a)

sAbs :: forall (t :: a). Sing t -> Sing (Apply AbsSym0 t :: a)

sSignum :: forall (t :: a). Sing t -> Sing (Apply SignumSym0 t :: a)

sFromInteger :: forall (t :: Nat). Sing t -> Sing (Apply FromIntegerSym0 t :: a)

(%-) :: forall (t :: a) (t :: a). (Apply (Apply (-@#@$) t) t :: a) ~ Apply (Apply TFHelper_6989586621679532557Sym0 t) t => Sing t -> Sing t -> Sing (Apply (Apply (-@#@$) t) t :: a) infixl 6

sNegate :: forall (t :: a). (Apply NegateSym0 t :: a) ~ Apply Negate_6989586621679532565Sym0 t => Sing t -> Sing (Apply NegateSym0 t :: a)

type family Subtract (a :: a) (a :: a) :: a where ...

Equations

Subtract x y = Apply (Apply (-@#@$) y) x

sSubtract :: forall a (t :: a) (t :: a). SNum a => Sing t -> Sing t -> Sing (Apply (Apply SubtractSym0 t) t :: a)

Defunctionalization symbols

data (+@#@$) :: forall a6989586621679532509. (~>) a6989586621679532509 ((~>) a6989586621679532509 a6989586621679532509) infixl 6

data (+@#@$$) (arg6989586621679532529 :: a6989586621679532509) :: (~>) a6989586621679532509 a6989586621679532509 infixl 6

type (+@#@$$$) (arg6989586621679532529 :: a6989586621679532509) (arg6989586621679532530 :: a6989586621679532509) = (+) arg6989586621679532529 arg6989586621679532530

data (-@#@$) :: forall a6989586621679532509. (~>) a6989586621679532509 ((~>) a6989586621679532509 a6989586621679532509) infixl 6

data (-@#@$$) (arg6989586621679532533 :: a6989586621679532509) :: (~>) a6989586621679532509 a6989586621679532509 infixl 6

type (-@#@$$$) (arg6989586621679532533 :: a6989586621679532509) (arg6989586621679532534 :: a6989586621679532509) = (-) arg6989586621679532533 arg6989586621679532534

data (*@#@$) :: forall a6989586621679532509. (~>) a6989586621679532509 ((~>) a6989586621679532509 a6989586621679532509) infixl 7

data (*@#@$$) (arg6989586621679532537 :: a6989586621679532509) :: (~>) a6989586621679532509 a6989586621679532509 infixl 7

type (*@#@$$$) (arg6989586621679532537 :: a6989586621679532509) (arg6989586621679532538 :: a6989586621679532509) = * arg6989586621679532537 arg6989586621679532538

data NegateSym0 :: forall a6989586621679532509. (~>) a6989586621679532509 a6989586621679532509

type NegateSym1 (arg6989586621679532541 :: a6989586621679532509) = Negate arg6989586621679532541

data AbsSym0 :: forall a6989586621679532509. (~>) a6989586621679532509 a6989586621679532509

type AbsSym1 (arg6989586621679532543 :: a6989586621679532509) = Abs arg6989586621679532543

data SignumSym0 :: forall a6989586621679532509. (~>) a6989586621679532509 a6989586621679532509

type SignumSym1 (arg6989586621679532545 :: a6989586621679532509) = Signum arg6989586621679532545

data FromIntegerSym0 :: forall a6989586621679532509. (~>) Nat a6989586621679532509

type FromIntegerSym1 (arg6989586621679532547 :: Nat) = FromInteger arg6989586621679532547

data SubtractSym0 :: forall a6989586621679537966. (~>) a6989586621679537966 ((~>) a6989586621679537966 a6989586621679537966)

data SubtractSym1 (a6989586621679537970 :: a6989586621679537966) :: (~>) a6989586621679537966 a6989586621679537966

type SubtractSym2 (a6989586621679537970 :: a6989586621679537966) (a6989586621679537971 :: a6989586621679537966) = Subtract a6989586621679537970 a6989586621679537971