Data.Singletons.Prelude.Tuple

Description

Defines functions and datatypes relating to the singleton for tuples, including a singletons version of all the definitions in Data.Tuple.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Tuple. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Singleton definitions

See Sing for more info.

type family Sing

The singleton kind-indexed type family.

data STuple0 z where

Constructors

STuple0 :: STuple0 ('() :: ())

data STuple2 z where

Constructors

STuple2 :: forall (a :: Type) (b :: Type) (n :: a) (n :: b). (Sing n) -> (Sing n) -> STuple2 ('(n, n) :: (a :: Type, b :: Type))

data STuple3 z where

Constructors

STuple3 :: forall (a :: Type) (b :: Type) (c :: Type) (n :: a) (n :: b) (n :: c). (Sing n) -> (Sing n) -> (Sing n) -> STuple3 ('(n, n, n) :: (a :: Type, b :: Type, c :: Type))

data STuple4 z where

Constructors

STuple4 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (n :: a) (n :: b) (n :: c) (n :: d). (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> STuple4 ('(n, n, n, n) :: (a :: Type, b :: Type, c :: Type, d :: Type))

data STuple5 z where

Constructors

STuple5 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (n :: a) (n :: b) (n :: c) (n :: d) (n :: e). (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> STuple5 ('(n, n, n, n, n) :: (a :: Type, b :: Type, c :: Type, d :: Type, e :: Type))

data STuple6 z where

Constructors

STuple6 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (f :: Type) (n :: a) (n :: b) (n :: c) (n :: d) (n :: e) (n :: f). (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> STuple6 ('(n, n, n, n, n, n) :: (a :: Type, b :: Type, c :: Type, d :: Type, e :: Type, f :: Type))

data STuple7 z where

Constructors

STuple7 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (f :: Type) (g :: Type) (n :: a) (n :: b) (n :: c) (n :: d) (n :: e) (n :: f) (n :: g). (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> (Sing n) -> STuple7 ('(n, n, n, n, n, n, n) :: (a :: Type, b :: Type, c :: Type, d :: Type, e :: Type, f :: Type, g :: Type))

Singletons from Data.Tuple

type family Fst a where ...

Equations

Fst '(x, _) = x

sFst :: forall a b (t :: (a, b)). Sing t -> Sing (Apply FstSym0 t :: a)

type family Snd a where ...

Equations

Snd '(_, y) = y

sSnd :: forall a b (t :: (a, b)). Sing t -> Sing (Apply SndSym0 t :: b)

type family Curry a a a where ...

Equations

Curry f x y = Apply f (Apply (Apply Tuple2Sym0 x) y)

sCurry :: forall a b c (t :: (~>) (a, b) c) (t :: a) (t :: b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply CurrySym0 t) t) t :: c)

type family Uncurry a a where ...

Equations

Uncurry f p = Apply (Apply f (Apply FstSym0 p)) (Apply SndSym0 p)

sUncurry :: forall a b c (t :: (~>) a ((~>) b c)) (t :: (a, b)). Sing t -> Sing t -> Sing (Apply (Apply UncurrySym0 t) t :: c)

type family Swap a where ...

Equations

Swap '(a, b) = Apply (Apply Tuple2Sym0 b) a

sSwap :: forall a b (t :: (a, b)). Sing t -> Sing (Apply SwapSym0 t :: (b, a))

Defunctionalization symbols

type Tuple0Sym0 = '() :: ()

data Tuple2Sym0 a6989586621679304484

data Tuple2Sym1 a6989586621679304484 a6989586621679304485

type Tuple2Sym2 (a6989586621679304484 :: a) (a6989586621679304485 :: b) = '(a6989586621679304484, a6989586621679304485) :: (a :: Type, b :: Type)

data Tuple3Sym0 a6989586621679304515

data Tuple3Sym1 a6989586621679304515 a6989586621679304516

data Tuple3Sym2 a6989586621679304515 a6989586621679304516 a6989586621679304517

type Tuple3Sym3 (a6989586621679304515 :: a) (a6989586621679304516 :: b) (a6989586621679304517 :: c) = '(a6989586621679304515, a6989586621679304516, a6989586621679304517) :: (a :: Type, b :: Type, c :: Type)

data Tuple4Sym0 a6989586621679304561

data Tuple4Sym1 a6989586621679304561 a6989586621679304562

data Tuple4Sym2 a6989586621679304561 a6989586621679304562 a6989586621679304563

data Tuple4Sym3 a6989586621679304561 a6989586621679304562 a6989586621679304563 a6989586621679304564

type Tuple4Sym4 (a6989586621679304561 :: a) (a6989586621679304562 :: b) (a6989586621679304563 :: c) (a6989586621679304564 :: d) = '(a6989586621679304561, a6989586621679304562, a6989586621679304563, a6989586621679304564) :: (a :: Type, b :: Type, c :: Type, d :: Type)

data Tuple5Sym0 a6989586621679304624

data Tuple5Sym1 a6989586621679304624 a6989586621679304625

data Tuple5Sym2 a6989586621679304624 a6989586621679304625 a6989586621679304626

data Tuple5Sym3 a6989586621679304624 a6989586621679304625 a6989586621679304626 a6989586621679304627

data Tuple5Sym4 a6989586621679304624 a6989586621679304625 a6989586621679304626 a6989586621679304627 a6989586621679304628

type Tuple5Sym5 (a6989586621679304624 :: a) (a6989586621679304625 :: b) (a6989586621679304626 :: c) (a6989586621679304627 :: d) (a6989586621679304628 :: e) = '(a6989586621679304624, a6989586621679304625, a6989586621679304626, a6989586621679304627, a6989586621679304628) :: (a :: Type, b :: Type, c :: Type, d :: Type, e :: Type)

data Tuple6Sym0 a6989586621679304706

data Tuple6Sym1 a6989586621679304706 a6989586621679304707

data Tuple6Sym2 a6989586621679304706 a6989586621679304707 a6989586621679304708

data Tuple6Sym3 a6989586621679304706 a6989586621679304707 a6989586621679304708 a6989586621679304709

data Tuple6Sym4 a6989586621679304706 a6989586621679304707 a6989586621679304708 a6989586621679304709 a6989586621679304710

data Tuple6Sym5 a6989586621679304706 a6989586621679304707 a6989586621679304708 a6989586621679304709 a6989586621679304710 a6989586621679304711

type Tuple6Sym6 (a6989586621679304706 :: a) (a6989586621679304707 :: b) (a6989586621679304708 :: c) (a6989586621679304709 :: d) (a6989586621679304710 :: e) (a6989586621679304711 :: f) = '(a6989586621679304706, a6989586621679304707, a6989586621679304708, a6989586621679304709, a6989586621679304710, a6989586621679304711) :: (a :: Type, b :: Type, c :: Type, d :: Type, e :: Type, f :: Type)

data Tuple7Sym0 a6989586621679304809

data Tuple7Sym1 a6989586621679304809 a6989586621679304810

data Tuple7Sym2 a6989586621679304809 a6989586621679304810 a6989586621679304811

data Tuple7Sym3 a6989586621679304809 a6989586621679304810 a6989586621679304811 a6989586621679304812

data Tuple7Sym4 a6989586621679304809 a6989586621679304810 a6989586621679304811 a6989586621679304812 a6989586621679304813

data Tuple7Sym5 a6989586621679304809 a6989586621679304810 a6989586621679304811 a6989586621679304812 a6989586621679304813 a6989586621679304814

data Tuple7Sym6 a6989586621679304809 a6989586621679304810 a6989586621679304811 a6989586621679304812 a6989586621679304813 a6989586621679304814 a6989586621679304815

type Tuple7Sym7 (a6989586621679304809 :: a) (a6989586621679304810 :: b) (a6989586621679304811 :: c) (a6989586621679304812 :: d) (a6989586621679304813 :: e) (a6989586621679304814 :: f) (a6989586621679304815 :: g) = '(a6989586621679304809, a6989586621679304810, a6989586621679304811, a6989586621679304812, a6989586621679304813, a6989586621679304814, a6989586621679304815) :: (a :: Type, b :: Type, c :: Type, d :: Type, e :: Type, f :: Type, g :: Type)

data FstSym0 a6989586621679360573

type FstSym1 (a6989586621679360573 :: (a, b)) = Fst a6989586621679360573 :: a

data SndSym0 a6989586621679360569

type SndSym1 (a6989586621679360569 :: (a, b)) = Snd a6989586621679360569 :: b

data CurrySym0 a6989586621679360561

data CurrySym1 a6989586621679360561 a6989586621679360562

data CurrySym2 a6989586621679360561 a6989586621679360562 a6989586621679360563

type CurrySym3 (a6989586621679360561 :: (~>) (a, b) c) (a6989586621679360562 :: a) (a6989586621679360563 :: b) = Curry a6989586621679360561 a6989586621679360562 a6989586621679360563 :: c

data UncurrySym0 a6989586621679360553

data UncurrySym1 a6989586621679360553 a6989586621679360554

type UncurrySym2 (a6989586621679360553 :: (~>) a ((~>) b c)) (a6989586621679360554 :: (a, b)) = Uncurry a6989586621679360553 a6989586621679360554 :: c

data SwapSym0 a6989586621679360547

type SwapSym1 (a6989586621679360547 :: (a, b)) = Swap a6989586621679360547 :: (b, a)