Data.Singletons.Prelude.Tuple

Description

Defines functions and datatypes relating to the singleton for tuples, including a singletons version of all the definitions in Data.Tuple.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Tuple. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Singleton definitions

See Sing for more info.

type family Sing :: k -> Type

The singleton kind-indexed type family.

data STuple0 :: () -> Type where

Constructors

STuple0 :: STuple0 '()

data STuple2 :: forall a b. (a, b) -> Type where

Constructors

STuple2 :: forall a b (n :: a) (n :: b). (Sing (n :: a)) -> (Sing (n :: b)) -> STuple2 '(n, n)

data STuple3 :: forall a b c. (a, b, c) -> Type where

Constructors

STuple3 :: forall a b c (n :: a) (n :: b) (n :: c). (Sing (n :: a)) -> (Sing (n :: b)) -> (Sing (n :: c)) -> STuple3 '(n, n, n)

data STuple4 :: forall a b c d. (a, b, c, d) -> Type where

Constructors

STuple4 :: forall a b c d (n :: a) (n :: b) (n :: c) (n :: d). (Sing (n :: a)) -> (Sing (n :: b)) -> (Sing (n :: c)) -> (Sing (n :: d)) -> STuple4 '(n, n, n, n)

data STuple5 :: forall a b c d e. (a, b, c, d, e) -> Type where

Constructors

STuple5 :: forall a b c d e (n :: a) (n :: b) (n :: c) (n :: d) (n :: e). (Sing (n :: a)) -> (Sing (n :: b)) -> (Sing (n :: c)) -> (Sing (n :: d)) -> (Sing (n :: e)) -> STuple5 '(n, n, n, n, n)

data STuple6 :: forall a b c d e f. (a, b, c, d, e, f) -> Type where

Constructors

STuple6 :: forall a b c d e f (n :: a) (n :: b) (n :: c) (n :: d) (n :: e) (n :: f). (Sing (n :: a)) -> (Sing (n :: b)) -> (Sing (n :: c)) -> (Sing (n :: d)) -> (Sing (n :: e)) -> (Sing (n :: f)) -> STuple6 '(n, n, n, n, n, n)

data STuple7 :: forall a b c d e f g. (a, b, c, d, e, f, g) -> Type where

Constructors

STuple7 :: forall a b c d e f g (n :: a) (n :: b) (n :: c) (n :: d) (n :: e) (n :: f) (n :: g). (Sing (n :: a)) -> (Sing (n :: b)) -> (Sing (n :: c)) -> (Sing (n :: d)) -> (Sing (n :: e)) -> (Sing (n :: f)) -> (Sing (n :: g)) -> STuple7 '(n, n, n, n, n, n, n)

Singletons from Data.Tuple

type family Fst (a :: (a, b)) :: a where ...

Equations

Fst '(x, _) = x

sFst :: forall a b (t :: (a, b)). Sing t -> Sing (Apply FstSym0 t :: a)

type family Snd (a :: (a, b)) :: b where ...

Equations

Snd '(_, y) = y

sSnd :: forall a b (t :: (a, b)). Sing t -> Sing (Apply SndSym0 t :: b)

type family Curry (a :: (~>) (a, b) c) (a :: a) (a :: b) :: c where ...

Equations

Curry f x y = Apply f (Apply (Apply Tuple2Sym0 x) y)

sCurry :: forall a b c (t :: (~>) (a, b) c) (t :: a) (t :: b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply CurrySym0 t) t) t :: c)

type family Uncurry (a :: (~>) a ((~>) b c)) (a :: (a, b)) :: c where ...

Equations

Uncurry f p = Apply (Apply f (Apply FstSym0 p)) (Apply SndSym0 p)

sUncurry :: forall a b c (t :: (~>) a ((~>) b c)) (t :: (a, b)). Sing t -> Sing t -> Sing (Apply (Apply UncurrySym0 t) t :: c)

type family Swap (a :: (a, b)) :: (b, a) where ...

Equations

Swap '(a, b) = Apply (Apply Tuple2Sym0 b) a

sSwap :: forall a b (t :: (a, b)). Sing t -> Sing (Apply SwapSym0 t :: (b, a))

Defunctionalization symbols

type Tuple0Sym0 = '()

data Tuple2Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 (a3530822107858468865 :: Type, b3530822107858468866 :: Type))

data Tuple2Sym1 (t6989586621679316808 :: a3530822107858468865 :: Type) :: forall (b3530822107858468866 :: Type). (~>) b3530822107858468866 (a3530822107858468865 :: Type, b3530822107858468866 :: Type)

type Tuple2Sym2 (t6989586621679316808 :: a3530822107858468865) (t6989586621679316809 :: b3530822107858468866) = '(t6989586621679316808, t6989586621679316809)

data Tuple3Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type)))

data Tuple3Sym1 (t6989586621679316839 :: a3530822107858468865 :: Type) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type))

data Tuple3Sym2 (t6989586621679316839 :: a3530822107858468865 :: Type) (t6989586621679316840 :: b3530822107858468866 :: Type) :: forall (c3530822107858468867 :: Type). (~>) c3530822107858468867 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type)

type Tuple3Sym3 (t6989586621679316839 :: a3530822107858468865) (t6989586621679316840 :: b3530822107858468866) (t6989586621679316841 :: c3530822107858468867) = '(t6989586621679316839, t6989586621679316840, t6989586621679316841)

data Tuple4Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type))))

data Tuple4Sym1 (t6989586621679316886 :: a3530822107858468865 :: Type) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type)))

data Tuple4Sym2 (t6989586621679316886 :: a3530822107858468865 :: Type) (t6989586621679316887 :: b3530822107858468866 :: Type) :: forall (c3530822107858468867 :: Type) (d3530822107858468868 :: Type). (~>) c3530822107858468867 ((~>) d3530822107858468868 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type))

data Tuple4Sym3 (t6989586621679316886 :: a3530822107858468865 :: Type) (t6989586621679316887 :: b3530822107858468866 :: Type) (t6989586621679316888 :: c3530822107858468867 :: Type) :: forall (d3530822107858468868 :: Type). (~>) d3530822107858468868 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type)

type Tuple4Sym4 (t6989586621679316886 :: a3530822107858468865) (t6989586621679316887 :: b3530822107858468866) (t6989586621679316888 :: c3530822107858468867) (t6989586621679316889 :: d3530822107858468868) = '(t6989586621679316886, t6989586621679316887, t6989586621679316888, t6989586621679316889)

data Tuple5Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type)))))

data Tuple5Sym1 (t6989586621679316951 :: a3530822107858468865 :: Type) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type))))

data Tuple5Sym2 (t6989586621679316951 :: a3530822107858468865 :: Type) (t6989586621679316952 :: b3530822107858468866 :: Type) :: forall (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type). (~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type)))

data Tuple5Sym3 (t6989586621679316951 :: a3530822107858468865 :: Type) (t6989586621679316952 :: b3530822107858468866 :: Type) (t6989586621679316953 :: c3530822107858468867 :: Type) :: forall (d3530822107858468868 :: Type) (e3530822107858468869 :: Type). (~>) d3530822107858468868 ((~>) e3530822107858468869 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type))

data Tuple5Sym4 (t6989586621679316951 :: a3530822107858468865 :: Type) (t6989586621679316952 :: b3530822107858468866 :: Type) (t6989586621679316953 :: c3530822107858468867 :: Type) (t6989586621679316954 :: d3530822107858468868 :: Type) :: forall (e3530822107858468869 :: Type). (~>) e3530822107858468869 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type)

type Tuple5Sym5 (t6989586621679316951 :: a3530822107858468865) (t6989586621679316952 :: b3530822107858468866) (t6989586621679316953 :: c3530822107858468867) (t6989586621679316954 :: d3530822107858468868) (t6989586621679316955 :: e3530822107858468869) = '(t6989586621679316951, t6989586621679316952, t6989586621679316953, t6989586621679316954, t6989586621679316955)

data Tuple6Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type))))))

data Tuple6Sym1 (t6989586621679317036 :: a3530822107858468865 :: Type) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type)))))

data Tuple6Sym2 (t6989586621679317036 :: a3530822107858468865 :: Type) (t6989586621679317037 :: b3530822107858468866 :: Type) :: forall (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type))))

data Tuple6Sym3 (t6989586621679317036 :: a3530822107858468865 :: Type) (t6989586621679317037 :: b3530822107858468866 :: Type) (t6989586621679317038 :: c3530822107858468867 :: Type) :: forall (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type)))

data Tuple6Sym4 (t6989586621679317036 :: a3530822107858468865 :: Type) (t6989586621679317037 :: b3530822107858468866 :: Type) (t6989586621679317038 :: c3530822107858468867 :: Type) (t6989586621679317039 :: d3530822107858468868 :: Type) :: forall (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) e3530822107858468869 ((~>) f3530822107858468870 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type))

data Tuple6Sym5 (t6989586621679317036 :: a3530822107858468865 :: Type) (t6989586621679317037 :: b3530822107858468866 :: Type) (t6989586621679317038 :: c3530822107858468867 :: Type) (t6989586621679317039 :: d3530822107858468868 :: Type) (t6989586621679317040 :: e3530822107858468869 :: Type) :: forall (f3530822107858468870 :: Type). (~>) f3530822107858468870 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type)

type Tuple6Sym6 (t6989586621679317036 :: a3530822107858468865) (t6989586621679317037 :: b3530822107858468866) (t6989586621679317038 :: c3530822107858468867) (t6989586621679317039 :: d3530822107858468868) (t6989586621679317040 :: e3530822107858468869) (t6989586621679317041 :: f3530822107858468870) = '(t6989586621679317036, t6989586621679317037, t6989586621679317038, t6989586621679317039, t6989586621679317040, t6989586621679317041)

data Tuple7Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type, g3530822107858468871 :: Type)))))))

data Tuple7Sym1 (t6989586621679317143 :: a3530822107858468865 :: Type) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type, g3530822107858468871 :: Type))))))

data Tuple7Sym2 (t6989586621679317143 :: a3530822107858468865 :: Type) (t6989586621679317144 :: b3530822107858468866 :: Type) :: forall (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type, g3530822107858468871 :: Type)))))

data Tuple7Sym3 (t6989586621679317143 :: a3530822107858468865 :: Type) (t6989586621679317144 :: b3530822107858468866 :: Type) (t6989586621679317145 :: c3530822107858468867 :: Type) :: forall (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type, g3530822107858468871 :: Type))))

data Tuple7Sym4 (t6989586621679317143 :: a3530822107858468865 :: Type) (t6989586621679317144 :: b3530822107858468866 :: Type) (t6989586621679317145 :: c3530822107858468867 :: Type) (t6989586621679317146 :: d3530822107858468868 :: Type) :: forall (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type, g3530822107858468871 :: Type)))

data Tuple7Sym5 (t6989586621679317143 :: a3530822107858468865 :: Type) (t6989586621679317144 :: b3530822107858468866 :: Type) (t6989586621679317145 :: c3530822107858468867 :: Type) (t6989586621679317146 :: d3530822107858468868 :: Type) (t6989586621679317147 :: e3530822107858468869 :: Type) :: forall (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) f3530822107858468870 ((~>) g3530822107858468871 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type, g3530822107858468871 :: Type))

data Tuple7Sym6 (t6989586621679317143 :: a3530822107858468865 :: Type) (t6989586621679317144 :: b3530822107858468866 :: Type) (t6989586621679317145 :: c3530822107858468867 :: Type) (t6989586621679317146 :: d3530822107858468868 :: Type) (t6989586621679317147 :: e3530822107858468869 :: Type) (t6989586621679317148 :: f3530822107858468870 :: Type) :: forall (g3530822107858468871 :: Type). (~>) g3530822107858468871 (a3530822107858468865 :: Type, b3530822107858468866 :: Type, c3530822107858468867 :: Type, d3530822107858468868 :: Type, e3530822107858468869 :: Type, f3530822107858468870 :: Type, g3530822107858468871 :: Type)

type Tuple7Sym7 (t6989586621679317143 :: a3530822107858468865) (t6989586621679317144 :: b3530822107858468866) (t6989586621679317145 :: c3530822107858468867) (t6989586621679317146 :: d3530822107858468868) (t6989586621679317147 :: e3530822107858468869) (t6989586621679317148 :: f3530822107858468870) (t6989586621679317149 :: g3530822107858468871) = '(t6989586621679317143, t6989586621679317144, t6989586621679317145, t6989586621679317146, t6989586621679317147, t6989586621679317148, t6989586621679317149)

data FstSym0 :: forall a6989586621679371873 b6989586621679371874. (~>) (a6989586621679371873, b6989586621679371874) a6989586621679371873

type FstSym1 (a6989586621679371975 :: (a6989586621679371873, b6989586621679371874)) = Fst a6989586621679371975

data SndSym0 :: forall a6989586621679371871 b6989586621679371872. (~>) (a6989586621679371871, b6989586621679371872) b6989586621679371872

type SndSym1 (a6989586621679371972 :: (a6989586621679371871, b6989586621679371872)) = Snd a6989586621679371972

data CurrySym0 :: forall a6989586621679371868 b6989586621679371869 c6989586621679371870. (~>) ((~>) (a6989586621679371868, b6989586621679371869) c6989586621679371870) ((~>) a6989586621679371868 ((~>) b6989586621679371869 c6989586621679371870))

data CurrySym1 (a6989586621679371963 :: (~>) (a6989586621679371868, b6989586621679371869) c6989586621679371870) :: (~>) a6989586621679371868 ((~>) b6989586621679371869 c6989586621679371870)

data CurrySym2 (a6989586621679371963 :: (~>) (a6989586621679371868, b6989586621679371869) c6989586621679371870) (a6989586621679371964 :: a6989586621679371868) :: (~>) b6989586621679371869 c6989586621679371870

type CurrySym3 (a6989586621679371963 :: (~>) (a6989586621679371868, b6989586621679371869) c6989586621679371870) (a6989586621679371964 :: a6989586621679371868) (a6989586621679371965 :: b6989586621679371869) = Curry a6989586621679371963 a6989586621679371964 a6989586621679371965

data UncurrySym0 :: forall a6989586621679371865 b6989586621679371866 c6989586621679371867. (~>) ((~>) a6989586621679371865 ((~>) b6989586621679371866 c6989586621679371867)) ((~>) (a6989586621679371865, b6989586621679371866) c6989586621679371867)

data UncurrySym1 (a6989586621679371957 :: (~>) a6989586621679371865 ((~>) b6989586621679371866 c6989586621679371867)) :: (~>) (a6989586621679371865, b6989586621679371866) c6989586621679371867

type UncurrySym2 (a6989586621679371957 :: (~>) a6989586621679371865 ((~>) b6989586621679371866 c6989586621679371867)) (a6989586621679371958 :: (a6989586621679371865, b6989586621679371866)) = Uncurry a6989586621679371957 a6989586621679371958

data SwapSym0 :: forall a6989586621679371863 b6989586621679371864. (~>) (a6989586621679371863, b6989586621679371864) (b6989586621679371864, a6989586621679371863)

type SwapSym1 (a6989586621679371953 :: (a6989586621679371863, b6989586621679371864)) = Swap a6989586621679371953