Data.Singletons.Prelude.Tuple

Description

Defines functions and datatypes relating to the singleton for tuples, including a singletons version of all the definitions in Data.Tuple.

Because many of these definitions are produced by Template Haskell, it is not possible to create proper Haddock documentation. Please look up the corresponding operation in Data.Tuple. Also, please excuse the apparent repeated variable names. This is due to an interaction between Template Haskell and Haddock.

Singleton definitions

See Sing for more info.

data family Sing :: k -> Type

The singleton kind-indexed data family.

type STuple0 = (Sing :: () -> Type)

type STuple2 = (Sing :: (a, b) -> Type)

type STuple3 = (Sing :: (a, b, c) -> Type)

type STuple4 = (Sing :: (a, b, c, d) -> Type)

type STuple5 = (Sing :: (a, b, c, d, e) -> Type)

type STuple6 = (Sing :: (a, b, c, d, e, f) -> Type)

type STuple7 = (Sing :: (a, b, c, d, e, f, g) -> Type)

Singletons from Data.Tuple

type family Fst (a :: (a, b)) :: a where ...

Equations

Fst '(x, _) = x

sFst :: forall a b (t :: (a, b)). Sing t -> Sing (Apply FstSym0 t :: a)

type family Snd (a :: (a, b)) :: b where ...

Equations

Snd '(_, y) = y

sSnd :: forall a b (t :: (a, b)). Sing t -> Sing (Apply SndSym0 t :: b)

type family Curry (a :: (~>) (a, b) c) (a :: a) (a :: b) :: c where ...

Equations

Curry f x y = Apply f (Apply (Apply Tuple2Sym0 x) y)

sCurry :: forall a b c (t :: (~>) (a, b) c) (t :: a) (t :: b). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply CurrySym0 t) t) t :: c)

type family Uncurry (a :: (~>) a ((~>) b c)) (a :: (a, b)) :: c where ...

Equations

Uncurry f p = Apply (Apply f (Apply FstSym0 p)) (Apply SndSym0 p)

sUncurry :: forall a b c (t :: (~>) a ((~>) b c)) (t :: (a, b)). Sing t -> Sing t -> Sing (Apply (Apply UncurrySym0 t) t :: c)

type family Swap (a :: (a, b)) :: (b, a) where ...

Equations

Swap '(a, b) = Apply (Apply Tuple2Sym0 b) a

sSwap :: forall a b (t :: (a, b)). Sing t -> Sing (Apply SwapSym0 t :: (b, a))

Defunctionalization symbols

type Tuple0Sym0 = '()

data Tuple2Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type)))

data Tuple2Sym1 (t6989586621679317691 :: (a3530822107858468865 :: Type)) :: forall (b3530822107858468866 :: Type). (~>) b3530822107858468866 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type))

type Tuple2Sym2 (t6989586621679317691 :: a3530822107858468865) (t6989586621679317692 :: b3530822107858468866) = '(t6989586621679317691, t6989586621679317692)

data Tuple3Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type))))

data Tuple3Sym1 (t6989586621679317722 :: (a3530822107858468865 :: Type)) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type)))

data Tuple3Sym2 (t6989586621679317722 :: (a3530822107858468865 :: Type)) (t6989586621679317723 :: (b3530822107858468866 :: Type)) :: forall (c3530822107858468867 :: Type). (~>) c3530822107858468867 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type))

type Tuple3Sym3 (t6989586621679317722 :: a3530822107858468865) (t6989586621679317723 :: b3530822107858468866) (t6989586621679317724 :: c3530822107858468867) = '(t6989586621679317722, t6989586621679317723, t6989586621679317724)

data Tuple4Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type)))))

data Tuple4Sym1 (t6989586621679317769 :: (a3530822107858468865 :: Type)) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type))))

data Tuple4Sym2 (t6989586621679317769 :: (a3530822107858468865 :: Type)) (t6989586621679317770 :: (b3530822107858468866 :: Type)) :: forall (c3530822107858468867 :: Type) (d3530822107858468868 :: Type). (~>) c3530822107858468867 ((~>) d3530822107858468868 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type)))

data Tuple4Sym3 (t6989586621679317769 :: (a3530822107858468865 :: Type)) (t6989586621679317770 :: (b3530822107858468866 :: Type)) (t6989586621679317771 :: (c3530822107858468867 :: Type)) :: forall (d3530822107858468868 :: Type). (~>) d3530822107858468868 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type))

type Tuple4Sym4 (t6989586621679317769 :: a3530822107858468865) (t6989586621679317770 :: b3530822107858468866) (t6989586621679317771 :: c3530822107858468867) (t6989586621679317772 :: d3530822107858468868) = '(t6989586621679317769, t6989586621679317770, t6989586621679317771, t6989586621679317772)

data Tuple5Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type))))))

data Tuple5Sym1 (t6989586621679317834 :: (a3530822107858468865 :: Type)) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type)))))

data Tuple5Sym2 (t6989586621679317834 :: (a3530822107858468865 :: Type)) (t6989586621679317835 :: (b3530822107858468866 :: Type)) :: forall (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type). (~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type))))

data Tuple5Sym3 (t6989586621679317834 :: (a3530822107858468865 :: Type)) (t6989586621679317835 :: (b3530822107858468866 :: Type)) (t6989586621679317836 :: (c3530822107858468867 :: Type)) :: forall (d3530822107858468868 :: Type) (e3530822107858468869 :: Type). (~>) d3530822107858468868 ((~>) e3530822107858468869 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type)))

data Tuple5Sym4 (t6989586621679317834 :: (a3530822107858468865 :: Type)) (t6989586621679317835 :: (b3530822107858468866 :: Type)) (t6989586621679317836 :: (c3530822107858468867 :: Type)) (t6989586621679317837 :: (d3530822107858468868 :: Type)) :: forall (e3530822107858468869 :: Type). (~>) e3530822107858468869 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type))

type Tuple5Sym5 (t6989586621679317834 :: a3530822107858468865) (t6989586621679317835 :: b3530822107858468866) (t6989586621679317836 :: c3530822107858468867) (t6989586621679317837 :: d3530822107858468868) (t6989586621679317838 :: e3530822107858468869) = '(t6989586621679317834, t6989586621679317835, t6989586621679317836, t6989586621679317837, t6989586621679317838)

data Tuple6Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type)))))))

data Tuple6Sym1 (t6989586621679317919 :: (a3530822107858468865 :: Type)) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type))))))

data Tuple6Sym2 (t6989586621679317919 :: (a3530822107858468865 :: Type)) (t6989586621679317920 :: (b3530822107858468866 :: Type)) :: forall (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type)))))

data Tuple6Sym3 (t6989586621679317919 :: (a3530822107858468865 :: Type)) (t6989586621679317920 :: (b3530822107858468866 :: Type)) (t6989586621679317921 :: (c3530822107858468867 :: Type)) :: forall (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type))))

data Tuple6Sym4 (t6989586621679317919 :: (a3530822107858468865 :: Type)) (t6989586621679317920 :: (b3530822107858468866 :: Type)) (t6989586621679317921 :: (c3530822107858468867 :: Type)) (t6989586621679317922 :: (d3530822107858468868 :: Type)) :: forall (e3530822107858468869 :: Type) (f3530822107858468870 :: Type). (~>) e3530822107858468869 ((~>) f3530822107858468870 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type)))

data Tuple6Sym5 (t6989586621679317919 :: (a3530822107858468865 :: Type)) (t6989586621679317920 :: (b3530822107858468866 :: Type)) (t6989586621679317921 :: (c3530822107858468867 :: Type)) (t6989586621679317922 :: (d3530822107858468868 :: Type)) (t6989586621679317923 :: (e3530822107858468869 :: Type)) :: forall (f3530822107858468870 :: Type). (~>) f3530822107858468870 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type))

type Tuple6Sym6 (t6989586621679317919 :: a3530822107858468865) (t6989586621679317920 :: b3530822107858468866) (t6989586621679317921 :: c3530822107858468867) (t6989586621679317922 :: d3530822107858468868) (t6989586621679317923 :: e3530822107858468869) (t6989586621679317924 :: f3530822107858468870) = '(t6989586621679317919, t6989586621679317920, t6989586621679317921, t6989586621679317922, t6989586621679317923, t6989586621679317924)

data Tuple7Sym0 :: forall (a3530822107858468865 :: Type) (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) a3530822107858468865 ((~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type), (g3530822107858468871 :: Type))))))))

data Tuple7Sym1 (t6989586621679318026 :: (a3530822107858468865 :: Type)) :: forall (b3530822107858468866 :: Type) (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) b3530822107858468866 ((~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type), (g3530822107858468871 :: Type)))))))

data Tuple7Sym2 (t6989586621679318026 :: (a3530822107858468865 :: Type)) (t6989586621679318027 :: (b3530822107858468866 :: Type)) :: forall (c3530822107858468867 :: Type) (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) c3530822107858468867 ((~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type), (g3530822107858468871 :: Type))))))

data Tuple7Sym3 (t6989586621679318026 :: (a3530822107858468865 :: Type)) (t6989586621679318027 :: (b3530822107858468866 :: Type)) (t6989586621679318028 :: (c3530822107858468867 :: Type)) :: forall (d3530822107858468868 :: Type) (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) d3530822107858468868 ((~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type), (g3530822107858468871 :: Type)))))

data Tuple7Sym4 (t6989586621679318026 :: (a3530822107858468865 :: Type)) (t6989586621679318027 :: (b3530822107858468866 :: Type)) (t6989586621679318028 :: (c3530822107858468867 :: Type)) (t6989586621679318029 :: (d3530822107858468868 :: Type)) :: forall (e3530822107858468869 :: Type) (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) e3530822107858468869 ((~>) f3530822107858468870 ((~>) g3530822107858468871 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type), (g3530822107858468871 :: Type))))

data Tuple7Sym5 (t6989586621679318026 :: (a3530822107858468865 :: Type)) (t6989586621679318027 :: (b3530822107858468866 :: Type)) (t6989586621679318028 :: (c3530822107858468867 :: Type)) (t6989586621679318029 :: (d3530822107858468868 :: Type)) (t6989586621679318030 :: (e3530822107858468869 :: Type)) :: forall (f3530822107858468870 :: Type) (g3530822107858468871 :: Type). (~>) f3530822107858468870 ((~>) g3530822107858468871 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type), (g3530822107858468871 :: Type)))

data Tuple7Sym6 (t6989586621679318026 :: (a3530822107858468865 :: Type)) (t6989586621679318027 :: (b3530822107858468866 :: Type)) (t6989586621679318028 :: (c3530822107858468867 :: Type)) (t6989586621679318029 :: (d3530822107858468868 :: Type)) (t6989586621679318030 :: (e3530822107858468869 :: Type)) (t6989586621679318031 :: (f3530822107858468870 :: Type)) :: forall (g3530822107858468871 :: Type). (~>) g3530822107858468871 ((a3530822107858468865 :: Type), (b3530822107858468866 :: Type), (c3530822107858468867 :: Type), (d3530822107858468868 :: Type), (e3530822107858468869 :: Type), (f3530822107858468870 :: Type), (g3530822107858468871 :: Type))

type Tuple7Sym7 (t6989586621679318026 :: a3530822107858468865) (t6989586621679318027 :: b3530822107858468866) (t6989586621679318028 :: c3530822107858468867) (t6989586621679318029 :: d3530822107858468868) (t6989586621679318030 :: e3530822107858468869) (t6989586621679318031 :: f3530822107858468870) (t6989586621679318032 :: g3530822107858468871) = '(t6989586621679318026, t6989586621679318027, t6989586621679318028, t6989586621679318029, t6989586621679318030, t6989586621679318031, t6989586621679318032)

data FstSym0 :: forall a6989586621679376647 b6989586621679376648. (~>) (a6989586621679376647, b6989586621679376648) a6989586621679376647

type FstSym1 (a6989586621679376743 :: (a6989586621679376647, b6989586621679376648)) = Fst a6989586621679376743

data SndSym0 :: forall a6989586621679376645 b6989586621679376646. (~>) (a6989586621679376645, b6989586621679376646) b6989586621679376646

type SndSym1 (a6989586621679376740 :: (a6989586621679376645, b6989586621679376646)) = Snd a6989586621679376740

data CurrySym0 :: forall a6989586621679376642 b6989586621679376643 c6989586621679376644. (~>) ((~>) (a6989586621679376642, b6989586621679376643) c6989586621679376644) ((~>) a6989586621679376642 ((~>) b6989586621679376643 c6989586621679376644))

data CurrySym1 (a6989586621679376731 :: (~>) (a6989586621679376642, b6989586621679376643) c6989586621679376644) :: (~>) a6989586621679376642 ((~>) b6989586621679376643 c6989586621679376644)

data CurrySym2 (a6989586621679376731 :: (~>) (a6989586621679376642, b6989586621679376643) c6989586621679376644) (a6989586621679376732 :: a6989586621679376642) :: (~>) b6989586621679376643 c6989586621679376644

type CurrySym3 (a6989586621679376731 :: (~>) (a6989586621679376642, b6989586621679376643) c6989586621679376644) (a6989586621679376732 :: a6989586621679376642) (a6989586621679376733 :: b6989586621679376643) = Curry a6989586621679376731 a6989586621679376732 a6989586621679376733

data UncurrySym0 :: forall a6989586621679376639 b6989586621679376640 c6989586621679376641. (~>) ((~>) a6989586621679376639 ((~>) b6989586621679376640 c6989586621679376641)) ((~>) (a6989586621679376639, b6989586621679376640) c6989586621679376641)

data UncurrySym1 (a6989586621679376746 :: (~>) a6989586621679376639 ((~>) b6989586621679376640 c6989586621679376641)) :: (~>) (a6989586621679376639, b6989586621679376640) c6989586621679376641

type UncurrySym2 (a6989586621679376746 :: (~>) a6989586621679376639 ((~>) b6989586621679376640 c6989586621679376641)) (a6989586621679376747 :: (a6989586621679376639, b6989586621679376640)) = Uncurry a6989586621679376746 a6989586621679376747

data SwapSym0 :: forall a6989586621679376637 b6989586621679376638. (~>) (a6989586621679376637, b6989586621679376638) (b6989586621679376638, a6989586621679376637)

type SwapSym1 (a6989586621679376727 :: (a6989586621679376637, b6989586621679376638)) = Swap a6989586621679376727