Data.Singletons.TH

Description

This module contains everything you need to derive your own singletons via Template Haskell.

TURN ON -XScopedTypeVariables IN YOUR MODULE IF YOU WANT THIS TO WORK.

Primary Template Haskell generation functions

singletons :: DsMonad q => q [Dec] -> q [Dec]

Make promoted and singleton versions of all declarations given, retaining the original declarations. See http://www.cis.upenn.edu/~eir/packages/singletons/README.html for further explanation.

singletonsOnly :: DsMonad q => q [Dec] -> q [Dec]

Make promoted and singleton versions of all declarations given, discarding the original declarations. Note that a singleton based on a datatype needs the original datatype, so this will fail if it sees any datatype declarations. Classes, instances, and functions are all fine.

genSingletons :: DsMonad q => [Name] -> q [Dec]

Generate singleton definitions from a type that is already defined. For example, the singletons package itself uses

$(genSingletons [''Bool, ''Maybe, ''Either, ''[]])

to generate singletons for Prelude types.

promote :: DsMonad q => q [Dec] -> q [Dec]

Promote every declaration given to the type level, retaining the originals.

promoteOnly :: DsMonad q => q [Dec] -> q [Dec]

Promote each declaration, discarding the originals. Note that a promoted datatype uses the same definition as an original datatype, so this will not work with datatypes. Classes, instances, and functions are all fine.

genDefunSymbols :: DsMonad q => [Name] -> q [Dec]

Generate defunctionalization symbols for existing type family

genPromotions :: DsMonad q => [Name] -> q [Dec]

Generate promoted definitions from a type that is already defined. This is generally only useful with classes.

Functions to generate equality instances

promoteEqInstances :: DsMonad q => [Name] -> q [Dec]

Produce instances for '(:==)' (type-level equality) from the given types

promoteEqInstance :: DsMonad q => Name -> q [Dec]

Produce an instance for '(:==)' (type-level equality) from the given type

singEqInstances :: DsMonad q => [Name] -> q [Dec]

Create instances of SEq and type-level '(:==)' for each type in the list

singEqInstance :: DsMonad q => Name -> q [Dec]

Create instance of SEq and type-level '(:==)' for the given type

singEqInstancesOnly :: DsMonad q => [Name] -> q [Dec]

Create instances of SEq (only -- no instance for '(:==)', which SEq generally relies on) for each type in the list

singEqInstanceOnly :: DsMonad q => Name -> q [Dec]

Create instances of SEq (only -- no instance for '(:==)', which SEq generally relies on) for the given type

singDecideInstances :: DsMonad q => [Name] -> q [Dec]

Create instances of SDecide for each type in the list.

singDecideInstance :: DsMonad q => Name -> q [Dec]

Create instance of SDecide for the given type.

Functions to generate Ord instances

promoteOrdInstances :: DsMonad q => [Name] -> q [Dec]

Produce instances for POrd from the given types

promoteOrdInstance :: DsMonad q => Name -> q [Dec]

Produce an instance for POrd from the given type

singOrdInstances :: DsMonad q => [Name] -> q [Dec]

Create instances of SOrd for the given types

singOrdInstance :: DsMonad q => Name -> q [Dec]

Create instance of SOrd for the given type

Functions to generate Bounded instances

promoteBoundedInstances :: DsMonad q => [Name] -> q [Dec]

Produce instances for PBounded from the given types

promoteBoundedInstance :: DsMonad q => Name -> q [Dec]

Produce an instance for PBounded from the given type

singBoundedInstances :: DsMonad q => [Name] -> q [Dec]

Create instances of SBounded for the given types

singBoundedInstance :: DsMonad q => Name -> q [Dec]

Create instance of SBounded for the given type

Functions to generate Enum instances

promoteEnumInstances :: DsMonad q => [Name] -> q [Dec]

Produce instances for PEnum from the given types

promoteEnumInstance :: DsMonad q => Name -> q [Dec]

Produce an instance for PEnum from the given type

singEnumInstances :: DsMonad q => [Name] -> q [Dec]

Create instances of SEnum for the given types

singEnumInstance :: DsMonad q => Name -> q [Dec]

Create instance of SEnum for the given type

Utility functions

cases

Arguments

:: DsMonad q

=> Name
The head of the type of the scrutinee. (Like ''Maybe or ''Bool.)

-> q Exp
The scrutinee, in a Template Haskell quote

-> q Exp
The body, in a Template Haskell quote

-> q Exp

The function cases generates a case expression where each right-hand side is identical. This may be useful if the type-checker requires knowledge of which constructor is used to satisfy equality or type-class constraints, but where each constructor is treated the same.

sCases

Arguments

:: DsMonad q

=> Name
The head of the type the scrutinee's type is based on. (Like ''Maybe or ''Bool.)

-> q Exp
The scrutinee, in a Template Haskell quote

-> q Exp
The body, in a Template Haskell quote

-> q Exp

The function sCases generates a case expression where each right-hand side is identical. This may be useful if the type-checker requires knowledge of which constructor is used to satisfy equality or type-class constraints, but where each constructor is treated the same. For sCases, unlike cases, the scrutinee is a singleton. But make sure to pass in the name of the original datatype, preferring ''Maybe over ''SMaybe.

Basic singleton definitions

data family Sing a

The singleton kind-indexed data family.

module Data.Singletons

Auxiliary definitions

These definitions might be mentioned in code generated by Template Haskell, so they must be in scope.

class (kproxy ~ KProxy) => PEq kproxy

The promoted analogue of Eq. If you supply no definition for '(:==)', then it defaults to a use of '(==)', from Data.Type.Equality.

Associated Types

type x :== y :: Bool infix 4

type x :/= y :: Bool infix 4

type family If cond tru fls :: k

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If k True tru fls = tru

If k False tru fls = fls

sIf :: Sing a -> Sing b -> Sing c -> Sing (If a b c)

Conditional over singletons

type family a :&& a :: Bool infixr 3

Equations

False :&& _z_1627785679 = FalseSym0

True :&& x = x

class (kparam ~ KProxy) => SEq kparam where

The singleton analogue of Eq. Unlike the definition for Eq, it is required that instances define a body for '(%:==)'. You may also supply a body for '(%:/=)'.

Minimal complete definition

(%:==)

Methods

(%:==) :: forall a b. Sing a -> Sing b -> Sing (a :== b) infix 4

Boolean equality on singletons

(%:/=) :: forall a b. Sing a -> Sing b -> Sing (a :/= b) infix 4

Boolean disequality on singletons

class (PEq (KProxy :: KProxy a), kproxy ~ KProxy) => POrd kproxy

Associated Types

type Compare arg arg :: Ordering

type arg :< arg :: Bool infix 4

type arg :<= arg :: Bool infix 4

type arg :> arg :: Bool infix 4

type arg :>= arg :: Bool infix 4

type Max arg arg :: a

type Min arg arg :: a

class (SEq (KProxy :: KProxy a), kproxy ~ KProxy) => SOrd kproxy where

Minimal complete definition

Nothing

Methods

sCompare :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply CompareSym0 t) t :: Ordering)

(%:<) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:<$) t) t :: Bool) infix 4

(%:<=) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:<=$) t) t :: Bool) infix 4

(%:>) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:>$) t) t :: Bool) infix 4

(%:>=) :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply (:>=$) t) t :: Bool) infix 4

sMax :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply MaxSym0 t) t :: a)

sMin :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply MinSym0 t) t :: a)

type family ThenCmp a a :: Ordering

Equations

ThenCmp EQ x = x

ThenCmp LT _z_1627819454 = LTSym0

ThenCmp GT _z_1627819457 = GTSym0

sThenCmp :: forall t t. Sing t -> Sing t -> Sing (Apply (Apply ThenCmpSym0 t) t :: Ordering)

type family Foldl a a a :: b

Equations

Foldl f z0 xs0 = Apply (Apply (Let1627751476LgoSym3 f z0 xs0) z0) xs0

sFoldl :: forall t t t. Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldlSym0 t) t) t :: b)

type family Any :: k

The type constructor Any is type to which you can unsafely coerce any lifted type, and back.

It is lifted, and hence represented by a pointer

It does not claim to be a data type, and that's important for the code generator, because the code gen may enter a data value but never enters a function value.

It's also used to instantiate un-constrained type variables after type checking. For example, length has type

length :: forall a. [a] -> Int

and the list datacon for the empty list has type

[] :: forall a. [a]

In order to compose these two terms as length [] a type application is required, but there is no constraint on the choice. In this situation GHC uses Any:

length (Any *) ([] (Any *))

Above, we print kinds explicitly, as if with -fprint-explicit-kinds.

Note that Any is kind polymorphic; its kind is thus forall k. k.

class (kparam ~ KProxy) => SDecide kparam where

Members of the SDecide "kind" class support decidable equality. Instances of this class are generated alongside singleton definitions for datatypes that derive an Eq instance.

Methods

(%~) :: forall a b. Sing a -> Sing b -> Decision (a :~: b)

Compute a proof or disproof of equality, given two singletons.

data a :~: b :: k -> k -> * where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: (:~:) k b b

data Void :: *

Uninhabited data type

Since: 4.8.0.0

type Refuted a = a -> Void

Because we can never create a value of type Void, a function that type-checks at a -> Void shows that objects of type a can never exist. Thus, we say that a is Refuted

data Decision a

A Decision about a type a is either a proof of existence or a proof that a cannot exist.

Constructors

Proved a
Witness for a

Disproved (Refuted a)
Proof that no a exists

data Proxy t :: k -> *

A concrete, poly-kinded proxy type

Constructors

Proxy

data KProxy t :: * -> *

A concrete, promotable proxy type, for use at the kind level There are no instances for this because it is intended at the kind level only

Constructors

KProxy

data SomeSing kproxy where

An existentially-quantified singleton. This type is useful when you want a singleton type, but there is no way of knowing, at compile-time, what the type index will be. To make use of this type, you will generally have to use a pattern-match:

foo :: Bool -> ...
foo b = case toSing b of
          SomeSing sb -> {- fancy dependently-typed code with sb -}

An example like the one above may be easier to write using withSomeSing.

Constructors

SomeSing :: Sing (a :: k) -> SomeSing (KProxy :: KProxy k)

type family Error str :: k

The promotion of error. This version is more poly-kinded for easier use.

data ErrorSym0 l

type TrueSym0 = True

type FalseSym0 = False

type LTSym0 = LT

type EQSym0 = EQ

type GTSym0 = GT

type Tuple0Sym0 = `()`

data Tuple2Sym0 l

data Tuple2Sym1 l l

type Tuple2Sym2 t t = `(t, t)`

data Tuple3Sym0 l

data Tuple3Sym1 l l

data Tuple3Sym2 l l l

type Tuple3Sym3 t t t = `(t, t, t)`

data Tuple4Sym0 l

data Tuple4Sym1 l l

data Tuple4Sym2 l l l

data Tuple4Sym3 l l l l

type Tuple4Sym4 t t t t = `(t, t, t, t)`

data Tuple5Sym0 l

data Tuple5Sym1 l l

data Tuple5Sym2 l l l

data Tuple5Sym3 l l l l

data Tuple5Sym4 l l l l l

type Tuple5Sym5 t t t t t = `(t, t, t, t, t)`

data Tuple6Sym0 l

data Tuple6Sym1 l l

data Tuple6Sym2 l l l

data Tuple6Sym3 l l l l

data Tuple6Sym4 l l l l l

data Tuple6Sym5 l l l l l l

type Tuple6Sym6 t t t t t t = `(t, t, t, t, t, t)`

data Tuple7Sym0 l

data Tuple7Sym1 l l

data Tuple7Sym2 l l l

data Tuple7Sym3 l l l l

data Tuple7Sym4 l l l l l

data Tuple7Sym5 l l l l l l

data Tuple7Sym6 l l l l l l l

type Tuple7Sym7 t t t t t t t = `(t, t, t, t, t, t, t)`

data CompareSym0 l

data ThenCmpSym0 l

data FoldlSym0 l

class SuppressUnusedWarnings t where

This class (which users should never see) is to be instantiated in order to use an otherwise-unused data constructor, such as the "kind-inference" data constructor for defunctionalization symbols.

Methods

suppressUnusedWarnings :: Proxy t -> ()