Data.Singletons.TypeError

Description

Defines a drop-in replacement for TypeError (from GHC.TypeLits) that can be used at the value level as well. Since this is a drop-in replacement, it is not recommended to import all of GHC.TypeLits and Data.Singletons.TypeError at the same time, as many of the definitons in the latter deliberately clash with the former.

Documentation

type family TypeError (a :: PErrorMessage) :: b where ...

A drop-in replacement for TypeError. This also exists at the value-level as typeError.

Equations

TypeError a = TypeError (ConvertPErrorMessage a)

sTypeError :: HasCallStack => Sing err -> Sing (TypeError err)

The singleton for typeError.

Note that this is not quite as expressive as TypeError, as it is unable to handle ShowType constructors at all.

typeError :: HasCallStack => ErrorMessage -> a

The value-level counterpart to TypeError.

Note that this is not quite as expressive as TypeError, as it is unable to print the contents of ShowType constructors (it will simply print "<type>" in their place).

data ErrorMessage' s

A description of a custom type error.

This is a variation on ErrorMessage that is parameterized over what text type is used in the Text constructor. Instantiating it with Text gives you ErrorMessage, and instantiating it with Symbol gives you PErrorMessage.

Constructors

Text s
Show the text as is.

ShowType t
Pretty print the type. ShowType :: k -> ErrorMessage

(ErrorMessage' s) :<>: (ErrorMessage' s) infixl 6
Put two pieces of error message next to each other.

(ErrorMessage' s) :$$: (ErrorMessage' s) infixl 5
Stack two pieces of error message on top of each other.

type ErrorMessage = ErrorMessage' Text

A value-level ErrorMessage' which uses Text as its text type.

type PErrorMessage = ErrorMessage' Symbol

A type-level ErrorMessage' which uses Symbol as its text kind.

data family Sing :: k -> Type infixl 6 :%<>:infixl 5 :%$$:

The singleton kind-indexed data family.

type SErrorMessage = (Sing :: PErrorMessage -> Type)

type family ConvertPErrorMessage (a :: PErrorMessage) :: ErrorMessage where ...

Convert a PErrorMessage to a ErrorMessage from GHC.TypeLits.

Equations

ConvertPErrorMessage (Text t) = Text t

ConvertPErrorMessage (ShowType ty) = ShowType ty

ConvertPErrorMessage (e1 :<>: e2) = ConvertPErrorMessage e1 :<>: ConvertPErrorMessage e2

ConvertPErrorMessage (e1 :$$: e2) = ConvertPErrorMessage e1 :$$: ConvertPErrorMessage e2

showErrorMessage :: ErrorMessage -> String

Convert an ErrorMessage into a human-readable String.

Defunctionalization symbols

data TextSym0 :: forall (s6989586621681354866 :: Type). (~>) s6989586621681354866 (ErrorMessage' (s6989586621681354866 :: Type))

type TextSym1 (t6989586621681355652 :: s6989586621681354866) = Text t6989586621681355652

data ShowTypeSym0 :: forall (s6989586621681354866 :: Type) t6989586621681354867. (~>) t6989586621681354867 (ErrorMessage' (s6989586621681354866 :: Type))

type ShowTypeSym1 (t6989586621681355654 :: t6989586621681354867) = ShowType t6989586621681355654

data (:<>:@#@$) :: forall (s6989586621681354866 :: Type). (~>) (ErrorMessage' s6989586621681354866) ((~>) (ErrorMessage' s6989586621681354866) (ErrorMessage' (s6989586621681354866 :: Type))) infixl 6

data (:<>:@#@$$) (t6989586621681355656 :: ErrorMessage' (s6989586621681354866 :: Type)) :: (~>) (ErrorMessage' s6989586621681354866) (ErrorMessage' (s6989586621681354866 :: Type)) infixl 6

type (:<>:@#@$$$) (t6989586621681355656 :: ErrorMessage' s6989586621681354866) (t6989586621681355657 :: ErrorMessage' s6989586621681354866) = (:<>:) t6989586621681355656 t6989586621681355657

data (:$$:@#@$) :: forall (s6989586621681354866 :: Type). (~>) (ErrorMessage' s6989586621681354866) ((~>) (ErrorMessage' s6989586621681354866) (ErrorMessage' (s6989586621681354866 :: Type))) infixl 5

data (:$$:@#@$$) (t6989586621681355660 :: ErrorMessage' (s6989586621681354866 :: Type)) :: (~>) (ErrorMessage' s6989586621681354866) (ErrorMessage' (s6989586621681354866 :: Type)) infixl 5

type (:$$:@#@$$$) (t6989586621681355660 :: ErrorMessage' s6989586621681354866) (t6989586621681355661 :: ErrorMessage' s6989586621681354866) = (:$$:) t6989586621681355660 t6989586621681355661

data TypeErrorSym0 :: forall b6989586621681354856. (~>) PErrorMessage b6989586621681354856

type TypeErrorSym1 (a6989586621681354857 :: PErrorMessage) = TypeError a6989586621681354857