Data.Singletons.TypeLits

Description

Defines and exports singletons useful for the Nat and Symbol kinds. This exports the internal, unsafe constructors. Use Data.Singletons.TypeLits for a safe interface.

Documentation

data Nat :: *

(Kind) This is the kind of type-level natural numbers.

data Symbol :: *

(Kind) This is the kind of type-level symbols. Declared here because class IP needs it

data family Sing (a :: k)

The singleton kind-indexed data family.

type SNat x = Sing x

Kind-restricted synonym for Sing for Nats

type SSymbol x = Sing x

Kind-restricted synonym for Sing for Symbols

withKnownNat :: Sing n -> (KnownNat n => r) -> r

Given a singleton for Nat, call something requiring a KnownNat instance.

withKnownSymbol :: Sing n -> (KnownSymbol n => r) -> r

Given a singleton for Symbol, call something requiring a KnownSymbol instance.

type family Error (str :: k0) :: k

The promotion of error. This version is more poly-kinded for easier use.

data ErrorSym0 l

type ErrorSym1 t = Error t

sError :: Sing (str :: Symbol) -> a

The singleton for error

class KnownNat n

This class gives the integer associated with a type-level natural. There are instances of the class for every concrete literal: 0, 1, 2, etc.

Since: 4.7.0.0

Minimal complete definition

natSing

natVal :: KnownNat n => proxy n -> Integer

Since: 4.7.0.0

class KnownSymbol n

This class gives the string associated with a type-level symbol. There are instances of the class for every concrete literal: "hello", etc.

Since: 4.7.0.0

Minimal complete definition

symbolSing

symbolVal :: KnownSymbol n => proxy n -> String

Since: 4.7.0.0

type (:^) a b = a ^ b infixr 8

data (:^$) l

data l :^$$ l

type (:^$$$) t t = (:^) t t

Orphan instances

Num Nat

Methods

(+) :: Nat -> Nat -> Nat

(-) :: Nat -> Nat -> Nat

(*) :: Nat -> Nat -> Nat

negate :: Nat -> Nat

abs :: Nat -> Nat

signum :: Nat -> Nat

fromInteger :: Integer -> Nat