Data.Singletons.TypeLits

Description

Defines and exports singletons useful for the Nat and Symbol kinds.

Documentation

data Nat

(Kind) This is the kind of type-level natural numbers.

data Symbol

(Kind) This is the kind of type-level symbols. Declared here because class IP needs it

type family Sing

The singleton kind-indexed type family.

data SNat n

Constructors

KnownNat n => SNat

data SSymbol n

Constructors

KnownSymbol n => SSym

withKnownNat :: Sing n -> (KnownNat n => r) -> r

Given a singleton for Nat, call something requiring a KnownNat instance.

withKnownSymbol :: Sing n -> (KnownSymbol n => r) -> r

Given a singleton for Symbol, call something requiring a KnownSymbol instance.

type family Error str where ...

The promotion of error. This version is more poly-kinded for easier use.

sError :: HasCallStack => Sing (str :: Symbol) -> a

The singleton for error

type family ErrorWithoutStackTrace str where ...

The promotion of errorWithoutStackTrace. This version is more poly-kinded for easier use.

sErrorWithoutStackTrace :: Sing (str :: Symbol) -> a

The singleton for errorWithoutStackTrace.

type family Undefined where ...

The promotion of undefined.

sUndefined :: HasCallStack => a

The singleton for undefined.

class KnownNat (n :: Nat)

This class gives the integer associated with a type-level natural. There are instances of the class for every concrete literal: 0, 1, 2, etc.

Since: base-4.7.0.0

Minimal complete definition

natSing

natVal :: forall (n :: Nat) proxy. KnownNat n => proxy n -> Natural

Since: base-4.10.0.0

class KnownSymbol (n :: Symbol)

This class gives the string associated with a type-level symbol. There are instances of the class for every concrete literal: "hello", etc.

Since: base-4.7.0.0

Minimal complete definition

symbolSing

symbolVal :: forall (n :: Symbol) proxy. KnownSymbol n => proxy n -> String

Since: base-4.7.0.0

type family (a :: Nat) ^ (b :: Nat) :: Nat where ... infixr 8

Exponentiation of type-level naturals.

Since: base-4.7.0.0

(%^) :: Sing a -> Sing b -> Sing (a ^ b) infixr 8

The singleton analogue of (^) for Nats.

type family (a :: Nat) <=? (b :: Nat) :: Bool where ... infix 4

Comparison of type-level naturals, as a function. NOTE: The functionality for this function should be subsumed by CmpNat, so this might go away in the future. Please let us know, if you encounter discrepancies between the two.

(%<=?) :: Sing a -> Sing b -> Sing (a <=? b) infix 4

The singleton analogue of <=?

Note that, because of historical reasons in GHC's Nat API, <=? is incompatible (unification-wise) with <= and the PEq, SEq, POrd, and SOrd instances for Nat. (a <=? b) ~ 'True does not imply anything about a <= b or any other PEq / POrd relationships.

(Be aware that <= in the paragraph above refers to <= from the POrd typeclass, exported from Data.Singletons.Prelude.Ord, and not the <= from GHC.TypeNats. The latter is simply a type alias for (a <=? b) ~ 'True.)

This is provided here for the sake of completeness and for compatibility with libraries with APIs built around <=?. New code should use CmpNat, exposed through this library through the POrd and SOrd instances for Nat.

type family Log2 (a :: Nat) :: Nat where ...

Log base 2 (round down) of natural numbers. Log 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0

sLog2 :: Sing x -> Sing (Log2 x)

type family Div (a :: Nat) (b :: Nat) :: Nat where ... infixl 7

Division (round down) of natural numbers. Div x 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0

sDiv :: Sing x -> Sing y -> Sing (Div x y) infixl 7

type family Mod (a :: Nat) (b :: Nat) :: Nat where ... infixl 7

Modulus of natural numbers. Mod x 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0

sMod :: Sing x -> Sing y -> Sing (Mod x y) infixl 7

type family DivMod a a where ...

Equations

DivMod x y = Apply (Apply Tuple2Sym0 (Apply (Apply DivSym0 x) y)) (Apply (Apply ModSym0 x) y)

sDivMod :: Sing x -> Sing y -> Sing (DivMod x y)

type family Quot a a where ... infixl 7

Equations

Quot a_6989586621679493742 a_6989586621679493744 = Apply (Apply DivSym0 a_6989586621679493742) a_6989586621679493744

sQuot :: Sing x -> Sing y -> Sing (Quot x y) infixl 7

type family Rem a a where ... infixl 7

Equations

Rem a_6989586621679493731 a_6989586621679493733 = Apply (Apply ModSym0 a_6989586621679493731) a_6989586621679493733

sRem :: Sing x -> Sing y -> Sing (Rem x y) infixl 7

type family QuotRem a a where ...

Equations

QuotRem a_6989586621679493753 a_6989586621679493755 = Apply (Apply DivModSym0 a_6989586621679493753) a_6989586621679493755

sQuotRem :: Sing x -> Sing y -> Sing (QuotRem x y)

Defunctionalization symbols

data ErrorSym0 a6989586621679472537

type ErrorSym1 (a6989586621679472537 :: k0) = Error a6989586621679472537 :: k

data ErrorWithoutStackTraceSym0 a6989586621679472770

type ErrorWithoutStackTraceSym1 (a6989586621679472770 :: k0) = ErrorWithoutStackTrace a6989586621679472770 :: k

type UndefinedSym0 = Undefined :: k

data KnownNatSym0 a6989586621679492252

type KnownNatSym1 (a6989586621679492252 :: Nat) = KnownNat a6989586621679492252 :: Constraint

data KnownSymbolSym0 a6989586621679492254

type KnownSymbolSym1 (a6989586621679492254 :: Symbol) = KnownSymbol a6989586621679492254 :: Constraint

data (^@#@$) a6989586621679473144 infixr 8

data a6989586621679473144 ^@#@$$ a6989586621679473145 infixr 8

type (^@#@$$$) (a6989586621679473144 :: Nat) (a6989586621679473145 :: Nat) = (^) a6989586621679473144 a6989586621679473145 :: Nat infixr 8

data (<=?@#@$) a6989586621679473434 infix 4

data a6989586621679473434 <=?@#@$$ a6989586621679473435 infix 4

type (<=?@#@$$$) (a6989586621679473434 :: Nat) (a6989586621679473435 :: Nat) = (<=?) a6989586621679473434 a6989586621679473435 :: Bool infix 4

data Log2Sym0 a6989586621679492734

type Log2Sym1 (a6989586621679492734 :: Nat) = Log2 a6989586621679492734 :: Nat

data DivSym0 a6989586621679492941 infixl 7

data DivSym1 a6989586621679492941 a6989586621679492942 infixl 7

type DivSym2 (a6989586621679492941 :: Nat) (a6989586621679492942 :: Nat) = Div a6989586621679492941 a6989586621679492942 :: Nat infixl 7

data ModSym0 a6989586621679493275 infixl 7

data ModSym1 a6989586621679493275 a6989586621679493276 infixl 7

type ModSym2 (a6989586621679493275 :: Nat) (a6989586621679493276 :: Nat) = Mod a6989586621679493275 a6989586621679493276 :: Nat infixl 7

data DivModSym0 a6989586621679493767

data DivModSym1 a6989586621679493767 a6989586621679493768

type DivModSym2 (a6989586621679493767 :: Nat) (a6989586621679493768 :: Nat) = DivMod a6989586621679493767 a6989586621679493768 :: (Nat, Nat)

data QuotSym0 a6989586621679493749 infixl 7

data QuotSym1 a6989586621679493749 a6989586621679493750 infixl 7

type QuotSym2 (a6989586621679493749 :: Nat) (a6989586621679493750 :: Nat) = Quot a6989586621679493749 a6989586621679493750 :: Nat infixl 7

data RemSym0 a6989586621679493738 infixl 7

data RemSym1 a6989586621679493738 a6989586621679493739 infixl 7

type RemSym2 (a6989586621679493738 :: Nat) (a6989586621679493739 :: Nat) = Rem a6989586621679493738 a6989586621679493739 :: Nat infixl 7

data QuotRemSym0 a6989586621679493760

data QuotRemSym1 a6989586621679493760 a6989586621679493761

type QuotRemSym2 (a6989586621679493760 :: Nat) (a6989586621679493761 :: Nat) = QuotRem a6989586621679493760 a6989586621679493761 :: (Nat, Nat)

Orphan instances

Enum Nat

Instance details

Methods

succ :: Nat -> Nat

pred :: Nat -> Nat

toEnum :: Int -> Nat

fromEnum :: Nat -> Int

enumFrom :: Nat -> [Nat]

enumFromThen :: Nat -> Nat -> [Nat]

enumFromTo :: Nat -> Nat -> [Nat]

enumFromThenTo :: Nat -> Nat -> Nat -> [Nat]

Eq Nat

Instance details

Methods

(==) :: Nat -> Nat -> Bool

(/=) :: Nat -> Nat -> Bool

Eq Symbol
This bogus instance is helpful for people who want to define functions over Symbols that will only be used at the type level or as singletons.

Instance details

Methods

(==) :: Symbol -> Symbol -> Bool

(/=) :: Symbol -> Symbol -> Bool

Num Nat
This bogus Num instance is helpful for people who want to define functions over Nats that will only be used at the type level or as singletons. A correct SNum instance for Nat singletons exists.

Instance details

Methods

(+) :: Nat -> Nat -> Nat

(-) :: Nat -> Nat -> Nat

(*) :: Nat -> Nat -> Nat

negate :: Nat -> Nat

abs :: Nat -> Nat

signum :: Nat -> Nat

fromInteger :: Integer -> Nat

Ord Nat

Instance details

Methods

compare :: Nat -> Nat -> Ordering

(<) :: Nat -> Nat -> Bool

(<=) :: Nat -> Nat -> Bool

(>) :: Nat -> Nat -> Bool

(>=) :: Nat -> Nat -> Bool

max :: Nat -> Nat -> Nat

min :: Nat -> Nat -> Nat

Ord Symbol

Instance details

Methods

compare :: Symbol -> Symbol -> Ordering

(<) :: Symbol -> Symbol -> Bool

(<=) :: Symbol -> Symbol -> Bool

(>) :: Symbol -> Symbol -> Bool

(>=) :: Symbol -> Symbol -> Bool

max :: Symbol -> Symbol -> Symbol

min :: Symbol -> Symbol -> Symbol

Show Nat

Instance details

Methods

showsPrec :: Int -> Nat -> ShowS

show :: Nat -> String

showList :: [Nat] -> ShowS

Show Symbol

Instance details

Methods

showsPrec :: Int -> Symbol -> ShowS

show :: Symbol -> String

showList :: [Symbol] -> ShowS

IsString Symbol

Instance details

Methods

fromString :: String -> Symbol

Semigroup Symbol

Instance details

Methods

(<>) :: Symbol -> Symbol -> Symbol

sconcat :: NonEmpty Symbol -> Symbol

stimes :: Integral b => b -> Symbol -> Symbol

Monoid Symbol

Instance details

Methods

mempty :: Symbol

mappend :: Symbol -> Symbol -> Symbol

mconcat :: [Symbol] -> Symbol