Data.Singletons.TypeLits

Description

Defines and exports singletons useful for the Nat and Symbol kinds.

Documentation

data Nat

(Kind) This is the kind of type-level natural numbers.

data Symbol

(Kind) This is the kind of type-level symbols. Declared here because class IP needs it

type family Sing :: k -> Type

The singleton kind-indexed type family.

data SNat (n :: Nat)

Constructors

KnownNat n => SNat

data SSymbol (n :: Symbol)

Constructors

KnownSymbol n => SSym

withKnownNat :: Sing n -> (KnownNat n => r) -> r

Given a singleton for Nat, call something requiring a KnownNat instance.

withKnownSymbol :: Sing n -> (KnownSymbol n => r) -> r

Given a singleton for Symbol, call something requiring a KnownSymbol instance.

type family Error (str :: k0) :: k where ...

The promotion of error. This version is more poly-kinded for easier use.

sError :: HasCallStack => Sing (str :: Symbol) -> a

The singleton for error

type family ErrorWithoutStackTrace (str :: k0) :: k where ...

The promotion of errorWithoutStackTrace. This version is more poly-kinded for easier use.

sErrorWithoutStackTrace :: Sing (str :: Symbol) -> a

The singleton for errorWithoutStackTrace.

type family Undefined :: k where ...

The promotion of undefined.

sUndefined :: HasCallStack => a

The singleton for undefined.

class KnownNat (n :: Nat)

This class gives the integer associated with a type-level natural. There are instances of the class for every concrete literal: 0, 1, 2, etc.

Since: base-4.7.0.0

Minimal complete definition

natSing

natVal :: forall (n :: Nat) proxy. KnownNat n => proxy n -> Natural

Since: base-4.10.0.0

class KnownSymbol (n :: Symbol)

This class gives the string associated with a type-level symbol. There are instances of the class for every concrete literal: "hello", etc.

Since: base-4.7.0.0

Minimal complete definition

symbolSing

symbolVal :: forall (n :: Symbol) proxy. KnownSymbol n => proxy n -> String

Since: base-4.7.0.0

type family (a :: Nat) ^ (b :: Nat) :: Nat where ... infixr 8

Exponentiation of type-level naturals.

Since: base-4.7.0.0

(%^) :: Sing a -> Sing b -> Sing (a ^ b) infixr 8

The singleton analogue of (^) for Nats.

type family (a :: Nat) <=? (b :: Nat) :: Bool where ... infix 4

Comparison of type-level naturals, as a function. NOTE: The functionality for this function should be subsumed by CmpNat, so this might go away in the future. Please let us know, if you encounter discrepancies between the two.

(%<=?) :: Sing a -> Sing b -> Sing (a <=? b) infix 4

The singleton analogue of <=?

Note that, because of historical reasons in GHC's Nat API, <=? is incompatible (unification-wise) with <= and the PEq, SEq, POrd, and SOrd instances for Nat. (a <=? b) ~ 'True does not imply anything about a <= b or any other PEq / POrd relationships.

(Be aware that <= in the paragraph above refers to <= from the POrd typeclass, exported from Data.Singletons.Prelude.Ord, and not the <= from GHC.TypeNats. The latter is simply a type alias for (a <=? b) ~ 'True.)

This is provided here for the sake of completeness and for compatibility with libraries with APIs built around <=?. New code should use CmpNat, exposed through this library through the POrd and SOrd instances for Nat.

type family Log2 (a :: Nat) :: Nat where ...

Log base 2 (round down) of natural numbers. Log 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0

sLog2 :: Sing x -> Sing (Log2 x)

type family Div (a :: Nat) (b :: Nat) :: Nat where ... infixl 7

Division (round down) of natural numbers. Div x 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0

sDiv :: Sing x -> Sing y -> Sing (Div x y) infixl 7

type family Mod (a :: Nat) (b :: Nat) :: Nat where ... infixl 7

Modulus of natural numbers. Mod x 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0

sMod :: Sing x -> Sing y -> Sing (Mod x y) infixl 7

type family DivMod (a :: Nat) (a :: Nat) :: (Nat, Nat) where ...

Equations

DivMod x y = Apply (Apply Tuple2Sym0 (Apply (Apply DivSym0 x) y)) (Apply (Apply ModSym0 x) y)

sDivMod :: Sing x -> Sing y -> Sing (DivMod x y)

type family Quot (a :: Nat) (a :: Nat) :: Nat where ... infixl 7

Equations

Quot a_6989586621679509656 a_6989586621679509658 = Apply (Apply DivSym0 a_6989586621679509656) a_6989586621679509658

sQuot :: Sing x -> Sing y -> Sing (Quot x y) infixl 7

type family Rem (a :: Nat) (a :: Nat) :: Nat where ... infixl 7

Equations

Rem a_6989586621679509646 a_6989586621679509648 = Apply (Apply ModSym0 a_6989586621679509646) a_6989586621679509648

sRem :: Sing x -> Sing y -> Sing (Rem x y) infixl 7

type family QuotRem (a :: Nat) (a :: Nat) :: (Nat, Nat) where ...

Equations

QuotRem a_6989586621679509666 a_6989586621679509668 = Apply (Apply DivModSym0 a_6989586621679509666) a_6989586621679509668

sQuotRem :: Sing x -> Sing y -> Sing (QuotRem x y)

Defunctionalization symbols

data ErrorSym0 :: forall k06989586621679487392 k6989586621679487393. (~>) k06989586621679487392 k6989586621679487393

type ErrorSym1 (str6989586621679487394 :: k06989586621679487392) = Error str6989586621679487394

data ErrorWithoutStackTraceSym0 :: forall k06989586621679488482 k6989586621679488483. (~>) k06989586621679488482 k6989586621679488483

type ErrorWithoutStackTraceSym1 (str6989586621679488484 :: k06989586621679488482) = ErrorWithoutStackTrace str6989586621679488484

type UndefinedSym0 = Undefined

data KnownNatSym0 :: (~>) Nat Constraint

type KnownNatSym1 (n6989586621679487549 :: Nat) = KnownNat n6989586621679487549

data KnownSymbolSym0 :: (~>) Symbol Constraint

type KnownSymbolSym1 (n6989586621679487615 :: Symbol) = KnownSymbol n6989586621679487615

data (^@#@$) :: (~>) Nat ((~>) Nat Nat) infixr 8

data (^@#@$$) (a3530822107858468865 :: Nat) :: (~>) Nat Nat infixr 8

type (^@#@$$$) (a3530822107858468865 :: Nat) (b3530822107858468866 :: Nat) = (^) a3530822107858468865 b3530822107858468866

data (<=?@#@$) :: (~>) Nat ((~>) Nat Bool) infix 4

data (<=?@#@$$) (a3530822107858468865 :: Nat) :: (~>) Nat Bool infix 4

type (<=?@#@$$$) (a3530822107858468865 :: Nat) (b3530822107858468866 :: Nat) = (<=?) a3530822107858468865 b3530822107858468866

data Log2Sym0 :: (~>) Nat Nat

type Log2Sym1 (a3530822107858468865 :: Nat) = Log2 a3530822107858468865

data DivSym0 :: (~>) Nat ((~>) Nat Nat) infixl 7

data DivSym1 (a3530822107858468865 :: Nat) :: (~>) Nat Nat infixl 7

type DivSym2 (a3530822107858468865 :: Nat) (b3530822107858468866 :: Nat) = Div a3530822107858468865 b3530822107858468866

data ModSym0 :: (~>) Nat ((~>) Nat Nat) infixl 7

data ModSym1 (a3530822107858468865 :: Nat) :: (~>) Nat Nat infixl 7

type ModSym2 (a3530822107858468865 :: Nat) (b3530822107858468866 :: Nat) = Mod a3530822107858468865 b3530822107858468866

data DivModSym0 :: (~>) Nat ((~>) Nat (Nat, Nat))

data DivModSym1 (a6989586621679509676 :: Nat) :: (~>) Nat (Nat, Nat)

type DivModSym2 (a6989586621679509676 :: Nat) (a6989586621679509677 :: Nat) = DivMod a6989586621679509676 a6989586621679509677

data QuotSym0 :: (~>) Nat ((~>) Nat Nat) infixl 7

data QuotSym1 (a6989586621679509660 :: Nat) :: (~>) Nat Nat infixl 7

type QuotSym2 (a6989586621679509660 :: Nat) (a6989586621679509661 :: Nat) = Quot a6989586621679509660 a6989586621679509661

data RemSym0 :: (~>) Nat ((~>) Nat Nat) infixl 7

data RemSym1 (a6989586621679509650 :: Nat) :: (~>) Nat Nat infixl 7

type RemSym2 (a6989586621679509650 :: Nat) (a6989586621679509651 :: Nat) = Rem a6989586621679509650 a6989586621679509651

data QuotRemSym0 :: (~>) Nat ((~>) Nat (Nat, Nat))

data QuotRemSym1 (a6989586621679509670 :: Nat) :: (~>) Nat (Nat, Nat)

type QuotRemSym2 (a6989586621679509670 :: Nat) (a6989586621679509671 :: Nat) = QuotRem a6989586621679509670 a6989586621679509671

Orphan instances

Enum Nat

Instance details

Methods

succ :: Nat -> Nat

pred :: Nat -> Nat

toEnum :: Int -> Nat

fromEnum :: Nat -> Int

enumFrom :: Nat -> [Nat]

enumFromThen :: Nat -> Nat -> [Nat]

enumFromTo :: Nat -> Nat -> [Nat]

enumFromThenTo :: Nat -> Nat -> Nat -> [Nat]

Eq Nat

Instance details

Methods

(==) :: Nat -> Nat -> Bool

(/=) :: Nat -> Nat -> Bool

Eq Symbol
This bogus instance is helpful for people who want to define functions over Symbols that will only be used at the type level or as singletons.

Instance details

Methods

(==) :: Symbol -> Symbol -> Bool

(/=) :: Symbol -> Symbol -> Bool

Num Nat
This bogus Num instance is helpful for people who want to define functions over Nats that will only be used at the type level or as singletons. A correct SNum instance for Nat singletons exists.

Instance details

Methods

(+) :: Nat -> Nat -> Nat

(-) :: Nat -> Nat -> Nat

(*) :: Nat -> Nat -> Nat

negate :: Nat -> Nat

abs :: Nat -> Nat

signum :: Nat -> Nat

fromInteger :: Integer -> Nat

Ord Nat

Instance details

Methods

compare :: Nat -> Nat -> Ordering

(<) :: Nat -> Nat -> Bool

(<=) :: Nat -> Nat -> Bool

(>) :: Nat -> Nat -> Bool

(>=) :: Nat -> Nat -> Bool

max :: Nat -> Nat -> Nat

min :: Nat -> Nat -> Nat

Ord Symbol

Instance details

Methods

compare :: Symbol -> Symbol -> Ordering

(<) :: Symbol -> Symbol -> Bool

(<=) :: Symbol -> Symbol -> Bool

(>) :: Symbol -> Symbol -> Bool

(>=) :: Symbol -> Symbol -> Bool

max :: Symbol -> Symbol -> Symbol

min :: Symbol -> Symbol -> Symbol

Show Nat

Instance details

Methods

showsPrec :: Int -> Nat -> ShowS

show :: Nat -> String

showList :: [Nat] -> ShowS

Show Symbol

Instance details

Methods

showsPrec :: Int -> Symbol -> ShowS

show :: Symbol -> String

showList :: [Symbol] -> ShowS

IsString Symbol

Instance details

Methods

fromString :: String -> Symbol

Semigroup Symbol

Instance details

Methods

(<>) :: Symbol -> Symbol -> Symbol

sconcat :: NonEmpty Symbol -> Symbol

stimes :: Integral b => b -> Symbol -> Symbol

Monoid Symbol

Instance details

Methods

mempty :: Symbol

mappend :: Symbol -> Symbol -> Symbol

mconcat :: [Symbol] -> Symbol