Data.Singletons.TypeRepStar

Description

This module defines singleton instances making Typeable the singleton for the kind *. The definitions don't fully line up with what is expected within the singletons library, so expect unusual results!

Documentation

data family Sing (a :: k)

The singleton kind-indexed data family.

Here is the definition of the singleton for *:

data instance Sing (a :: *) where
  STypeRep :: Typeable a => Sing a

Instances for SingI, SingKind, SEq, SDecide, and TestCoercion are also supplied.

Orphan instances

SingKind Type

Associated Types

type DemoteRep Type :: *

Methods

fromSing :: Sing Type a -> DemoteRep Type

toSing :: DemoteRep Type -> SomeSing Type

SDecide Type

Methods

(%~) :: Sing Type a -> Sing Type b -> Decision ((Type :~: a) b)

SEq Type

Methods

(%:==) :: Sing Type a -> Sing Type b -> Sing Bool ((Type :== a) b)

(%:/=) :: Sing Type a -> Sing Type b -> Sing Bool ((Type :/= a) b)

Typeable * a => SingI * a

Methods

sing :: Sing a a

PEq Type (Proxy * Type)

Associated Types

type ((Proxy * Type) :== (x :: Proxy * Type)) (y :: Proxy * Type) :: Bool

type ((Proxy * Type) :/= (x :: Proxy * Type)) (y :: Proxy * Type) :: Bool

TestCoercion * (Sing *)

Methods

testCoercion :: f a -> f b -> Maybe (Coercion (Sing *) a b)