Data.Singletons.TypeRepTYPE

Description

This module defines singleton instances making TypeRep the singleton for the kind TYPE rep (for some RuntimeRep rep), an instantiation of which is the famous kind Type. The definitions don't fully line up with what is expected within the singletons library, so expect unusual results!

Documentation

data family Sing :: k -> Type

The singleton kind-indexed data family.

Here is the definition of the singleton for TYPE rep:

newtype instance Sing :: forall (rep :: RuntimeRep). TYPE rep -> Type where
  STypeRep :: forall (rep :: RuntimeRep) (a :: TYPE rep). TypeRep a -> Sing a

Instances for SingI, SingKind, SEq, SDecide, and TestCoercion are also supplied.

data SomeTypeRepTYPE :: RuntimeRep -> Type where

A variant of SomeTypeRep whose underlying TypeRep is restricted to kind TYPE rep (for some RuntimeRep rep).

Constructors

SomeTypeRepTYPE :: forall (rep :: RuntimeRep) (a :: TYPE rep). !(TypeRep a) -> SomeTypeRepTYPE rep

Orphan instances

SingKind (TYPE rep)

Instance details

Associated Types

type Demote (TYPE rep) = (r :: Type)

Methods

fromSing :: Sing a -> Demote (TYPE rep)

toSing :: Demote (TYPE rep) -> SomeSing (TYPE rep)

SDecide (TYPE rep)

Instance details

Methods

(%~) :: Sing a -> Sing b -> Decision (a :~: b)

PEq (TYPE rep)

Instance details

Associated Types

type x == y :: Bool

type x /= y :: Bool

SEq (TYPE rep)

Instance details

Methods

(%==) :: Sing a -> Sing b -> Sing (a == b)

(%/=) :: Sing a -> Sing b -> Sing (a /= b)

Typeable a => SingI (a :: TYPE rep)

Instance details

Methods

sing :: Sing a

Eq (Sing a)

Instance details

Methods

(==) :: Sing a -> Sing a -> Bool

(/=) :: Sing a -> Sing a -> Bool

Ord (Sing a)

Instance details

Methods

compare :: Sing a -> Sing a -> Ordering

(<) :: Sing a -> Sing a -> Bool

(<=) :: Sing a -> Sing a -> Bool

(>) :: Sing a -> Sing a -> Bool

(>=) :: Sing a -> Sing a -> Bool

max :: Sing a -> Sing a -> Sing a

min :: Sing a -> Sing a -> Sing a

Show (Sing a)

Instance details

Methods

showsPrec :: Int -> Sing a -> ShowS

show :: Sing a -> String

showList :: [Sing a] -> ShowS